Add a seasonally adjust model to a multi-processing

Function to add a new seasonally adjust object (class "SA" or "jSA") in a workspace object.

add_sa_item(workspace, multiprocessing, sa_obj, name)

Arguments

workspace
the workspace to add the seasonally adjust model.

multiprocessing
the name or index of the multiprocessing to add the seasonally adjust model.

sa_obj
the seasonally adjust object to export.

name
the name of the seasonally adjust model in the multiprocessing. By default the name of the sa_obj is used.

See also

Examples

# \donttest{
dir <- tempdir()
spec_x13 <- x13_spec(spec = "RSA5c", easter.enabled = FALSE)
sa_x13 <- x13(ipi_c_eu[, "FR"], spec = spec_x13)
spec_ts <- tramoseats_spec(spec = "RSA5")
sa_ts <- jtramoseats(ipi_c_eu[, "FR"], spec = spec_ts)

wk <- new_workspace()
new_multiprocessing(wk, "sa1")
add_sa_item(wk, "sa1", sa_x13, "X13")
add_sa_item(wk, "sa1", sa_ts, "TramoSeats")

save_workspace(wk, file.path(dir, "workspace.xml"))
# }