Count the number of objects inside a workspace or multiprocessing

Generics functions to count the number of multiprocessing (respectively sa_item) inside a workspace (respectively multiprocessing).

count(x)

Arguments

x
the workspace or the codemultiprocessing.

See also

Examples

wk <- new_workspace()
mp <- new_multiprocessing(wk, "sa1")
count(wk) # 1 multiprocessing inside the workspace wk

#> [1] 1

count(mp) # 0 sa_item inside the multiprocessing mp

#> [1] 0