Get the seasonally adjusted model from a workspace

Generics functions to get seasonally adjusted model(s) from workspace, multiprocessing or sa_item object. get_model returns a "SA" objects while get_jmodel returns the Java objects of the models.

get_jmodel(x, workspace, userdefined = NULL, progress_bar = TRUE)

get_model(x, workspace, userdefined = NULL, progress_bar = TRUE)

Arguments

x
the object to get the seasonally adjusted model.

workspace
the workspace object where models are stored. If x is a workspace object this parameter is not used.

userdefined
vector with characters for additional output variables. (see x13 or tramoseats).

progress_bar
boolean: if TRUE a progress bar is printed.

Value

get_model() returns a seasonally adjust object (class c("SA", "X13") or c("SA", "TRAMO_SEATS") or list of seasonally adjust objects:

if x is a sa_item object, get_model(x) returns a "SA" object (or a jSA object with get_jmodel(x));

if x is a multiprocessing object, get_ts(x) returns list of length the number of sa_items, each element containing a "SA" object (or a jSA object with get_jmodel(x));

if x is a workspace object, get_ts(x) returns list of length the number of multiprocessing, each element containing a list of a "SA" object (or a jSA object with get_jmodel(x)).

See also

Examples

# \donttest{
spec_x13 <- x13_spec(spec = "RSA5c", easter.enabled = FALSE)
sa_x13 <- x13(ipi_c_eu[, "FR"], spec = spec_x13)
spec_ts <- tramoseats_spec(spec = "RSA5")
sa_ts <- tramoseats(ipi_c_eu[, "FR"], spec = spec_ts)

wk <- new_workspace()
mp <- new_multiprocessing(wk, "sa1")
add_sa_item(wk, "sa1", sa_x13, "X13")
add_sa_item(wk, "sa1", sa_ts, "TramoSeats")

compute(wk) # It's important to compute the workspace to get the SA model
sa_item1 <- get_object(mp, 1)

get_model(sa_item1, wk) # Extract the model of the sa_item1: its the object sa_x13

#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5393      0.048
#> BTheta(1)  -0.5076      0.050
#> 
#>              Estimate Std. Error
#> Monday       -0.07677      0.168
#> Tuesday       0.06278      0.167
#> Wednesday     0.20386      0.168
#> Thursday     -0.56900      0.168
#> Friday        0.48643      0.168
#> Saturday      0.16823      0.168
#> Leap year    -0.04202      0.516
#> LS (11-2008) -8.47862      1.328
#> LS (1-2009)  -7.24743      1.328
#> LS (5-2008)  -5.17890      1.271
#> 
#> 
#> Residual standard error: 1.692 on 310 degrees of freedom
#> Log likelihood = -630.1, aic =  1286 aicc =  1287, bic(corrected for length) = 1.266
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.070
#> M(2)    0.044
#> M(3)    0.908
#> M(4)    0.621
#> M(5)    0.702
#> M(6)    0.225
#> M(7)    0.075
#> M(8)    0.212
#> M(9)    0.057
#> M(10)   0.186
#> M(11)   0.160
#> Q       0.298
#> Q-M2    0.330
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t           s          i
#> Jan 2017  97.4 100.8343 100.7013  -3.4343306  0.1330766
#> Feb 2017  97.5 100.3650 101.0249  -2.8649615 -0.6599802
#> Mar 2017 112.0 102.2535 101.4647   9.7464542  0.7888578
#> Apr 2017 103.0 101.0035 101.9216   1.9964916 -0.9181391
#> May 2017 100.4 103.0243 102.3451  -2.6243031  0.6791988
#> Jun 2017 111.2 102.5088 102.7541   8.6911964 -0.2453078
#> Jul 2017 103.4 103.6337 103.1399  -0.2336808  0.4937705
#> Aug 2017  79.3 103.2251 103.5264 -23.9251325 -0.3013005
#> Sep 2017 109.7 103.2506 103.9215   6.4494326 -0.6709239
#> Oct 2017 114.0 107.0447 104.3063   6.9552773  2.7383859
#> Nov 2017 107.7 105.5642 104.6206   2.1358354  0.9435229
#> Dec 2017 101.4 104.5036 104.8670  -3.1036149 -0.3633398
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f           i_f
#> Jan 2018 102.12660 104.9825 105.0158  -2.8558846 -0.0333189370
#> Feb 2018 102.18467 105.1651 105.1322  -2.9804296  0.0328501611
#> Mar 2018 114.75482 104.9927 105.2650   9.7620667 -0.2722128214
#> Apr 2018 107.12347 105.4543 105.4551   1.6692056 -0.0008294846
#> May 2018 102.76596 105.8226 105.7251  -3.0566117  0.0974629726
#> Jun 2018 115.36179 106.0357 106.0400   9.3261226 -0.0043589056
#> Jul 2018 106.15595 106.5844 106.3555  -0.4284186  0.2288559841
#> Aug 2018  82.90307 106.3248 106.6641 -23.4217780 -0.3392719434
#> Sep 2018 112.28236 106.6926 106.9366   5.5897789 -0.2440563282
#> Oct 2018 115.41162 107.9129 107.1741   7.4986863  0.7388293101
#> Nov 2018 109.85751 107.2445 107.3673   2.6130640 -0.1228820216
#> Dec 2018 103.52884 107.3040 107.5048  -3.7751434 -0.2008278423
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.648
#>  Seasonal     49.383
#>  Irregular     0.419
#>  TD & Hol.     0.039
#>  Others       50.184
#>  Total       101.672
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.011
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.970
#>  f-test on i (seasonal dummies)         0.887
#>  Residual seasonality (entire series)   0.967
#>  Residual seasonality (last 3 years)    0.931
#>  f-test on sa (td)                      0.974
#>  f-test on i (td)                       0.999
#> 
#> 
#> Additional output variables

# To get all the models of the multiprocessing mp:
get_model(mp, wk)

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> $X13
#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5393      0.048
#> BTheta(1)  -0.5076      0.050
#> 
#>              Estimate Std. Error
#> Monday       -0.07677      0.168
#> Tuesday       0.06278      0.167
#> Wednesday     0.20386      0.168
#> Thursday     -0.56900      0.168
#> Friday        0.48643      0.168
#> Saturday      0.16823      0.168
#> Leap year    -0.04202      0.516
#> LS (11-2008) -8.47862      1.328
#> LS (1-2009)  -7.24743      1.328
#> LS (5-2008)  -5.17890      1.271
#> 
#> 
#> Residual standard error: 1.692 on 310 degrees of freedom
#> Log likelihood = -630.1, aic =  1286 aicc =  1287, bic(corrected for length) = 1.266
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.070
#> M(2)    0.044
#> M(3)    0.908
#> M(4)    0.621
#> M(5)    0.702
#> M(6)    0.225
#> M(7)    0.075
#> M(8)    0.212
#> M(9)    0.057
#> M(10)   0.186
#> M(11)   0.160
#> Q       0.298
#> Q-M2    0.330
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t           s          i
#> Jan 2017  97.4 100.8343 100.7013  -3.4343306  0.1330766
#> Feb 2017  97.5 100.3650 101.0249  -2.8649615 -0.6599802
#> Mar 2017 112.0 102.2535 101.4647   9.7464542  0.7888578
#> Apr 2017 103.0 101.0035 101.9216   1.9964916 -0.9181391
#> May 2017 100.4 103.0243 102.3451  -2.6243031  0.6791988
#> Jun 2017 111.2 102.5088 102.7541   8.6911964 -0.2453078
#> Jul 2017 103.4 103.6337 103.1399  -0.2336808  0.4937705
#> Aug 2017  79.3 103.2251 103.5264 -23.9251325 -0.3013005
#> Sep 2017 109.7 103.2506 103.9215   6.4494326 -0.6709239
#> Oct 2017 114.0 107.0447 104.3063   6.9552773  2.7383859
#> Nov 2017 107.7 105.5642 104.6206   2.1358354  0.9435229
#> Dec 2017 101.4 104.5036 104.8670  -3.1036149 -0.3633398
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f           i_f
#> Jan 2018 102.12660 104.9825 105.0158  -2.8558846 -0.0333189370
#> Feb 2018 102.18467 105.1651 105.1322  -2.9804296  0.0328501611
#> Mar 2018 114.75482 104.9927 105.2650   9.7620667 -0.2722128214
#> Apr 2018 107.12347 105.4543 105.4551   1.6692056 -0.0008294846
#> May 2018 102.76596 105.8226 105.7251  -3.0566117  0.0974629726
#> Jun 2018 115.36179 106.0357 106.0400   9.3261226 -0.0043589056
#> Jul 2018 106.15595 106.5844 106.3555  -0.4284186  0.2288559841
#> Aug 2018  82.90307 106.3248 106.6641 -23.4217780 -0.3392719434
#> Sep 2018 112.28236 106.6926 106.9366   5.5897789 -0.2440563282
#> Oct 2018 115.41162 107.9129 107.1741   7.4986863  0.7388293101
#> Nov 2018 109.85751 107.2445 107.3673   2.6130640 -0.1228820216
#> Dec 2018 103.52884 107.3040 107.5048  -3.7751434 -0.2008278423
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.648
#>  Seasonal     49.383
#>  Irregular     0.419
#>  TD & Hol.     0.039
#>  Others       50.184
#>  Total       101.672
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.011
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.970
#>  f-test on i (seasonal dummies)         0.887
#>  Residual seasonality (entire series)   0.967
#>  Residual seasonality (last 3 years)    0.931
#>  f-test on sa (td)                      0.974
#>  f-test on i (td)                       0.999
#> 
#> 
#> Additional output variables
#> 
#> $TramoSeats
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.3555      0.255
#> Phi(2)      0.1344      0.130
#> Theta(1)   -0.1867      0.255
#> BTheta(1)  -0.6117      0.047
#> 
#>                Estimate Std. Error
#> Monday       -0.0022678      0.002
#> Tuesday       0.0006417      0.002
#> Wednesday     0.0032444      0.002
#> Thursday     -0.0055635      0.002
#> Friday        0.0033124      0.002
#> Saturday      0.0019048      0.002
#> Easter [6]   -0.0106133      0.003
#> LS (11-2008) -0.0795302      0.013
#> LS (1-2009)  -0.0711422      0.013
#> AO (8-2001)   0.0447823      0.013
#> 
#> 
#> Residual standard error: 0.01594 on 308 degrees of freedom
#> Log likelihood = 875.5, aic =  1262 aicc =  1263, bic(corrected for length) = -8.027
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 + 0.355520 B + 0.134366 B^2 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.186667 B - 0.611719 B^12 + 0.114188 B^13 
#> 
#> 
#> SA
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.314513 B + 0.340411 B^2 
#> Innovation variance:  0.4651916 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.040093 B - 0.959907 B^2 
#> Innovation variance:  0.04852415 
#> 
#> Seasonal
#> AR :  1 + 0.355520 B + 0.134366 B^2 
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 0.715509 B + 0.458526 B^2 + 0.307968 B^3 + 0.130783 B^4 - 0.050030 B^5 - 0.216796 B^6 - 0.353788 B^7 - 0.443393 B^8 - 0.467164 B^9 - 0.373994 B^10 - 0.163801 B^11 - 0.411499 B^12 - 0.132321 B^13 
#> Innovation variance:  0.1608041 
#> 
#> Irregular
#> Innovation variance:  0.2049349 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2017  97.4 100.9340 100.9645 0.9649872 0.9996976
#> Feb 2017  97.5 100.7675 101.2596 0.9675741 0.9951403
#> Mar 2017 112.0 102.0785 101.7061 1.0971949 1.0036614
#> Apr 2017 103.0 101.8555 102.1727 1.0112369 0.9968946
#> May 2017 100.4 103.0213 102.6101 0.9745555 1.0040080
#> Jun 2017 111.2 102.8013 102.9761 1.0816988 0.9983021
#> Jul 2017 103.4 103.3075 103.3373 1.0008951 0.9997120
#> Aug 2017  79.3 103.7806 103.7465 0.7641122 1.0003281
#> Sep 2017 109.7 103.7197 104.2758 1.0576581 0.9946676
#> Oct 2017 114.0 105.6630 104.8765 1.0789015 1.0074997
#> Nov 2017 107.7 105.4786 105.1828 1.0210600 1.0028123
#> Dec 2017 101.4 104.9682 105.2367 0.9660069 0.9974489
#> 
#> Forecasts:
#>               y_f     sa_f      t_f       s_f       i_f
#> Jan 2018 102.2106 105.3440 105.3449 0.9702559 0.9999915
#> Feb 2018 102.3328 105.5134 105.5143 0.9698565 0.9999915
#> Mar 2018 114.1449 105.6830 105.6839 1.0800682 0.9999915
#> Apr 2018 108.0526 105.8530 105.8539 1.0207800 0.9999915
#> May 2018 102.6903 106.0232 106.0241 0.9685646 0.9999915
#> Jun 2018 115.9079 106.1936 106.1945 1.0914767 0.9999915
#> Jul 2018 106.2462 106.3644 106.3653 0.9988891 0.9999915
#> Aug 2018  81.4640 106.5354 106.5363 0.7646660 0.9999915
#> Sep 2018 112.8928 106.7067 106.7076 1.0579730 0.9999915
#> Oct 2018 115.3561 106.8783 106.8792 1.0793227 0.9999915
#> Nov 2018 109.2863 107.0501 107.0510 1.0208889 0.9999915
#> Dec 2018 103.2836 107.2222 107.2231 0.9632669 0.9999915
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.179
#>  Seasonal     66.777
#>  Irregular     0.105
#>  TD & Hol.     0.067
#>  Others       30.524
#>  Total        99.651
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.187
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.977
#>  f-test on sa (td)                      0.930
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables
#> 

# To get all the models of the workspace wk:
get_model(wk)

#> Multiprocessing 1 on 1:
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> $sa1
#> $sa1$X13
#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5393      0.048
#> BTheta(1)  -0.5076      0.050
#> 
#>              Estimate Std. Error
#> Monday       -0.07677      0.168
#> Tuesday       0.06278      0.167
#> Wednesday     0.20386      0.168
#> Thursday     -0.56900      0.168
#> Friday        0.48643      0.168
#> Saturday      0.16823      0.168
#> Leap year    -0.04202      0.516
#> LS (11-2008) -8.47862      1.328
#> LS (1-2009)  -7.24743      1.328
#> LS (5-2008)  -5.17890      1.271
#> 
#> 
#> Residual standard error: 1.692 on 310 degrees of freedom
#> Log likelihood = -630.1, aic =  1286 aicc =  1287, bic(corrected for length) = 1.266
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.070
#> M(2)    0.044
#> M(3)    0.908
#> M(4)    0.621
#> M(5)    0.702
#> M(6)    0.225
#> M(7)    0.075
#> M(8)    0.212
#> M(9)    0.057
#> M(10)   0.186
#> M(11)   0.160
#> Q       0.298
#> Q-M2    0.330
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t           s          i
#> Jan 2017  97.4 100.8343 100.7013  -3.4343306  0.1330766
#> Feb 2017  97.5 100.3650 101.0249  -2.8649615 -0.6599802
#> Mar 2017 112.0 102.2535 101.4647   9.7464542  0.7888578
#> Apr 2017 103.0 101.0035 101.9216   1.9964916 -0.9181391
#> May 2017 100.4 103.0243 102.3451  -2.6243031  0.6791988
#> Jun 2017 111.2 102.5088 102.7541   8.6911964 -0.2453078
#> Jul 2017 103.4 103.6337 103.1399  -0.2336808  0.4937705
#> Aug 2017  79.3 103.2251 103.5264 -23.9251325 -0.3013005
#> Sep 2017 109.7 103.2506 103.9215   6.4494326 -0.6709239
#> Oct 2017 114.0 107.0447 104.3063   6.9552773  2.7383859
#> Nov 2017 107.7 105.5642 104.6206   2.1358354  0.9435229
#> Dec 2017 101.4 104.5036 104.8670  -3.1036149 -0.3633398
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f           i_f
#> Jan 2018 102.12660 104.9825 105.0158  -2.8558846 -0.0333189370
#> Feb 2018 102.18467 105.1651 105.1322  -2.9804296  0.0328501611
#> Mar 2018 114.75482 104.9927 105.2650   9.7620667 -0.2722128214
#> Apr 2018 107.12347 105.4543 105.4551   1.6692056 -0.0008294846
#> May 2018 102.76596 105.8226 105.7251  -3.0566117  0.0974629726
#> Jun 2018 115.36179 106.0357 106.0400   9.3261226 -0.0043589056
#> Jul 2018 106.15595 106.5844 106.3555  -0.4284186  0.2288559841
#> Aug 2018  82.90307 106.3248 106.6641 -23.4217780 -0.3392719434
#> Sep 2018 112.28236 106.6926 106.9366   5.5897789 -0.2440563282
#> Oct 2018 115.41162 107.9129 107.1741   7.4986863  0.7388293101
#> Nov 2018 109.85751 107.2445 107.3673   2.6130640 -0.1228820216
#> Dec 2018 103.52884 107.3040 107.5048  -3.7751434 -0.2008278423
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.648
#>  Seasonal     49.383
#>  Irregular     0.419
#>  TD & Hol.     0.039
#>  Others       50.184
#>  Total       101.672
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.011
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.970
#>  f-test on i (seasonal dummies)         0.887
#>  Residual seasonality (entire series)   0.967
#>  Residual seasonality (last 3 years)    0.931
#>  f-test on sa (td)                      0.974
#>  f-test on i (td)                       0.999
#> 
#> 
#> Additional output variables
#> 
#> $sa1$TramoSeats
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.3555      0.255
#> Phi(2)      0.1344      0.130
#> Theta(1)   -0.1867      0.255
#> BTheta(1)  -0.6117      0.047
#> 
#>                Estimate Std. Error
#> Monday       -0.0022678      0.002
#> Tuesday       0.0006417      0.002
#> Wednesday     0.0032444      0.002
#> Thursday     -0.0055635      0.002
#> Friday        0.0033124      0.002
#> Saturday      0.0019048      0.002
#> Easter [6]   -0.0106133      0.003
#> LS (11-2008) -0.0795302      0.013
#> LS (1-2009)  -0.0711422      0.013
#> AO (8-2001)   0.0447823      0.013
#> 
#> 
#> Residual standard error: 0.01594 on 308 degrees of freedom
#> Log likelihood = 875.5, aic =  1262 aicc =  1263, bic(corrected for length) = -8.027
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 + 0.355520 B + 0.134366 B^2 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.186667 B - 0.611719 B^12 + 0.114188 B^13 
#> 
#> 
#> SA
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.314513 B + 0.340411 B^2 
#> Innovation variance:  0.4651916 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.040093 B - 0.959907 B^2 
#> Innovation variance:  0.04852415 
#> 
#> Seasonal
#> AR :  1 + 0.355520 B + 0.134366 B^2 
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 0.715509 B + 0.458526 B^2 + 0.307968 B^3 + 0.130783 B^4 - 0.050030 B^5 - 0.216796 B^6 - 0.353788 B^7 - 0.443393 B^8 - 0.467164 B^9 - 0.373994 B^10 - 0.163801 B^11 - 0.411499 B^12 - 0.132321 B^13 
#> Innovation variance:  0.1608041 
#> 
#> Irregular
#> Innovation variance:  0.2049349 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2017  97.4 100.9340 100.9645 0.9649872 0.9996976
#> Feb 2017  97.5 100.7675 101.2596 0.9675741 0.9951403
#> Mar 2017 112.0 102.0785 101.7061 1.0971949 1.0036614
#> Apr 2017 103.0 101.8555 102.1727 1.0112369 0.9968946
#> May 2017 100.4 103.0213 102.6101 0.9745555 1.0040080
#> Jun 2017 111.2 102.8013 102.9761 1.0816988 0.9983021
#> Jul 2017 103.4 103.3075 103.3373 1.0008951 0.9997120
#> Aug 2017  79.3 103.7806 103.7465 0.7641122 1.0003281
#> Sep 2017 109.7 103.7197 104.2758 1.0576581 0.9946676
#> Oct 2017 114.0 105.6630 104.8765 1.0789015 1.0074997
#> Nov 2017 107.7 105.4786 105.1828 1.0210600 1.0028123
#> Dec 2017 101.4 104.9682 105.2367 0.9660069 0.9974489
#> 
#> Forecasts:
#>               y_f     sa_f      t_f       s_f       i_f
#> Jan 2018 102.2106 105.3440 105.3449 0.9702559 0.9999915
#> Feb 2018 102.3328 105.5134 105.5143 0.9698565 0.9999915
#> Mar 2018 114.1449 105.6830 105.6839 1.0800682 0.9999915
#> Apr 2018 108.0526 105.8530 105.8539 1.0207800 0.9999915
#> May 2018 102.6903 106.0232 106.0241 0.9685646 0.9999915
#> Jun 2018 115.9079 106.1936 106.1945 1.0914767 0.9999915
#> Jul 2018 106.2462 106.3644 106.3653 0.9988891 0.9999915
#> Aug 2018  81.4640 106.5354 106.5363 0.7646660 0.9999915
#> Sep 2018 112.8928 106.7067 106.7076 1.0579730 0.9999915
#> Oct 2018 115.3561 106.8783 106.8792 1.0793227 0.9999915
#> Nov 2018 109.2863 107.0501 107.0510 1.0208889 0.9999915
#> Dec 2018 103.2836 107.2222 107.2231 0.9632669 0.9999915
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.179
#>  Seasonal     66.777
#>  Irregular     0.105
#>  TD & Hol.     0.067
#>  Others       30.524
#>  Total        99.651
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.187
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.977
#>  f-test on sa (td)                      0.930
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables
#> 
#> 

# }