Functions around 'jSA' objects

get_dictionary returns the indicators that can be extracted from "jSA" objects, get_indicators extract a list of indicators and jSA2R returns the corresponding "SA".

get_jspec(x, ...)

get_dictionary(x)

get_indicators(x, ...)

jSA2R(x, userdefined = NULL)

Arguments

x
a "jSA" object.

...
characters containing the names of the indicators to extract.

userdefined
userdefined vector with characters for additional output variables (see user_defined_variables). Only used for "SA" objects.

Value

get_dictionary a vector of characters, get_indicators returns a list containing the indicators that are extracted, jSA2R returns a "SA" or a "regarima" object and get_jspec returns a Java object.

Details

A "jSA" object is a list with three elements:

"result": the Java object with the results of a seasonal adjustment or a pre-adjustment method.

"spec": the Java object with the specification of a seasonal adjustment or a pre-adjustment method.

"dictionary": the Java object with dictionnary of a seasonal adjustment or a pre-adjustment method. In particular, it contains all the user-defined regressors.

get_dictionary returns the list of indicators that can be extracted from a jSA object by the function get_indicators.

jSA2R returns the corresponding formatted seasonal adjustment ("SA" object) or RegARIMA ("regarima" object) model.

get_jspec returns the Java object that contains the specification from an object "jSA", "X13", "TRAMO_SEATS" or "sa_item".

Examples

myseries <- ipi_c_eu[, "FR"]
mysa <- jx13(myseries, spec = "RSA5c")
get_dictionary(mysa)

#>   [1] "y"                                              
#>   [2] "y_f"                                            
#>   [3] "t"                                              
#>   [4] "t_f"                                            
#>   [5] "sa"                                             
#>   [6] "sa_f"                                           
#>   [7] "s"                                              
#>   [8] "s_f"                                            
#>   [9] "i"                                              
#>  [10] "i_f"                                            
#>  [11] "mode"                                           
#>  [12] "preprocessing.model.span.start"                 
#>  [13] "preprocessing.model.span.end"                   
#>  [14] "preprocessing.model.span.n"                     
#>  [15] "preprocessing.model.espan.start"                
#>  [16] "preprocessing.model.espan.end"                  
#>  [17] "preprocessing.model.espan.n"                    
#>  [18] "preprocessing.model.log"                        
#>  [19] "preprocessing.model.adjust"                     
#>  [20] "preprocessing.model.y"                          
#>  [21] "preprocessing.model.y_f"                        
#>  [22] "preprocessing.model.y_ef"                       
#>  [23] "preprocessing.model.yc"                         
#>  [24] "preprocessing.model.yc_f"                       
#>  [25] "preprocessing.model.yc_ef"                      
#>  [26] "preprocessing.model.l"                          
#>  [27] "preprocessing.model.y_lin"                      
#>  [28] "preprocessing.model.y_lin_f"                    
#>  [29] "preprocessing.model.ycal"                       
#>  [30] "preprocessing.model.ycal_f"                     
#>  [31] "preprocessing.model.det"                        
#>  [32] "preprocessing.model.det_f"                      
#>  [33] "preprocessing.model.l_f"                        
#>  [34] "preprocessing.model.l_b"                        
#>  [35] "preprocessing.model.cal"                        
#>  [36] "preprocessing.model.cal_f"                      
#>  [37] "preprocessing.model.tde"                        
#>  [38] "preprocessing.model.tde_f"                      
#>  [39] "preprocessing.model.mhe"                        
#>  [40] "preprocessing.model.mhe_f"                      
#>  [41] "preprocessing.model.ee"                         
#>  [42] "preprocessing.model.ee_f"                       
#>  [43] "preprocessing.model.omhe"                       
#>  [44] "preprocessing.model.omhe_f"                     
#>  [45] "preprocessing.model.out(*)"                     
#>  [46] "preprocessing.model.out_f"                      
#>  [47] "preprocessing.model.out_i"                      
#>  [48] "preprocessing.model.out_i_f"                    
#>  [49] "preprocessing.model.out_t"                      
#>  [50] "preprocessing.model.out_t_f"                    
#>  [51] "preprocessing.model.out_s"                      
#>  [52] "preprocessing.model.out_s_f"                    
#>  [53] "preprocessing.model.reg"                        
#>  [54] "preprocessing.model.reg_f"                      
#>  [55] "preprocessing.model.reg_t"                      
#>  [56] "preprocessing.model.reg_t_f"                    
#>  [57] "preprocessing.model.reg_s"                      
#>  [58] "preprocessing.model.reg_s_f"                    
#>  [59] "preprocessing.model.reg_i"                      
#>  [60] "preprocessing.model.reg_i_f"                    
#>  [61] "preprocessing.model.reg_sa"                     
#>  [62] "preprocessing.model.reg_sa_f"                   
#>  [63] "preprocessing.model.reg_y"                      
#>  [64] "preprocessing.model.reg_y_f"                    
#>  [65] "preprocessing.model.reg_u"                      
#>  [66] "preprocessing.model.reg_u_f"                    
#>  [67] "preprocessing.model.fullresiduals"              
#>  [68] "preprocessing.model.lp"                         
#>  [69] "preprocessing.model.ntd"                        
#>  [70] "preprocessing.model.nmh"                        
#>  [71] "preprocessing.model.td(*)"                      
#>  [72] "preprocessing.model.easter"                     
#>  [73] "preprocessing.model.nout"                       
#>  [74] "preprocessing.model.noutao"                     
#>  [75] "preprocessing.model.noutls"                     
#>  [76] "preprocessing.model.nouttc"                     
#>  [77] "preprocessing.model.noutso"                     
#>  [78] "preprocessing.model.coefficients"               
#>  [79] "preprocessing.model.description"                
#>  [80] "preprocessing.model.covar"                      
#>  [81] "preprocessing.model.pcovar"                     
#>  [82] "preprocessing.model.fcasts(?)"                  
#>  [83] "preprocessing.model.bcasts(?)"                  
#>  [84] "preprocessing.model.lin_fcasts(?)"              
#>  [85] "preprocessing.model.lin_bcasts(?)"              
#>  [86] "preprocessing.model.efcasts(?)"                 
#>  [87] "preprocessing.arima.parameters"                 
#>  [88] "preprocessing.arima.p"                          
#>  [89] "preprocessing.arima.d"                          
#>  [90] "preprocessing.arima.q"                          
#>  [91] "preprocessing.arima.bp"                         
#>  [92] "preprocessing.arima.bd"                         
#>  [93] "preprocessing.arima.bq"                         
#>  [94] "preprocessing.likelihood.neffectiveobs"         
#>  [95] "preprocessing.likelihood.np"                    
#>  [96] "preprocessing.likelihood.logvalue"              
#>  [97] "preprocessing.likelihood.adjustedlogvalue"      
#>  [98] "preprocessing.likelihood.ssqerr"                
#>  [99] "preprocessing.likelihood.aic"                   
#> [100] "preprocessing.likelihood.aicc"                  
#> [101] "preprocessing.likelihood.bic"                   
#> [102] "preprocessing.likelihood.bicc"                  
#> [103] "preprocessing.likelihood.ser"                   
#> [104] "preprocessing.likelihood.ser-ml"                
#> [105] "preprocessing.residuals.res"                    
#> [106] "preprocessing.residuals.mean"                   
#> [107] "preprocessing.residuals.skewness"               
#> [108] "preprocessing.residuals.kurtosis"               
#> [109] "preprocessing.residuals.dh"                     
#> [110] "preprocessing.residuals.lb"                     
#> [111] "preprocessing.residuals.lb2"                    
#> [112] "preprocessing.residuals.seaslb"                 
#> [113] "preprocessing.residuals.bp"                     
#> [114] "preprocessing.residuals.bp2"                    
#> [115] "preprocessing.residuals.seasbp"                 
#> [116] "preprocessing.residuals.nruns"                  
#> [117] "preprocessing.residuals.lruns"                  
#> [118] "mstats.M(*)"                                    
#> [119] "mstats.Q"                                       
#> [120] "mstats.Q-M2"                                    
#> [121] "decomposition.a1"                               
#> [122] "decomposition.a1a"                              
#> [123] "decomposition.a1b"                              
#> [124] "decomposition.a6"                               
#> [125] "decomposition.a7"                               
#> [126] "decomposition.a8"                               
#> [127] "decomposition.a8t"                              
#> [128] "decomposition.a8s"                              
#> [129] "decomposition.a8i"                              
#> [130] "decomposition.a9"                               
#> [131] "decomposition.a9sa"                             
#> [132] "decomposition.a9u"                              
#> [133] "decomposition.a9ser"                            
#> [134] "decomposition.b1"                               
#> [135] "decomposition.b2"                               
#> [136] "decomposition.b3"                               
#> [137] "decomposition.b4"                               
#> [138] "decomposition.b5"                               
#> [139] "decomposition.b6"                               
#> [140] "decomposition.b7"                               
#> [141] "decomposition.b8"                               
#> [142] "decomposition.b9"                               
#> [143] "decomposition.b10"                              
#> [144] "decomposition.b11"                              
#> [145] "decomposition.b12"                              
#> [146] "decomposition.b13"                              
#> [147] "decomposition.b14"                              
#> [148] "decomposition.b15"                              
#> [149] "decomposition.b16"                              
#> [150] "decomposition.b17"                              
#> [151] "decomposition.b18"                              
#> [152] "decomposition.b19"                              
#> [153] "decomposition.b20"                              
#> [154] "decomposition.c1"                               
#> [155] "decomposition.c2"                               
#> [156] "decomposition.c3"                               
#> [157] "decomposition.c4"                               
#> [158] "decomposition.c5"                               
#> [159] "decomposition.c6"                               
#> [160] "decomposition.c7"                               
#> [161] "decomposition.c8"                               
#> [162] "decomposition.c9"                               
#> [163] "decomposition.c10"                              
#> [164] "decomposition.c11"                              
#> [165] "decomposition.c12"                              
#> [166] "decomposition.c13"                              
#> [167] "decomposition.c14"                              
#> [168] "decomposition.c15"                              
#> [169] "decomposition.c16"                              
#> [170] "decomposition.c17"                              
#> [171] "decomposition.c18"                              
#> [172] "decomposition.c19"                              
#> [173] "decomposition.c20"                              
#> [174] "decomposition.d1"                               
#> [175] "decomposition.d2"                               
#> [176] "decomposition.d3"                               
#> [177] "decomposition.d4"                               
#> [178] "decomposition.d5"                               
#> [179] "decomposition.d6"                               
#> [180] "decomposition.d7"                               
#> [181] "decomposition.d8"                               
#> [182] "decomposition.d9"                               
#> [183] "decomposition.d10"                              
#> [184] "decomposition.d10a"                             
#> [185] "decomposition.d10b"                             
#> [186] "decomposition.d11"                              
#> [187] "decomposition.d11a"                             
#> [188] "decomposition.d12"                              
#> [189] "decomposition.d12a"                             
#> [190] "decomposition.d13"                              
#> [191] "decomposition.d14"                              
#> [192] "decomposition.d15"                              
#> [193] "decomposition.d16"                              
#> [194] "decomposition.d16a"                             
#> [195] "decomposition.d16b"                             
#> [196] "decomposition.d18"                              
#> [197] "decomposition.d19"                              
#> [198] "decomposition.d20"                              
#> [199] "decomposition.e1"                               
#> [200] "decomposition.e2"                               
#> [201] "decomposition.e3"                               
#> [202] "decomposition.e11"                              
#> [203] "decomposition.d9filter"                         
#> [204] "decomposition.slen"                             
#> [205] "decomposition.d12filter"                        
#> [206] "decomposition.tlen"                             
#> [207] "diagnostics.qs"                                 
#> [208] "diagnostics.ftest"                              
#> [209] "diagnostics.qs.on.i"                            
#> [210] "diagnostics.ftest.on.i"                         
#> [211] "diagnostics.combined.all.kruskalwallis"         
#> [212] "diagnostics.combined.all.stable"                
#> [213] "diagnostics.combined.all.evolutive"             
#> [214] "diagnostics.combined.all.summary"               
#> [215] "diagnostics.combined.all.stable.ssm"            
#> [216] "diagnostics.combined.all.stable.ssr"            
#> [217] "diagnostics.combined.all.stable.ssq"            
#> [218] "diagnostics.combined.all.evolutive.ssm"         
#> [219] "diagnostics.combined.all.evolutive.ssr"         
#> [220] "diagnostics.combined.all.evolutive.ssq"         
#> [221] "diagnostics.combined.end.kruskalwallis"         
#> [222] "diagnostics.combined.end.stable"                
#> [223] "diagnostics.combined.end.evolutive"             
#> [224] "diagnostics.combined.end.summary"               
#> [225] "diagnostics.combined.end.stable.ssm"            
#> [226] "diagnostics.combined.end.stable.ssr"            
#> [227] "diagnostics.combined.end.stable.ssq"            
#> [228] "diagnostics.combined.end.evolutive.ssm"         
#> [229] "diagnostics.combined.end.evolutive.ssr"         
#> [230] "diagnostics.combined.end.evolutive.ssq"         
#> [231] "diagnostics.combined.residual.all.kruskalwallis"
#> [232] "diagnostics.combined.residual.all.stable"       
#> [233] "diagnostics.combined.residual.all.evolutive"    
#> [234] "diagnostics.combined.residual.all.summary"      
#> [235] "diagnostics.combined.residual.all.stable.ssm"   
#> [236] "diagnostics.combined.residual.all.stable.ssr"   
#> [237] "diagnostics.combined.residual.all.stable.ssq"   
#> [238] "diagnostics.combined.residual.all.evolutive.ssm"
#> [239] "diagnostics.combined.residual.all.evolutive.ssr"
#> [240] "diagnostics.combined.residual.all.evolutive.ssq"
#> [241] "diagnostics.combined.residual.end.kruskalwallis"
#> [242] "diagnostics.combined.residual.end.stable"       
#> [243] "diagnostics.combined.residual.end.evolutive"    
#> [244] "diagnostics.combined.residual.end.summary"      
#> [245] "diagnostics.combined.residual.end.stable.ssm"   
#> [246] "diagnostics.combined.residual.end.stable.ssr"   
#> [247] "diagnostics.combined.residual.end.stable.ssq"   
#> [248] "diagnostics.combined.residual.end.evolutive.ssm"
#> [249] "diagnostics.combined.residual.end.evolutive.ssr"
#> [250] "diagnostics.combined.residual.end.evolutive.ssq"
#> [251] "diagnostics.residual.all"                       
#> [252] "diagnostics.residual.end"                       
#> [253] "diagnostics.residualtd"                         
#> [254] "diagnostics.residualtd.on.i"                    
#> [255] "diagnostics.variancedecomposition"              
#> [256] "diagnostics.logstat"                            
#> [257] "diagnostics.levelstat"                          
#> [258] "diagnostics.fcast-insample-mean"                
#> [259] "diagnostics.fcast-outsample-mean"               
#> [260] "diagnostics.fcast-outsample-variance"           
#> [261] "diagnostics.seas-lin-f"                         
#> [262] "diagnostics.seas-lin-qs"                        
#> [263] "diagnostics.seas-lin-kw"                        
#> [264] "diagnostics.seas-lin-friedman"                  
#> [265] "diagnostics.seas-lin-periodogram"               
#> [266] "diagnostics.seas-lin-spectralpeaks"             
#> [267] "diagnostics.seas-si-combined"                   
#> [268] "diagnostics.seas-si-evolutive"                  
#> [269] "diagnostics.seas-si-stable"                     
#> [270] "diagnostics.seas-res-f"                         
#> [271] "diagnostics.seas-res-qs"                        
#> [272] "diagnostics.seas-res-kw"                        
#> [273] "diagnostics.seas-res-friedman"                  
#> [274] "diagnostics.seas-res-periodogram"               
#> [275] "diagnostics.seas-res-spectralpeaks"             
#> [276] "diagnostics.seas-res-combined"                  
#> [277] "diagnostics.seas-res-combined3"                 
#> [278] "diagnostics.seas-res-evolutive"                 
#> [279] "diagnostics.seas-res-stable"                    
#> [280] "diagnostics.seas-i-f"                           
#> [281] "diagnostics.seas-i-qs"                          
#> [282] "diagnostics.seas-i-kw"                          
#> [283] "diagnostics.seas-i-periodogram"                 
#> [284] "diagnostics.seas-i-spectralpeaks"               
#> [285] "diagnostics.seas-i-combined"                    
#> [286] "diagnostics.seas-i-combined3"                   
#> [287] "diagnostics.seas-i-evolutive"                   
#> [288] "diagnostics.seas-i-stable"                      
#> [289] "diagnostics.seas-sa-f"                          
#> [290] "diagnostics.seas-sa-qs"                         
#> [291] "diagnostics.seas-sa-kw"                         
#> [292] "diagnostics.seas-sa-friedman"                   
#> [293] "diagnostics.seas-sa-periodogram"                
#> [294] "diagnostics.seas-sa-spectralpeaks"              
#> [295] "diagnostics.seas-sa-combined"                   
#> [296] "diagnostics.seas-sa-combined3"                  
#> [297] "diagnostics.seas-sa-evolutive"                  
#> [298] "diagnostics.seas-sa-stable"                     
#> [299] "diagnostics.seas-sa-ac1"                        
#> [300] "diagnostics.td-sa-all"                          
#> [301] "diagnostics.td-sa-last"                         
#> [302] "diagnostics.td-i-all"                           
#> [303] "diagnostics.td-i-last"                          
#> [304] "diagnostics.td-res-all"                         
#> [305] "diagnostics.td-res-last"                        
#> [306] "diagnostics.ic-ratio-henderson"                 
#> [307] "diagnostics.ic-ratio"                           
#> [308] "diagnostics.msr-global"                         
#> [309] "diagnostics.msr(*)"                             

get_indicators(mysa, "decomposition.b2", "decomposition.d10")

#> $decomposition.b2
#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1990                                                              79.97462
#> 1991  79.14862  79.04963  78.99237  78.91587  78.84040  78.83051  78.79194
#> 1992  78.39537  78.30364  78.18578  77.98102  77.83824  77.71292  77.49157
#> 1993  75.63152  75.42812  75.23374  75.02702  74.68760  74.45731  74.49803
#> 1994  75.58056  75.80685  76.07165  76.43317  76.90807  77.48759  78.00594
#> 1995  79.67865  79.84856  79.96898  80.08516  80.15470  80.14449  80.01277
#> 1996  79.87305  79.93776  80.00881  80.06726  80.09337  80.05535  80.03103
#> 1997  81.78207  82.14395  82.54838  83.06921  83.64832  84.18463  84.77391
#> 1998  87.18772  87.45463  87.64767  87.82840  88.02275  88.19737  88.26743
#> 1999  89.12811  89.29357  89.52737  89.85595  90.12343  90.44858  90.91915
#> 2000  93.15528  93.46995  93.73492  93.99584  94.34625  94.67860  94.85021
#> 2001  96.01423  96.19393  96.34311  96.32015  96.17443  95.83465  95.52199
#> 2002  94.85331  94.72636  94.62625  94.54485  94.45775  94.33409  94.19881
#> 2003  93.16948  92.93725  92.74572  92.72874  92.74192  92.76649  92.82594
#> 2004  93.90506  94.09264  94.33509  94.56694  94.61034  94.70668  95.02891
#> 2005  95.34220  95.23999  95.27797  95.20535  95.17711  95.34188  95.35981
#> 2006  95.73987  95.92935  96.03012  96.16480  96.35532  96.37650  96.31721
#> 2007  97.11664  97.40735  97.49180  97.54672  97.68144  97.69115  97.71251
#> 2008  97.29330  96.97743  96.67454  96.24924  95.81061  95.50841  95.21348
#> 2009  93.50425  93.61072  93.91161  94.32009  94.73157  95.01935  95.19477
#> 2010  97.93777  98.27051  98.50321  98.73558  98.88051  99.19761  99.85763
#> 2011 101.50507 101.66585 101.78420 101.91088 102.10923 102.27491 102.22593
#> 2012 101.29138 101.34710 101.27825 100.99088 100.65898 100.24148  99.73959
#> 2013  99.16473  98.96056  98.74540  98.75089  98.78916  98.77496  98.82147
#> 2014  98.53073  98.43840  98.46187  98.44796  98.35148  98.35856  98.42127
#> 2015  98.77256  98.89925  99.16798  99.34505  99.64424  99.90024 100.08318
#> 2016 100.55658 100.51763 100.49617 100.41641 100.33403 100.42854 100.60895
#> 2017 101.51210 101.77592 101.95835 102.37202 102.97272 103.33969 103.42918
#> 2018 104.09336 104.32718 104.33023 104.26352 104.17977 104.04852 104.05398
#> 2019 104.71826 104.64668 104.63213 104.72727 104.59758 104.33984 104.16076
#> 2020 103.13056 102.99825 102.93924 102.79111 102.73058 102.82356          
#>            Aug       Sep       Oct       Nov       Dec
#> 1990  79.88637  79.75331  79.60255  79.27117  79.14708
#> 1991  78.66279  78.51635  78.50491  78.66901  78.62753
#> 1992  77.21611  76.86421  76.44879  76.13205  75.86018
#> 1993  74.54497  74.59243  74.70886  74.91183  75.25456
#> 1994  78.45504  78.89075  79.25911  79.44773  79.51012
#> 1995  79.84336  79.75705  79.73598  79.68898  79.75737
#> 1996  80.16408  80.41654  80.74690  81.13476  81.47006
#> 1997  85.35393  85.81787  86.14614  86.50563  86.86130
#> 1998  88.28782  88.34363  88.46148  88.60545  88.87151
#> 1999  91.32505  91.75361  92.21914  92.69148  92.96897
#> 2000  95.12216  95.53473  95.78038  95.72474  95.78908
#> 2001  95.37372  95.27617  95.21538  95.02681  94.89767
#> 2002  94.10017  93.92435  93.80742  93.77313  93.48948
#> 2003  92.83068  92.78754  92.82196  93.07753  93.53793
#> 2004  95.27492  95.28723  95.27325  95.39017  95.45809
#> 2005  95.29182  95.39824  95.50294  95.49433  95.57216
#> 2006  96.43441  96.65738  96.79719  96.80975  96.85173
#> 2007  97.78121  97.73813  97.77959  97.82212  97.61691
#> 2008  94.90751  94.53212  94.05450  93.71311  93.57891
#> 2009  95.38081  95.77379  96.31931  96.91942  97.47400
#> 2010 100.60105 101.20657 101.58272 101.66742 101.51752
#> 2011 101.87558 101.53650 101.40151 101.21265 101.17229
#> 2012  99.52277  99.44490  99.30087  99.25040  99.24269
#> 2013  98.89299  98.91060  98.91831  98.90678  98.72517
#> 2014  98.40316  98.40589  98.45685  98.48577  98.68098
#> 2015 100.17191 100.13409 100.15728 100.44506 100.63478
#> 2016 100.75358 100.98357 101.12819 101.16486 101.27517
#> 2017 103.50553 103.72541 103.90457 103.77602 103.76088
#> 2018 104.26155 104.30166 104.36675 104.79180 104.89003
#> 2019 103.95965 103.78985 103.61911 103.32917 103.20402
#> 2020                                                  
#> 
#> $decomposition.d10
#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1990 0.9644547 0.9714325 1.0762842 1.0041756 0.9698028 1.0954001 0.9898772
#> 1991 0.9652359 0.9703241 1.0748282 1.0043128 0.9701846 1.0977884 0.9894657
#> 1992 0.9679342 0.9679487 1.0724116 1.0044563 0.9696323 1.1012546 0.9892535
#> 1993 0.9701194 0.9643322 1.0707193 1.0049583 0.9678437 1.1058002 0.9895270
#> 1994 0.9711967 0.9618657 1.0693752 1.0049716 0.9654400 1.1097305 0.9913342
#> 1995 0.9707449 0.9608291 1.0689122 1.0044753 0.9639580 1.1121288 0.9941917
#> 1996 0.9701204 0.9612747 1.0687983 1.0029850 0.9638816 1.1123414 0.9976440
#> 1997 0.9692164 0.9613130 1.0699359 1.0019306 0.9647918 1.1115761 1.0002197
#> 1998 0.9673060 0.9625328 1.0727944 1.0005434 0.9659114 1.1096601 1.0025046
#> 1999 0.9661874 0.9646965 1.0764350 0.9988820 0.9666756 1.1080544 1.0043854
#> 2000 0.9651029 0.9675089 1.0820650 0.9988429 0.9652574 1.1054602 1.0063410
#> 2001 0.9644688 0.9694746 1.0869853 1.0007530 0.9630149 1.1043665 1.0073828
#> 2002 0.9639129 0.9703722 1.0912496 1.0041276 0.9612648 1.1041208 1.0064592
#> 2003 0.9651601 0.9696085 1.0916213 1.0069277 0.9628082 1.1056341 1.0038122
#> 2004 0.9663075 0.9680630 1.0909302 1.0087409 0.9646354 1.1072559 1.0022105
#> 2005 0.9684257 0.9662826 1.0883481 1.0085907 0.9670721 1.1094631 1.0024004
#> 2006 0.9702722 0.9665237 1.0861748 1.0066444 0.9676616 1.1097367 1.0040599
#> 2007 0.9722191 0.9672872 1.0840815 1.0045275 0.9675832 1.1091177 1.0064931
#> 2008 0.9738124 0.9706968 1.0834729 1.0033964 0.9642393 1.1056935 1.0102485
#> 2009 0.9752336 0.9735429 1.0833882 1.0039083 0.9597313 1.1015408 1.0150948
#> 2010 0.9759092 0.9767347 1.0841452 1.0048317 0.9549416 1.0976489 1.0189895
#> 2011 0.9744818 0.9783557 1.0856236 1.0077979 0.9508887 1.0941373 1.0229273
#> 2012 0.9713780 0.9806077 1.0875245 1.0106084 0.9471420 1.0920065 1.0258097
#> 2013 0.9683207 0.9804459 1.0878107 1.0144037 0.9459366 1.0898123 1.0276119
#> 2014 0.9667386 0.9782536 1.0873919 1.0156572 0.9491468 1.0880471 1.0252101
#> 2015 0.9655882 0.9738288 1.0856571 1.0152665 0.9549415 1.0863193 1.0223335
#> 2016 0.9659380 0.9702547 1.0838374 1.0120864 0.9616657 1.0851689 1.0193265
#> 2017 0.9668104 0.9675213 1.0818391 1.0088257 0.9670456 1.0837093 1.0188957
#> 2018 0.9675564 0.9665172 1.0815811 1.0055891 0.9712372 1.0826396 1.0188539
#> 2019 0.9670835 0.9669790 1.0816991 1.0040065 0.9727806 1.0820762 1.0202668
#>            Aug       Sep       Oct       Nov       Dec
#> 1990 0.7077229 1.0630958 1.1030932 1.0347723 1.0201040
#> 1991 0.7074773 1.0628806 1.1021286 1.0334840 1.0223099
#> 1992 0.7073151 1.0626788 1.1003074 1.0320550 1.0260625
#> 1993 0.7075246 1.0622549 1.0981933 1.0300526 1.0299468
#> 1994 0.7079630 1.0624356 1.0970389 1.0270519 1.0322240
#> 1995 0.7090349 1.0621382 1.0964988 1.0246286 1.0319500
#> 1996 0.7105314 1.0622110 1.0969351 1.0227299 1.0293472
#> 1997 0.7135033 1.0610156 1.0968049 1.0226596 1.0250700
#> 1998 0.7169127 1.0600752 1.0964715 1.0225404 1.0190716
#> 1999 0.7208317 1.0594015 1.0942980 1.0233521 1.0115071
#> 2000 0.7234503 1.0606436 1.0930737 1.0235810 1.0030682
#> 2001 0.7249512 1.0632095 1.0919944 1.0229975 0.9950834
#> 2002 0.7249346 1.0681810 1.0912202 1.0213470 0.9886186
#> 2003 0.7249415 1.0731995 1.0891056 1.0198842 0.9848760
#> 2004 0.7253249 1.0767398 1.0871555 1.0187975 0.9824805
#> 2005 0.7269456 1.0775126 1.0839436 1.0188162 0.9812265
#> 2006 0.7306867 1.0772605 1.0807742 1.0209987 0.9781708
#> 2007 0.7356957 1.0755284 1.0773262 1.0234475 0.9744806
#> 2008 0.7411604 1.0730002 1.0758646 1.0254440 0.9708594
#> 2009 0.7462273 1.0710840 1.0733465 1.0257610 0.9687856
#> 2010 0.7513335 1.0699289 1.0722147 1.0243324 0.9675168
#> 2011 0.7559329 1.0693953 1.0708716 1.0209949 0.9672789
#> 2012 0.7604925 1.0676525 1.0719003 1.0177180 0.9673293
#> 2013 0.7647995 1.0665090 1.0710511 1.0159277 0.9687475
#> 2014 0.7683126 1.0642692 1.0730056 1.0174651 0.9706789
#> 2015 0.7708120 1.0613500 1.0745442 1.0207859 0.9723384
#> 2016 0.7713617 1.0576105 1.0790199 1.0247424 0.9720864
#> 2017 0.7713717 1.0549979 1.0823394 1.0273251 0.9700639
#> 2018 0.7701564 1.0529094 1.0875719 1.0287700 0.9665878
#> 2019 0.7691657 1.0521801 1.0908568 1.0286887 0.9634841
#> 

# To convert to the R object
jSA2R(mysa)

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.02578      0.118
#> Phi(2)     0.15940      0.078
#> Theta(1)  -0.51733      0.113
#> BTheta(1) -0.70304      0.041
#> 
#>               Estimate Std. Error
#> Monday        0.004967      0.002
#> Tuesday       0.008415      0.002
#> Wednesday     0.010614      0.002
#> Thursday      0.002060      0.002
#> Friday        0.008274      0.002
#> Saturday     -0.016328      0.002
#> Easter [1]   -0.021145      0.004
#> AO (5-2011)   0.128134      0.017
#> LS (11-2008) -0.087008      0.017
#> LS (1-2009)  -0.068435      0.017
#> 
#> 
#> Residual standard error: 0.02041 on 332 degrees of freedom
#> Log likelihood = 853.7, aic =  1528 aicc =  1530, bic(corrected for length) = -7.547
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.097
#> M(2)    0.053
#> M(3)    0.995
#> M(4)    0.421
#> M(5)    0.950
#> M(6)    0.154
#> M(7)    0.074
#> M(8)    0.211
#> M(9)    0.064
#> M(10)   0.267
#> M(11)   0.250
#> Q       0.322
#> Q-M2    0.355
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.1945 105.1440 0.9876940 1.0004808
#> Feb 2019 101.9 106.3206 105.5014 0.9584217 1.0077654
#> Mar 2019 111.0 104.7345 105.7233 1.0598229 0.9906473
#> Apr 2019 107.4 106.1448 105.7281 1.0118257 1.0039413
#> May 2019 105.5 106.2039 105.5978 0.9933724 1.0057396
#> Jun 2019 105.8 101.1898 105.3259 1.0455597 0.9607304
#> Jul 2019 110.1 105.3543 104.8985 1.0450450 1.0043457
#> Aug 2019  78.7 102.9339 104.3791 0.7645685 0.9861536
#> Sep 2019 108.5 104.4720 103.8115 1.0385554 1.0063631
#> Oct 2019 116.8 104.8375 103.2992 1.1141052 1.0148919
#> Nov 2019 103.8 101.7212 102.9938 1.0204364 0.9876441
#> Dec 2019  97.7 101.8723 102.8805 0.9590443 0.9902000
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.79694 103.1662 102.8951 0.9867273 1.0026346
#> Feb 2020 101.02102 103.3357 102.9515 0.9776003 1.0037323
#> Mar 2020 112.10997 103.5043 102.9465 1.0831427 1.0054189
#> Apr 2020 104.50015 103.4303 102.8832 1.0103439 1.0053179
#> May 2020  96.60823 101.9351 102.7280 0.9477426 0.9922819
#> Jun 2020 112.13668 102.1960 102.5441 1.0972706 0.9966052
#> Jul 2020 106.78914 102.3841 102.4931 1.0430249 0.9989358
#> Aug 2020  76.62863 102.6907 102.5880 0.7462084 1.0010004
#> Sep 2020 110.85790 103.4529 102.7454 1.0715783 1.0068865
#> Oct 2020 111.29327 102.4307 102.8914 1.0865224 0.9955221
#> Nov 2020 104.21972 102.6701 103.0257 1.0150931 0.9965489
#> Dec 2020 101.55551 103.4907 103.1115 0.9813009 1.0036778
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.941
#>  Seasonal     62.123
#>  Irregular     0.842
#>  TD & Hol.     2.428
#>  Others       32.787
#>  Total       100.121
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.283
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          0.909
#>  qs test on i                           0.746
#>  f-test on sa (seasonal dummies)        0.974
#>  f-test on i (seasonal dummies)         0.947
#>  Residual seasonality (entire series)   0.932
#>  Residual seasonality (last 3 years)    0.940
#>  f-test on sa (td)                      0.985
#>  f-test on i (td)                       0.983
#> 
#> 
#> Additional output variables