RegARIMA model specification, pre-adjustment in X13

Function to create (and/or modify) a c("regarima_spec","X13") class object with the RegARIMA model specification for the X13 method. The object can be created from a predefined 'JDemetra+' model specification (a character), a previous specification (c("regarima_spec","X13") object) or a X13 RegARIMA model (c("regarima","X13")).

regarima_spec_x13(
  spec = c("RG5c", "RG0", "RG1", "RG2c", "RG3", "RG4c"),
  preliminary.check = NA,
  estimate.from = NA_character_,
  estimate.to = NA_character_,
  estimate.first = NA_integer_,
  estimate.last = NA_integer_,
  estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_,
  estimate.tol = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.adjust = c(NA, "None", "LeapYear", "LengthOfPeriod"),
  transform.aicdiff = NA_integer_,
  usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA,
  usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA,
  usrdef.varEnabled = NA,
  usrdef.var = NA,
  usrdef.varType = NA,
  usrdef.varCoef = NA,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined", "None"),
  tradingdays.autoadjust = NA,
  tradingdays.leapyear = c(NA, "LeapYear", "LengthOfPeriod", "None"),
  tradingdays.stocktd = NA_integer_,
  tradingdays.test = c(NA, "Remove", "Add", "None"),
  easter.enabled = NA,
  easter.julian = NA,
  easter.duration = NA_integer_,
  easter.test = c(NA, "Add", "Remove", "None"),
  outlier.enabled = NA,
  outlier.from = NA_character_,
  outlier.to = NA_character_,
  outlier.first = NA_integer_,
  outlier.last = NA_integer_,
  outlier.exclFirst = NA_integer_,
  outlier.exclLast = NA_integer_,
  outlier.ao = NA,
  outlier.tc = NA,
  outlier.ls = NA,
  outlier.so = NA,
  outlier.usedefcv = NA,
  outlier.cv = NA_integer_,
  outlier.method = c(NA, "AddOne", "AddAll"),
  outlier.tcrate = NA_integer_,
  automdl.enabled = NA,
  automdl.acceptdefault = NA,
  automdl.cancel = NA_integer_,
  automdl.ub1 = NA_integer_,
  automdl.ub2 = NA_integer_,
  automdl.mixed = NA,
  automdl.balanced = NA,
  automdl.armalimit = NA_integer_,
  automdl.reducecv = NA_integer_,
  automdl.ljungboxlimit = NA_integer_,
  automdl.ubfinal = NA_integer_,
  arima.mu = NA,
  arima.p = NA_integer_,
  arima.d = NA_integer_,
  arima.q = NA_integer_,
  arima.bp = NA_integer_,
  arima.bd = NA_integer_,
  arima.bq = NA_integer_,
  arima.coefEnabled = NA,
  arima.coef = NA,
  arima.coefType = NA,
  fcst.horizon = NA_integer_
)

Arguments

spec
model specification. It can be a character of predefined 'JDemetra+' model specification (see Details), an object of class c("regarima_spec","X13") or an object of class c("regarima", "X13"). The default is "RG5c".

preliminary.check
boolean to check the quality of the input series and exclude highly problematic ones: e.g. these with a number of identical observations and/or missing values above pre-specified threshold values.

The time span of the series to be used for the estimation of the RegARIMA model coefficients (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: estimate.from, estimate.to, estimate.first, estimate.last, estimate.exclFirst and estimate.exclLast; where estimate.from and estimate.to have priority over remaining span control variables, estimate.last and estimate.first have priority over estimate.exclFirst and estimate.exclLast, and estimate.last has priority over estimate.first.

estimate.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with estimate.to.

estimate.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with estimate.from.

estimate.first
numeric specifying the number of periods considered at the beginning of the series.

estimate.last
numeric specifying the number of periods considered at the end of the series.

estimate.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with estimate.exclLast.

estimate.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with estimate.exclFirst.

estimate.tol
numeric, convergence tolerance. The absolute changes in the log-likelihood function are compared to this value to check for the convergence of the estimation iterations.

transform.function
transformation of the input series: "None" - no transformation of the series; "Log" - takes the log of the series; "Auto" - the program tests for the log-level specification.

transform.adjust
pre-adjustment of the input series for length of period or leap year effects: "None" - no adjustment; "LeapYear" - leap year effect; "LengthOfPeriod" - length of period. Modifications of this variable are taken into account only when transform.function is set to "Log".

transform.aicdiff
numeric defining the difference in AICC needed to accept no transformation when the automatic transformation selection is chosen (considered only when transform.function is set to "Auto").

Control variables for the pre-specified outliers. The pre-specified outliers are used in the model only if they are enabled (usrdef.outliersEnabled=TRUE) and the outliers' type (usrdef.outliersType) and date (usrdef.outliersDate) are provided.

usrdef.outliersEnabled
logicals. If TRUE the program uses the pre-specified outliers.

usrdef.outliersType
vector defining the outliers' type. Possible types are: ("AO") - additive, ("LS") - level shift, ("TC") - transitory change, ("SO") - seasonal outlier. E.g.: usrdef.outliersType =c("AO","AO","LS").

usrdef.outliersDate
vector defining the outliers' date. The dates should be characters in format "YYYY-MM-DD". E.g.: usrdef.outliersDate=c("2009-10-01","2005-02-01","2003-04-01").

usrdef.outliersCoef
vector providing fixed coefficients for the outliers. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined. E.g.: usrdef.outliersCoef=c(200,170,20).

Control variables for the user-defined variables:

usrdef.varEnabled
logicals. If TRUE the program uses the user-defined variables.

usrdef.var
time series (ts) or matrix of time series (mts) with the user-defined variables.

usrdef.varType
vector of character(s) defining the user-defined variables component type. Possible types are: "Undefined", "Series", "Trend", "Seasonal", "SeasonallyAdjusted", "Irregular", "Calendar". The type "Calendar" has to be used with tradingdays.option = "UserDefined" to use user-defined calendar regressors. If not specified, the program will assign the "Undefined" type.

usrdef.varCoef
vector providing fixed coefficients for the user-defined variables. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined.

tradingdays.option
defines the type of the trading days regression variables: "TradingDays" - six day-of-the-week regression variables; "WorkingDays" - one working/non-working day contrast variable; "None" - no correction for trading days and working days effects; "UserDefined" - user-defined trading days regressors (regressors have to be defined by the usrdef.var argument with usrdef.varType set to "Calendar" and usrdef.varEnabled = TRUE). "None" has also to be chosen for the "day-of-week effects" correction (tradingdays.stocktd to be modified accordingly).

tradingdays.autoadjust
logicals. If TRUE the program corrects automatically for the leap year effect. Modifications of this variable are taken into account only when transform.function is set to "Auto".

tradingdays.leapyear
option for including the leap-year effect in the model: "LeapYear" - leap year effect; "LengthOfPeriod" - length of period, "None" - no effect included. The leap-year effect can be pre-specified in the model only if the input series was not pre-adjusted (transform.adjust set to "None") and the automatic correction for the leap-year effect was not selected (tradingdays.autoadjust set to FALSE).

tradingdays.stocktd
numeric indicating the day of the month when inventories and other stock are reported (to denote the last day of the month set the variable to 31). Modifications of this variable are taken into account only when tradingdays.option is set to "None".

tradingdays.test
defines the pre-tests for the significance of the trading day regression variables based on the AICC statistics: "Add" - the trading day variables are not included in the initial regression model but can be added to the RegARIMA model after the test; "Remove" - the trading day variables belong to the initial regression model but can be removed from the RegARIMA model after the test; "None" - the trading day variables are not pre-tested and are included in the model.

easter.enabled
logicals. If TRUE the program considers the Easter effect in the model.

easter.julian
logicals. If TRUE the program uses the Julian Easter (expressed in Gregorian calendar).

easter.duration
numeric indicating the duration of the Easter effect (length in days, between 1 and 20).

easter.test
defines the pre-tests for the significance of the Easter effect based on the t-statistic (Easter effect is considered as significant if the t-statistic is greater than 1.96): "Add" - the Easter effect variable is not included in the initial regression model but can be added to the RegARIMA model after the test; "Remove" - the Easter effect variable belong to the initial regression model but can be removed from the RegARIMA model after the test; "None" - the Easter effect variable is not pre-tested and is included in the model.

outlier.enabled
logicals. If TRUE the automatic detection of outliers is enabled in the defined time span.

The time span of the series to be searched for outliers (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: outlier.from, outlier.to, outlier.first, outlier.last, outlier.exclFirst and outlier.exclLast; where outlier.from and outlier.to have priority over remaining span control variables, outlier.last and outlier.first have priority over outlier.exclFirst and outlier.exclLast, and outlier.last has priority over outlier.first.

outlier.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with outlier.to.

outlier.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with outlier.from.

outlier.first
numeric specifying the number of periods considered at the beginning of the series.

outlier.last
numeric specifying the number of periods considered at the end of the series.

outlier.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with outlier.exclLast.

outlier.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with outlier.exclFirst.

outlier.ao
logicals. If TRUE the automatic detection of additive outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.tc
logicals. If TRUE the automatic detection of transitory changes is enabled (outlier.enabled must be also set to TRUE).

outlier.ls
logicals. If TRUE the automatic detection of level shifts is enabled (outlier.enabled must be also set to TRUE).

outlier.so
logicals. If TRUE the automatic detection of seasonal outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.usedefcv
logicals. If TRUE the critical value for the outliers' detection procedure is automatically determined by the number of observations in the outlier detection time span. If FALSE the procedure uses the inputted critical value (outlier.cv).

outlier.cv
numeric. Inputted critical value for the outliers' detection procedure. The modification of this variable is taken into account only when outlier.usedefcv is set to FALSE.

outlier.method
determines how the program successively adds detected outliers to the model. At present only the AddOne method is supported.

outlier.tcrate
numeric. The rate of decay for the transitory change outlier.

automdl.enabled
logicals. If TRUE the automatic modelling of the ARIMA model is enabled. If FALSE the parameters of the ARIMA model can be specified.

Control variables for the automatic modelling of the ARIMA model (automdl.enabled is set to TRUE):

automdl.acceptdefault
logicals. If TRUE the default model (ARIMA(0,1,1)(0,1,1)) may be chosen in the first step of the automatic model identification. If the Ljung-Box Q statistics for the residuals is acceptable, the default model is accepted and no further attempt will be made to identify any other.

automdl.cancel
numeric, cancelation limit. If the difference in moduli of an AR and an MA roots (when estimating ARIMA(1,0,1)(1,0,1) models in the second step of the automatic identification of the differencing orders) is smaller than cancelation limit, the two roots are assumed equal and cancel out.

automdl.ub1
numeric, first unit root limit. It is the threshold value for the initial unit root test in the automatic differencing procedure. When one of the roots in the estimation of the ARIMA(2,0,0)(1,0,0) plus mean model, performed in the first step of the automatic model identification procedure, is larger than first unit root limit in modulus, it is set equal to unity.

automdl.ub2
numeric, second unit root limit. When one of the roots in the estimation of the ARIMA(1,0,1)(1,0,1) plus mean model, which is performed in the second step of the automatic model identification procedure, is larger than second unit root limit in modulus, it is checked if there is a common factor in the corresponding AR and MA polynomials of the ARMA model that can be cancelled (see automdl.cancel). If there is no cancellation, the AR root is set equal to unity (i.e. the differencing order changes).

automdl.mixed
logicals. The variable controls whether ARIMA models with non-seasonal AR and MA terms or seasonal AR and MA terms will be considered in the automatic model identification procedure. If FALSE a model with AR and MA terms in both the seasonal and non-seasonal parts of the model can be acceptable, provided there are not AR and MA terms in either the seasonal or non-seasonal.

automdl.balanced
logicals. If TRUE, the automatic model identification procedure will have a preference for balanced models (i.e. models for which the order of the combined AR and differencing operator is equal to the order of the combined MA operator).

automdl.armalimit
numeric, arma limit. It is the threshold value for t-statistics of ARMA coefficients and constant term used for the final test of model parsimony. If the highest order ARMA coefficient has a t-value less than this value in magnitude, the order of the model is reduced. Also if the constant term has a t-value less than arma limit in magnitude, it is removed from the set of regressors.

automdl.reducecv
numeric, ReduceCV. The percentage by which the outlier's critical value will be reduced when an identified model is found to have a Ljung-Box statistic with an unacceptable confidence coefficient. The parameter should be between 0 and 1, and will only be active when automatic outlier identification is enabled. The reduced critical value will be set to (1-ReduceCV)xCV, where CV is the original critical value.

automdl.ljungboxlimit
numeric, Ljung Box limit. Acceptance criterion for the confidence intervals of the Ljung-Box Q statistic. If the LjungBox Q statistics for the residuals of a final model is greater than Ljung Box limit, the model is rejected, the outlier critical value is reduced, and model and outlier identification (if specified) is redone with a reduced value.

automdl.ubfinal
numeric, final unit root limit. The threshold value for the final unit root test. If the magnitude of an AR root for the final model is less than the final unit root limit, a unit root is assumed, the order of the AR polynomial is reduced by one, and the appropriate order of the differencing (non-seasonal, seasonal) is increased. The parameter value should be greater than one.

Control variables for the non-automatic modelling of the ARIMA model (automdl.enabled is set to FALSE):

arima.mu
logicals. If TRUE, the mean is considered as part of the ARIMA model.

arima.p
numeric. The order of the non-seasonal autoregressive (AR) polynomial.

arima.d
numeric. Regular differencing order.

arima.q
numeric. The order of the non-seasonal moving average (MA) polynomial.

arima.bp
numeric. The order of the seasonal autoregressive (AR) polynomial.

arima.bd
numeric. Seasonal differencing order.

arima.bq
numeric. The order of the seasonal moving average (MA) polynomial.

Control variables for the user-defined ARMA coefficients. Coefficients can be defined for the regular and seasonal autoregressive (AR) polynomials and moving average (MA) polynomials. The model considers the coefficients only if the procedure for their estimation (arima.coefType) is provided, and the number of provided coefficients matches the sum of (regular and seasonal) AR and MA orders (p,q,bp,bq).

arima.coefEnabled
logicals. If TRUE the program uses the user-defined ARMA coefficients.

arima.coef
vector providing the coefficients for the regular and seasonal AR and MA polynominals. The length of the vector must equal the sum of the regular and seasonal AR and MA orders. The coefficients shall be provided in the order: regular AR (Phi - p elements), regular MA (Theta - q elements), seasonal AR (BPhi - bp elements) and seasonal MA (BTheta - bq elements). E.g.: arima.coef=c(0.6,0.7) with arima.p=1, arima.q=0,arima.bp=1 and arima.bq=0.

arima.coefType
vector defining ARMA coefficients estimation procedure. Possible procedures are: "Undefined" - no use of user-defined input (i.e. coefficients are estimated), "Fixed" - fixes the coefficients at the value provided by the user, "Initial" - the value defined by the user is used as initial condition. For orders for which the coefficients shall not be defined, the arima.coef can be set to NA or 0 or the arima.coefType can be set to "Undefined". E.g.: arima.coef = c(-0.8,-0.6,NA), arima.coefType = c("Fixed","Fixed","Undefined").

fcst.horizon
numeric, forecasts horizon. Length of the forecasts generated by the RegARIMA model in periods (positive values) or years (negative values). By default the program generates two years forecasts (fcst.horizon set to -2).

Value

A list of class c("regarima_spec","X13") with the below components. Each component refers to different part of the RegARIMA model specification, mirroring the arguments of the function (for details see arguments description). Each of the lowest-level component (except span, pre-specified outliers, user-defined variables and pre-specified ARMA coefficients) is structured within a data frame with columns denoting different variables of the model specification and rows referring to: first row - base specification, as provided within the argument spec; second row - user modifications as specified by the remaining arguments of the function (e.g.: arima.d); and third row - final model specification, values that will be used in the function regarima. The final specification (third row) shall include user modifications (row two) unless they were wrongly specified. The pre-specified outliers, user-defined variables and pre-specified ARMA coefficients consist of a list with the Predefined (base model specification) and Final values.

estimate

data frame. Variables referring to: span - time span for the model estimation, tolerance - argument estimate.tol. The final values can be also accessed with the function s_estimate.

transform

data frame. Variables referring to: tfunction - argument transform.function, adjust - argument transform.adjust, aicdiff - argument transform.aicdiff. The final values can be also accessed with the function s_transform.

regression

list including the information on the user-defined variables (userdef), trading.days effect and easter effect. The user-defined part includes: specification - data frame with the information if pre-specified outliers (outlier) and user-defined variables (variables) are included in the model and if fixed coefficients are used (outlier.coef and variables.coef). The final values can be also accessed with the function s_usrdef; outliers - matrixes with the outliers (Predefined and Final). The final outliers can be also accessed with the function s_preOut; and variables - list with the Predefined and Final user-defined variables (series) and its description (description) including the information on the variable type and values of fixed coefficients. The final user-defined variables can be also accessed with the function s_preVar. Within the data frame trading.days variables refer to: option - argument tradingdays.option, autoadjust - argument tradingdays.autoadjust, leapyear - argument tradingdays.leapyear, stocktd - argument tradingdays.stocktd, test - argument tradingdays.test. The final trading.days values can be also accessed with the function s_td. Within the data frame easter variables refer to: enabled - argument easter.enabled, julian - argument easter.julian, duration - argument easter.duration, test - argument easter.test. The final easter values can be also accessed with the function s_easter.

outliers

data frame. Variables referring to: enabled - argument outlier.enabled, span - time span for the outliers' detection, ao - argument outlier.ao, tc - argument outlier.tc, ls - argument outlier.ls, so - argument outlier.so, usedefcv - argument outlier.usedefcv, cv - argument outlier.cv, method - argument outlier.method, tcrate - argument outlier.tcrate. The final values can be also accessed with the function s_out.

arima

list containing a data frame with the ARIMA settings (specification) and matrixes with the information on the pre-specified ARMA coefficients (coefficients). The matrix Predefined refers to the pre-defined model specification and matrix Final to the final specification. Both matrixes contain the value of the ARMA coefficients and the procedure for its estimation. Within the data frame specification the variable enabled refer to the argument automdl.enabled and all the remaining variables (automdl.acceptdefault, automdl.cancel, automdl.ub1, automdl.ub2, automdl.mixed, automdl.balanced, automdl.armalimit, automdl.reducecv, automdl.ljungboxlimit, automdl.ubfinal, arima.mu, arima.p, arima.d, arima.q, arima.bp, arima.bd, arima.bq) to the respective function arguments. The final values of the specification can be also accessed with the function s_arima and final pre-specified ARMA coefficients with the function s_arimaCoef.

forecast

data frame with the forecast horizon (argument fcst.horizon). The final value can be also accessed with the function s_fcst.

span

matrix containing the final time span for the model estimation and outliers' detection. Contains the same information as the variable span in the data frames estimate and outliers. The matrix can be also accessed with the function s_span.

Details

The available predefined 'JDemetra+' model specifications are described in the table below.

Identifier |	Log/level detection |	Outliers detection |	Calendar effects |	ARIMA	RG0 |	NA |

NA |	NA |	Airline(+mean)	RG1 |	automatic |	AO/LS/TC |	NA |

Airline(+mean)	RG2c |	automatic |	AO/LS/TC |	2 td vars + Easter |	Airline(+mean)	RG3 |

automatic |	AO/LS/TC |	NA |	automatic	RG4c |	automatic |	AO/LS/TC |

2 td vars + Easter |	automatic	RG5c |	automatic |	AO/LS/TC |	7 td vars + Easter |	automatic

References

Info on 'JDemetra+', usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en

Examples

# \donttest{
myseries <- ipi_c_eu[, "FR"]
myspec1 <- regarima_spec_x13(spec = "RG5c")
myreg1 <- regarima(myseries, spec = myspec1)

 # Modify a pre-specified model specification
myspec2 <- regarima_spec_x13(spec = "RG5c",
                             tradingdays.option = "WorkingDays")
myreg2 <- regarima(myseries, spec = myspec2)

 # Modify the model specification from a "regarima" object
myspec3 <- regarima_spec_x13(myreg1, tradingdays.option = "WorkingDays")
myreg3 <- regarima(myseries, myspec3)

 # Modify the model specification from a "regarima_spec" object
myspec4 <- regarima_spec_x13(myspec1, tradingdays.option = "WorkingDays")
myreg4 <- regarima(myseries, myspec4)

 # Pre-specified outliers
myspec1 <- regarima_spec_x13(spec = "RG5c", usrdef.outliersEnabled = TRUE,
              usrdef.outliersType = c("LS", "AO"),
              usrdef.outliersDate = c("2008-10-01", "2002-01-01"),
              usrdef.outliersCoef = c(36, 14),
              transform.function = "None")

myreg1 <- regarima(myseries, myspec1)
myreg1

#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.1057      0.114
#> Phi(2)      0.2141      0.076
#> Theta(1)   -0.4709      0.111
#> BTheta(1)  -0.6642      0.043
#> 
#>               Estimate Std. Error
#> Monday         0.58039      0.234
#> Tuesday        0.82241      0.235
#> Wednesday      1.10804      0.234
#> Thursday       0.09871      0.232
#> Friday         0.89917      0.233
#> Saturday      -1.62006      0.231
#> Leap year      2.14549      0.731
#> Easter [1]    -2.34335      0.456
#> AO (9-2008)   32.07187      2.931
#> LS (9-2008)  -57.03600      2.631
#> AO (5-2011)   13.22101      1.841
#> TC (9-2008)   23.33640      3.967
#> TC (12-2001) -20.56517      2.911
#> AO (12-2001)  17.24940      2.981
#> TC (2-2002)   10.70687      1.939
#> 
#> Fixed outliers: 
#>              Coefficients
#> LS (10-2008)           36
#> AO (1-2002)            14
#> 
#> 
#> Residual standard error: 2.185 on 327 degrees of freedom
#> Log likelihood = -767.3, aic =  1575 aicc =  1577, bic(corrected for length) = 1.883
#> 

s_preOut(myreg1)

#>   type       date coeff
#> 1   LS 2008-10-01    36
#> 2   AO 2002-01-01    14

 # User-defined variables
var1 <- ts(rnorm(length(myseries))*10, start = start(myseries),
           frequency = 12)
var2 <- ts(rnorm(length(myseries))*100, start = start(myseries),
           frequency = 12)
var <- ts.union(var1, var2)

myspec1 <- regarima_spec_x13(spec = "RG5c", usrdef.varEnabled = TRUE,
                             usrdef.var = var)
myreg1 <- regarima(myseries, myspec1)
myreg1

#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)    -0.01674      0.118
#> Phi(2)     0.15656      0.077
#> Theta(1)  -0.53112      0.112
#> BTheta(1) -0.71657      0.041
#> 
#>                Estimate Std. Error
#> r.var1       -8.812e-05      0.000
#> r.var2        1.179e-05      0.000
#> Monday        4.985e-03      0.002
#> Tuesday       8.384e-03      0.002
#> Wednesday     1.115e-02      0.002
#> Thursday      1.387e-03      0.002
#> Friday        8.397e-03      0.002
#> Saturday     -1.641e-02      0.002
#> Easter [1]   -2.156e-02      0.004
#> AO (5-2011)   1.302e-01      0.018
#> LS (11-2008) -1.155e-01      0.016
#> 
#> 
#> Residual standard error: 0.02081 on 331 degrees of freedom
#> Log likelihood = 846.8, aic =  1544 aicc =  1546, bic(corrected for length) = -7.492
#> 

myspec2 <- regarima_spec_x13(spec = "RG5c", usrdef.varEnabled = TRUE,
                             usrdef.var = var1, usrdef.varCoef = 2,
                             transform.function = "None")
myreg2 <- regarima(myseries, myspec2)
s_preVar(myreg2)

#> $series
#>               Jan          Feb          Mar          Apr          May
#> 1990  11.94158303   9.13057148  -7.86752993  -4.10929700   4.76373682
#> 1991 -13.02671671   6.34857535   9.99655161  -3.37748953  -0.86033607
#> 1992  -6.61257061   5.48811286  -5.39367800  -1.71318426  -7.66734605
#> 1993  -0.47539083   7.86925425  11.66684112  -2.71823871  -3.13704285
#> 1994 -11.14492926  -8.42058813 -15.04294209  -2.84026455   0.42869041
#> 1995  -5.78524902  -1.69109048 -19.19232520 -15.34266387 -11.14761221
#> 1996   9.71396469  -5.79663020  16.04179789   2.25973337  -5.14857122
#> 1997 -15.55643970   0.93501283  -3.66949421  -1.29408799   4.07795520
#> 1998 -16.09575292 -10.49710115  20.52033974   1.76436411  11.28307358
#> 1999  -2.81725199  -7.91679461   0.01653727 -11.87530126   3.62417465
#> 2000   4.95847824  12.99750970 -16.15985506 -12.50367945  15.82132306
#> 2001  17.40218394  -3.24455939  -4.47511333   4.16082526 -12.95865368
#> 2002  -7.00069022  18.60931509  18.03924959 -13.56043492  -3.14751072
#> 2003 -14.00912333  14.45014910  -4.23481621  -7.33749597  -7.08484753
#> 2004  17.47150257  -0.81704668   1.24141586   3.92236141   3.78419970
#> 2005  12.07632181   0.32001384   1.80397550   0.77160580   5.58395338
#> 2006   8.15566796   1.20032668   8.81789848 -17.98743117  -9.57051078
#> 2007   2.26820843  -5.91868492  -4.05414494  14.14224665   2.41625825
#> 2008  -3.44214729   6.45775407 -15.34500551   5.99672514  -5.04358892
#> 2009  20.20798363 -18.38661562   1.72937164  -7.63245040 -11.18492914
#> 2010  -0.17532122  -3.90664205  -4.02775110  12.76706963   8.21075754
#> 2011  12.13732488  15.03810074 -11.69425021   6.13377167  14.83931321
#> 2012   7.34408890  -5.75309458   0.74897554   0.59301839   8.48264703
#> 2013  -7.59094766  -5.09303603  -1.56039669   1.32582442  -9.74084022
#> 2014  -9.75148693  -0.27117079   4.46671886  -3.29112249   1.13049080
#> 2015  28.39266231   9.61819160   5.08432001   2.59523192  -7.73404126
#> 2016  13.22082615  12.00558995  -7.81874565  -4.41476451  -2.46028129
#> 2017   6.34250029   8.65538237  -4.30593090   0.13030894   0.51249629
#> 2018  -2.76989004   1.36469589  10.67207722   0.01828103   1.40092252
#> 2019 -10.62911432  16.33683158   5.40830359  14.56857971 -19.33784243
#>               Jun          Jul          Aug          Sep          Oct
#> 1990   1.85915868 -13.24716253  11.72371055   2.31492772   4.83071904
#> 1991 -17.18817584  -9.29121572   8.13676046   5.26413546  10.11957390
#> 1992  -6.06574846  16.47312821   8.41069348   3.38175097  -4.04883076
#> 1993  -2.86300150  -8.03440263  -0.94555455  -0.74046506   7.14490683
#> 1994  -0.08866413 -29.49083784   0.20358254  -0.98446980   5.89109098
#> 1995  15.97811635  -6.39805078  15.66690553 -14.49155712  -7.91503959
#> 1996  -8.23788240   3.34415250  -0.93490765   3.04062078  -4.76507534
#> 1997  -5.83968954  -1.93449890  -2.69546670   0.73658232   3.57236138
#> 1998   4.35001744   5.48833492   6.47418304   8.78463454   3.50797869
#> 1999  -5.49435319   6.92949838  -0.60843916 -11.93540930  -1.19908059
#> 2000  12.52529118  -2.09395701  -6.39033687   3.98902937  13.41973500
#> 2001   6.82633090   4.96913333  14.39660656  12.07358558   2.38514694
#> 2002 -10.19265181  -6.12249656  -2.89068749  15.06983292   3.67722694
#> 2003  14.75092459   8.45004185  12.93994431   2.98161142  -4.05682490
#> 2004   1.66175673  11.53387369   0.14748302   0.21440845  -8.26352170
#> 2005  -1.66854367  -2.27980619   2.42666056  -9.28482969  15.15406678
#> 2006  -4.74058606 -18.60461283   2.53690073  -2.14401713 -23.70181451
#> 2007 -14.36087478  -3.14552745  -4.80542811   8.36254248  -0.52349287
#> 2008  -4.06909625  -4.60935206  22.14908426  14.86967820  -3.55098858
#> 2009  11.97977461   9.81193912  11.91749205   6.32833568   5.32825947
#> 2010   4.19785146  17.11661220 -22.15632715  -1.84878197   6.71168243
#> 2011  -4.43773736   2.68589741 -12.28216829   4.78528762  16.69859573
#> 2012  25.75396587   0.94259964  -5.19949532   7.74428621  12.28289259
#> 2013 -12.50949663  -0.46336042  11.37387771  12.64449891  -5.48742076
#> 2014  10.51229395   3.26977400   7.87037444   4.41559619 -10.19304514
#> 2015 -18.05527531   1.78383232  -5.25777309   4.00186316  -1.22587586
#> 2016   3.65792190  -0.34132882   2.96227003  -2.21059772   1.87706159
#> 2017   5.38638152   5.19672613 -24.22911275  -8.06290253  -4.84679992
#> 2018  -5.20598538  23.65811636 -16.96474415   6.82585584  13.61561976
#> 2019  -8.80400798  -7.97730162  -8.86130974  -2.94827785  -8.86651693
#>               Nov          Dec
#> 1990  -5.35319580  13.78135819
#> 1991   8.31379809   4.15134145
#> 1992   9.00009039  11.90666655
#> 1993   1.24718366  12.96594339
#> 1994  -4.25751629   6.34658985
#> 1995  -5.04480733   4.01826700
#> 1996  -2.41311717   8.24155833
#> 1997   5.50428418   0.38401793
#> 1998   0.49879720   8.35749446
#> 1999  -7.08107950 -16.16663973
#> 2000  -0.37607422   5.43670114
#> 2001  -2.98837878  13.84621050
#> 2002  -0.20034941  -9.79921959
#> 2003  -1.38807578  -2.22592731
#> 2004   0.47399133  -2.93107372
#> 2005  -8.11598824  13.11247408
#> 2006   0.03258078   6.69745568
#> 2007   9.51957625  14.94512798
#> 2008   8.76203737  -2.14696646
#> 2009  -8.20121900  -1.59325676
#> 2010  -7.95140242 -15.66257835
#> 2011  -0.36455545  -4.40936271
#> 2012  18.80606885  -1.25241723
#> 2013   6.82391430  -9.29773363
#> 2014  -1.58698326 -16.37701292
#> 2015  -0.71345032  20.95734365
#> 2016  -4.38062172  -0.88733579
#> 2017  -6.53077284  -2.93764247
#> 2018  19.05955757  -6.72838542
#> 2019   0.99084673   7.20546895
#> 
#> $description
#>              type coeff
#> userdef Undefined     2
#> 

 # Pre-specified ARMA coefficients
myspec1 <- regarima_spec_x13(spec = "RG5c", automdl.enabled =FALSE,
             arima.p = 1, arima.q = 1, arima.bp = 0, arima.bq = 1,
             arima.coefEnabled = TRUE, arima.coef = c(-0.8, -0.6, 0),
             arima.coefType = c(rep("Fixed", 2), "Undefined"))

s_arimaCoef(myspec1)

#>                Type Value
#> Phi(1)        Fixed  -0.8
#> Theta(1)      Fixed  -0.6
#> BTheta(1) Undefined   0.0

myreg1 <- regarima(myseries, myspec1)
myreg1

#> y = regression model + arima (1, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     -0.8000      0.000
#> Theta(1)   -0.6000      0.000
#> BTheta(1)  -0.6933      0.041
#> 
#>              Estimate Std. Error
#> Monday       0.005835      0.002
#> Tuesday      0.007964      0.002
#> Wednesday    0.010841      0.002
#> Thursday     0.001883      0.002
#> Friday       0.009833      0.002
#> Saturday    -0.018229      0.002
#> Easter [1]  -0.020470      0.004
#> AO (5-2011)  0.144191      0.016
#> 
#> 
#> Residual standard error: 0.02585 on 337 degrees of freedom
#> Log likelihood =   772, aic =  1681 aicc =  1682, bic(corrected for length) = -7.159
#> 

# }