Saving and loading a model specification, SA and pre-adjustment in X13 and TRAMO-SEATS

save_spec saves a SA or RegARIMA model specification. load_spec loads the previously saved model specification.

save_spec(object, file = file.path(tempdir(), "spec.RData"))

load_spec(file = "spec.RData")

Arguments

object
object of one of the classes: c("SA_spec","X13"), c("SA_spec","TRAMO_SEATS"), c("SA","X13"), c("SA","TRAMO_SEATS"), c("regarima_spec","X13"), c("regarima_spec","TRAMO_SEATS"), c("regarima","X13"), c("regarima","TRAMO_SEATS").

file
(path and) name of the file where the model specification will be saved (have been saved).

Value

load_spec returns an object of class "SA_spec" or "regarima_spec".

Details

save_spec saves the final model specification of a "SA_spec", "SA", "regarima_spec" or "regarima" class object. load_spec loads the previously saved model specification. It creates a c("SA_spec","X13"), c("sA_spec","TRAMO_SEATS"), c("regarima_spec","X13") or c("regarima_spec","TRAMO_SEATS") class object, in line with the class of the previously saved model specification.

References

Info on JDemtra+, usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en

Examples

# \donttest{
myseries <- ipi_c_eu[, "FR"]
myreg1 <- regarima_x13(myseries, spec = "RG5c")
myspec2 <- regarima_spec_x13(myreg1, estimate.from = "2005-10-01", outlier.from = "2010-03-01")
myreg2 <- regarima(myseries, myspec2)

myreg3 <- regarima_tramoseats(myseries, spec = "TRfull")
myspec4 <-regarima_spec_tramoseats(myreg3, tradingdays.mauto = "Unused",
                                  tradingdays.option ="WorkingDays",
                                  easter.type = "Standard",
                                  automdl.enabled = FALSE, arima.mu = TRUE)
myreg4 <-regarima(myseries, myspec4)

myspec6 <- x13_spec("RSA5c")
mysa6 <- x13(myseries, myspec6)

myspec7 <- tramoseats_spec("RSAfull")
mysa7 <- tramoseats(myseries, myspec7)

dir <- tempdir()

 # Save the model specification from a c("regarima_spec","X13") class object
save_spec(myspec2, file.path(dir, "specx13.RData"))
 # Save the model specification from a c("regarima","X13") class object
save_spec(myreg2, file.path(dir,"regx13.RData"))
 # Save the model specification from a c("regarima_spec","TRAMO_SEATS") class object
save_spec(myspec4, file.path(dir,"specTS.RData"))
 # Save the model specification from a c("regarima","TRAMO_SEATS") class object
save_spec(myreg4, file.path(dir,"regTS.RData"))
 # Save model from a c("SA_spec","X13") class object
save_spec(myspec6, file.path(dir,"specFullx13.RData"))
 # Save model from a c("SA","X13") class object
save_spec(mysa6, file.path(dir,"sax13.RData"))
 # Save model from a c("SA_spec","TRAMO_SEATS") class object
save_spec(myspec7, file.path(dir,"specFullTS.RData"))
 # Save model from a c("SA","TRAMO_SEATS") class object
save_spec(mysa7, file.path(dir,"saTS.RData"))

 # Load the model specification
myspec2a <- load_spec(file.path(dir,"specx13.RData"))
myspec2b <- load_spec(file.path(dir,"regx13.RData"))
myspec4a <- load_spec(file.path(dir,"specTS.RData"))
myspec4b <- load_spec(file.path(dir,"regTS.RData"))
myspec6a <- load_spec(file.path(dir,"specFullx13.RData"))
myspec6b <- load_spec(file.path(dir,"sax13.RData"))
myspec7a <- load_spec(file.path(dir,"specFullTS.RData"))
myspec7b <- load_spec(file.path(dir,"saTS.RData"))

regarima(myseries, myspec2a)

#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.2939      0.079
#> BTheta(1)  -0.7651      0.058
#> 
#>              Estimate Std. Error
#> Monday       0.010967      0.003
#> Tuesday      0.012055      0.003
#> Wednesday    0.006858      0.003
#> Thursday     0.001519      0.003
#> Friday       0.013811      0.003
#> Saturday    -0.020566      0.003
#> Easter [1]  -0.017599      0.006
#> AO (5-2011)  0.128270      0.018
#> 
#> 
#> Residual standard error: 0.0229 on 147 degrees of freedom
#> Log likelihood = 367.2, aic = 749.1 aicc = 750.9, bic(corrected for length) = -7.233
#> 

x13(myseries, myspec6a)

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.02578      0.118
#> Phi(2)     0.15940      0.078
#> Theta(1)  -0.51733      0.113
#> BTheta(1) -0.70304      0.041
#> 
#>               Estimate Std. Error
#> Monday        0.004967      0.002
#> Tuesday       0.008415      0.002
#> Wednesday     0.010614      0.002
#> Thursday      0.002060      0.002
#> Friday        0.008274      0.002
#> Saturday     -0.016328      0.002
#> Easter [1]   -0.021145      0.004
#> AO (5-2011)   0.128134      0.017
#> LS (11-2008) -0.087008      0.017
#> LS (1-2009)  -0.068435      0.017
#> 
#> 
#> Residual standard error: 0.02041 on 332 degrees of freedom
#> Log likelihood = 853.7, aic =  1528 aicc =  1530, bic(corrected for length) = -7.547
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.097
#> M(2)    0.053
#> M(3)    0.995
#> M(4)    0.421
#> M(5)    0.950
#> M(6)    0.154
#> M(7)    0.074
#> M(8)    0.211
#> M(9)    0.064
#> M(10)   0.267
#> M(11)   0.250
#> Q       0.322
#> Q-M2    0.355
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.1945 105.1440 0.9876940 1.0004808
#> Feb 2019 101.9 106.3206 105.5014 0.9584217 1.0077654
#> Mar 2019 111.0 104.7345 105.7233 1.0598229 0.9906473
#> Apr 2019 107.4 106.1448 105.7281 1.0118257 1.0039413
#> May 2019 105.5 106.2039 105.5978 0.9933724 1.0057396
#> Jun 2019 105.8 101.1898 105.3259 1.0455597 0.9607304
#> Jul 2019 110.1 105.3543 104.8985 1.0450450 1.0043457
#> Aug 2019  78.7 102.9339 104.3791 0.7645685 0.9861536
#> Sep 2019 108.5 104.4720 103.8115 1.0385554 1.0063631
#> Oct 2019 116.8 104.8375 103.2992 1.1141052 1.0148919
#> Nov 2019 103.8 101.7212 102.9938 1.0204364 0.9876441
#> Dec 2019  97.7 101.8723 102.8805 0.9590443 0.9902000
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.79694 103.1662 102.8951 0.9867273 1.0026346
#> Feb 2020 101.02102 103.3357 102.9515 0.9776003 1.0037323
#> Mar 2020 112.10997 103.5043 102.9465 1.0831427 1.0054189
#> Apr 2020 104.50015 103.4303 102.8832 1.0103439 1.0053179
#> May 2020  96.60823 101.9351 102.7280 0.9477426 0.9922819
#> Jun 2020 112.13668 102.1960 102.5441 1.0972706 0.9966052
#> Jul 2020 106.78914 102.3841 102.4931 1.0430249 0.9989358
#> Aug 2020  76.62863 102.6907 102.5880 0.7462084 1.0010004
#> Sep 2020 110.85790 103.4529 102.7454 1.0715783 1.0068865
#> Oct 2020 111.29327 102.4307 102.8914 1.0865224 0.9955221
#> Nov 2020 104.21972 102.6701 103.0257 1.0150931 0.9965489
#> Dec 2020 101.55551 103.4907 103.1115 0.9813009 1.0036778
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.941
#>  Seasonal     62.123
#>  Irregular     0.842
#>  TD & Hol.     2.428
#>  Others       32.787
#>  Total       100.121
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.283
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          0.909
#>  qs test on i                           0.746
#>  f-test on sa (seasonal dummies)        0.974
#>  f-test on i (seasonal dummies)         0.947
#>  Residual seasonality (entire series)   0.932
#>  Residual seasonality (last 3 years)    0.940
#>  f-test on sa (td)                      0.985
#>  f-test on i (td)                       0.983
#> 
#> 
#> Additional output variables

tramoseats(myseries, myspec7a)

#> RegARIMA
#> y = regression model + arima (2, 1, 0, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.3858      0.053
#> Phi(2)      0.2428      0.053
#> BTheta(1)  -0.7134      0.040
#> 
#>              Estimate Std. Error
#> Week days    0.007131      0.000
#> Leap year    0.022154      0.007
#> Easter [6]  -0.021360      0.004
#> AO (5-2011)  0.128991      0.017
#> AO (5-2000)  0.057730      0.017
#> 
#> 
#> Residual standard error: 0.02174 on 338 degrees of freedom
#> Log likelihood = 831.8, aic =  1560 aicc =  1560, bic(corrected for length) = -7.523
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 + 0.385785 B + 0.242834 B^2 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.713428 B^12 
#> 
#> 
#> SA
#> AR :  1 + 0.385785 B + 0.242834 B^2 
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 0.975113 B + 0.003731 B^2 - 0.003172 B^3 + 0.002186 B^4 
#> Innovation variance:  0.7469663 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.027738 B - 0.972262 B^2 
#> Innovation variance:  0.06986498 
#> 
#> Seasonal
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 1.353435 B + 1.194117 B^2 + 1.260347 B^3 + 1.124065 B^4 + 0.886062 B^5 + 0.682894 B^6 + 0.440009 B^7 + 0.271828 B^8 + 0.020068 B^9 - 0.054641 B^10 - 0.222194 B^11 
#> Innovation variance:  0.03026107 
#> 
#> Transitory
#> AR :  1 + 0.385785 B + 0.242834 B^2 
#> MA :  1 - 0.316265 B - 0.683735 B^2 
#> Innovation variance:  0.05225072 
#> 
#> Irregular
#> Innovation variance:  0.2217707 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.1743 104.9429 0.9878842 1.0022048
#> Feb 2019 101.9 105.9207 105.1941 0.9620400 1.0069073
#> Mar 2019 111.0 104.4311 105.2847 1.0629021 0.9918924
#> Apr 2019 107.4 105.5933 105.4042 1.0171096 1.0017943
#> May 2019 105.5 107.2997 105.1676 0.9832269 1.0202734
#> Jun 2019 105.8 101.9020 104.5972 1.0382528 0.9742326
#> Jul 2019 110.1 106.0606 104.2943 1.0380862 1.0169351
#> Aug 2019  78.7 103.0620 104.0531 0.7636179 0.9904748
#> Sep 2019 108.5 103.7505 103.8759 1.0457777 0.9987934
#> Oct 2019 116.8 105.1043 103.5817 1.1112766 1.0147004
#> Nov 2019 103.8 102.2380 102.8763 1.0152784 0.9937947
#> Dec 2019  97.7 101.4426 102.3135 0.9631060 0.9914880
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.14796 102.4076 102.1524 0.9876994 1.0024982
#> Feb 2020  98.73586 102.2044 102.1359 0.9660638 1.0006702
#> Mar 2020 111.14762 102.0245 102.1194 1.0894214 0.9990702
#> Apr 2020 103.85947 102.1162 102.1029 1.0170713 1.0001303
#> May 2020  95.54759 102.0976 102.0864 0.9358456 1.0001098
#> Jun 2020 111.32747 102.0557 102.0699 1.0908507 0.9998602
#> Jul 2020 105.98668 102.0495 102.0535 1.0385811 0.9999614
#> Aug 2020  75.98545 102.0352 102.0370 0.7446983 0.9999830
#> Sep 2020 109.40398 102.0154 102.0205 1.0724264 0.9999501
#> Oct 2020 110.59660 101.9997 102.0040 1.0842840 0.9999575
#> Nov 2020 103.53656 101.9837 101.9875 1.0152266 0.9999627
#> Dec 2020 100.67280 101.9669 101.9711 0.9873092 0.9999589
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         5.694
#>  Seasonal     88.840
#>  Irregular     0.820
#>  TD & Hol.     3.412
#>  Others        0.469
#>  Total        99.234
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.344
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.984
#>  f-test on sa (td)                      0.135
#>  f-test on i (td)                       0.256
#> 
#> 
#> Additional output variables

regarima(myseries, myspec4a)

#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5960      0.044
#> BTheta(1)  -0.7032      0.041
#> 
#>                Estimate Std. Error
#> Mean          2.029e-05      0.000
#> Monday        5.066e-03      0.002
#> Tuesday       8.286e-03      0.002
#> Wednesday     1.075e-02      0.002
#> Thursday      1.973e-03      0.002
#> Friday        7.949e-03      0.002
#> Saturday     -1.622e-02      0.002
#> Leap year     2.031e-02      0.007
#> Easter [6]   -1.884e-02      0.004
#> AO (5-2011)   1.251e-01      0.017
#> LS (11-2008) -8.905e-02      0.017
#> LS (1-2009)  -6.888e-02      0.017
#> 
#> 
#> Residual standard error: 0.0205 on 332 degrees of freedom
#> Log likelihood = 852.3, aic =  1531 aicc =  1532, bic(corrected for length) = -7.539
#> 

x13(myseries, myspec6b)

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.02578      0.118
#> Phi(2)     0.15940      0.078
#> Theta(1)  -0.51733      0.113
#> BTheta(1) -0.70304      0.041
#> 
#>               Estimate Std. Error
#> Monday        0.004967      0.002
#> Tuesday       0.008415      0.002
#> Wednesday     0.010614      0.002
#> Thursday      0.002060      0.002
#> Friday        0.008274      0.002
#> Saturday     -0.016328      0.002
#> Easter [1]   -0.021145      0.004
#> AO (5-2011)   0.128134      0.017
#> LS (11-2008) -0.087008      0.017
#> LS (1-2009)  -0.068435      0.017
#> 
#> 
#> Residual standard error: 0.02041 on 332 degrees of freedom
#> Log likelihood = 853.7, aic =  1528 aicc =  1530, bic(corrected for length) = -7.547
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.097
#> M(2)    0.053
#> M(3)    0.995
#> M(4)    0.421
#> M(5)    0.950
#> M(6)    0.154
#> M(7)    0.074
#> M(8)    0.211
#> M(9)    0.064
#> M(10)   0.267
#> M(11)   0.250
#> Q       0.322
#> Q-M2    0.355
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.1945 105.1440 0.9876940 1.0004808
#> Feb 2019 101.9 106.3206 105.5014 0.9584217 1.0077654
#> Mar 2019 111.0 104.7345 105.7233 1.0598229 0.9906473
#> Apr 2019 107.4 106.1448 105.7281 1.0118257 1.0039413
#> May 2019 105.5 106.2039 105.5978 0.9933724 1.0057396
#> Jun 2019 105.8 101.1898 105.3259 1.0455597 0.9607304
#> Jul 2019 110.1 105.3543 104.8985 1.0450450 1.0043457
#> Aug 2019  78.7 102.9339 104.3791 0.7645685 0.9861536
#> Sep 2019 108.5 104.4720 103.8115 1.0385554 1.0063631
#> Oct 2019 116.8 104.8375 103.2992 1.1141052 1.0148919
#> Nov 2019 103.8 101.7212 102.9938 1.0204364 0.9876441
#> Dec 2019  97.7 101.8723 102.8805 0.9590443 0.9902000
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.79694 103.1662 102.8951 0.9867273 1.0026346
#> Feb 2020 101.02102 103.3357 102.9515 0.9776003 1.0037323
#> Mar 2020 112.10997 103.5043 102.9465 1.0831427 1.0054189
#> Apr 2020 104.50015 103.4303 102.8832 1.0103439 1.0053179
#> May 2020  96.60823 101.9351 102.7280 0.9477426 0.9922819
#> Jun 2020 112.13668 102.1960 102.5441 1.0972706 0.9966052
#> Jul 2020 106.78914 102.3841 102.4931 1.0430249 0.9989358
#> Aug 2020  76.62863 102.6907 102.5880 0.7462084 1.0010004
#> Sep 2020 110.85790 103.4529 102.7454 1.0715783 1.0068865
#> Oct 2020 111.29327 102.4307 102.8914 1.0865224 0.9955221
#> Nov 2020 104.21972 102.6701 103.0257 1.0150931 0.9965489
#> Dec 2020 101.55551 103.4907 103.1115 0.9813009 1.0036778
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.941
#>  Seasonal     62.123
#>  Irregular     0.842
#>  TD & Hol.     2.428
#>  Others       32.787
#>  Total       100.121
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.283
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          0.909
#>  qs test on i                           0.746
#>  f-test on sa (seasonal dummies)        0.974
#>  f-test on i (seasonal dummies)         0.947
#>  Residual seasonality (entire series)   0.932
#>  Residual seasonality (last 3 years)    0.940
#>  f-test on sa (td)                      0.985
#>  f-test on i (td)                       0.983
#> 
#> 
#> Additional output variables

tramoseats(myseries, myspec7b)

#> RegARIMA
#> y = regression model + arima (2, 1, 0, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.3858      0.053
#> Phi(2)      0.2428      0.053
#> BTheta(1)  -0.7134      0.040
#> 
#>              Estimate Std. Error
#> Week days    0.007131      0.000
#> Leap year    0.022154      0.007
#> Easter [6]  -0.021360      0.004
#> AO (5-2011)  0.128991      0.017
#> AO (5-2000)  0.057730      0.017
#> 
#> 
#> Residual standard error: 0.02174 on 338 degrees of freedom
#> Log likelihood = 831.8, aic =  1560 aicc =  1560, bic(corrected for length) = -7.523
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 + 0.385785 B + 0.242834 B^2 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.713428 B^12 
#> 
#> 
#> SA
#> AR :  1 + 0.385785 B + 0.242834 B^2 
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 0.975113 B + 0.003731 B^2 - 0.003172 B^3 + 0.002186 B^4 
#> Innovation variance:  0.7469663 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.027738 B - 0.972262 B^2 
#> Innovation variance:  0.06986498 
#> 
#> Seasonal
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 1.353435 B + 1.194117 B^2 + 1.260347 B^3 + 1.124065 B^4 + 0.886062 B^5 + 0.682894 B^6 + 0.440009 B^7 + 0.271828 B^8 + 0.020068 B^9 - 0.054641 B^10 - 0.222194 B^11 
#> Innovation variance:  0.03026107 
#> 
#> Transitory
#> AR :  1 + 0.385785 B + 0.242834 B^2 
#> MA :  1 - 0.316265 B - 0.683735 B^2 
#> Innovation variance:  0.05225072 
#> 
#> Irregular
#> Innovation variance:  0.2217707 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.1743 104.9429 0.9878842 1.0022048
#> Feb 2019 101.9 105.9207 105.1941 0.9620400 1.0069073
#> Mar 2019 111.0 104.4311 105.2847 1.0629021 0.9918924
#> Apr 2019 107.4 105.5933 105.4042 1.0171096 1.0017943
#> May 2019 105.5 107.2997 105.1676 0.9832269 1.0202734
#> Jun 2019 105.8 101.9020 104.5972 1.0382528 0.9742326
#> Jul 2019 110.1 106.0606 104.2943 1.0380862 1.0169351
#> Aug 2019  78.7 103.0620 104.0531 0.7636179 0.9904748
#> Sep 2019 108.5 103.7505 103.8759 1.0457777 0.9987934
#> Oct 2019 116.8 105.1043 103.5817 1.1112766 1.0147004
#> Nov 2019 103.8 102.2380 102.8763 1.0152784 0.9937947
#> Dec 2019  97.7 101.4426 102.3135 0.9631060 0.9914880
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.14796 102.4076 102.1524 0.9876994 1.0024982
#> Feb 2020  98.73586 102.2044 102.1359 0.9660638 1.0006702
#> Mar 2020 111.14762 102.0245 102.1194 1.0894214 0.9990702
#> Apr 2020 103.85947 102.1162 102.1029 1.0170713 1.0001303
#> May 2020  95.54759 102.0976 102.0864 0.9358456 1.0001098
#> Jun 2020 111.32747 102.0557 102.0699 1.0908507 0.9998602
#> Jul 2020 105.98668 102.0495 102.0535 1.0385811 0.9999614
#> Aug 2020  75.98545 102.0352 102.0370 0.7446983 0.9999830
#> Sep 2020 109.40398 102.0154 102.0205 1.0724264 0.9999501
#> Oct 2020 110.59660 101.9997 102.0040 1.0842840 0.9999575
#> Nov 2020 103.53656 101.9837 101.9875 1.0152266 0.9999627
#> Dec 2020 100.67280 101.9669 101.9711 0.9873092 0.9999589
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         5.694
#>  Seasonal     88.840
#>  Irregular     0.820
#>  TD & Hol.     3.412
#>  Others        0.469
#>  Total        99.234
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.344
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.984
#>  f-test on sa (td)                      0.135
#>  f-test on i (td)                       0.256
#> 
#> 
#> Additional output variables

# }