Access model specification, SA and pre-adjustment in X13 and TRAMO-SEATS

Below functions access different parts of the final model specification as included in the "SA", "regarima", "SA_spec" and "regarima_spec" S3 class objects.

s_estimate(object = NA)

s_transform(object = NA)

s_usrdef(object = NA)

s_preOut(object = NA)

s_preVar(object = NA)

s_td(object = NA)

s_easter(object = NA)

s_out(object = NA)

s_arima(object = NA)

s_arimaCoef(object = NA)

s_fcst(object = NA)

s_span(object = NA)

s_x11(object = NA)

s_seats(object = NA)

Arguments

object
object of one of the classes: c("SA","X13"), c("SA","TRAMO_SEATS"), c("SA_spec","X13"), c("SA_spec","TRAMO_SEATS"), c("regarima","X13"), c("regarima","TRAMO_SEATS"), c("regarima_spec","X13"), c("regarima_spec","TRAMO_SEATS").

Value

s_estimate returns a data.frame with the estimate variables

s_transform returns a data.frame with the transform varaibles

s_usrdef returns a data.frame with the user-defined regressors (outliers and variables) model specification, indicating if those variables are included in the model and if coefficients are pre-specified

s_preOut returns a data.frame with the pre-specified outliers

s_preVar returns a list with the information on the user-defined variables, including: series - the time series and description - data.frame with the variable type and coefficients

s_td returns a data.frame with the trading.days variables

s_easter returns a data.frame with the easter variables

s_out returns a data.frame with the outliers detection variables

s_arima returns a data.frame with the arima variables

s_arimaCoef returns a data.frame with the user-specified ARMA coefficients

s_fcst returns a data.frame with the forecast horizon

s_span returns a data.frame with the span variables

s_x11 returns a data.frame with the x11 variables

s_seats returns a data.frame with the seats variables

References

Info on 'JDemetra+', usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en

Examples

# \donttest{
myseries <- ipi_c_eu[, "FR"]
myreg1 <- regarima_x13(myseries, spec = "RG5c")
myspec1 <- regarima_spec_x13(myreg1,
             estimate.from = "2005-10-01",
             outlier.from = "2010-03-01")

s_estimate(myreg1)

#>  preliminary.check span tolerance
#>               TRUE  All     1e-07

s_estimate(myspec1)

#>  preliminary.check            span tolerance
#>               TRUE From 2005-10-01     1e-07

s_transform(myreg1)

#>  tfunction adjust aicdiff
#>       Auto   None      -2

s_transform(myspec1)

#>  tfunction adjust aicdiff
#>       Auto   None      -2

s_usrdef(myreg1)

#>   outlier outlier.coef variables variables.coef
#> 1   FALSE        FALSE     FALSE          FALSE

s_usrdef(myspec1)

#>  outlier outlier.coef variables variables.coef
#>    FALSE        FALSE     FALSE          FALSE

myspec2 <- regarima_spec_x13(myreg1, usrdef.outliersEnabled = TRUE,
             usrdef.outliersType = c("LS", "AO"),
             usrdef.outliersDate = c("2009-10-01", "2005-02-01"))
myreg2 <- regarima(myseries, myspec2)

s_preOut(myreg2)

#>   type       date coeff
#> 1   LS 2009-10-01    NA
#> 2   AO 2005-02-01    NA

s_preOut(myspec2)

#>   type       date coeff
#> 1   LS 2009-10-01    NA
#> 2   AO 2005-02-01    NA

var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
var3 <- ts.union(var1, var2)
myspec3 <- regarima_spec_x13(spec = "RG5c",
                             usrdef.varEnabled = TRUE,
                             usrdef.var = var3)
myreg3 <- regarima(myseries, myspec3)

s_preVar(myspec3)

#> $series
#>                  var1         var2
#> Jan 1990   7.17057189  -34.8512400
#> Feb 1990   4.05856898  -54.5470652
#> Mar 1990 -14.99181516   -2.0519678
#> Apr 1990 -12.23578511  -33.8849889
#> May 1990  14.25347975   18.5592100
#> Jun 1990   5.35483990   34.5862657
#> Jul 1990  -4.79124096  -79.3332374
#> Aug 1990   7.08755575  -74.8631612
#> Sep 1990  -0.10940710 -184.6666644
#> Oct 1990 -13.92609039 -155.6824415
#> Nov 1990   6.24947544    0.2354941
#> Dec 1990   6.80719557  147.0206284
#> Jan 1991  22.51140001  -77.7484495
#> Feb 1991 -18.42120983   70.4613433
#> Mar 1991  -0.88932029   69.6606398
#> Apr 1991  -6.54185827   93.1972929
#> May 1991  -3.71089697  261.6178688
#> Jun 1991   7.32142682 -147.7492275
#> Jul 1991   1.58182934    7.4677229
#> Aug 1991  10.05473158  -64.0770175
#> Sep 1991  -2.97361367   11.4925465
#> Oct 1991  -2.35435164   74.7021300
#> Nov 1991  -2.63170847 -111.8479256
#> Dec 1991  -3.20464012   34.1004819
#> Jan 1992  19.30659239   12.5031186
#> Feb 1992 -21.87774508   31.1369721
#> Mar 1992   7.00974005   42.6443804
#> Apr 1992  -3.76301495 -126.9579318
#> May 1992   3.35181016   86.9811052
#> Jun 1992  -2.93525659   -2.3574630
#> Jul 1992   0.11142179   64.6068260
#> Aug 1992  11.98017338  123.2672999
#> Sep 1992  -4.67733450 -128.0410543
#> Oct 1992   9.16382866  -17.6086170
#> Nov 1992  -7.06060331  114.8912321
#> Dec 1992   7.86935538   10.6409425
#> Jan 1993   3.99883945  105.2599358
#> Feb 1993  -6.38705322    6.3508668
#> Mar 1993  -3.43121408   76.0288668
#> Apr 1993 -11.18278418  -58.3391234
#> May 1993   5.78376820  -83.3176971
#> Jun 1993   1.13186948  -95.5279258
#> Jul 1993   4.40206334   46.8197456
#> Aug 1993 -20.02320938  148.9361598
#> Sep 1993  10.41633836   84.5637803
#> Oct 1993 -10.13748921   -9.0330359
#> Nov 1993  -5.04802858  -76.0974092
#> Dec 1993   0.89953043  -42.2703485
#> Jan 1994 -12.92000906 -111.6990551
#> Feb 1994 -17.30530080 -137.3205437
#> Mar 1994  11.32424775  -46.4627202
#> Apr 1994   9.55398518  -75.1982369
#> May 1994  -7.14008278  -40.7974543
#> Jun 1994  -1.30393939   17.7182569
#> Jul 1994   9.50804885   99.9326344
#> Aug 1994   4.70794223   11.6269256
#> Sep 1994   3.34838125 -183.5065315
#> Oct 1994 -15.49423258  -29.8695682
#> Nov 1994  -6.21150422   31.4652720
#> Dec 1994 -16.15618228 -108.2466160
#> Jan 1995   8.53465374   22.1793014
#> Feb 1995  18.88819161   -4.9851303
#> Mar 1995  -8.67441159  110.7534869
#> Apr 1995   7.14420496  -99.3360619
#> May 1995   6.14427022  192.4843474
#> Jun 1995   3.05098743  -55.6762404
#> Jul 1995  -9.25999799   32.9474662
#> Aug 1995  13.59534396  175.8286351
#> Sep 1995   2.03698336  127.3909790
#> Oct 1995  12.88543508  -58.3885195
#> Nov 1995  -2.12971834  -32.7692378
#> Dec 1995   4.87549814   14.5241262
#> Jan 1996 -10.67709165   21.4101138
#> Feb 1996 -11.80488666    1.1554669
#> Mar 1996  21.59175760  121.9699084
#> Apr 1996  -4.17812648 -103.0643641
#> May 1996   0.87654958   76.4734975
#> Jun 1996  -5.43709068  -51.4982604
#> Jul 1996  -1.19867022  174.0986972
#> Aug 1996  18.00624858  100.5845241
#> Sep 1996  -7.94247350 -110.7352945
#> Oct 1996  -3.60360559  -62.7882790
#> Nov 1996   6.31460646 -147.9501948
#> Dec 1996 -16.63279352   15.4343173
#> Jan 1997  -4.88005245 -229.9484409
#> Feb 1997  -0.20752219  -43.8727684
#> Mar 1997   9.33871038  -64.9680607
#> Apr 1997  -2.32670878   42.4040947
#> May 1997 -10.19630064    4.7049728
#> Jun 1997 -16.80328321 -128.5044066
#> Jul 1997  22.81394917    4.5532358
#> Aug 1997 -10.42285303 -144.9429700
#> Sep 1997   7.12307192 -163.3704403
#> Oct 1997  12.20747981  -51.7284962
#> Nov 1997   3.36018683  -99.1151430
#> Dec 1997  -2.79147937  -72.8485045
#> Jan 1998   0.76422131   45.1951634
#> Feb 1998  -3.54509030   88.7181335
#> Mar 1998  14.32658547  126.4573401
#> Apr 1998   0.02671474 -129.1419391
#> May 1998   3.44012511   18.2510306
#> Jun 1998  -8.30323380   -8.9809675
#> Jul 1998   4.82115591  -11.2629276
#> Aug 1998  -6.88660369   16.2801338
#> Sep 1998 -10.93359720   55.5861006
#> Oct 1998   8.41060978   85.3394680
#> Nov 1998  14.54136238  -19.4899361
#> Dec 1998   1.37240272   -4.3411909
#> Jan 1999   3.25544925  -52.9438316
#> Feb 1999  12.23728019  -59.8510599
#> Mar 1999 -10.53314136   14.9297682
#> Apr 1999  18.40332640  125.5820916
#> May 1999 -22.31484275  -23.4369528
#> Jun 1999  -5.95844719  142.2340276
#> Jul 1999  -2.30472754   29.2127542
#> Aug 1999  17.56315261  -73.6886751
#> Sep 1999   4.75705566  -12.9996552
#> Oct 1999   7.82644366  220.0885652
#> Nov 1999 -11.46245649  161.3081286
#> Dec 1999   9.50180682  -26.7499583
#> Jan 2000  -3.45227990   12.7603168
#> Feb 2000 -24.73551488  -37.0892736
#> Mar 2000  13.76264164  -45.2402281
#> Apr 2000  24.77145067  119.7478248
#> May 2000 -13.37234198   96.1339696
#> Jun 2000  20.24155203  109.9244552
#> Jul 2000   8.59593181    5.3534542
#> Aug 2000   7.28471344   -7.1082557
#> Sep 2000  17.02971138  -26.5212622
#> Oct 2000 -13.49671966 -108.7472787
#> Nov 2000   0.85960972   11.2995341
#> Dec 2000   2.04207089  -36.8761608
#> Jan 2001  -9.79050445 -181.3818422
#> Feb 2001   5.94650797  -72.7994368
#> Mar 2001  10.39317096  -32.3343456
#> Apr 2001 -12.78834279   11.8178061
#> May 2001  -3.19978820   52.2721179
#> Jun 2001 -18.72828606  -39.4712044
#> Jul 2001   0.91322657  -73.8262361
#> Aug 2001   7.12363770  131.9700103
#> Sep 2001  -1.10119494  -54.7717472
#> Oct 2001 -10.51575816  211.6530743
#> Nov 2001  -8.91401342   23.8889683
#> Dec 2001  -6.19138043  -51.5969374
#> Jan 2002   2.55199046  115.8752180
#> Feb 2002  -4.39315917  -63.4357812
#> Mar 2002  14.68149212   86.2710827
#> Apr 2002   8.32506126 -168.1573026
#> May 2002  -9.13787407 -163.6792119
#> Jun 2002  -2.72595426  143.0290161
#> Jul 2002   7.51086555  153.9376495
#> Aug 2002  -0.76856255   -7.7210862
#> Sep 2002   1.47101528   72.2232423
#> Oct 2002  -3.21238201   76.2012389
#> Nov 2002 -17.03140073 -110.5216744
#> Dec 2002  -7.61038444 -146.1168585
#> Jan 2003  -7.59025393   29.5561895
#> Feb 2003   2.92608648   65.8526506
#> Mar 2003   3.32236386   64.5142725
#> Apr 2003  -2.31934768   62.3653458
#> May 2003  -8.15807168  231.5849301
#> Jun 2003   2.16074797  -38.1598877
#> Jul 2003 -11.38865785    3.1688164
#> Aug 2003   1.40105007   77.6625400
#> Sep 2003  -6.62133906 -198.3047831
#> Oct 2003  -5.34391766   -0.5446710
#> Nov 2003 -14.12830881   70.2411713
#> Dec 2003  12.17203500  122.3164016
#> Jan 2004   4.25531888  -14.1397453
#> Feb 2004  -1.20928613  -49.5379010
#> Mar 2004   6.51430723   84.0522626
#> Apr 2004   4.90205565   56.0050141
#> May 2004   7.05192456  113.7045939
#> Jun 2004   8.05650671   47.3801310
#> Jul 2004  -1.79990583  -15.0440355
#> Aug 2004  10.72861877  213.8995235
#> Sep 2004  -0.82523465   21.7996549
#> Oct 2004  10.72598758 -189.5969181
#> Nov 2004  -0.02079741  -14.2738393
#> Dec 2004  -8.59905321  146.2425734
#> Jan 2005   7.48074130  107.4657122
#> Feb 2005  13.56515448   65.5056672
#> Mar 2005   4.00001599  -12.3133658
#> Apr 2005   6.36256611  -92.9721805
#> May 2005 -18.66035907  151.6556529
#> Jun 2005  -5.80927235   -7.6792959
#> Jul 2005  10.93896134  -36.9791935
#> Aug 2005  -5.32691097   60.6232668
#> Sep 2005   6.83984092  152.6737997
#> Oct 2005  -6.84263667   -8.1510851
#> Nov 2005  16.25943271  142.8358850
#> Dec 2005  -6.31734454  101.3571835
#> Jan 2006 -13.53779986  -67.6577827
#> Feb 2006   3.87122228  -23.6565665
#> Mar 2006  -9.67667515   50.5599115
#> Apr 2006  -3.58785321 -252.7752928
#> May 2006  -5.86397434  -70.0668673
#> Jun 2006  16.12913478    9.1607577
#> Jul 2006   5.05671343   85.2173585
#> Aug 2006   0.24106090  112.3909322
#> Sep 2006   2.95359647   82.7869298
#> Oct 2006  12.75350360 -136.8112810
#> Nov 2006   2.04477180   18.9759677
#> Dec 2006   1.86747913  196.0755931
#> Jan 2007   7.09749842  -80.6017715
#> Feb 2007   3.75846729   24.7350268
#> Mar 2007   4.05413237   50.7828231
#> Apr 2007   3.30985753  -49.5801751
#> May 2007  15.93760511  -90.2102368
#> Jun 2007   5.55075757  101.2532538
#> Jul 2007 -12.09478318  163.6540567
#> Aug 2007  -2.25557951  -82.5520511
#> Sep 2007  -3.42381835 -130.1284005
#> Oct 2007 -17.36559496  -48.4042607
#> Nov 2007   1.63560903  -19.1561941
#> Dec 2007   2.98293513 -120.3991350
#> Jan 2008   1.75870029 -155.4565833
#> Feb 2008 -14.05601011  -14.9289700
#> Mar 2008  -2.15255891  -98.8248953
#> Apr 2008  -5.02680266 -130.0530931
#> May 2008  16.83881702   76.4001921
#> Jun 2008 -17.25036625  -76.4220023
#> Jul 2008  -2.52554758  105.9842219
#> Aug 2008 -16.04868007   57.4400444
#> Sep 2008  11.53376302  145.4354816
#> Oct 2008 -12.66363058  192.6606505
#> Nov 2008  -2.50127880   37.7693133
#> Dec 2008  -5.38988584   45.7817317
#> Jan 2009  -6.11159385  -31.4478166
#> Feb 2009   8.07837347  -19.6337224
#> Mar 2009  13.57967557   53.2207247
#> Apr 2009  10.00928446 -129.6930815
#> May 2009 -12.58705305  -11.6459881
#> Jun 2009 -13.95429138 -179.1435814
#> Jul 2009   3.26775093  103.7638450
#> Aug 2009  -5.66269784  -75.1737315
#> Sep 2009  -0.40093207  144.3320207
#> Oct 2009  10.07425936   73.4189087
#> Nov 2009 -13.73520862 -105.4557986
#> Dec 2009   3.30858231    3.0424727
#> Jan 2010  11.43623543  -62.0591656
#> Feb 2010 -13.81713624  -19.3502015
#> Mar 2010  -8.44069552   39.3606874
#> Apr 2010 -18.36583479   -3.3226158
#> May 2010  10.44170831  -52.8722573
#> Jun 2010   2.15210107  307.1007147
#> Jul 2010   8.00349614 -142.4723774
#> Aug 2010   7.14933031   33.0701264
#> Sep 2010  -3.17171209 -175.9479023
#> Oct 2010 -22.52045124 -101.6384201
#> Nov 2010  15.05248197  103.6862040
#> Dec 2010  -4.19147872    3.4712279
#> Jan 2011  -2.08179987 -169.3691356
#> Feb 2011  10.25008622  -37.4082127
#> Mar 2011 -21.09352248 -125.8362453
#> Apr 2011   5.81155983  -60.8105630
#> May 2011   4.27476044 -136.8707861
#> Jun 2011   3.09712183  -94.2163446
#> Jul 2011  -3.48680801   73.9857730
#> Aug 2011   4.72684011    2.2094341
#> Sep 2011   7.02742649   54.8882732
#> Oct 2011 -15.43797498 -135.0698058
#> Nov 2011 -11.39419424  -64.4467633
#> Dec 2011   3.70802634  104.7844734
#> Jan 2012   2.49291812  -32.1393364
#> Feb 2012  12.40415284    3.4325883
#> Mar 2012  -3.71594045  106.3948370
#> Apr 2012   9.08843016 -158.7608154
#> May 2012   7.72073385 -158.3462404
#> Jun 2012   3.66647135   11.0531327
#> Jul 2012   2.34063054   24.6298837
#> Aug 2012  -8.24637494    5.7149917
#> Sep 2012   7.93613760 -108.3303495
#> Oct 2012   8.15375576   21.6301111
#> Nov 2012  -6.40412559  -82.3685525
#> Dec 2012   1.54668564 -150.0780397
#> Jan 2013 -13.19062104  122.6567010
#> Feb 2013  -5.59444239  -21.5867461
#> Mar 2013 -12.78683713  173.3784762
#> Apr 2013  -1.61854635 -143.2535684
#> May 2013 -12.12698055  -17.6478952
#> Jun 2013  13.71345622   75.6314868
#> Jul 2013   0.82627704  -98.0007712
#> Aug 2013 -12.65853773  -38.6424379
#> Sep 2013  11.74275795   38.1516254
#> Oct 2013   6.95991675  134.5768539
#> Nov 2013   8.00070376   98.3748677
#> Dec 2013 -10.02196891    8.6637298
#> Jan 2014  -6.77115667   46.9669925
#> Feb 2014   7.20963599  -89.2900170
#> Mar 2014   9.60168144    0.2229617
#> Apr 2014  -1.45665667   15.0747679
#> May 2014  15.05255540  -57.4554552
#> Jun 2014   6.78180406   23.5160435
#> Jul 2014  -2.65552196  126.5858408
#> Aug 2014  -4.58132537 -139.0298039
#> Sep 2014  17.38431533   11.5795724
#> Oct 2014  -0.13105949   89.2577296
#> Nov 2014  -7.94353354    3.8752879
#> Dec 2014   2.55333393   -3.8372698
#> Jan 2015  -8.78777648   81.1453346
#> Feb 2015  -5.02384014    3.3486905
#> Mar 2015   9.18235108   66.3318866
#> Apr 2015  -4.15556110  -22.0091795
#> May 2015  -4.97454127   -1.5736545
#> Jun 2015   1.22983974  208.7713501
#> Jul 2015  -9.94093841  149.6586647
#> Aug 2015   4.85937859  -30.9764025
#> Sep 2015  -5.51041758  158.4290933
#> Oct 2015   1.66600822 -114.7278380
#> Nov 2015   0.95564919   93.7994138
#> Dec 2015   9.37361737 -117.6473378
#> Jan 2016  -7.78592244  140.2133012
#> Feb 2016  -3.46014963  -17.7175377
#> Mar 2016   1.45927947  -65.0604263
#> Apr 2016  -6.79378277 -122.9364408
#> May 2016  -6.13698253   25.1325895
#> Jun 2016  22.02935147 -156.3215351
#> Jul 2016  -3.28772531   81.3079991
#> Aug 2016  17.51776850  -11.0848678
#> Sep 2016  -9.71651737  -76.1172061
#> Oct 2016 -16.89546894   82.9512582
#> Nov 2016 -14.81386178 -105.5171964
#> Dec 2016  24.26894783   22.2967963
#> Jan 2017  -2.28460020  244.1195038
#> Feb 2017  11.13283521  200.0641700
#> Mar 2017  -4.37063852  -75.4805825
#> Apr 2017  12.64884610 -130.3238961
#> May 2017  14.36597731   24.6026158
#> Jun 2017  -9.18664322   54.8893707
#> Jul 2017 -19.14602512   48.6414289
#> Aug 2017   1.14969531  -85.7244463
#> Sep 2017  15.10348121  -20.8889627
#> Oct 2017  -0.09052613   -2.1828351
#> Nov 2017  -1.35856744   37.1728862
#> Dec 2017  -3.36250668 -230.0614875
#> Jan 2018   3.57796527  -68.6391490
#> Feb 2018   3.40498580   82.3954712
#> Mar 2018  30.97230704   67.1560953
#> Apr 2018   6.77352741  -13.6811085
#> May 2018   3.85994964  -32.4855419
#> Jun 2018  -8.38842632  116.5364420
#> Jul 2018  -5.30027590  -90.3741440
#> Aug 2018 -22.93182084   98.5612115
#> Sep 2018  18.55263642  -12.6323527
#> Oct 2018  -6.24714964 -144.3967173
#> Nov 2018   1.75872582    3.2154077
#> Dec 2018 -13.65788853 -211.7716447
#> Jan 2019 -13.45518474  166.5816928
#> Feb 2019   9.47265271   57.5671055
#> Mar 2019  -1.15739671   -0.2235961
#> Apr 2019  -4.29404494  -72.2014992
#> May 2019   0.79631976   44.8672361
#> Jun 2019  10.06463994    5.1736240
#> Jul 2019  17.54518747 -125.3649454
#> Aug 2019  -1.19361311   55.8374959
#> Sep 2019  16.43998625   29.8702402
#> Oct 2019 -13.10997765  -41.5528828
#> Nov 2019   3.62701424 -162.5374469
#> Dec 2019  13.69665491   29.9647475
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 

s_preVar(myreg3)

#> $series
#>                  var1         var2
#> Jan 1990   7.17057189  -34.8512400
#> Feb 1990   4.05856898  -54.5470652
#> Mar 1990 -14.99181516   -2.0519678
#> Apr 1990 -12.23578511  -33.8849889
#> May 1990  14.25347975   18.5592100
#> Jun 1990   5.35483990   34.5862657
#> Jul 1990  -4.79124096  -79.3332374
#> Aug 1990   7.08755575  -74.8631612
#> Sep 1990  -0.10940710 -184.6666644
#> Oct 1990 -13.92609039 -155.6824415
#> Nov 1990   6.24947544    0.2354941
#> Dec 1990   6.80719557  147.0206284
#> Jan 1991  22.51140001  -77.7484495
#> Feb 1991 -18.42120983   70.4613433
#> Mar 1991  -0.88932029   69.6606398
#> Apr 1991  -6.54185827   93.1972929
#> May 1991  -3.71089697  261.6178688
#> Jun 1991   7.32142682 -147.7492275
#> Jul 1991   1.58182934    7.4677229
#> Aug 1991  10.05473158  -64.0770175
#> Sep 1991  -2.97361367   11.4925465
#> Oct 1991  -2.35435164   74.7021300
#> Nov 1991  -2.63170847 -111.8479256
#> Dec 1991  -3.20464012   34.1004819
#> Jan 1992  19.30659239   12.5031186
#> Feb 1992 -21.87774508   31.1369721
#> Mar 1992   7.00974005   42.6443804
#> Apr 1992  -3.76301495 -126.9579318
#> May 1992   3.35181016   86.9811052
#> Jun 1992  -2.93525659   -2.3574630
#> Jul 1992   0.11142179   64.6068260
#> Aug 1992  11.98017338  123.2672999
#> Sep 1992  -4.67733450 -128.0410543
#> Oct 1992   9.16382866  -17.6086170
#> Nov 1992  -7.06060331  114.8912321
#> Dec 1992   7.86935538   10.6409425
#> Jan 1993   3.99883945  105.2599358
#> Feb 1993  -6.38705322    6.3508668
#> Mar 1993  -3.43121408   76.0288668
#> Apr 1993 -11.18278418  -58.3391234
#> May 1993   5.78376820  -83.3176971
#> Jun 1993   1.13186948  -95.5279258
#> Jul 1993   4.40206334   46.8197456
#> Aug 1993 -20.02320938  148.9361598
#> Sep 1993  10.41633836   84.5637803
#> Oct 1993 -10.13748921   -9.0330359
#> Nov 1993  -5.04802858  -76.0974092
#> Dec 1993   0.89953043  -42.2703485
#> Jan 1994 -12.92000906 -111.6990551
#> Feb 1994 -17.30530080 -137.3205437
#> Mar 1994  11.32424775  -46.4627202
#> Apr 1994   9.55398518  -75.1982369
#> May 1994  -7.14008278  -40.7974543
#> Jun 1994  -1.30393939   17.7182569
#> Jul 1994   9.50804885   99.9326344
#> Aug 1994   4.70794223   11.6269256
#> Sep 1994   3.34838125 -183.5065315
#> Oct 1994 -15.49423258  -29.8695682
#> Nov 1994  -6.21150422   31.4652720
#> Dec 1994 -16.15618228 -108.2466160
#> Jan 1995   8.53465374   22.1793014
#> Feb 1995  18.88819161   -4.9851303
#> Mar 1995  -8.67441159  110.7534869
#> Apr 1995   7.14420496  -99.3360619
#> May 1995   6.14427022  192.4843474
#> Jun 1995   3.05098743  -55.6762404
#> Jul 1995  -9.25999799   32.9474662
#> Aug 1995  13.59534396  175.8286351
#> Sep 1995   2.03698336  127.3909790
#> Oct 1995  12.88543508  -58.3885195
#> Nov 1995  -2.12971834  -32.7692378
#> Dec 1995   4.87549814   14.5241262
#> Jan 1996 -10.67709165   21.4101138
#> Feb 1996 -11.80488666    1.1554669
#> Mar 1996  21.59175760  121.9699084
#> Apr 1996  -4.17812648 -103.0643641
#> May 1996   0.87654958   76.4734975
#> Jun 1996  -5.43709068  -51.4982604
#> Jul 1996  -1.19867022  174.0986972
#> Aug 1996  18.00624858  100.5845241
#> Sep 1996  -7.94247350 -110.7352945
#> Oct 1996  -3.60360559  -62.7882790
#> Nov 1996   6.31460646 -147.9501948
#> Dec 1996 -16.63279352   15.4343173
#> Jan 1997  -4.88005245 -229.9484409
#> Feb 1997  -0.20752219  -43.8727684
#> Mar 1997   9.33871038  -64.9680607
#> Apr 1997  -2.32670878   42.4040947
#> May 1997 -10.19630064    4.7049728
#> Jun 1997 -16.80328321 -128.5044066
#> Jul 1997  22.81394917    4.5532358
#> Aug 1997 -10.42285303 -144.9429700
#> Sep 1997   7.12307192 -163.3704403
#> Oct 1997  12.20747981  -51.7284962
#> Nov 1997   3.36018683  -99.1151430
#> Dec 1997  -2.79147937  -72.8485045
#> Jan 1998   0.76422131   45.1951634
#> Feb 1998  -3.54509030   88.7181335
#> Mar 1998  14.32658547  126.4573401
#> Apr 1998   0.02671474 -129.1419391
#> May 1998   3.44012511   18.2510306
#> Jun 1998  -8.30323380   -8.9809675
#> Jul 1998   4.82115591  -11.2629276
#> Aug 1998  -6.88660369   16.2801338
#> Sep 1998 -10.93359720   55.5861006
#> Oct 1998   8.41060978   85.3394680
#> Nov 1998  14.54136238  -19.4899361
#> Dec 1998   1.37240272   -4.3411909
#> Jan 1999   3.25544925  -52.9438316
#> Feb 1999  12.23728019  -59.8510599
#> Mar 1999 -10.53314136   14.9297682
#> Apr 1999  18.40332640  125.5820916
#> May 1999 -22.31484275  -23.4369528
#> Jun 1999  -5.95844719  142.2340276
#> Jul 1999  -2.30472754   29.2127542
#> Aug 1999  17.56315261  -73.6886751
#> Sep 1999   4.75705566  -12.9996552
#> Oct 1999   7.82644366  220.0885652
#> Nov 1999 -11.46245649  161.3081286
#> Dec 1999   9.50180682  -26.7499583
#> Jan 2000  -3.45227990   12.7603168
#> Feb 2000 -24.73551488  -37.0892736
#> Mar 2000  13.76264164  -45.2402281
#> Apr 2000  24.77145067  119.7478248
#> May 2000 -13.37234198   96.1339696
#> Jun 2000  20.24155203  109.9244552
#> Jul 2000   8.59593181    5.3534542
#> Aug 2000   7.28471344   -7.1082557
#> Sep 2000  17.02971138  -26.5212622
#> Oct 2000 -13.49671966 -108.7472787
#> Nov 2000   0.85960972   11.2995341
#> Dec 2000   2.04207089  -36.8761608
#> Jan 2001  -9.79050445 -181.3818422
#> Feb 2001   5.94650797  -72.7994368
#> Mar 2001  10.39317096  -32.3343456
#> Apr 2001 -12.78834279   11.8178061
#> May 2001  -3.19978820   52.2721179
#> Jun 2001 -18.72828606  -39.4712044
#> Jul 2001   0.91322657  -73.8262361
#> Aug 2001   7.12363770  131.9700103
#> Sep 2001  -1.10119494  -54.7717472
#> Oct 2001 -10.51575816  211.6530743
#> Nov 2001  -8.91401342   23.8889683
#> Dec 2001  -6.19138043  -51.5969374
#> Jan 2002   2.55199046  115.8752180
#> Feb 2002  -4.39315917  -63.4357812
#> Mar 2002  14.68149212   86.2710827
#> Apr 2002   8.32506126 -168.1573026
#> May 2002  -9.13787407 -163.6792119
#> Jun 2002  -2.72595426  143.0290161
#> Jul 2002   7.51086555  153.9376495
#> Aug 2002  -0.76856255   -7.7210862
#> Sep 2002   1.47101528   72.2232423
#> Oct 2002  -3.21238201   76.2012389
#> Nov 2002 -17.03140073 -110.5216744
#> Dec 2002  -7.61038444 -146.1168585
#> Jan 2003  -7.59025393   29.5561895
#> Feb 2003   2.92608648   65.8526506
#> Mar 2003   3.32236386   64.5142725
#> Apr 2003  -2.31934768   62.3653458
#> May 2003  -8.15807168  231.5849301
#> Jun 2003   2.16074797  -38.1598877
#> Jul 2003 -11.38865785    3.1688164
#> Aug 2003   1.40105007   77.6625400
#> Sep 2003  -6.62133906 -198.3047831
#> Oct 2003  -5.34391766   -0.5446710
#> Nov 2003 -14.12830881   70.2411713
#> Dec 2003  12.17203500  122.3164016
#> Jan 2004   4.25531888  -14.1397453
#> Feb 2004  -1.20928613  -49.5379010
#> Mar 2004   6.51430723   84.0522626
#> Apr 2004   4.90205565   56.0050141
#> May 2004   7.05192456  113.7045939
#> Jun 2004   8.05650671   47.3801310
#> Jul 2004  -1.79990583  -15.0440355
#> Aug 2004  10.72861877  213.8995235
#> Sep 2004  -0.82523465   21.7996549
#> Oct 2004  10.72598758 -189.5969181
#> Nov 2004  -0.02079741  -14.2738393
#> Dec 2004  -8.59905321  146.2425734
#> Jan 2005   7.48074130  107.4657122
#> Feb 2005  13.56515448   65.5056672
#> Mar 2005   4.00001599  -12.3133658
#> Apr 2005   6.36256611  -92.9721805
#> May 2005 -18.66035907  151.6556529
#> Jun 2005  -5.80927235   -7.6792959
#> Jul 2005  10.93896134  -36.9791935
#> Aug 2005  -5.32691097   60.6232668
#> Sep 2005   6.83984092  152.6737997
#> Oct 2005  -6.84263667   -8.1510851
#> Nov 2005  16.25943271  142.8358850
#> Dec 2005  -6.31734454  101.3571835
#> Jan 2006 -13.53779986  -67.6577827
#> Feb 2006   3.87122228  -23.6565665
#> Mar 2006  -9.67667515   50.5599115
#> Apr 2006  -3.58785321 -252.7752928
#> May 2006  -5.86397434  -70.0668673
#> Jun 2006  16.12913478    9.1607577
#> Jul 2006   5.05671343   85.2173585
#> Aug 2006   0.24106090  112.3909322
#> Sep 2006   2.95359647   82.7869298
#> Oct 2006  12.75350360 -136.8112810
#> Nov 2006   2.04477180   18.9759677
#> Dec 2006   1.86747913  196.0755931
#> Jan 2007   7.09749842  -80.6017715
#> Feb 2007   3.75846729   24.7350268
#> Mar 2007   4.05413237   50.7828231
#> Apr 2007   3.30985753  -49.5801751
#> May 2007  15.93760511  -90.2102368
#> Jun 2007   5.55075757  101.2532538
#> Jul 2007 -12.09478318  163.6540567
#> Aug 2007  -2.25557951  -82.5520511
#> Sep 2007  -3.42381835 -130.1284005
#> Oct 2007 -17.36559496  -48.4042607
#> Nov 2007   1.63560903  -19.1561941
#> Dec 2007   2.98293513 -120.3991350
#> Jan 2008   1.75870029 -155.4565833
#> Feb 2008 -14.05601011  -14.9289700
#> Mar 2008  -2.15255891  -98.8248953
#> Apr 2008  -5.02680266 -130.0530931
#> May 2008  16.83881702   76.4001921
#> Jun 2008 -17.25036625  -76.4220023
#> Jul 2008  -2.52554758  105.9842219
#> Aug 2008 -16.04868007   57.4400444
#> Sep 2008  11.53376302  145.4354816
#> Oct 2008 -12.66363058  192.6606505
#> Nov 2008  -2.50127880   37.7693133
#> Dec 2008  -5.38988584   45.7817317
#> Jan 2009  -6.11159385  -31.4478166
#> Feb 2009   8.07837347  -19.6337224
#> Mar 2009  13.57967557   53.2207247
#> Apr 2009  10.00928446 -129.6930815
#> May 2009 -12.58705305  -11.6459881
#> Jun 2009 -13.95429138 -179.1435814
#> Jul 2009   3.26775093  103.7638450
#> Aug 2009  -5.66269784  -75.1737315
#> Sep 2009  -0.40093207  144.3320207
#> Oct 2009  10.07425936   73.4189087
#> Nov 2009 -13.73520862 -105.4557986
#> Dec 2009   3.30858231    3.0424727
#> Jan 2010  11.43623543  -62.0591656
#> Feb 2010 -13.81713624  -19.3502015
#> Mar 2010  -8.44069552   39.3606874
#> Apr 2010 -18.36583479   -3.3226158
#> May 2010  10.44170831  -52.8722573
#> Jun 2010   2.15210107  307.1007147
#> Jul 2010   8.00349614 -142.4723774
#> Aug 2010   7.14933031   33.0701264
#> Sep 2010  -3.17171209 -175.9479023
#> Oct 2010 -22.52045124 -101.6384201
#> Nov 2010  15.05248197  103.6862040
#> Dec 2010  -4.19147872    3.4712279
#> Jan 2011  -2.08179987 -169.3691356
#> Feb 2011  10.25008622  -37.4082127
#> Mar 2011 -21.09352248 -125.8362453
#> Apr 2011   5.81155983  -60.8105630
#> May 2011   4.27476044 -136.8707861
#> Jun 2011   3.09712183  -94.2163446
#> Jul 2011  -3.48680801   73.9857730
#> Aug 2011   4.72684011    2.2094341
#> Sep 2011   7.02742649   54.8882732
#> Oct 2011 -15.43797498 -135.0698058
#> Nov 2011 -11.39419424  -64.4467633
#> Dec 2011   3.70802634  104.7844734
#> Jan 2012   2.49291812  -32.1393364
#> Feb 2012  12.40415284    3.4325883
#> Mar 2012  -3.71594045  106.3948370
#> Apr 2012   9.08843016 -158.7608154
#> May 2012   7.72073385 -158.3462404
#> Jun 2012   3.66647135   11.0531327
#> Jul 2012   2.34063054   24.6298837
#> Aug 2012  -8.24637494    5.7149917
#> Sep 2012   7.93613760 -108.3303495
#> Oct 2012   8.15375576   21.6301111
#> Nov 2012  -6.40412559  -82.3685525
#> Dec 2012   1.54668564 -150.0780397
#> Jan 2013 -13.19062104  122.6567010
#> Feb 2013  -5.59444239  -21.5867461
#> Mar 2013 -12.78683713  173.3784762
#> Apr 2013  -1.61854635 -143.2535684
#> May 2013 -12.12698055  -17.6478952
#> Jun 2013  13.71345622   75.6314868
#> Jul 2013   0.82627704  -98.0007712
#> Aug 2013 -12.65853773  -38.6424379
#> Sep 2013  11.74275795   38.1516254
#> Oct 2013   6.95991675  134.5768539
#> Nov 2013   8.00070376   98.3748677
#> Dec 2013 -10.02196891    8.6637298
#> Jan 2014  -6.77115667   46.9669925
#> Feb 2014   7.20963599  -89.2900170
#> Mar 2014   9.60168144    0.2229617
#> Apr 2014  -1.45665667   15.0747679
#> May 2014  15.05255540  -57.4554552
#> Jun 2014   6.78180406   23.5160435
#> Jul 2014  -2.65552196  126.5858408
#> Aug 2014  -4.58132537 -139.0298039
#> Sep 2014  17.38431533   11.5795724
#> Oct 2014  -0.13105949   89.2577296
#> Nov 2014  -7.94353354    3.8752879
#> Dec 2014   2.55333393   -3.8372698
#> Jan 2015  -8.78777648   81.1453346
#> Feb 2015  -5.02384014    3.3486905
#> Mar 2015   9.18235108   66.3318866
#> Apr 2015  -4.15556110  -22.0091795
#> May 2015  -4.97454127   -1.5736545
#> Jun 2015   1.22983974  208.7713501
#> Jul 2015  -9.94093841  149.6586647
#> Aug 2015   4.85937859  -30.9764025
#> Sep 2015  -5.51041758  158.4290933
#> Oct 2015   1.66600822 -114.7278380
#> Nov 2015   0.95564919   93.7994138
#> Dec 2015   9.37361737 -117.6473378
#> Jan 2016  -7.78592244  140.2133012
#> Feb 2016  -3.46014963  -17.7175377
#> Mar 2016   1.45927947  -65.0604263
#> Apr 2016  -6.79378277 -122.9364408
#> May 2016  -6.13698253   25.1325895
#> Jun 2016  22.02935147 -156.3215351
#> Jul 2016  -3.28772531   81.3079991
#> Aug 2016  17.51776850  -11.0848678
#> Sep 2016  -9.71651737  -76.1172061
#> Oct 2016 -16.89546894   82.9512582
#> Nov 2016 -14.81386178 -105.5171964
#> Dec 2016  24.26894783   22.2967963
#> Jan 2017  -2.28460020  244.1195038
#> Feb 2017  11.13283521  200.0641700
#> Mar 2017  -4.37063852  -75.4805825
#> Apr 2017  12.64884610 -130.3238961
#> May 2017  14.36597731   24.6026158
#> Jun 2017  -9.18664322   54.8893707
#> Jul 2017 -19.14602512   48.6414289
#> Aug 2017   1.14969531  -85.7244463
#> Sep 2017  15.10348121  -20.8889627
#> Oct 2017  -0.09052613   -2.1828351
#> Nov 2017  -1.35856744   37.1728862
#> Dec 2017  -3.36250668 -230.0614875
#> Jan 2018   3.57796527  -68.6391490
#> Feb 2018   3.40498580   82.3954712
#> Mar 2018  30.97230704   67.1560953
#> Apr 2018   6.77352741  -13.6811085
#> May 2018   3.85994964  -32.4855419
#> Jun 2018  -8.38842632  116.5364420
#> Jul 2018  -5.30027590  -90.3741440
#> Aug 2018 -22.93182084   98.5612115
#> Sep 2018  18.55263642  -12.6323527
#> Oct 2018  -6.24714964 -144.3967173
#> Nov 2018   1.75872582    3.2154077
#> Dec 2018 -13.65788853 -211.7716447
#> Jan 2019 -13.45518474  166.5816928
#> Feb 2019   9.47265271   57.5671055
#> Mar 2019  -1.15739671   -0.2235961
#> Apr 2019  -4.29404494  -72.2014992
#> May 2019   0.79631976   44.8672361
#> Jun 2019  10.06463994    5.1736240
#> Jul 2019  17.54518747 -125.3649454
#> Aug 2019  -1.19361311   55.8374959
#> Sep 2019  16.43998625   29.8702402
#> Oct 2019 -13.10997765  -41.5528828
#> Nov 2019   3.62701424 -162.5374469
#> Dec 2019  13.69665491   29.9647475
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 

s_td(myreg1)

#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove

s_td(myspec1)

#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove

s_easter(myreg1)

#>  enabled julian duration test
#>     TRUE  FALSE        8  Add

s_easter(myspec1)

#>  enabled julian duration test
#>     TRUE  FALSE        8  Add

s_out(myreg1)

#>  enabled span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE  All TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7

s_out(myspec1)

#>  enabled            span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE From 2010-03-01 TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7

s_arima(myreg1)

#>  enabled automdl.acceptdefault automdl.cancel automdl.ub1 automdl.ub2
#>     TRUE                 FALSE            0.1    1.041667        0.88
#>  automdl.mixed automdl.balanced automdl.armalimit automdl.reducecv
#>           TRUE            FALSE                 1          0.14286
#>  automdl.ljungboxlimit automdl.ubfinal arima.mu arima.p arima.d arima.q
#>                   0.95            1.05    FALSE       0       1       1
#>  arima.bp arima.bd arima.bq arima.coef
#>         0        1        1      FALSE

s_arima(myspec1)

#>  enabled automdl.acceptdefault automdl.cancel automdl.ub1 automdl.ub2
#>     TRUE                 FALSE            0.1    1.041667        0.88
#>  automdl.mixed automdl.balanced automdl.armalimit automdl.reducecv
#>           TRUE            FALSE                 1          0.14286
#>  automdl.ljungboxlimit automdl.ubfinal arima.mu arima.p arima.d arima.q
#>                   0.95            1.05    FALSE       0       1       1
#>  arima.bp arima.bd arima.bq arima.coef
#>         0        1        1      FALSE

myspec4 <- regarima_spec_x13(myreg1, automdl.enabled = FALSE,
             arima.coefEnabled = TRUE,
             arima.p = 1,arima.q = 1, arima.bp = 1, arima.bq = 1,
             arima.coef = rep(0.2, 4),
             arima.coefType = rep("Initial", 4))
myreg4 <- regarima(myseries, myspec4)

s_arimaCoef(myreg4)

#>              Type Value
#> Phi(1)    Initial   0.2
#> Theta(1)  Initial   0.2
#> BPhi(1)   Initial   0.2
#> BTheta(1) Initial   0.2

s_arimaCoef(myspec4)

#>              Type Value
#> Phi(1)    Initial   0.2
#> Theta(1)  Initial   0.2
#> BPhi(1)   Initial   0.2
#> BTheta(1) Initial   0.2

s_fcst(myreg1)

#>  horizon
#>       -2

s_fcst(myspec1)

#>  horizon
#>       -2

s_span(myreg1)

#>          type         d0         d1 n0 n1
#> estimate  All 1900-01-01 2020-12-31  0  0
#> outlier   All 1900-01-01 2020-12-31  0  0

s_span(myspec1)

#>          type         d0         d1 n0 n1
#> estimate From 2005-10-01 2020-12-31  0  0
#> outlier  From 2010-03-01 2020-12-31  0  0

myspec5 <- x13_spec(spec = "RSA5c", x11.seasonalComp = FALSE)
mysa5 <- x13(myseries, myspec5)

s_x11(mysa5)

#>   x11.mode x11.seasonalComp x11.lsigma x11.usigma x11.trendAuto x11.trendma
#>  Undefined            FALSE        1.5        2.5          TRUE          13
#>  x11.seasonalma x11.fcasts x11.bcasts x11.calendarSigma x11.sigmaVector
#>             Msr         -1          0              None              NA
#>  x11.excludeFcasts
#>              FALSE

s_x11(myspec5)

#>   x11.mode x11.seasonalComp x11.lsigma x11.usigma x11.trendAuto x11.trendma
#>  Undefined            FALSE        1.5        2.5          TRUE          13
#>  x11.seasonalma x11.fcasts x11.bcasts x11.calendarSigma x11.sigmaVector
#>             Msr         -1          0              None              NA
#>  x11.excludeFcasts
#>              FALSE

myspec6 <- tramoseats_spec(spec = "RSAfull", seats.approx = "Noisy")
mysa6 <- tramoseats(myseries, myspec6)

s_seats(mysa6)

#>  seats.predictionLength seats.approx seats.maBoundary seats.trendBoundary
#>                      -1        Noisy             0.95                 0.5
#>  seats.seasdBoundary seats.seasdBoundary1 seats.seasTol seats.method
#>                  0.8                  0.8             2       Burman

s_seats(mysa6)

#>  seats.predictionLength seats.approx seats.maBoundary seats.trendBoundary
#>                      -1        Noisy             0.95                 0.5
#>  seats.seasdBoundary seats.seasdBoundary1 seats.seasTol seats.method
#>                  0.8                  0.8             2       Burman

# }