TRAMO-SEATS model specification, SA/TRAMO-SEATS

Function to create (and/or modify) a c("SA_spec", "TRAMO_SEATS") class object with the SA model specification for the TRAMO-SEATS method. It can be done from a pre-defined 'JDemetra+' model specification (a character), a previous specification (c("SA_spec", "TRAMO_SEATS") object) or a seasonal adjustment model (c("SA", "TRAMO_SEATS") object).

tramoseats_spec(
  spec = c("RSAfull", "RSA0", "RSA1", "RSA2", "RSA3", "RSA4", "RSA5"),
  preliminary.check = NA,
  estimate.from = NA_character_,
  estimate.to = NA_character_,
  estimate.first = NA_integer_,
  estimate.last = NA_integer_,
  estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_,
  estimate.tol = NA_integer_,
  estimate.eml = NA,
  estimate.urfinal = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.fct = NA_integer_,
  usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA,
  usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA,
  usrdef.varEnabled = NA,
  usrdef.var = NA,
  usrdef.varType = NA,
  usrdef.varCoef = NA,
  tradingdays.mauto = c(NA, "Unused", "FTest", "WaldTest"),
  tradingdays.pftd = NA_integer_,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined", "None"),
  tradingdays.leapyear = NA,
  tradingdays.stocktd = NA_integer_,
  tradingdays.test = c(NA, "Separate_T", "Joint_F", "None"),
  easter.type = c(NA, "Unused", "Standard", "IncludeEaster", "IncludeEasterMonday"),
  easter.julian = NA,
  easter.duration = NA_integer_,
  easter.test = NA,
  outlier.enabled = NA,
  outlier.from = NA_character_,
  outlier.to = NA_character_,
  outlier.first = NA_integer_,
  outlier.last = NA_integer_,
  outlier.exclFirst = NA_integer_,
  outlier.exclLast = NA_integer_,
  outlier.ao = NA,
  outlier.tc = NA,
  outlier.ls = NA,
  outlier.so = NA,
  outlier.usedefcv = NA,
  outlier.cv = NA_integer_,
  outlier.eml = NA,
  outlier.tcrate = NA_integer_,
  automdl.enabled = NA,
  automdl.acceptdefault = NA,
  automdl.cancel = NA_integer_,
  automdl.ub1 = NA_integer_,
  automdl.ub2 = NA_integer_,
  automdl.armalimit = NA_integer_,
  automdl.reducecv = NA_integer_,
  automdl.ljungboxlimit = NA_integer_,
  automdl.compare = NA,
  arima.mu = NA,
  arima.p = NA_integer_,
  arima.d = NA_integer_,
  arima.q = NA_integer_,
  arima.bp = NA_integer_,
  arima.bd = NA_integer_,
  arima.bq = NA_integer_,
  arima.coefEnabled = NA,
  arima.coef = NA,
  arima.coefType = NA,
  fcst.horizon = NA_integer_,
  seats.predictionLength = NA_integer_,
  seats.approx = c(NA, "None", "Legacy", "Noisy"),
  seats.trendBoundary = NA_integer_,
  seats.seasdBoundary = NA_integer_,
  seats.seasdBoundary1 = NA_integer_,
  seats.seasTol = NA_integer_,
  seats.maBoundary = NA_integer_,
  seats.method = c(NA, "Burman", "KalmanSmoother", "McElroyMatrix")
)

Arguments

spec
model specification X13. It can be a character of predefined TRAMO-SEATS 'JDemetra+' model specification (see Details), an object of class c("SA_spec","TRAMO_SEATS") or an object of class c("SA", "TRAMO_SEATS"). The default is "RSAfull".

preliminary.check
boolean to check the quality of the input series and exclude highly problematic ones: e.g. these with a number of identical observations and/or missing values above pre-specified threshold values.

The time span of the series to be used for the estimation of the RegArima model coefficients (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: estimate.from, estimate.to, estimate.first, estimate.last, estimate.exclFirst and estimate.exclLast; where estimate.from and estimate.to have priority over remaining span control variables, estimate.last and estimate.first have priority over estimate.exclFirst and estimate.exclLast, and estimate.last has priority over estimate.first.

estimate.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with estimate.to.

estimate.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with estimate.from.

estimate.first
numeric specifying the number of periods considered at the beginning of the series.

estimate.last
numeric specifying the number of periods considered at the end of the series.

estimate.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with estimate.exclLast.

estimate.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with estimate.exclFirst.

estimate.tol
numeric, convergence tolerance. The absolute changes in the log-likelihood function are compared to this value to check for the convergence of the estimation iterations.

estimate.eml
logicals, exact maximum likelihood estimation. If TRUE the program performs an exact maximum likelihood estimation. If FASLE the Unconditional Least Squares method is used.

estimate.urfinal
numeric, final unit root limit. The threshold value for the final unit root test for identification of differencing orders. If the magnitude of an AR root for the final model is less than this number, a unit root is assumed, the order of the AR polynomial is reduced by one, and the appropriate order of the differencing (non-seasonal, seasonal) is increased.

transform.function
transformation of the input series: "None" - no transformation of the series; "Log" - takes the log of the series; "Auto" - the program tests for the log-level specification.

transform.fct
numeric controlling the bias in the log/level pre-test: transform.fct > 1 favours levels, transform.fct< 1 favours logs. Considered only when transform.function is set to "Auto".

Control variables for the pre-specified outliers. The pre-specified outliers are used in the model only if they are enabled (usrdef.outliersEnabled=TRUE) and the outliers' type (usrdef.outliersType) and date (usrdef.outliersDate) are provided.

usrdef.outliersEnabled
logicals. If TRUE the program uses the pre-specified outliers.

usrdef.outliersType
vector defining the outliers' type. Possible types are: ("AO") - additive, ("LS") - level shift, ("TC") - transitory change, ("SO") - seasonal outlier. E.g.: usrdef.outliersType =c("AO","AO","LS").

usrdef.outliersDate
vector defining the outliers' date. The dates should be characters in format "YYYY-MM-DD". E.g.: usrdef.outliersDate=c("2009-10-01","2005-02-01","2003-04-01").

usrdef.outliersCoef
vector providing fixed coefficients for the outliers. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined. E.g.: usrdef.outliersCoef=c(200,170,20).

Control variables for the user-defined variables:

usrdef.varEnabled
logicals. If TRUE the program uses the user-defined variables.

usrdef.var
time series (ts) or matrix of time series (mts) with the user-defined variables.

usrdef.varType
vector of character(s) defining the user-defined variables component type. Possible types are: "Undefined", "Series", "Trend", "Seasonal", "SeasonallyAdjusted", "Irregular", "Calendar". The type "Calendar" has to be used with tradingdays.option = "UserDefined" to use user-defined calendar regressors. If not specified, the program will assign the "Undefined" type.

usrdef.varCoef
vector providing fixed coefficients for the user-defined variables. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined.

tradingdays.mauto
defines whether the calendar effects should be added to the model manually ("Unused") or automatically. In the automatic selection, the choice of the number of calendar variables can be based on the F-Test ("FTest") or the Wald Test ("WaldTest"); the model with higher F value is chosen, provided that it is higher than tradingdays.pftd).

tradingdays.pftd
numeric. P-value applied in the test specified by the automatic parameter (tradingdays.mauto) to assess the significance of the pre-tested calendar effects variables and whether they should be included in the RegArima model.

Control variables for the manual selection of calendar effects variables (tradingdays.mauto is set to "Unused"):

tradingdays.option
defines the type of the trading days regression variables: "TradingDays" - six day-of-the-week regression variables; "WorkingDays" - one working/non-working day contrast variable; "None" - no correction for trading days and working days effects; "UserDefined" - user-defined trading days regressors (regressors have to be defined by the usrdef.var argument with usrdef.varType set to "Calendar" and usrdef.varEnabled = TRUE). "None" has also to be chosen for the "day-of-week effects" correction (tradingdays.stocktd to be modified accordingly).

tradingdays.leapyear
logicals. Specifies if the leap-year correction should be included. If TRUE the model includes the leap-year effect.

tradingdays.stocktd
numeric indicating the day of the month when inventories and other stock are reported (to denote the last day of the month set the variable to 31). Modifications of this variable are taken into account only when tradingdays.option is set to "None".

tradingdays.test
defines the pre-tests of the trading day effects: "None" - calendar variables are used in the model without pre-testing; "Separate_T" - a t-test is applied to each trading day variable separately and the trading day variables are included in the RegArima model if at least one t-statistic is greater than 2.6 or if two t-statistics are greater than 2.0 (in absolute terms); "Joint_F" - a joint F-test of significance of all the trading day variables. The trading day effect is significant if the F statistic is greater than 0.95.

easter.type
specifies the presence and the length of the Easter effect: "Unused" - Easter effect is not considered; "Standard" - influences the period of n days strictly before Easter Sunday; "IncludeEaster" - influences the entire period (n) up to and including Easter Sunday; "IncludeEasterMonday" - influences the entire period (n) up to and including Easter Monday.

easter.julian
logicals. If TRUE the program uses the Julian Easter (expressed in Gregorian calendar).

easter.duration
numeric indicating the duration of the Easter effect (length in days, between 1 and 15).

easter.test
logicals. If TRUE the program performs a t-test for the significance of the Easter effect. The Easter effect is considered as significant if the modulus of t-statistic is greater than 1.96.

outlier.enabled
logicals. If TRUE the automatic detection of outliers is enabled in the defined time span.

The time span of the series to be searched for outliers (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: outlier.from, outlier.to, outlier.first, outlier.last, outlier.exclFirst and outlier.exclLast; where outlier.from and outlier.to have priority over remaining span control variables, outlier.last and outlier.first have priority over outlier.exclFirst and outlier.exclLast, and outlier.last has priority over outlier.first.

outlier.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with outlier.to.

outlier.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with outlier.from.

outlier.first
numeric specifying the number of periods considered at the beginning of the series.

outlier.last
numeric specifying the number of periods considered at the end of the series.

outlier.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with outlier.exclLast.

outlier.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with outlier.exclFirst.

outlier.ao
logicals. If TRUE the automatic detection of additive outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.tc
logicals. If TRUE the automatic detection of transitory changes is enabled (outlier.enabled must be also set to TRUE).

outlier.ls
logicals. If TRUE the automatic detection of level shifts is enabled (outlier.enabled must be also set to TRUE).

outlier.so
logicals. If TRUE the automatic detection of seasonal outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.usedefcv
logicals. If TRUE the critical value for the outliers' detection procedure is automatically determined by the number of observations in the outlier detection time span. If FALSE the procedure uses the inputted critical value (outlier.cv).

outlier.cv
numeric. Inputted critical value for the outliers' detection procedure. The modification of this variable is taken in to account only when outlier.usedefcv is set to FALSE.

outlier.eml
logicals, exact likelihood estimation method.Controls the method applied for a parameter estimation in the intermediate steps of the automatic detection and correction of outliers. If TRUE an exact likelihood estimation method is used, when FALSE the fast Hannan-Rissanen method is used.

outlier.tcrate
numeric. The rate of decay for the transitory change outlier.

automdl.enabled
logicals. If TRUE the automatic modelling of the ARIMA model is enabled. If FALSE the parameters of the ARIMA model can be specified.

Control variables for the automatic modelling of the ARIMA model (automdl.enabled is set to TRUE):

automdl.acceptdefault
logicals. If TRUE the default model (ARIMA(0,1,1)(0,1,1)) may be chosen in the first step of the automatic model identification. If the Ljung-Box Q statistics for the residuals is acceptable, the default model is accepted and no further attempt will be made to identify any other.

automdl.cancel
numeric, cancelation limit. If the difference in moduli of an AR and an MA roots (when estimating ARIMA(1,0,1)(1,0,1) models in the second step of the automatic identification of the differencing orders) is smaller than cancelation limit, the two roots are assumed equal and cancel out.

automdl.ub1
numeric, first unit root limit. It is the threshold value for the initial unit root test in the automatic differencing procedure. When one of the roots in the estimation of the ARIMA(2,0,0)(1,0,0) plus mean model, performed in the first step of the automatic model identification procedure, is larger than first unit root limit in modulus, it is set equal to unity.

automdl.ub2
numeric, second unit root limit. When one of the roots in the estimation of the ARIMA(1,0,1)(1,0,1) plus mean model, which is performed in the second step of the automatic model identification procedure, is larger than second unit root limit in modulus, it is checked if there is a common factor in the corresponding AR and MA polynomials of the ARMA model that can be cancelled (see automdl.cancel). If there is no cancellation, the AR root is set equal to unity (i.e. the differencing order changes).

automdl.armalimit
numeric, arma limit. It is the threshold value for t-statistics of ARMA coefficients and constant term used for the final test of model parsimony. If the highest order ARMA coefficient has a t-value less than this value in magnitude, the order of the model is reduced. Also if the constant term has a t-value less than arma limit in magnitude, it is removed from the set of regressors.

automdl.reducecv
numeric, ReduceCV. The percentage by which the outlier's critical value will be reduced when an identified model is found to have a Ljung-Box statistic with an unacceptable confidence coefficient. The parameter should be between 0 and 1, and will only be active when automatic outlier identification is enabled. The reduced critical value will be set to (1-ReduceCV)xCV, where CV is the original critical value.

automdl.ljungboxlimit
numeric, Ljung Box limit. Acceptance criterion for the confidence intervals of the Ljung-Box Q statistic. If the LjungBox Q statistics for the residuals of a final model is greater than Ljung Box limit, the model is rejected, the outlier critical value is reduced, and model and outlier identification (if specified) is redone with a reduced value.

automdl.compare
logicals. If TRUE the program compares the model identified by the automatic procedure to the default model (ARIMA(0,1,1)(0,1,1)) and the model with the best fit is selected. Criteria considered are residual diagnostics, the model structure and the number of outliers.

Control variables for the non-automatic modelling of the ARIMA model (automdl.enabled is set to FALSE):

arima.mu
logicals. If TRUE, the mean is considered as part of the ARIMA model.

arima.p
numeric. The order of the non-seasonal autoregressive (AR) polynomial.

arima.d
numeric. Regular differencing order.

arima.q
numeric. The order of the non-seasonal moving average (MA) polynomial.

arima.bp
numeric. The order of the seasonal autoregressive (AR) polynomial.

arima.bd
numeric. Seasonal differencing order.

arima.bq
numeric. The order of the seasonal moving average (MA) polynomial.

Control variables for the user-defined ARMA coefficients. Coefficients can be defined for the regular and seasonal autoregressive (AR) polynomials and moving average (MA) polynomials. The model considers the coefficients only if the procedure for their estimation (arima.coefType) is provided, and the number of provided coefficients matches the sum of (regular and seasonal) AR and MA orders (p,q,bp,bq).

arima.coefEnabled
logicals. If TRUE the program uses the user-defined ARMA coefficients.

arima.coef
vector providing the coefficients for the regular and seasonal AR and MA polynominals. The length of the vector must equal the sum of the regular and seasonal AR and MA orders. The coefficients shall be provided in the order: regular AR (Phi - p elements), regular MA (Theta - q elements), seasonal AR (BPhi - bp elements) and seasonal MA (BTheta - bq elements). E.g.: arima.coef=c(0.6,0.7) with arima.p=1, arima.q=0,arima.bp=1 and arima.bq=0.

arima.coefType
vector defining ARMA coefficients estimation procedure. Possible procedures are: "Undefined" - no use of user-defined input (i.e. coefficients are estimated), "Fixed" - fixes the coefficients at the value provided by the user, "Initial" - the value defined by the user is used as initial condition. For orders for which the coefficients shall not be defined, the arima.coef can be set to NA or 0 or the arima.coefType can be set to "Undefined". E.g.: arima.coef = c(-0.8,-0.6,NA), arima.coefType = c("Fixed","Fixed","Undefined").

fcst.horizon
numeric, forecasts horizon. Length of the forecasts generated by the RegARIMA model in periods (positive values) or years (negative values). By default the program generates two years forecasts (fcst.horizon set to -2).

seats.predictionLength
integer, number of forecasts used in the decomposition. Negative values correspond to numbers of years.

seats.approx
character, approximation mode. When the ARIMA model estimated by TRAMO does not accept an admissible decomposition, SEATS: "None" - performs an approximation; "Legacy" - replaces the model with a decomposable one; "Noisy" - estimates a new model by adding a white noise to the non-admissible model estimated by TRAMO.

seats.trendBoundary
numeric, trend boundary. The boundary from which an AR root is integrated in the trend component. If the modulus of the inverse real root is greater than Trend boundary, the AR root is integrated in the trend component. Below this value the root is integrated in the transitory component.

seats.seasdBoundary
numeric, seasonal boundary. Boundary from which a negative AR root is integrated in the seasonal component.

seats.seasdBoundary1
numeric, seasonal boundary (unique). Boundary from which a negative AR root is integrated in the seasonal component when the root is the unique seasonal root.

seats.seasTol
numeric, seasonal tolerance. The tolerance (measured in degrees) to allocate the AR non-real roots to the seasonal component (if the modulus of the inverse complex AR root is greater than Trend boundary and the frequency of this root differs from one of the seasonal frequencies by less than Seasonal tolerance) or the transitory component (otherwise).

seats.maBoundary
numeric, MA unit root boundary. When the modulus of an estimated MA root falls in the range (xl, 1), it is set to xl.

seats.method
character, estimation method of the unobserved components. The choice can be made from: "Burman" (default, may result in a significant underestimation of the standard deviations of the components as it may become numerically unstable when some roots of the MA polynomial are near 1); "KalmanSmoother" (it is not disturbed by the (quasi-) unit roots in MA); "McElroyMatrix" (has the same stability issues as the Burman's algorithm).

Value

A two-elements list of class c("SA_spec", "TRAMO_SEATS"): (1) object of class c("regarima_spec", "TRAMO_SEATS") with the RegARIMA model specification, (2) object of class c("seats_spec", "data.frame") with the SEATS algorithm specification. Each component refers to different part of the SA model specification, mirroring the arguments of the function (for details see arguments description). Each of the lowest-level component (except span, pre-specified outliers, user-defined variables and pre-specified ARMA coefficients) is structured within a data frame with columns denoting different variables of the model specification and rows referring to: first row - base specification, as provided within the argument spec; second row - user modifications as specified by the remaining arguments of the function (e.g.: arima.d); and third row - final model specification. The final specification (third row) shall include user modifications (row two) unless they were wrongly specified. The pre-specified outliers, user-defined variables and pre-specified ARMA coefficients consist of a list with the Predefined (base model specification) and Final values.

regarima

object of class c("regarima_spec", "TRAMO_SEATS"). See Value of the function regarima_spec_tramoseats

seats

data.frame of class c("seats_spec", "data.frame"), containing the seats variables in line with the names of the arguments variables. The final values can be also accessed with the function s_seats.

Details

The available predefined 'JDemetra+' model specifications are described in the table below.

Identifier |	Log/level detection |	Outliers detection |	Calendar effects |	ARIMA	RSA0 |	NA |	NA |

NA |	Airline(+mean)	RSA1 |	automatic |	AO/LS/TC |	NA |	Airline(+mean)	RSA2 |

automatic |	AO/LS/TC |	2 td vars + Easter |	Airline(+mean)	RSA3 |	automatic |	AO/LS/TC |	NA |

automatic	RSA4 |	automatic |	AO/LS/TC |	2 td vars + Easter |	automatic	RSA5 |	automatic |

AO/LS/TC |	7 td vars + Easter |	automatic	RSAfull |	automatic |	AO/LS/TC |	automatic |	automatic

References

Info on 'JDemetra+', usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en BOX G.E.P. and JENKINS G.M. (1970), "Time Series Analysis: Forecasting and Control", Holden-Day, San Francisco.

BOX G.E.P., JENKINS G.M., REINSEL G.C. and LJUNG G.M. (2015), "Time Series Analysis: Forecasting and Control", John Wiley & Sons, Hoboken, N. J., 5th edition.

See also

Examples

# \donttest{
myseries <- ipi_c_eu[, "FR"]
myspec1 <- tramoseats_spec(spec = c("RSAfull"))
mysa1 <- tramoseats(myseries, spec = myspec1)

# Modify a pre-specified model specification
myspec2 <- tramoseats_spec(spec = "RSAfull", tradingdays.mauto = "Unused",
                           tradingdays.option = "WorkingDays",
                           easter.type = "Standard",
                           automdl.enabled = FALSE, arima.mu = TRUE)
mysa2 <- tramoseats(myseries, spec = myspec2)

# Modify the model specification from a "SA" object
myspec3 <- tramoseats_spec(mysa1, tradingdays.mauto = "Unused",
                           tradingdays.option = "WorkingDays",
                           easter.type = "Standard", automdl.enabled = FALSE, arima.mu = TRUE)
mysa3 <- tramoseats(myseries, myspec3)

# Modify the model specification from a "SA_spec" object
myspec4 <- tramoseats_spec(myspec1, tradingdays.mauto = "Unused",
                           tradingdays.option = "WorkingDays",
                           easter.type = "Standard", automdl.enabled = FALSE, arima.mu = TRUE)
mysa4 <- tramoseats(myseries, myspec4)

# Pre-specified outliers
myspec5 <- tramoseats_spec(spec = "RSAfull",
                           usrdef.outliersEnabled = TRUE,
                           usrdef.outliersType = c("LS", "LS"),
                           usrdef.outliersDate = c("2008-10-01", "2003-01-01"),
                           usrdef.outliersCoef = c(10,-8), transform.function = "None")
s_preOut(myspec5)

#>   type       date coeff
#> 1   LS 2008-10-01    10
#> 2   LS 2003-01-01    -8

mysa5 <- tramoseats(myseries, myspec5)
mysa5

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.1743      0.135
#> Phi(2)      0.2190      0.080
#> Theta(1)   -0.3606      0.136
#> BTheta(1)  -0.6922      0.042
#> 
#>              Estimate Std. Error
#> Week days      0.6908      0.035
#> Leap year      2.1822      0.777
#> Easter [6]    -2.5110      0.487
#> AO (5-2011)   13.0416      1.924
#> LS (11-2008) -16.9663      1.813
#> AO (12-2002)  -7.4300      1.923
#> AO (5-2000)    6.2458      1.930
#> 
#> Fixed outliers: 
#>              Coefficients
#> LS (10-2008)           10
#> LS (1-2003)            -8
#> 
#> 
#> Residual standard error: 2.355 on 335 degrees of freedom
#> Log likelihood = -793.8, aic =  1612 aicc =  1612, bic(corrected for length) = 1.899
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 + 0.174285 B + 0.218991 B^2 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.360552 B - 0.692185 B^12 + 0.249569 B^13 
#> 
#> 
#> SA
#> AR :  1 + 0.174285 B + 0.218991 B^2 
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.333358 B + 0.355207 B^2 - 0.005654 B^3 + 0.003008 B^4 
#> Innovation variance:  0.7296772 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.030169 B - 0.969831 B^2 
#> Innovation variance:  0.03807013 
#> 
#> Seasonal
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 1.152296 B + 0.790031 B^2 + 0.777066 B^3 + 0.664651 B^4 + 0.456747 B^5 + 0.277007 B^6 + 0.084998 B^7 - 0.031090 B^8 - 0.223274 B^9 - 0.205265 B^10 - 0.309062 B^11 
#> Innovation variance:  0.03242449 
#> 
#> Transitory
#> AR :  1 + 0.174285 B + 0.218991 B^2 
#> MA :  1 - 0.405638 B - 0.594362 B^2 
#> Innovation variance:  0.07749037 
#> 
#> Irregular
#> Innovation variance:  0.2572587 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t          s           i
#> Jan 2019 103.9 104.9598 104.7857  -1.059847  0.17416262
#> Feb 2019 101.9 105.6229 104.9286  -3.722888  0.69428512
#> Mar 2019 111.0 104.0205 105.0002   6.979527 -0.97968117
#> Apr 2019 107.4 105.7484 105.0635   1.651620  0.68487980
#> May 2019 105.5 107.0266 104.8880  -1.526553  2.13854133
#> Jun 2019 105.8 101.5481 104.5265   4.251930 -2.97842095
#> Jul 2019 110.1 106.1764 104.3032   3.923564  1.87320128
#> Aug 2019  78.7 103.0492 104.0832 -24.349202 -1.03403013
#> Sep 2019 108.5 103.8249 103.8702   4.675067 -0.04522098
#> Oct 2019 116.8 105.4772 103.5816  11.322804  1.89556484
#> Nov 2019 103.8 102.0162 103.0914   1.783805 -1.07518992
#> Dec 2019  97.7 101.5509 102.7398  -3.850883 -1.18888075
#> 
#> Forecasts:
#>                y_f     sa_f      t_f        s_f           i_f
#> Jan 2020 101.92042 102.9773 102.6542  -1.056921  0.3231389686
#> Feb 2020  99.58001 102.8345 102.6544  -3.254505  0.1801537511
#> Mar 2020 111.91343 102.5524 102.6546   9.361008 -0.1021627628
#> Apr 2020 104.29214 102.6331 102.6548   1.658999 -0.0216466343
#> May 2020  96.39031 102.6811 102.6550  -6.290825  0.0261454438
#> Jun 2020 111.67696 102.6554 102.6552   9.021587  0.0001836581
#> Jul 2020 106.63810 102.6496 102.6554   3.988470 -0.0057576330
#> Aug 2020  75.87278 102.6566 102.6556 -26.783772  0.0009632512
#> Sep 2020 109.77010 102.6569 102.6558   7.113213  0.0010929910
#> Oct 2020 111.61062 102.6556 102.6560   8.955025 -0.0004014358
#> Nov 2020 104.41843 102.6560 102.6562   1.762409 -0.0001693911
#> Dec 2020 101.20000 102.6565 102.6564  -1.456521  0.0001174334
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.770
#>  Seasonal     60.583
#>  Irregular     1.016
#>  TD & Hol.     2.653
#>  Others       31.106
#>  Total        98.127
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.205
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.966
#>  f-test on sa (td)                      0.112
#>  f-test on i (td)                       0.279
#> 
#> 
#> Additional output variables

s_preOut(mysa5)

#>   type       date coeff
#> 1   LS 2008-10-01    10
#> 2   LS 2003-01-01    -8

# User-defined calendar regressors
var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
var<- ts.union(var1, var2)

myspec6 <- tramoseats_spec(spec = "RSAfull", tradingdays.option = "UserDefined",
                           usrdef.varEnabled = TRUE, usrdef.var = var,
                           usrdef.varType = c("Calendar", "Calendar"))

#> Warning: With tradingdays.option = "UserDefined", the parameters tradingdays.leapyear and tradingdays.stocktd are ignored.

s_preVar(myspec6)

#> $series
#>                   var1          var2
#> Jan 1990  -7.305080266  -45.50780017
#> Feb 1990  -3.958412463 -100.26900477
#> Mar 1990 -14.158085153   42.80695718
#> Apr 1990  -1.089389180   20.25623016
#> May 1990   0.742999101   18.15832754
#> Jun 1990   4.501901453   56.56938676
#> Jul 1990   3.873458182  -27.76569799
#> Aug 1990  -2.582950137 -136.42484813
#> Sep 1990  15.234906027  129.92072495
#> Oct 1990   5.513434382 -199.28465574
#> Nov 1990   3.299017734  -10.37160872
#> Dec 1990   1.729226955   29.97662573
#> Jan 1991   4.266415362  -40.30004095
#> Feb 1991  -8.136995848   33.84909572
#> Mar 1991   6.513550941 -147.84270394
#> Apr 1991   9.199817959   -4.65149259
#> May 1991   3.901017877    3.16881860
#> Jun 1991   9.567240981   28.15326072
#> Jul 1991  -1.699735633   80.31952537
#> Aug 1991  -3.799779162  -20.88583488
#> Sep 1991   1.442079570  -33.67281762
#> Oct 1991  -7.182840890  -45.63097517
#> Nov 1991   2.061086480  102.42603097
#> Dec 1991  -3.673343612  145.73951507
#> Jan 1992   7.884428769  -22.49305874
#> Feb 1992  -3.109527953   29.00685098
#> Mar 1992  -8.591400992   25.92765437
#> Apr 1992  10.852796628    0.98922998
#> May 1992  -7.847793241  -18.46739254
#> Jun 1992 -12.209602337  165.42465847
#> Jul 1992 -10.566914319   43.68543558
#> Aug 1992  -9.968905540   84.73161118
#> Sep 1992  -6.726050850   99.62823002
#> Oct 1992   6.652683671  106.74524423
#> Nov 1992  22.360776830   -5.33425785
#> Dec 1992  -6.017981691   99.29262251
#> Jan 1993  -2.916177420  -49.93526848
#> Feb 1993  -8.409748620 -111.63933292
#> Mar 1993  20.316995153  -97.45167645
#> Apr 1993   4.293721252  278.74035400
#> May 1993   8.110175212  -77.33134688
#> Jun 1993  -1.742999487  348.05397484
#> Jul 1993  -3.163714832  -61.74219736
#> Aug 1993  -4.930018949   52.07824425
#> Sep 1993   1.907784743   59.25402530
#> Oct 1993 -11.037449839  -90.03468488
#> Nov 1993  -1.064133137  147.42813210
#> Dec 1993  14.589927128  108.51326027
#> Jan 1994  -3.074742982 -170.66874385
#> Feb 1994   7.571814113 -116.86962998
#> Mar 1994 -17.108253899 -141.94533339
#> Apr 1994   5.953096078   64.59709494
#> May 1994 -20.679666505  -42.71327890
#> Jun 1994  -0.857397560  118.98281468
#> Jul 1994   0.322865498  -13.98591424
#> Aug 1994   2.896863973   18.93959765
#> Sep 1994   8.089380688    5.75844052
#> Oct 1994   3.262624714  -58.72461267
#> Nov 1994 -17.038574793  -24.14145797
#> Dec 1994   5.255045893   76.38003907
#> Jan 1995  -8.030338499  -67.80967332
#> Feb 1995  -2.376557208   83.69941394
#> Mar 1995   8.553490586    4.19394664
#> Apr 1995  -4.694870449   97.93095986
#> May 1995   1.180127840  149.60956543
#> Jun 1995  12.474415549  138.11671565
#> Jul 1995  -6.883972985   17.71868508
#> Aug 1995 -13.835695751  -60.45588857
#> Sep 1995  -0.429324153   -0.64125536
#> Oct 1995  -2.092342298 -103.09176045
#> Nov 1995 -14.669934544  -59.66193013
#> Dec 1995  22.357783800  123.00053344
#> Jan 1996  -6.251449635 -212.52639407
#> Feb 1996 -11.679137854   54.65412848
#> Mar 1996   1.495288314  101.14553407
#> Apr 1996   8.613199004  136.71746024
#> May 1996 -10.462034070   70.78379701
#> Jun 1996   1.722600112  -12.25158540
#> Jul 1996  -3.070207510  -79.94091347
#> Aug 1996   9.033456097  -17.93600041
#> Sep 1996   9.852486936   10.94528569
#> Oct 1996  16.438284835   23.09265529
#> Nov 1996   8.145597243 -126.96631620
#> Dec 1996   8.251131605   62.05258113
#> Jan 1997  -3.457739223  -64.23485204
#> Feb 1997   4.111473348   66.23578138
#> Mar 1997  -6.432143830   24.58842772
#> Apr 1997   9.420594767  -84.78936872
#> May 1997   4.695917838  -43.47674107
#> Jun 1997  -6.621122380   88.31759490
#> Jul 1997  -3.094711022   58.63186803
#> Aug 1997   7.945742636  -25.66920407
#> Sep 1997  -9.148509563  122.01357048
#> Oct 1997  -4.655203093  232.58879673
#> Nov 1997   8.196693629 -130.70423363
#> Dec 1997  15.210357381    0.02714681
#> Jan 1998   2.653547903  114.25115700
#> Feb 1998   6.190669397  -22.19401352
#> Mar 1998   4.710808367  -20.83145925
#> Apr 1998   4.631442134   74.99227118
#> May 1998  -1.601613628   11.76056177
#> Jun 1998  -0.356854719    9.90212037
#> Jul 1998  -5.746848158  136.16884687
#> Aug 1998   8.453500168  -83.89882385
#> Sep 1998  -5.817661225 -106.49372190
#> Oct 1998  -5.023571057  211.31680157
#> Nov 1998  10.880762646   51.06671868
#> Dec 1998  -3.929530194  120.02630345
#> Jan 1999  -6.604611191  104.54600553
#> Feb 1999  -1.490251268   72.04436071
#> Mar 1999  -8.754852976   50.29684073
#> Apr 1999  16.990015707   95.58412135
#> May 1999   0.681902616   49.21280559
#> Jun 1999   5.331875802  -15.59494684
#> Jul 1999  -0.809489931  -43.30621776
#> Aug 1999  24.467814279 -122.34203505
#> Sep 1999   3.838990336  -29.03433729
#> Oct 1999 -10.521715779   34.02319889
#> Nov 1999 -17.222338269   23.45520557
#> Dec 1999  -6.308192495    1.18239369
#> Jan 2000  -9.535959793   60.47329872
#> Feb 2000  -0.799824652   73.75097630
#> Mar 2000  10.626959023  -89.81034882
#> Apr 2000  13.510568359   17.18213563
#> May 2000   2.332902205   91.32347891
#> Jun 2000   3.338987914 -110.51534547
#> Jul 2000  -7.647702581   39.34501445
#> Aug 2000   7.693844601 -136.98426711
#> Sep 2000  -2.150129615   73.94716153
#> Oct 2000   3.833805853   63.83709370
#> Nov 2000  -5.424953612 -228.03538615
#> Dec 2000 -20.945782198 -201.12246984
#> Jan 2001   3.780500587  206.87405597
#> Feb 2001  -7.059342281 -131.50525575
#> Mar 2001  -5.871045298 -110.97503996
#> Apr 2001   1.000863593    5.48967197
#> May 2001  12.020663588  -86.25873906
#> Jun 2001   6.598439960   93.07754912
#> Jul 2001  -6.267453556   57.11704007
#> Aug 2001  -5.297761060  123.48319265
#> Sep 2001   8.765031871   46.20031264
#> Oct 2001  -0.501382118  -73.42081944
#> Nov 2001  22.803903528  132.33606241
#> Dec 2001  13.798105639   45.72947704
#> Jan 2002  -5.736360343  -73.64807522
#> Feb 2002 -14.168374281  214.36128315
#> Mar 2002 -12.449074589  -33.81593680
#> Apr 2002   9.000824800   -7.09703999
#> May 2002  19.894302565   -8.82100204
#> Jun 2002 -27.869752864   35.50907169
#> Jul 2002  -0.231353042 -120.48084555
#> Aug 2002   2.041608395  145.60005038
#> Sep 2002   0.442285247   -0.04656092
#> Oct 2002   6.270974878  -11.76236571
#> Nov 2002  -9.472495541   18.41482138
#> Dec 2002 -11.228269438  -86.12007223
#> Jan 2003  -0.203236495   77.93108467
#> Feb 2003  12.250255435  -45.40641716
#> Mar 2003   0.830589847 -202.86071844
#> Apr 2003 -18.859151704   13.74981227
#> May 2003  10.569271276  -65.31652495
#> Jun 2003   6.129121886   72.00350979
#> Jul 2003   5.896500909   80.84364277
#> Aug 2003  -0.664513411   96.24521304
#> Sep 2003  -6.081078807   10.31380723
#> Oct 2003   9.473045743   68.63601515
#> Nov 2003   3.931699575  109.51989336
#> Dec 2003  -3.852059312   18.18922153
#> Jan 2004  -6.092631219  -19.68754162
#> Feb 2004  -3.843851460  -84.80613844
#> Mar 2004  -9.014236692 -187.74430764
#> Apr 2004   8.729458934 -258.61413063
#> May 2004   5.618212586   10.80672274
#> Jun 2004 -17.572785220 -183.83419463
#> Jul 2004  -8.121160950  -12.30852580
#> Aug 2004   2.386905873   14.74878788
#> Sep 2004  25.060017195 -143.06464098
#> Oct 2004  -9.805871650  -10.59880850
#> Nov 2004  15.661250555   50.31819093
#> Dec 2004  -7.588674355  251.26804489
#> Jan 2005   4.938451044  -44.13813498
#> Feb 2005  -3.271504519 -217.89991293
#> Mar 2005  -4.348790398  -67.81489679
#> Apr 2005  15.197231060   72.88472658
#> May 2005   0.036778033 -124.34327365
#> Jun 2005  -4.023399502  -58.82719769
#> Jul 2005 -14.954664332  110.85391313
#> Aug 2005  -9.459855132  -50.23799876
#> Sep 2005   9.064295082   34.72986863
#> Oct 2005 -18.138676228  174.81467948
#> Nov 2005  -9.517754317   46.28553413
#> Dec 2005  11.268520993   76.78697074
#> Jan 2006   1.398890649   45.10162329
#> Feb 2006   1.003399767  -35.25772249
#> Mar 2006 -12.045286294   36.64544857
#> Apr 2006  -0.834582964  -48.81248718
#> May 2006  -7.970822469   87.33900144
#> Jun 2006  11.015245786  179.87806220
#> Jul 2006   1.084104063  105.07952142
#> Aug 2006 -11.905067990  217.52306395
#> Sep 2006   2.799212761  -18.35628629
#> Oct 2006   7.774077731  -68.67930022
#> Nov 2006  -2.210897240    2.52921842
#> Dec 2006  -1.376768354   77.70312224
#> Jan 2007   1.152300433  135.80460668
#> Feb 2007  15.332695729   31.19454244
#> Mar 2007  -6.557744602   99.11976849
#> Apr 2007   8.583787007  -38.18986462
#> May 2007 -11.876705517  -88.02615153
#> Jun 2007   5.395403804    2.30683213
#> Jul 2007  -2.663621627   70.83967956
#> Aug 2007   7.581722346  147.10711674
#> Sep 2007 -10.314020292  142.31851069
#> Oct 2007  10.249315104  -73.17572527
#> Nov 2007  -1.452615921 -122.78623180
#> Dec 2007   1.092238302  -84.65572264
#> Jan 2008 -14.785987947  164.14782427
#> Feb 2008 -22.791912223 -253.69827268
#> Mar 2008   6.000768904  151.49668186
#> Apr 2008  -3.465781995  134.44652874
#> May 2008  -2.851536719   34.22179953
#> Jun 2008  -6.476521580  -32.46704812
#> Jul 2008 -12.722147004  -95.09212602
#> Aug 2008  -1.751731809   12.25963233
#> Sep 2008 -22.013195775  220.95975457
#> Oct 2008   5.798759794  130.49280363
#> Nov 2008   0.627964711  -52.55140662
#> Dec 2008  17.751245181  115.16359225
#> Jan 2009  13.836946999   82.61668965
#> Feb 2009  16.309602864  115.62448319
#> Mar 2009  -8.083912981   41.09726943
#> Apr 2009   0.955654515   -5.60486435
#> May 2009   9.168081903   78.01771348
#> Jun 2009   4.738435430 -192.31777839
#> Jul 2009  -3.952457160  -27.36494881
#> Aug 2009   4.293985438   60.37177315
#> Sep 2009   8.509307909 -131.73483134
#> Oct 2009  -1.703946179   95.55886301
#> Nov 2009  -0.007914038    1.28535381
#> Dec 2009   8.606414955   29.14360940
#> Jan 2010   6.403423802 -201.68055585
#> Feb 2010   1.068762896 -159.64884036
#> Mar 2010 -14.604777644   79.05133025
#> Apr 2010   4.551713753   -5.59537712
#> May 2010  12.720846841  278.46565380
#> Jun 2010 -19.151875806   39.54117043
#> Jul 2010   2.355890946  269.79624187
#> Aug 2010  10.559143097  105.66101329
#> Sep 2010   4.316772438  178.18693119
#> Oct 2010 -17.459432119  131.86880388
#> Nov 2010  26.345754763  -72.40831537
#> Dec 2010  -0.308035714  -26.32911657
#> Jan 2011 -22.561887639  -34.85907549
#> Feb 2011  18.368225390  -21.32519004
#> Mar 2011   4.985355661   74.86664662
#> Apr 2011   1.562468467 -300.45498345
#> May 2011   5.565016905  -20.57101950
#> Jun 2011   6.871069775  -14.35167183
#> Jul 2011 -33.041822297  131.09571969
#> Aug 2011   8.783299631   76.45732607
#> Sep 2011  14.611756864   56.44685721
#> Oct 2011  -1.094066808  -17.53213851
#> Nov 2011  13.256063791   -0.25746856
#> Dec 2011  11.406805737   58.87499380
#> Jan 2012  -9.303642146   17.53171579
#> Feb 2012  -1.946233382  139.69229710
#> Mar 2012   9.240072693   82.54166479
#> Apr 2012  -2.837713588   55.47305107
#> May 2012   6.438188432   54.53779827
#> Jun 2012  12.049299753  -53.63845039
#> Jul 2012  -9.864334772 -183.92675089
#> Aug 2012  13.629994009  -71.23421939
#> Sep 2012  -6.371366390   -9.45664557
#> Oct 2012 -12.527051445   24.30159689
#> Nov 2012  -5.751377304  102.04461058
#> Dec 2012 -16.763008174   92.32619914
#> Jan 2013   2.628120182   26.31799404
#> Feb 2013 -14.208832929  -68.75861883
#> Mar 2013   7.738467161   22.08611800
#> Apr 2013  -6.301913980  -55.41037774
#> May 2013 -25.232229681   55.18463880
#> Jun 2013  10.456761237  -16.18329852
#> Jul 2013   7.513402330  135.14035346
#> Aug 2013   7.981608534   -0.53331422
#> Sep 2013 -15.676418337 -240.51334407
#> Oct 2013 -11.474768636   98.98093589
#> Nov 2013   4.355867220 -123.30316799
#> Dec 2013  10.532004191    2.74105624
#> Jan 2014   9.635066048  -39.10052334
#> Feb 2014   6.939309379  240.18101774
#> Mar 2014  -2.300709516   55.41959989
#> Apr 2014  -9.355586095   56.07759451
#> May 2014  -9.026203778 -143.07888081
#> Jun 2014  -0.308600940   25.65309367
#> Jul 2014  -8.686782246   44.75129842
#> Aug 2014  21.877875041   90.91545502
#> Sep 2014  15.385301545  -65.76117921
#> Oct 2014 -18.820726709   25.47812926
#> Nov 2014   9.158750045 -136.29594240
#> Dec 2014  -4.017400140 -115.16580441
#> Jan 2015 -12.568740002    9.23561855
#> Feb 2015  11.843968692  -52.61090352
#> Mar 2015  10.085724636  -94.12824980
#> Apr 2015 -10.661521156 -116.03851187
#> May 2015 -10.802139977  146.32167885
#> Jun 2015 -15.265903705 -122.70357371
#> Jul 2015 -13.704004261  128.91223057
#> Aug 2015 -12.935694331   52.19834145
#> Sep 2015 -29.610297132    3.86469118
#> Oct 2015  11.255956830    0.54809002
#> Nov 2015   0.040275376  205.09627114
#> Dec 2015   1.626711541   58.40034881
#> Jan 2016 -14.839223699  -71.96403767
#> Feb 2016  11.648603184   31.72919743
#> Mar 2016 -12.724193004   16.67156540
#> Apr 2016 -11.137786657  250.26785326
#> May 2016 -13.069370448   40.04087093
#> Jun 2016  -1.742082303  107.21119854
#> Jul 2016  -6.832515233    5.50009631
#> Aug 2016  -6.215682860   -6.13638601
#> Sep 2016  -1.551632320 -107.64435907
#> Oct 2016 -10.864935828   50.43042930
#> Nov 2016   5.502220067   53.62245498
#> Dec 2016  13.947806779 -135.49772917
#> Jan 2017 -15.883192579   22.75063071
#> Feb 2017  20.675835289 -252.94187083
#> Mar 2017 -16.971130469  -81.85311433
#> Apr 2017   8.605076312 -158.95299935
#> May 2017  -3.796434887 -108.24933233
#> Jun 2017  18.476301068   20.58043597
#> Jul 2017  10.376561791 -108.11823198
#> Aug 2017 -14.695747682  -18.87925541
#> Sep 2017 -27.769482100   31.27564538
#> Oct 2017  -0.769797878    7.47169985
#> Nov 2017  -9.544920721  -60.43448541
#> Dec 2017   7.086794196    4.52466495
#> Jan 2018  17.770868082  164.85396371
#> Feb 2018   8.663143079  -75.56829840
#> Mar 2018   1.530058172  -28.81793527
#> Apr 2018  -0.463890794  -73.01505059
#> May 2018 -16.728474148   52.41485203
#> Jun 2018 -10.898906119  -59.17005292
#> Jul 2018  -0.761757047 -118.39119861
#> Aug 2018 -13.058317158  120.03531317
#> Sep 2018  -0.623845628  -90.80713269
#> Oct 2018   3.423715685 -154.88176869
#> Nov 2018  -5.489306150   40.80449528
#> Dec 2018   9.150613267   87.23428206
#> Jan 2019  -0.876039035 -186.04776905
#> Feb 2019  -0.255794707   20.61176045
#> Mar 2019 -12.543574063 -106.55260009
#> Apr 2019  18.801248228   44.42791106
#> May 2019  -1.409493686   39.00637392
#> Jun 2019  -1.066911374  -11.94045102
#> Jul 2019   7.245550904  147.49626484
#> Aug 2019   4.288180835 -129.46582609
#> Sep 2019 -15.802157511 -146.76126825
#> Oct 2019   7.747558144  -21.13007304
#> Nov 2019 -11.642921245   41.96187254
#> Dec 2019   0.088543033  143.33403558
#> 
#> $description
#>          type coeff
#> var1 Calendar    NA
#> var2 Calendar    NA
#> 

mysa6 <- tramoseats(myseries, myspec6)

myspec7 <- tramoseats_spec(spec = "RSAfull", usrdef.varEnabled = TRUE,
                           usrdef.var = var, usrdef.varCoef = c(17,-1),
                           transform.function = "None")
mysa7 <- tramoseats(myseries, myspec7)

# Pre-specified ARMA coefficients
myspec8 <- tramoseats_spec(spec = "RSAfull",
                           arima.coefEnabled = TRUE, automdl.enabled = FALSE,
                           arima.p = 2, arima.q = 0,
                           arima.bp = 1, arima.bq = 1,
                           arima.coef = c(-0.12, -0.12, -0.3, -0.99),
                           arima.coefType = rep("Fixed", 4))
mysa8 <- tramoseats(myseries, myspec8)
mysa8

#> RegARIMA
#> y = regression model + arima (2, 1, 0, 1, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)       -0.12          0
#> Phi(2)       -0.12          0
#> BPhi(1)      -0.30          0
#> BTheta(1)    -0.99          0
#> 
#>               Estimate Std. Error
#> Mean        -0.0000713      0.000
#> Week days    0.0072617      0.000
#> Leap year    0.0211795      0.007
#> Easter [6]  -0.0203701      0.004
#> AO (5-2011)  0.1385801      0.016
#> LS (1-2009) -0.0445176      0.025
#> 
#> 
#> Residual standard error: 0.02564 on 340 degrees of freedom
#> Log likelihood = 763.7, aic =  1692 aicc =  1692, bic(corrected for length) = -7.226
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 - 0.120000 B - 0.120000 B^2 - 0.300000 B^12 + 0.036000 B^13 + 0.036000 B^14 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.950000 B^12 
#> 
#> 
#> SA
#> AR :  1 - 1.024538 B - 0.011455 B^2 + 0.108545 B^3 
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.941460 B + 0.995458 B^2 + 0.036596 B^3 - 0.098708 B^4 + 0.008921 B^5 
#> Innovation variance:  0.4957156 
#> 
#> Trend
#> AR :  1 - 0.904538 B 
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 0.722193 B - 0.998833 B^2 + 0.723360 B^3 
#> Innovation variance:  0.1026625 
#> 
#> Seasonal
#> AR :  1 + 0.904538 B + 0.818189 B^2 + 0.740083 B^3 + 0.669433 B^4 + 0.605527 B^5 + 0.547723 B^6 + 0.495436 B^7 + 0.448140 B^8 + 0.405360 B^9 + 0.366664 B^10 + 0.331661 B^11 
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 2.915179 B + 5.617938 B^2 + 8.595374 B^3 + 11.505684 B^4 + 14.135215 B^5 + 16.348945 B^6 + 18.036965 B^7 + 19.197199 B^8 + 19.965176 B^9 + 20.301074 B^10 + 20.149549 B^11 + 19.110304 B^12 + 17.206569 B^13 + 14.558551 B^14 + 11.650924 B^15 + 8.806927 B^16 + 6.228869 B^17 + 4.044510 B^18 + 2.357558 B^19 + 1.170033 B^20 + 0.354583 B^21 - 0.052292 B^22 
#> Innovation variance:  0.2584115 
#> 
#> Transitory
#> AR :  1 - 0.120000 B - 0.120000 B^2 
#> MA :  1 + 1.287572 B + 0.287572 B^2 
#> Innovation variance:  9.66767e-06 
#> 
#> Irregular
#> Innovation variance:  0.1228651 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.4635 105.0617 0.9851745 1.0038251
#> Feb 2019 101.9 105.8270 105.3195 0.9628924 1.0048182
#> Mar 2019 111.0 104.3239 105.0676 1.0639938 0.9929219
#> Apr 2019 107.4 105.4439 105.3998 1.0185506 1.0004185
#> May 2019 105.5 106.5980 105.0680 0.9896995 1.0145620
#> Jun 2019 105.8 101.9461 104.0220 1.0378038 0.9800435
#> Jul 2019 110.1 105.1131 104.1036 1.0474435 1.0096970
#> Aug 2019  78.7 104.2508 104.1201 0.7549103 1.0012553
#> Sep 2019 108.5 103.0526 103.5774 1.0528604 0.9949335
#> Oct 2019 116.8 103.7981 103.2007 1.1252618 1.0057889
#> Nov 2019 103.8 102.3024 102.3690 1.0146390 0.9993498
#> Dec 2019  97.7 101.1355 101.3758 0.9660306 0.9976301
#> 
#> Forecasts:
#>                y_f      sa_f       t_f       s_f       i_f
#> Jan 2020  98.23489 100.77276 100.77440 0.9791656 0.9999837
#> Feb 2020  95.69761 100.36667 100.36829 0.9591408 0.9999839
#> Mar 2020 107.19256  99.99264  99.99424 1.0784959 0.9999840
#> Apr 2020 100.42241  99.64737  99.64896 1.0140381 0.9999840
#> May 2020  92.77257  99.32792  99.32951 0.9393879 0.9999840
#> Jun 2020 108.12638  99.03165  99.03324 1.0951061 0.9999840
#> Jul 2020 101.90287  98.75621  98.75778 1.0390325 0.9999840
#> Aug 2020  70.97991  98.49946  98.50103 0.7273066 0.9999840
#> Sep 2020 105.77721  98.25951  98.26108 1.0821421 0.9999840
#> Oct 2020 106.59914  98.03465  98.03621 1.0930513 0.9999840
#> Nov 2020  98.57942  97.82334  97.82490 1.0135857 0.9999840
#> Dec 2020  97.66554  97.62420  97.62576 1.0019043 0.9999840
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         3.975
#>  Seasonal     85.839
#>  Irregular     0.263
#>  TD & Hol.     3.375
#>  Others        4.021
#>  Total        97.473
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.313
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.958
#>  f-test on sa (td)                      0.013
#>  f-test on i (td)                       0.072
#> 
#> 
#> Additional output variables

s_arimaCoef(myspec8)

#>            Type Value
#> Phi(1)    Fixed -0.12
#> Phi(2)    Fixed -0.12
#> BPhi(1)   Fixed -0.30
#> BTheta(1) Fixed -0.99

s_arimaCoef(mysa8)

#>            Type Value
#> Phi(1)    Fixed -0.12
#> Phi(2)    Fixed -0.12
#> BPhi(1)   Fixed -0.30
#> BTheta(1) Fixed -0.99

# }