TRAMO-SEATS model specification, SA/TRAMO-SEATS

tramoseats_spec_def creates (and modifies), from a predefined JDemetra+ model specification, a c("SA_spec","TRAMO_SEATS") class object with the SA model specification for the TRAMO-SEATS method.

tramoseats_spec creates (and/or modifies) a c("SA_spec","TRAMO_SEATS") class object with the SA model specification for the TRAMO-SEATS method. The object is created from a c("SA","TRAMO_SEATS") or c("SA_spec","TRAMO_SEATS") class object.

tramoseats_spec_def(spec = c("RSAfull", "RSA0", "RSA1", "RSA2", "RSA3",
  "RSA4", "RSA5"), estimate.from = NA_character_,
  estimate.to = NA_character_, estimate.first = NA_integer_,
  estimate.last = NA_integer_, estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_, estimate.tol = NA_integer_,
  estimate.eml = NA, estimate.urfinal = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.fct = NA_integer_, usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA, usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA, usrdef.varEnabled = NA, usrdef.var = NA,
  usrdef.varType = NA, usrdef.varCoef = NA, tradingdays.mauto = c(NA,
  "Unused", "FTest", "WaldTest"), tradingdays.pftd = NA_integer_,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined",
  "None"), tradingdays.leapyear = NA,
  tradingdays.stocktd = NA_integer_, tradingdays.test = c(NA,
  "Separate_T", "Joint_F", "None"), easter.type = c(NA, "Unused",
  "Standard", "IncludeEaster", "IncludeEasterMonday"),
  easter.julian = NA, easter.duration = NA_integer_,
  easter.test = NA, outlier.enabled = NA,
  outlier.from = NA_character_, outlier.to = NA_character_,
  outlier.first = NA_integer_, outlier.last = NA_integer_,
  outlier.exclFirst = NA_integer_, outlier.exclLast = NA_integer_,
  outlier.ao = NA, outlier.tc = NA, outlier.ls = NA,
  outlier.so = NA, outlier.usedefcv = NA, outlier.cv = NA_integer_,
  outlier.eml = NA, outlier.tcrate = NA_integer_,
  automdl.enabled = NA, automdl.acceptdefault = NA,
  automdl.cancel = NA_integer_, automdl.ub1 = NA_integer_,
  automdl.ub2 = NA_integer_, automdl.armalimit = NA_integer_,
  automdl.reducecv = NA_integer_, automdl.ljungboxlimit = NA_integer_,
  automdl.compare = NA, arima.mu = NA, arima.p = NA_integer_,
  arima.d = NA_integer_, arima.q = NA_integer_,
  arima.bp = NA_integer_, arima.bd = NA_integer_,
  arima.bq = NA_integer_, arima.coefEnabled = NA, arima.coef = NA,
  arima.coefType = NA, fcst.horizon = NA_integer_,
  seats.approx = c(NA, "None", "Legacy", "Noisy"),
  seats.trendBoundary = NA_integer_, seats.seasdBoundary = NA_integer_,
  seats.seasdBoundary1 = NA_integer_, seats.seasTol = NA_integer_,
  seats.maBoundary = NA_integer_, seats.method = c(NA, "Burman",
  "KalmanSmoother", "McElroyMatrix"))

tramoseats_spec(object, estimate.from = NA_character_,
  estimate.to = NA_character_, estimate.first = NA_integer_,
  estimate.last = NA_integer_, estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_, estimate.tol = NA_integer_,
  estimate.eml = NA, estimate.urfinal = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.fct = NA_integer_, usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA, usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA, usrdef.varEnabled = NA, usrdef.var = NA,
  usrdef.varType = NA, usrdef.varCoef = NA, tradingdays.mauto = c(NA,
  "Unused", "FTest", "WaldTest"), tradingdays.pftd = NA_integer_,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined",
  "None"), tradingdays.leapyear = NA,
  tradingdays.stocktd = NA_integer_, tradingdays.test = c(NA,
  "Separate_T", "Joint_F", "None"), easter.type = c(NA, "Unused",
  "Standard", "IncludeEaster", "IncludeEasterMonday"),
  easter.julian = NA, easter.duration = NA_integer_,
  easter.test = NA, outlier.enabled = NA,
  outlier.from = NA_character_, outlier.to = NA_character_,
  outlier.first = NA_integer_, outlier.last = NA_integer_,
  outlier.exclFirst = NA_integer_, outlier.exclLast = NA_integer_,
  outlier.ao = NA, outlier.tc = NA, outlier.ls = NA,
  outlier.so = NA, outlier.usedefcv = NA, outlier.cv = NA_integer_,
  outlier.eml = NA, outlier.tcrate = NA_integer_,
  automdl.enabled = NA, automdl.acceptdefault = NA,
  automdl.cancel = NA_integer_, automdl.ub1 = NA_integer_,
  automdl.ub2 = NA_integer_, automdl.armalimit = NA_integer_,
  automdl.reducecv = NA_integer_, automdl.ljungboxlimit = NA_integer_,
  automdl.compare = NA, arima.mu = NA, arima.p = NA_integer_,
  arima.d = NA_integer_, arima.q = NA_integer_,
  arima.bp = NA_integer_, arima.bd = NA_integer_,
  arima.bq = NA_integer_, arima.coefEnabled = NA, arima.coef = NA,
  arima.coefType = NA, fcst.horizon = NA_integer_,
  seats.approx = c(NA, "None", "Legacy", "Noisy"),
  seats.trendBoundary = NA_integer_, seats.seasdBoundary = NA_integer_,
  seats.seasdBoundary1 = NA_integer_, seats.seasTol = NA_integer_,
  seats.maBoundary = NA_integer_, seats.method = c(NA, "Burman",
  "KalmanSmoother", "McElroyMatrix"))

Arguments

spec
predefined JDemetra+ model specification (see Details). The default is "RSAfull".

The time span of the series to be used for the estimation of the RegArima model coefficients (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: estimate.from, estimate.to, estimate.first, estimate.last, estimate.exclFirst and estimate.exclLast; where estimate.from and estimate.to have priority over remaining span control variables, estimate.last and estimate.first have priority over estimate.exclFirst and estimate.exclLast, and estimate.last has priority over estimate.first.

estimate.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with estimate.to.

estimate.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with estimate.from.

estimate.first
numeric specifying the number of periods considered at the beginning of the series.

estimate.last
numeric specifying the number of periods considered at the end of the series.

estimate.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with estimate.exclLast.

estimate.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with estimate.exclFirst.

estimate.tol
numeric, convergence tolerance. The absolute changes in the log-likelihood function are compared to this value to check for the convergence of the estimation iterations.

estimate.eml
logicals, exact maximum likelihood estimation. If TRUE the program performs an exact maximum likelihood estimation. If FASLE the Unconditional Least Squares method is used.

estimate.urfinal
numeric, final unit root limit. The threshold value for the final unit root test for identification of differencing orders. If the magnitude of an AR root for the final model is less than this number, a unit root is assumed, the order of the AR polynomial is reduced by one, and the appropriate order of the differencing (non-seasonal, seasonal) is increased.

transform.function
transformation of the input series: "None" - no transformation of the series; "Log" - takes the log of the series; "Auto" - the program tests for the log-level specification.

transform.fct
numeric controlling the bias in the log/level pre-test: transform.fct > 1 favours levels, transform.fct< 1 favours logs. Considered only when transform.function is set to "Auto".

Control variables for the pre-specified outliers. The pre-specified outliers are used in the model only if they are enabled (usrdef.outliersEnabled=TRUE) and the outliers' type (usrdef.outliersType) and date (usrdef.outliersDate) are provided.

usrdef.outliersEnabled
logicals. If TRUE the program uses the pre-specified outliers.

usrdef.outliersType
vector defining the outliers' type. Possible types are: ("AO") - additive, ("LS") - level shift, ("TC") - transitory change, ("SO") - seasonal outlier. E.g.: usrdef.outliersType =c("AO","AO","LS").

usrdef.outliersDate
vector defining the outliers' date. The dates should be characters in format "YYYY-MM-DD". E.g.: usrdef.outliersDate=c("2009-10-01","2005-02-01","2003-04-01").

usrdef.outliersCoef
vector providing fixed coefficients for the outliers. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined. E.g.: usrdef.outliersCoef=c(200,170,20).

Control variables for the user-defined variables:

usrdef.varEnabled
logicals. If TRUE the program uses the user-defined variables.

usrdef.var
time series (ts) or matrix of time series (mts) with the user-defined variables.

usrdef.varType
vector of character(s) defining the user-defined variables component type. Possible types are: "Undefined", "Series", "Trend", "Seasonal", "SeasonallyAdjusted", "Irregular". If not specified, the program will assign the "Undefined" type.

usrdef.varCoef
vector providing fixed coefficients for the user-defined variables. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined.

tradingdays.mauto
defines whether the calendar effects should be added to the model manually ("Unused") or automatically. In the automatic selection, the choice of the number of calendar variables can be based on the F-Test ("FTest") or the Wald Test ("WaldTest"); the model with higher F value is chosen, provided that it is higher than tradingdays.pftd).

tradingdays.pftd
numeric. P-value applied in the test specified by the automatic parameter (tradingdays.mauto) to assess the significance of the pre-tested calendar effects variables and whether they should be included in the RegArima model.

Control variables for the manual selection of calendar effects variables (tradingdays.mauto is set to "Unused"):

tradingdays.option
defines the type of the trading days regression variables: "TradingDays" - six day-of-the-week regression variables; "WorkingDays" - one working/non-working day contrast variable; "None" - no correction for trading days and working days effects; "UserDefined" - user-defined trading days regressors (currently not working). "None" has also to be chosen for the "day-of-week effects" correction (tradingdays.stocktd to be modified accordingly).

tradingdays.leapyear
logicals. Specifies if the leap-year correction should be included. If TRUE the model includes the leap-year effect.

tradingdays.stocktd
numeric indicating the day of the month when inventories and other stock are reported (to denote the last day of the month set the variable to 31). Modifications of this variable are taken into account only when tradingdays.option is set to "None".

tradingdays.test
defines the pre-tests of the trading day effects: "None" - calendar variables are used in the model without pre-testing; "Separate_T" - a t-test is applied to each trading day variable separately and the trading day variables are included in the RegArima model if at least one t-statistic is greater than 2.6 or if two t-statistics are greater than 2.0 (in absolute terms); "Joint_F" - a joint F-test of significance of all the trading day variables. The trading day effect is significant if the F statistic is greater than 0.95.

easter.type
specifies the presence and the length of the Easter effect: "Unused" - Easter effect is not considered; "Standard" - influences the period of n days strictly before Easter Sunday; "IncludeEaster" - influences the entire period (n) up to and including Easter Sunday; "IncludeEasterMonday" - influences the entire period (n) up to and including Easter Monday.

easter.julian
logicals. If TRUE the program uses the Julian Easter (expressed in Gregorian calendar).

easter.duration
numeric indicating the duration of the Easter effect (length in days, between 1 and 15).

easter.test
logicals. If TRUE the program performs a t-test for the significance of the Easter effect. The Easter effect is considered as significant if the modulus of t-statistic is greater than 1.96.

outlier.enabled
logicals. If TRUE the automatic detection of outliers is enabled in the defined time span.

The time span of the series to be searched for outliers (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: outlier.from, outlier.to, outlier.first, outlier.last, outlier.exclFirst and outlier.exclLast; where outlier.from and outlier.to have priority over remaining span control variables, outlier.last and outlier.first have priority over outlier.exclFirst and outlier.exclLast, and outlier.last has priority over outlier.first.

outlier.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with outlier.to.

outlier.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with outlier.from.

outlier.first
numeric specifying the number of periods considered at the beginning of the series.

outlier.last
numeric specifying the number of periods considered at the end of the series.

outlier.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with outlier.exclLast.

outlier.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with outlier.exclFirst.

outlier.ao
logicals. If TRUE the automatic detection of additive outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.tc
logicals. If TRUE the automatic detection of transitory changes is enabled (outlier.enabled must be also set to TRUE).

outlier.ls
logicals. If TRUE the automatic detection of level shifts is enabled (outlier.enabled must be also set to TRUE).

outlier.so
logicals. If TRUE the automatic detection of seasonal outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.usedefcv
logicals. If TRUE the critical value for the outliers' detection procedure is automatically determined by the number of observations in the outlier detection time span. If FALSE the procedure uses the inputted critical value (outlier.cv).

outlier.cv
numeric. Inputted critical value for the outliers' detection procedure. The modification of this variable is taken in to account only when outlier.usedefcv is set to FALSE.

outlier.eml
logicals, exact likelihood estimation method.Controls the method applied for a parameter estimation in the intermediate steps of the automatic detection and correction of outliers. If TRUE an exact likelihood estimation method is used, when FALSE the fast Hannan-Rissanen method is used.

outlier.tcrate
numeric. The rate of decay for the transitory change outlier.

automdl.enabled
logicals. If TRUE the automatic modelling of the ARIMA model is enabled. If FALSE the parameters of the ARIMA model can be specified.

Control variables for the automatic modelling of the ARIMA model (automdl.enabled is set to TRUE):

automdl.acceptdefault
logicals. If TRUE the default model (ARIMA(0,1,1)(0,1,1)) may be chosen in the first step of the automatic model identification. If the Ljung-Box Q statistics for the residuals is acceptable, the default model is accepted and no further attempt will be made to identify any other.

automdl.cancel
numeric, cancelation limit. If the difference in moduli of an AR and an MA roots (when estimating ARIMA(1,0,1)(1,0,1) models in the second step of the automatic identification of the differencing orders) is smaller than cancelation limit, the two roots are assumed equal and cancel out.

automdl.ub1
numeric, first unit root limit. It is the threshold value for the initial unit root test in the automatic differencing procedure. When one of the roots in the estimation of the ARIMA(2,0,0)(1,0,0) plus mean model, performed in the first step of the automatic model identification procedure, is larger than first unit root limit in modulus, it is set equal to unity.

automdl.ub2
numeric, second unit root limit. When one of the roots in the estimation of the ARIMA(1,0,1)(1,0,1) plus mean model, which is performed in the second step of the automatic model identification procedure, is larger than second unit root limit in modulus, it is checked if there is a common factor in the corresponding AR and MA polynomials of the ARMA model that can be cancelled (see automdl.cancel). If there is no cancellation, the AR root is set equal to unity (i.e. the differencing order changes).

automdl.armalimit
numeric, arma limit. It is the threshold value for t-statistics of ARMA coefficients and constant term used for the final test of model parsimony. If the highest order ARMA coefficient has a t-value less than this value in magnitude, the order of the model is reduced. Also if the constant term has a t-value less than arma limit in magnitude, it is removed from the set of regressors.

automdl.reducecv
numeric, ReduceCV. The percentage by which the outlier's critical value will be reduced when an identified model is found to have a Ljung-Box statistic with an unacceptable confidence coefficient. The parameter should be between 0 and 1, and will only be active when automatic outlier identification is enabled. The reduced critical value will be set to (1-ReduceCV)xCV, where CV is the original critical value.

automdl.ljungboxlimit
numeric, Ljung Box limit. Acceptance criterion for the confidence intervals of the Ljung-Box Q statistic. If the LjungBox Q statistics for the residuals of a final model is greater than Ljung Box limit, the model is rejected, the outlier critical value is reduced, and model and outlier identification (if specified) is redone with a reduced value.

automdl.compare
logicals. If TRUE the program compares the model identified by the automatic procedure to the default model (ARIMA(0,1,1)(0,1,1)) and the model with the best fit is selected. Criteria considered are residual diagnostics, the model structure and the number of outliers.

Control variables for the non-automatic modelling of the ARIMA model (automdl.enabled is set to FALSE):

arima.mu
logicals. If TRUE, the mean is considered as part of the ARIMA model.

arima.p
numeric. The order of the non-seasonal autoregressive (AR) polynomial.

arima.d
numeric. Regular differencing order.

arima.q
numeric. The order of the non-seasonal moving average (MA) polynomial.

arima.bp
numeric. The order of the seasonal autoregressive (AR) polynomial.

arima.bd
numeric. Seasonal differencing order.

arima.bq
numeric. The order of the seasonal moving average (MA) polynomial.

Control variables for the user-defined ARMA coefficients. Coefficients can be defined for the regular and seasonal autoregressive (AR) polynomials and moving average (MA) polynomials. The model considers the coefficients only if the procedure for their estimation (arima.coefType) is provided, and the number of provided coefficients matches the sum of (regular and seasonal) AR and MA orders (p,q,bp,bq).

arima.coefEnabled
logicals. If TRUE the program uses the user-defined ARMA coefficients.

arima.coef
vector providing the coefficients for the regular and seasonal AR and MA polynominals. The length of the vector must equal the sum of the regular and seasonal AR and MA orders. The coefficients shall be provided in the order: regular AR (Phi - p elements), regular MA (Theta - q elements), seasonal AR (BPhi - bp elements) and seasonal MA (BTheta - bq elements). E.g.: arima.coef=c(0.6,0.7) with arima.p=1, arima.q=0,arima.bp=1 and arima.bq=0.

arima.coefType
vector defining ARMA coefficients estimation procedure. Possible procedures are: "Undefined" - no use of user-defined input (i.e. coefficients are estimated), "Fixed" - fixes the coefficients at the value provided by the user, "Initial" - the value defined by the user is used as initial condition. For orders for which the coefficients shall not be defined, the arima.coef can be set to NA or 0 or the arima.coefType can be set to "Undefined". E.g.: arima.coef = c(-0.8,-0.6,NA), arima.coefType = c("Fixed","Fixed","Undefined").

fcst.horizon
numeric, forecasts horizon. Length of the forecasts generated by the RegARIMA model in periods (positive values) or years (negative values). By default the program generates two years forecasts (fcst.horizon set to -2).

seats.approx
character, approximation mode. When the ARIMA model estimated by TRAMO does not accept an admissible decomposition, SEATS: "None" - performs an approximation; "Legacy" - replaces the model with a decomposable one; "Noisy" - estimates a new model by adding a white noise to the non-admissible model estimated by TRAMO.

seats.trendBoundary
numeric, trend boundary. The boundary from which an AR root is integrated in the trend component. If the modulus of the inverse real root is greater than Trend boundary, the AR root is integrated in the trend component. Below this value the root is integrated in the transitory component.

seats.seasdBoundary
numeric, seasonal boundary. Boundary from which a negative AR root is integrated in the seasonal component.

seats.seasdBoundary1
numeric, seasonal boundary (unique). Boundary from which a negative AR root is integrated in the seasonal component when the root is the unique seasonal root.

seats.seasTol
numeric, seasonal tolerance. The tolerance (measured in degrees) to allocate the AR non-real roots to the seasonal component (if the modulus of the inverse complex AR root is greater than Trend boundary and the frequency of this root differs from one of the seasonal frequencies by less than Seasonal tolerance) or the transitory component (otherwise).

seats.maBoundary
numeric, MA unit root boundary. When the modulus of an estimated MA root falls in the range (xl, 1), it is set to xl.

seats.method
character, estimation method of the unobserved components. The choice can be made from: "Burman" (default, may result in a significant underestimation of the standard deviations of the components as it may become numerically unstable when some roots of the MA polynomial are near 1); "KalmanSmoother" (it is not disturbed by the (quasi-) unit roots in MA); "McElroyMatrix" (has the same stability issues as the Burman's algorithm).

object
model specification, object of class c("SA_spec","TRAMO_SEATS") or c("SA","TRAMO_SEATS").

Value

A two-elements list of class c("SA_spec","TRAMO_SEATS"): (1) object of class c("regarima_spec","TRAMO_SEATS") with the RegARIMA model specification, (2) object of class c("seats_spec","data.frame") with the SEATS algorithm specification. Each component refers to different part of the SA model specification, mirroring the arguments of the function (for details see arguments description). Each of the lowest-level component (except span, pre-specified outliers, user-defined variables and pre-specified ARMA coefficients) is structured within a data frame with columns denoting different variables of the model specification and rows referring to: first row - base specification, as provided within the argument spec or object; second row - user modifications as specified by the remaining arguments of the function (e.g.: arima.d); and third row - final model specification. The final specification (third row) shall include user modifications (row two) unless they were wrongly specified. The pre-specified outliers, user-defined variables and pre-specified ARMA coefficients consist of a list with the Predefined (base model specification) and Final values.

regarima

object of class c("regarima_spec","TRAMO_SEATS"). See Value of the function regarima_spec_tramoseats

seats

data.frame of class c("seats_spec","data.frame"), containing the seats variables in line with the names of the arguments variables. The final values can be also accessed with the function s_seats.

Details

The available predefined JDemetra+ model specifications (for the function tramoseats_spec_def) are described in the table below.

Identifier	|		Log/level detection	|		Outliers detection

|		Calender effects	|		ARIMA

RSA0 |		NA	|		NA	|

NA	|	Airline(+mean)		RSA1 |	automatic |	AO/LS/TC |

NA	|	Airline(+mean)		RSA2 |	automatic |	AO/LS/TC |	2 td vars + Easter |

Airline(+mean)		RSA3 |	automatic |	AO/LS/TC |		NA	|

automatic		RSA4 |	automatic |	AO/LS/TC |	2 td vars + Easter |	automatic

RSA5 |	automatic |	AO/LS/TC |	7 td vars + Easter |	automatic		RSAfull |	automatic |

AO/LS/TC |	automatic |	automatic		Identifier	|		Log/level detection

References

Info on JDemetra+, usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en BOX G.E.P. and JENKINS G.M. (1970), "Time Series Analysis: Forecasting and Control", Holden-Day, San Francisco.

BOX G.E.P., JENKINS G.M., REINSEL G.C. and LJUNG G.M. (2015), "Time Series Analysis: Forecasting and Control", John Wiley & Sons, Hoboken, N. J., 5th edition.

See also

Examples

myseries <- ipi_c_eu[, "FR"]
myspec1 <- tramoseats_spec_def(spec = c("RSAfull"))
mysa1 <- tramoseats(myseries, spec=myspec1)

# Modify a pre-specified model specification
myspec2 <- tramoseats_spec_def(spec = c("RSAfull"), tradingdays.mauto = "Unused",
                              tradingdays.option = "WorkingDays",
                              easter.type = "Standard",
                              automdl.enabled = FALSE, arima.mu = TRUE)
mysa2 <- tramoseats(myseries, spec = myspec2)

# Modify the model specification from a "SA" object
myspec3 <- tramoseats_spec(mysa1, tradingdays.mauto = "Unused",
                           tradingdays.option = "WorkingDays",
                           easter.type = "Standard", automdl.enabled = FALSE, arima.mu = TRUE)
mysa3 <- tramoseats(myseries, myspec3)

# Modify the model specification from a "SA_spec" object
myspec4 <- tramoseats_spec(myspec1,tradingdays.mauto = "Unused",
                           tradingdays.option = "WorkingDays",
                           easter.type = "Standard", automdl.enabled = FALSE, arima.mu = TRUE)
mysa4 <- tramoseats(myseries, myspec4)

# Pre-specified outliers
myspec1 <- tramoseats_spec_def(spec=c("RSAfull"),
                              usrdef.outliersEnabled = TRUE,
                              usrdef.outliersType = c("LS","LS"),
                              usrdef.outliersDate = c("2008-10-01","2003-01-01"),
                              usrdef.outliersCoef = c(10,-8), transform.function = "None")
s_preOut(myspec1)

#>   type       date coeff
#> 1   LS 2008-10-01    10
#> 2   LS 2003-01-01    -8

mysa1 <- tramoseats(myseries, myspec1)
mysa1

#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5290      0.048
#> BTheta(1)  -0.5767      0.048
#> 
#>               Estimate Std. Error
#> Monday        -0.07228      0.192
#> Tuesday        0.11018      0.190
#> Wednesday      0.16251      0.191
#> Thursday      -0.55185      0.190
#> Friday         0.41498      0.192
#> Saturday       0.13543      0.191
#> Easter [6]    -1.12520      0.394
#> LS (11-2008) -12.25330      1.561
#> AO (12-2002)  -6.45816      1.493
#> LS (1-2009)   -7.07048      1.498
#> LS (9-2008)   -6.20208      1.500
#> 
#> Fixed outliers: 
#>              Coefficients
#> LS (10-2008)           10
#> LS (1-2003)            -8
#> 
#> 
#> Residual standard error: 1.858 on 323 degrees of freedom
#> Log likelihood =  -661, aic =  1350 aicc =  1351, bic(corrected for length) = 1.471
#> 
#> 
#> 
#> Decomposition
#> Model
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.529038 B - 0.576652 B^12 + 0.305071 B^13 
#> 
#> 
#> SA
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.493010 B + 0.513732 B^2 
#> Innovation variance:  0.6429039 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.044711 B - 0.955289 B^2 
#> Innovation variance:  0.03452403 
#> 
#> Seasonal
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 1.222150 B + 1.160157 B^2 + 1.000327 B^3 + 0.763420 B^4 + 0.512045 B^5 + 0.268153 B^6 + 0.048532 B^7 - 0.122082 B^8 - 0.270259 B^9 - 0.360991 B^10 - 0.581084 B^11 
#> Innovation variance:  0.04338372 
#> 
#> Irregular
#> Innovation variance:  0.3632607 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t           s           i
#> Jan 2017  97.4 100.7979 100.9584  -3.3979068 -0.16047894
#> Feb 2017  97.5 100.4892 101.2959  -2.9892168 -0.80664833
#> Mar 2017 112.0 102.5769 101.7208   9.4231475  0.85605254
#> Apr 2017 103.0 101.4948 102.1242   1.5051912 -0.62944043
#> May 2017 100.4 103.1439 102.5098  -2.7438926  0.63412880
#> Jun 2017 111.2 102.5950 102.8643   8.6049570 -0.26928561
#> Jul 2017 103.4 103.2819 103.2055   0.1181372  0.07637373
#> Aug 2017  79.3 103.3246 103.5825 -24.0245752 -0.25791384
#> Sep 2017 109.7 103.1983 104.0809   6.5016545 -0.88258093
#> Oct 2017 114.0 106.2017 104.6202   7.7982962  1.58152641
#> Nov 2017 107.7 105.4460 104.9030   2.2539632  0.54299390
#> Dec 2017 101.4 104.4912 104.9886  -3.0911545 -0.49745990
#> 
#> Forecasts:
#>                y_f     sa_f      t_f          s_f          i_f
#> Jan 2018 102.06576 105.1151 105.1151  -3.04936263 1.220712e-11
#> Feb 2018 102.22450 105.2868 105.2868  -3.06228987 1.148237e-11
#> Mar 2018 113.95694 105.4585 105.4585   8.49847169 1.313083e-11
#> Apr 2018 107.79414 105.6301 105.6301   2.16400028 1.155342e-11
#> May 2018 102.59676 105.8018 105.8018  -3.20504699 1.095657e-11
#> Jun 2018 115.21892 105.9735 105.9735   9.24544124 1.355716e-11
#> Jul 2018 106.21789 106.1452 106.1452   0.07273947 1.278977e-11
#> Aug 2018  82.55735 106.3168 106.3168 -23.75946902 1.050182e-11
#> Sep 2018 112.30108 106.4885 106.4885   5.81258745 1.290346e-11
#> Oct 2018 114.93423 106.6602 106.6602   8.27405932 1.314504e-11
#> Nov 2018 109.38136 106.8318 106.8318   2.54952421 1.145395e-11
#> Dec 2018 103.39058 107.0035 107.0035  -3.61293045 1.223555e-11
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.499
#>  Seasonal     45.062
#>  Irregular     0.357
#>  TD & Hol.     0.056
#>  Others       52.353
#>  Total        99.327
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.054
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.974
#>  f-test on sa (td)                      0.998
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables

s_preOut(mysa1)

#>   type       date coeff
#> 1   LS 2008-10-01    10
#> 2   LS 2003-01-01    -8

# User-defined variables
var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
var<- ts.union(var1, var2)

myspec1 <- tramoseats_spec_def(spec=c("RSAfull"),
                              usrdef.varEnabled = TRUE, usrdef.var = var)
s_preVar(myspec1)

#> $series
#>                  var1         var2
#> Jan 1990  -4.24674825   -1.7532122
#> Feb 1990 -11.43137603  -39.0664205
#> Mar 1990   2.33440224  -40.2775110
#> Apr 1990  -1.89236577  127.6706963
#> May 1990 -16.63411256   82.1075754
#> Jun 1990  19.15690770   41.9785146
#> Jul 1990  -8.16894436  171.1661220
#> Aug 1990   3.83650493 -221.5632715
#> Sep 1990  -4.58632251  -18.4878197
#> Oct 1990  -7.15703978   67.1168243
#> Nov 1990   4.29548541  -79.5140242
#> Dec 1990 -11.69852745 -156.6257835
#> Jan 1991  13.22088309  121.3732488
#> Feb 1991  13.65964636  150.3810074
#> Mar 1991  -2.49723710 -116.9425021
#> Apr 1991  -4.24721350   61.3377167
#> May 1991  10.98439335  148.3931321
#> Jun 1991   6.70923041  -44.3773736
#> Jul 1991   0.30954656   26.8589741
#> Aug 1991  23.68511961 -122.8216829
#> Sep 1991  25.85788113   47.8528762
#> Oct 1991  12.12280955  166.9859573
#> Nov 1991 -12.51196841   -3.6455545
#> Dec 1991   4.00266042  -44.0936271
#> Jan 1992   9.34707631   73.4408890
#> Feb 1992  -4.72669846  -57.5309458
#> Mar 1992  15.83013221    7.4897554
#> Apr 1992   5.97754220    5.9301839
#> May 1992   4.66084586   84.8264703
#> Jun 1992  -7.44367537  257.5396587
#> Jul 1992   9.59055258    9.4259964
#> Aug 1992   1.52637081  -51.9949532
#> Sep 1992  -7.69368697   77.4428621
#> Oct 1992  13.93080739  122.8289259
#> Nov 1992   9.18668766  188.0606885
#> Dec 1992  10.16087064  -12.5241723
#> Jan 1993   5.34635439  -75.9094766
#> Feb 1993  -9.87628560  -50.9303603
#> Mar 1993  11.87792273  -15.6039669
#> Apr 1993  -5.17542212   13.2582442
#> May 1993  -2.59560251  -97.4084022
#> Jun 1993  -3.28064671 -125.0949663
#> Jul 1993   0.73432395   -4.6336042
#> Aug 1993  -2.47863020  113.7387771
#> Sep 1993 -13.73862256  126.4449891
#> Oct 1993  -0.40445818  -54.8742076
#> Nov 1993   4.21538239   68.2391430
#> Dec 1993   2.01597506  -92.9773363
#> Jan 1994 -16.97191918  -97.5148693
#> Feb 1994   6.42287683   -2.7117079
#> Mar 1994  -9.95239613   44.6671886
#> Apr 1994   9.63813900  -32.9112249
#> May 1994 -16.56037231   11.3049080
#> Jun 1994  10.70861091  105.1229395
#> Jul 1994  -1.09026360   32.6977400
#> Aug 1994  18.99186393   78.7037444
#> Sep 1994 -11.37030733   44.1559619
#> Oct 1994  -2.79719758 -101.9304514
#> Nov 1994  -8.94129051  -15.8698326
#> Dec 1994   1.36701846 -163.7701292
#> Jan 1995  -7.49165418  283.9266231
#> Feb 1995   5.18199077   96.1819160
#> Mar 1995  -1.92337214   50.8432001
#> Apr 1995   0.28809812   25.9523192
#> May 1995   3.58590886  -77.3404126
#> Jun 1995  -0.28995035 -180.5527531
#> Jul 1995  11.47042072   17.8383232
#> Aug 1995   3.73588935  -52.5777309
#> Sep 1995   3.23339211   40.0186316
#> Oct 1995  -8.29819320  -12.2587586
#> Nov 1995  13.94462582   -7.1345032
#> Dec 1995  -1.91543577  209.5734365
#> Jan 1996   2.72277021  132.2082615
#> Feb 1996 -10.81659008  120.0558995
#> Mar 1996 -23.29399361  -78.1874565
#> Apr 1996  -5.49625959  -44.1476451
#> May 1996  -0.72579991  -24.6028129
#> Jun 1996  10.32288404   36.5792190
#> Jul 1996   2.15138526   -3.4132882
#> Aug 1996  -4.94340125   29.6227003
#> Sep 1996  15.12138281  -22.1059772
#> Oct 1996  -5.79463264   18.7706159
#> Nov 1996  16.74789634  -43.8062172
#> Dec 1996 -10.00980259   -8.8733579
#> Jan 1997  12.22702835   63.4250029
#> Feb 1997  10.77188166   86.5538237
#> Mar 1997  -6.11945460  -43.0593090
#> Apr 1997   5.06866968    1.3030894
#> May 1997   4.60068373    5.1249629
#> Jun 1997  14.84391861   53.8638152
#> Jul 1997   8.81963226   51.9672613
#> Aug 1997  -5.36702239 -242.2911275
#> Sep 1997  12.85553707  -80.6290253
#> Oct 1997   5.87848533  -48.4679992
#> Nov 1997 -13.08482026  -65.3077284
#> Dec 1997   3.16726325  -29.3764247
#> Jan 1998  11.94158303  -27.6989004
#> Feb 1998   9.13057148   13.6469589
#> Mar 1998  -7.86752993  106.7207722
#> Apr 1998  -4.10929700    0.1828103
#> May 1998   4.76373682   14.0092252
#> Jun 1998   1.85915868  -52.0598538
#> Jul 1998 -13.24716253  236.5811636
#> Aug 1998  11.72371055 -169.6474415
#> Sep 1998   2.31492772   68.2585584
#> Oct 1998   4.83071904  136.1561976
#> Nov 1998  -5.35319580  190.5955757
#> Dec 1998  13.78135819  -67.2838542
#> Jan 1999 -13.02671671 -106.2911432
#> Feb 1999   6.34857535  163.3683158
#> Mar 1999   9.99655161   54.0830359
#> Apr 1999  -3.37748953  145.6857971
#> May 1999  -0.86033607 -193.3784243
#> Jun 1999 -17.18817584  -88.0400798
#> Jul 1999  -9.29121572  -79.7730162
#> Aug 1999   8.13676046  -88.6130974
#> Sep 1999   5.26413546  -29.4827785
#> Oct 1999  10.11957390  -88.6651693
#> Nov 1999   8.31379809    9.9084673
#> Dec 1999   4.15134145   72.0546895
#> Jan 2000  -6.61257061  100.0469215
#> Feb 2000   5.48811286   52.7999513
#> Mar 2000  -5.39367800  116.3605931
#> Apr 2000  -1.71318426   17.3137014
#> May 2000  -7.66734605   57.7917253
#> Jun 2000  -6.06574846  -11.6008707
#> Jul 2000  16.47312821   90.4521145
#> Aug 2000   8.41069348   -1.6586710
#> Sep 2000   3.38175097  -92.2297409
#> Oct 2000  -4.04883076 -106.0114550
#> Nov 2000   9.00009039 -178.7199640
#> Dec 2000  11.90666655 -106.6551779
#> Jan 2001  -0.47539083 -112.9829171
#> Feb 2001   7.86925425   38.7267796
#> Mar 2001  11.66684112  127.0263382
#> Apr 2001  -2.71823871 -164.6117829
#> May 2001  -3.13704285  -12.3604069
#> Jun 2001  -2.86300150  -97.1640654
#> Jul 2001  -8.03440263  106.8587364
#> Aug 2001  -0.94555455  231.6252874
#> Sep 2001  -0.74046506 -168.2873392
#> Oct 2001   7.14490683   13.5875428
#> Nov 2001   1.24718366   49.8437767
#> Dec 2001  12.96594339  133.3462968
#> Jan 2002 -11.14492926   -2.2954568
#> Feb 2002  -8.42058813  -19.1523237
#> Mar 2002 -15.04294209  -44.0614893
#> Apr 2002  -2.84026455 -194.1405982
#> May 2002   0.42869041  141.5706719
#> Jun 2002  -0.08866413   59.6399657
#> Jul 2002 -29.49083784  -65.2922289
#> Aug 2002   0.20358254  -15.4144673
#> Sep 2002  -0.98446980  -44.3645660
#> Oct 2002   5.89109098  -51.8233778
#> Nov 2002  -4.25751629   20.6107064
#> Dec 2002   6.34658985  163.9453666
#> Jan 2003  -5.78524902 -156.4734714
#> Feb 2003  -1.69109048   51.1589918
#> Mar 2003 -19.19232520 -153.9661373
#> Apr 2003 -15.34266387   36.2795564
#> May 2003 -11.14761221   77.0735944
#> Jun 2003  15.97811635 -128.8986224
#> Jul 2003  -6.39805078  -71.3194370
#> Aug 2003  15.66690553  -10.2383386
#> Sep 2003 -14.49155712  -64.7611750
#> Oct 2003  -7.91503959  -37.3253836
#> Nov 2003  -5.04480733    8.3347213
#> Dec 2003   4.01826700   91.7216273
#> Jan 2004   9.71396469  -25.8953302
#> Feb 2004  -5.79663020   67.7025733
#> Mar 2004  16.04179789  -42.0100735
#> Apr 2004   2.25973337  104.7205425
#> May 2004  -5.14857122 -154.6576348
#> Jun 2004  -8.23788240 -104.6005681
#> Jul 2004   3.34415250 -227.2669757
#> Aug 2004  -0.93490765  -31.6651938
#> Sep 2004   3.04062078 -147.6190645
#> Oct 2004  -4.76507534  -10.0387827
#> Nov 2004  -2.41311717  -74.6222611
#> Dec 2004   8.24155833   20.3645883
#> Jan 2005 -15.55643970   49.7450361
#> Feb 2005   0.93501283 -101.4286644
#> Mar 2005  -3.66949421  114.5760347
#> Apr 2005  -1.29408799  -21.4418765
#> May 2005   4.07795520  167.4899343
#> Jun 2005  -5.83968954   11.6360148
#> Jul 2005  -1.93449890  142.6565587
#> Aug 2005  -2.69546670  -23.9667330
#> Sep 2005   0.73658232 -110.5473915
#> Oct 2005   3.57236138   32.2897729
#> Nov 2005   5.50428418    2.1746800
#> Dec 2005   0.38401793   31.7624820
#> Jan 2006 -16.09575292  -96.9700843
#> Feb 2006 -10.49710115  143.4467716
#> Mar 2006  20.52033974  -98.8804847
#> Apr 2006   1.76436411   39.6349598
#> May 2006  11.28307358   52.0057416
#> Jun 2006   4.35001744   71.6472580
#> Jul 2006   5.48833492   -6.1115805
#> Aug 2006   6.47418304  227.4376578
#> Sep 2006   8.78463454  -89.7637203
#> Oct 2006   3.50797869   61.8037813
#> Nov 2006   0.49879720    7.7563519
#> Dec 2006   8.35749446   71.3500681
#> Jan 2007  -2.81725199   52.6209825
#> Feb 2007  -7.91679461  -47.0555734
#> Mar 2007   0.01653727   60.8828550
#> Apr 2007 -11.87530126   77.3244763
#> May 2007   3.62417465  -81.9748202
#> Jun 2007  -5.49435319  -12.8422743
#> Jul 2007   6.92949838  -36.5022796
#> Aug 2007  -0.60843916   -7.4789419
#> Sep 2007 -11.93540930 -109.6987881
#> Oct 2007  -1.19908059   96.6713334
#> Nov 2007  -7.08107950   93.9926947
#> Dec 2007 -16.16663973 -160.0175495
#> Jan 2008   4.95847824   20.8536586
#> Feb 2008  12.99750970  112.1260348
#> Mar 2008 -16.15985506  -65.1995398
#> Apr 2008 -12.50367945 -109.0135269
#> May 2008  15.82132306  -71.2182080
#> Jun 2008  12.52529118 -136.2656668
#> Jul 2008  -2.09395701  -50.0690393
#> Aug 2008  -6.39033687  143.8472860
#> Sep 2008   3.98902937  -51.1637695
#> Oct 2008  13.41973500   -0.1967701
#> Nov 2008  -0.37607422  178.3134496
#> Dec 2008   5.43670114  103.1380979
#> Jan 2009  17.40218394  -58.0704688
#> Feb 2009  -3.24455939 -159.3084983
#> Mar 2009  -4.47511333  -48.9042542
#> Apr 2009   4.16082526  -75.2700133
#> May 2009 -12.95865368   66.5378803
#> Jun 2009   6.82633090   73.5663889
#> Jul 2009   4.96913333   73.0799344
#> Aug 2009  14.39660656  -52.8728468
#> Sep 2009  12.07358558 -121.2315122
#> Oct 2009   2.38514694  -61.9899029
#> Nov 2009  -2.98837878   78.0897501
#> Dec 2009  13.84621050    9.6229707
#> Jan 2010  -7.00069022 -156.1350265
#> Feb 2010  18.60931509   69.8374219
#> Mar 2010  18.03924959    3.8195755
#> Apr 2010 -13.56043492   23.4660661
#> May 2010  -3.14751072  102.1328475
#> Jun 2010 -10.19265181  147.5123570
#> Jul 2010  -6.12249656   -4.8398385
#> Aug 2010  -2.89068749  -59.4672262
#> Sep 2010  15.06983292  -47.4006086
#> Oct 2010   3.67722694  -32.6361283
#> Nov 2010  -0.20034941 -118.5661675
#> Dec 2010  -9.79921959  -97.7288119
#> Jan 2011 -14.00912333  -68.2677435
#> Feb 2011  14.45014910   -3.2527689
#> Mar 2011  -4.23481621  148.1481122
#> Apr 2011  -7.33749597    2.8254600
#> May 2011  -7.08484753  -68.8045915
#> Jun 2011  14.75092459  -89.4731112
#> Jul 2011   8.45004185  -53.3170591
#> Aug 2011  12.93994431   58.3858371
#> Sep 2011   2.98161142 -126.2859995
#> Oct 2011  -4.05682490   89.3197098
#> Nov 2011  -1.38807578  -89.7183505
#> Dec 2011  -2.22592731   55.4612733
#> Jan 2012  17.47150257 -116.4003290
#> Feb 2012  -0.81704668    1.8656670
#> Mar 2012   1.24141586   32.8821201
#> Apr 2012   3.92236141  -29.4946314
#> May 2012   3.78419970   64.1599904
#> Jun 2012   1.66175673   92.6677808
#> Jul 2012  11.53387369  177.2124003
#> Aug 2012   0.14748302  166.2092676
#> Sep 2012   0.21440845    2.7928905
#> Oct 2012  -8.26352170  161.5495109
#> Nov 2012   0.47399133 -134.7522413
#> Dec 2012  -2.93107372   19.3626828
#> Jan 2013  12.07632181   42.1505299
#> Feb 2013   0.32001384  -50.0133196
#> Mar 2013   1.80397550 -159.4426010
#> Apr 2013   0.77160580  141.7076086
#> May 2013   5.58395338  105.0564114
#> Jun 2013  -1.66854367  -58.4431908
#> Jul 2013  -2.27980619  -52.6714875
#> Aug 2013   2.42666056  -12.2240719
#> Sep 2013  -9.28482969  -58.6328284
#> Oct 2013  15.15406678  -24.7866875
#> Nov 2013  -8.11598824   80.1257151
#> Dec 2013  13.11247408  252.5108768
#> Jan 2014   8.15566796   27.0033039
#> Feb 2014   1.20032668  -88.6357031
#> Mar 2014   8.81789848  -76.9523848
#> Apr 2014 -17.98743117   45.4057021
#> May 2014  -9.57051078  -70.6657247
#> Jun 2014  -4.74058606  -94.9395463
#> Jul 2014 -18.60461283 -106.6400599
#> Aug 2014   2.53690073  -88.6451189
#> Sep 2014  -2.14401713   66.8592305
#> Oct 2014 -23.70181451 -206.7564333
#> Nov 2014   0.03258078  169.4146408
#> Dec 2014   6.69745568  -18.9770360
#> Jan 2015   2.26820843   42.2166994
#> Feb 2015  -5.91868492  -66.3115731
#> Mar 2015  -4.05414494  -66.2765455
#> Apr 2015  14.14224665   11.3292209
#> May 2015   2.41625825   94.2951075
#> Jun 2015 -14.36087478 -121.5873331
#> Jul 2015  -3.14552745   -4.7262742
#> Aug 2015  -4.80542811  -19.2044959
#> Sep 2015   8.36254248   63.2148567
#> Oct 2015  -0.52349287  -15.1904953
#> Nov 2015   9.51957625    1.5200375
#> Dec 2015  14.94512798   88.8718179
#> Jan 2016  -3.44214729    1.0334954
#> Feb 2016   6.45775407   34.4932320
#> Mar 2016 -15.34500551  -91.3178422
#> Apr 2016   5.99672514  -92.4398105
#> May 2016  -5.04358892  -37.2077190
#> Jun 2016  -4.06909625  -29.7797852
#> Jul 2016  -4.60935206  157.4491890
#> Aug 2016  22.14908426 -198.0661690
#> Sep 2016  14.86967820  -16.6246184
#> Oct 2016  -3.55098858   12.0218534
#> Nov 2016   8.76203737  -66.1734312
#> Dec 2016  -2.14696646  -53.0622574
#> Jan 2017  20.20798363  -30.1418705
#> Feb 2017 -18.38661562  -60.2301607
#> Mar 2017   1.72937164  -31.8208084
#> Apr 2017  -7.63245040   30.8196044
#> May 2017 -11.18492914   79.8650087
#> Jun 2017  11.97977461  175.0574467
#> Jul 2017   9.81193912   26.5330264
#> Aug 2017  11.91749205   31.7154641
#> Sep 2017   6.32833568  -14.8736377
#> Oct 2017   5.32825947   32.2218955
#> Nov 2017  -8.20121900  -32.4209133
#> Dec 2017  -1.59325676  -78.3359275
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 

mysa1 <- tramoseats(myseries, myspec1)

myspec2 <- tramoseats_spec_def(spec=c("RSAfull"), usrdef.varEnabled = TRUE,
                              usrdef.var = var, usrdef.varCoef = c(17,-1),
                              transform.function = "None")
mysa2 <- tramoseats(myseries, myspec2)

# Pre-specified ARMA coefficients
myspec1 <- tramoseats_spec_def(spec = c("RSAfull"),
                              arima.coefEnabled = TRUE, automdl.enabled = FALSE,
                              arima.p = 2, arima.q = 0,
                              arima.bp = 1, arima.bq = 1,
                              arima.coef = c(-0.12, -0.12, -0.3, -0.99),
                              arima.coefType = rep("Fixed", 4))
mysa1 <- tramoseats(myseries, myspec1)
mysa1

#> RegARIMA
#> y = regression model + arima (2, 1, 0, 1, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)       -0.12          0
#> Phi(2)       -0.12          0
#> BPhi(1)      -0.30          0
#> BTheta(1)    -0.99          0
#> 
#>                Estimate Std. Error
#> Mean         -3.039e-05      0.000
#> Monday       -2.430e-03      0.002
#> Tuesday       3.562e-04      0.002
#> Wednesday     4.238e-03      0.002
#> Thursday     -6.307e-03      0.002
#> Friday        4.267e-03      0.002
#> Saturday      5.834e-04      0.002
#> Easter [6]   -1.144e-02      0.003
#> LS (11-2008) -5.733e-02      0.020
#> LS (1-2009)  -4.622e-02      0.020
#> 
#> 
#> Residual standard error: 0.02054 on 323 degrees of freedom
#> Log likelihood = 781.7, aic =  1441 aicc =  1442, bic(corrected for length) = -7.591
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 - 0.120000 B - 0.120000 B^2 - 0.300000 B^12 + 0.036000 B^13 + 0.036000 B^14 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.950000 B^12 
#> 
#> 
#> SA
#> AR :  1 - 1.024538 B - 0.011455 B^2 + 0.108545 B^3 
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.941460 B + 0.995458 B^2 + 0.036596 B^3 - 0.098708 B^4 + 0.008921 B^5 
#> Innovation variance:  0.4957156 
#> 
#> Trend
#> AR :  1 - 0.904538 B 
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 0.722193 B - 0.998833 B^2 + 0.723360 B^3 
#> Innovation variance:  0.1026625 
#> 
#> Seasonal
#> AR :  1 + 0.904538 B + 0.818189 B^2 + 0.740083 B^3 + 0.669433 B^4 + 0.605527 B^5 + 0.547723 B^6 + 0.495436 B^7 + 0.448140 B^8 + 0.405360 B^9 + 0.366664 B^10 + 0.331661 B^11 
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 2.915179 B + 5.617938 B^2 + 8.595374 B^3 + 11.505684 B^4 + 14.135214 B^5 + 16.348944 B^6 + 18.036964 B^7 + 19.197197 B^8 + 19.965174 B^9 + 20.301073 B^10 + 20.149548 B^11 + 19.110302 B^12 + 17.206568 B^13 + 14.558550 B^14 + 11.650923 B^15 + 8.806927 B^16 + 6.228869 B^17 + 4.044510 B^18 + 2.357557 B^19 + 1.170033 B^20 + 0.354583 B^21 - 0.052292 B^22 
#> Innovation variance:  0.2584115 
#> 
#> Transitory
#> AR :  1 - 0.120000 B - 0.120000 B^2 
#> MA :  1 + 1.287572 B + 0.287572 B^2 
#> Innovation variance:  9.66767e-06 
#> 
#> Irregular
#> Innovation variance:  0.1228651 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2017  97.4 101.2668 101.0578 0.9618154 1.0020686
#> Feb 2017  97.5 101.2721 101.5876 0.9627533 0.9968935
#> Mar 2017 112.0 102.5609 102.1042 1.0920339 1.0044733
#> Apr 2017 103.0 102.0990 102.4310 1.0088251 0.9967584
#> May 2017 100.4 103.0163 102.5974 0.9746033 1.0040828
#> Jun 2017 111.2 102.2454 102.7571 1.0875793 0.9950202
#> Jul 2017 103.4 103.2943 103.3689 1.0010236 0.9992781
#> Aug 2017  79.3 104.7000 103.7753 0.7574017 1.0089114
#> Sep 2017 109.7 102.5846 103.6790 1.0693609 0.9894443
#> Oct 2017 114.0 104.5874 104.0358 1.0899973 1.0053026
#> Nov 2017 107.7 104.4425 104.1189 1.0311891 1.0031083
#> Dec 2017 101.4 103.2201 103.5819 0.9823665 0.9965070
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2018  99.15257 103.2978 103.2989 0.9635235 0.9999901
#> Feb 2018  98.49070 103.2623 103.2633 0.9585165 0.9999903
#> Mar 2018 110.09644 103.2323 103.2333 1.0713667 0.9999904
#> Apr 2018 103.83531 103.2074 103.2084 1.0108616 0.9999904
#> May 2018  99.26109 103.1870 103.1880 0.9666254 0.9999905
#> Jun 2018 112.97989 103.1708 103.1717 1.0975670 0.9999905
#> Jul 2018 102.73624 103.1582 103.1592 0.9999921 0.9999905
#> Aug 2018  76.60539 103.1489 103.1499 0.7477158 0.9999905
#> Sep 2018 109.12847 103.1426 103.1436 1.0619685 0.9999905
#> Oct 2018 112.33867 103.1390 103.1400 1.0926514 0.9999905
#> Nov 2018 105.35175 103.1377 103.1387 1.0250662 0.9999905
#> Dec 2018 101.11651 103.1387 103.1396 0.9811503 0.9999905
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.773
#>  Seasonal     77.478
#>  Irregular     0.138
#>  TD & Hol.     0.092
#>  Others       16.847
#>  Total        97.328
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.254
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.134
#>  f-test on sa (td)                      0.954
#>  f-test on i (td)                       0.987
#> 
#> 
#> Additional output variables

s_arimaCoef(myspec1)

#>            Type Value
#> Phi(1)    Fixed -0.12
#> Phi(2)    Fixed -0.12
#> BPhi(1)   Fixed -0.30
#> BTheta(1) Fixed -0.99

s_arimaCoef(mysa1)

#>            Type Value
#> Phi(1)    Fixed -0.12
#> Phi(2)    Fixed -0.12
#> BPhi(1)   Fixed -0.30
#> BTheta(1) Fixed -0.99