X-13ARIMA-SEATS model specification, SA/X13

Function to create (and/or modify) a c("SA_spec", "X13") class object with the SA model specification for the X13 method. It can be done from a pre-defined 'JDemetra+' model specification (a character), a previous specification (c("SA_spec", "X13") object) or a seasonal adjustment model (c("SA", "X13") object).

x13_spec(
  spec = c("RSA5c", "RSA0", "RSA1", "RSA2c", "RSA3", "RSA4c", "X11"),
  preliminary.check = NA,
  estimate.from = NA_character_,
  estimate.to = NA_character_,
  estimate.first = NA_integer_,
  estimate.last = NA_integer_,
  estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_,
  estimate.tol = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.adjust = c(NA, "None", "LeapYear", "LengthOfPeriod"),
  transform.aicdiff = NA_integer_,
  usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA,
  usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA,
  usrdef.varEnabled = NA,
  usrdef.var = NA,
  usrdef.varType = NA,
  usrdef.varCoef = NA,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined", "None"),
  tradingdays.autoadjust = NA,
  tradingdays.leapyear = c(NA, "LeapYear", "LengthOfPeriod", "None"),
  tradingdays.stocktd = NA_integer_,
  tradingdays.test = c(NA, "Remove", "Add", "None"),
  easter.enabled = NA,
  easter.julian = NA,
  easter.duration = NA_integer_,
  easter.test = c(NA, "Add", "Remove", "None"),
  outlier.enabled = NA,
  outlier.from = NA_character_,
  outlier.to = NA_character_,
  outlier.first = NA_integer_,
  outlier.last = NA_integer_,
  outlier.exclFirst = NA_integer_,
  outlier.exclLast = NA_integer_,
  outlier.ao = NA,
  outlier.tc = NA,
  outlier.ls = NA,
  outlier.so = NA,
  outlier.usedefcv = NA,
  outlier.cv = NA_integer_,
  outlier.method = c(NA, "AddOne", "AddAll"),
  outlier.tcrate = NA_integer_,
  automdl.enabled = NA,
  automdl.acceptdefault = NA,
  automdl.cancel = NA_integer_,
  automdl.ub1 = NA_integer_,
  automdl.ub2 = NA_integer_,
  automdl.mixed = NA,
  automdl.balanced = NA,
  automdl.armalimit = NA_integer_,
  automdl.reducecv = NA_integer_,
  automdl.ljungboxlimit = NA_integer_,
  automdl.ubfinal = NA_integer_,
  arima.mu = NA,
  arima.p = NA_integer_,
  arima.d = NA_integer_,
  arima.q = NA_integer_,
  arima.bp = NA_integer_,
  arima.bd = NA_integer_,
  arima.bq = NA_integer_,
  arima.coefEnabled = NA,
  arima.coef = NA,
  arima.coefType = NA,
  fcst.horizon = NA_integer_,
  x11.mode = c(NA, "Undefined", "Additive", "Multiplicative", "LogAdditive",
    "PseudoAdditive"),
  x11.seasonalComp = NA,
  x11.lsigma = NA_integer_,
  x11.usigma = NA_integer_,
  x11.trendAuto = NA,
  x11.trendma = NA_integer_,
  x11.seasonalma = NA_character_,
  x11.fcasts = NA_integer_,
  x11.bcasts = NA_integer_,
  x11.calendarSigma = NA,
  x11.sigmaVector = NA,
  x11.excludeFcasts = NA
)

Arguments

spec
model specification X13. It can be a character of predefined X13 'JDemetra+' model specification (see Details), an object of class c("SA_spec","X13") or an object of class c("SA", "X13"). The default is "RSA5c".

preliminary.check
boolean to check the quality of the input series and exclude highly problematic ones: e.g. these with a number of identical observations and/or missing values above pre-specified threshold values.

The time span of the series to be used for the estimation of the RegARIMA model coefficients (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: estimate.from, estimate.to, estimate.first, estimate.last, estimate.exclFirst and estimate.exclLast; where estimate.from and estimate.to have priority over remaining span control variables, estimate.last and estimate.first have priority over estimate.exclFirst and estimate.exclLast, and estimate.last has priority over estimate.first.

estimate.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with estimate.to.

estimate.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with estimate.from.

estimate.first
numeric specifying the number of periods considered at the beginning of the series.

estimate.last
numeric specifying the number of periods considered at the end of the series.

estimate.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with estimate.exclLast.

estimate.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with estimate.exclFirst.

estimate.tol
numeric, convergence tolerance. The absolute changes in the log-likelihood function are compared to this value to check for the convergence of the estimation iterations.

transform.function
transformation of the input series: "None" - no transformation of the series; "Log" - takes the log of the series; "Auto" - the program tests for the log-level specification.

transform.adjust
pre-adjustment of the input series for length of period or leap year effects: "None" - no adjustment; "LeapYear" - leap year effect; "LengthOfPeriod" - length of period. Modifications of this variable are taken into account only when transform.function is set to "Log".

transform.aicdiff
numeric defining the difference in AICC needed to accept no transformation when the automatic transformation selection is chosen (considered only when transform.function is set to "Auto").

Control variables for the pre-specified outliers. The pre-specified outliers are used in the model only if they are enabled (usrdef.outliersEnabled=TRUE) and the outliers' type (usrdef.outliersType) and date (usrdef.outliersDate) are provided.

usrdef.outliersEnabled
logicals. If TRUE the program uses the pre-specified outliers.

usrdef.outliersType
vector defining the outliers' type. Possible types are: ("AO") - additive, ("LS") - level shift, ("TC") - transitory change, ("SO") - seasonal outlier. E.g.: usrdef.outliersType =c("AO","AO","LS").

usrdef.outliersDate
vector defining the outliers' date. The dates should be characters in format "YYYY-MM-DD". E.g.: usrdef.outliersDate=c("2009-10-01","2005-02-01","2003-04-01").

usrdef.outliersCoef
vector providing fixed coefficients for the outliers. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined. E.g.: usrdef.outliersCoef=c(200,170,20).

Control variables for the user-defined variables:

usrdef.varEnabled
logicals. If TRUE the program uses the user-defined variables.

usrdef.var
time series (ts) or matrix of time series (mts) with the user-defined variables.

usrdef.varType
vector of character(s) defining the user-defined variables component type. Possible types are: "Undefined", "Series", "Trend", "Seasonal", "SeasonallyAdjusted", "Irregular", "Calendar". The type "Calendar" has to be used with tradingdays.option = "UserDefined" to use user-defined calendar regressors. If not specified, the program will assign the "Undefined" type.

usrdef.varCoef
vector providing fixed coefficients for the user-defined variables. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined.

tradingdays.option
defines the type of the trading days regression variables: "TradingDays" - six day-of-the-week regression variables; "WorkingDays" - one working/non-working day contrast variable; "None" - no correction for trading days and working days effects; "UserDefined" - user-defined trading days regressors (regressors have to be defined by the usrdef.var argument with usrdef.varType set to "Calendar" and usrdef.varEnabled = TRUE). "None" has also to be chosen for the "day-of-week effects" correction (tradingdays.stocktd to be modified accordingly).

tradingdays.autoadjust
logicals. If TRUE the program corrects automatically for the leap year effect. Modifications of this variable are taken into account only when transform.function is set to "Auto".

tradingdays.leapyear
option for including the leap-year effect in the model: "LeapYear" - leap year effect; "LengthOfPeriod" - length of period, "None" - no effect included. The leap-year effect can be pre-specified in the model only if the input series was not pre-adjusted (transform.adjust set to "None") and the automatic correction for the leap-year effect was not selected (tradingdays.autoadjust set to FALSE).

tradingdays.stocktd
numeric indicating the day of the month when inventories and other stock are reported (to denote the last day of the month set the variable to 31). Modifications of this variable are taken into account only when tradingdays.option is set to "None".

tradingdays.test
defines the pre-tests for the significance of the trading day regression variables based on the AICC statistics: "Add" - the trading day variables are not included in the initial regression model but can be added to the RegARIMA model after the test; "Remove" - the trading day variables belong to the initial regression model but can be removed from the RegARIMA model after the test; "None" - the trading day variables are not pre-tested and are included in the model.

easter.enabled
logicals. If TRUE the program considers the Easter effect in the model.

easter.julian
logicals. If TRUE the program uses the Julian Easter (expressed in Gregorian calendar).

easter.duration
numeric indicating the duration of the Easter effect (length in days, between 1 and 20).

easter.test
defines the pre-tests for the significance of the Easter effect based on the t-statistic (Easter effect is considered as significant if the t-statistic is greater than 1.96): "Add" - the Easter effect variable is not included in the initial regression model but can be added to the RegARIMA model after the test; "Remove" - the Easter effect variable belong to the initial regression model but can be removed from the RegARIMA model after the test; "None" - the Easter effect variable is not pre-tested and is included in the model.

outlier.enabled
logicals. If TRUE the automatic detection of outliers is enabled in the defined time span.

The time span of the series to be searched for outliers (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: outlier.from, outlier.to, outlier.first, outlier.last, outlier.exclFirst and outlier.exclLast; where outlier.from and outlier.to have priority over remaining span control variables, outlier.last and outlier.first have priority over outlier.exclFirst and outlier.exclLast, and outlier.last has priority over outlier.first.

outlier.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with outlier.to.

outlier.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with outlier.from.

outlier.first
numeric specifying the number of periods considered at the beginning of the series.

outlier.last
numeric specifying the number of periods considered at the end of the series.

outlier.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with outlier.exclLast.

outlier.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with outlier.exclFirst.

outlier.ao
logicals. If TRUE the automatic detection of additive outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.tc
logicals. If TRUE the automatic detection of transitory changes is enabled (outlier.enabled must be also set to TRUE).

outlier.ls
logicals. If TRUE the automatic detection of level shifts is enabled (outlier.enabled must be also set to TRUE).

outlier.so
logicals. If TRUE the automatic detection of seasonal outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.usedefcv
logicals. If TRUE the critical value for the outliers' detection procedure is automatically determined by the number of observations in the outlier detection time span. If FALSE the procedure uses the inputted critical value (outlier.cv).

outlier.cv
numeric. Inputted critical value for the outliers' detection procedure. The modification of this variable is taken into account only when outlier.usedefcv is set to FALSE.

outlier.method
determines how the program successively adds detected outliers to the model. At present only the AddOne method is supported.

outlier.tcrate
numeric. The rate of decay for the transitory change outlier.

automdl.enabled
logicals. If TRUE the automatic modelling of the ARIMA model is enabled. If FALSE the parameters of the ARIMA model can be specified.

Control variables for the automatic modelling of the ARIMA model (automdl.enabled is set to TRUE):

automdl.acceptdefault
logicals. If TRUE the default model (ARIMA(0,1,1)(0,1,1)) may be chosen in the first step of the automatic model identification. If the Ljung-Box Q statistics for the residuals is acceptable, the default model is accepted and no further attempt will be made to identify any other.

automdl.cancel
numeric, cancelation limit. If the difference in moduli of an AR and an MA roots (when estimating ARIMA(1,0,1)(1,0,1) models in the second step of the automatic identification of the differencing orders) is smaller than cancelation limit, the two roots are assumed equal and cancel out.

automdl.ub1
numeric, first unit root limit. It is the threshold value for the initial unit root test in the automatic differencing procedure. When one of the roots in the estimation of the ARIMA(2,0,0)(1,0,0) plus mean model, performed in the first step of the automatic model identification procedure, is larger than first unit root limit in modulus, it is set equal to unity.

automdl.ub2
numeric, second unit root limit. When one of the roots in the estimation of the ARIMA(1,0,1)(1,0,1) plus mean model, which is performed in the second step of the automatic model identification procedure, is larger than second unit root limit in modulus, it is checked if there is a common factor in the corresponding AR and MA polynomials of the ARMA model that can be cancelled (see automdl.cancel). If there is no cancellation, the AR root is set equal to unity (i.e. the differencing order changes).

automdl.mixed
logicals. The variable controls whether ARIMA models with non-seasonal AR and MA terms or seasonal AR and MA terms will be considered in the automatic model identification procedure. If FALSE a model with AR and MA terms in both the seasonal and non-seasonal parts of the model can be acceptable, provided there are not AR and MA terms in either the seasonal or non-seasonal.

automdl.balanced
logicals. If TRUE, the automatic model identification procedure will have a preference for balanced models (i.e. models for which the order of the combined AR and differencing operator is equal to the order of the combined MA operator).

automdl.armalimit
numeric, arma limit. It is the threshold value for t-statistics of ARMA coefficients and constant term used for the final test of model parsimony. If the highest order ARMA coefficient has a t-value less than this value in magnitude, the order of the model is reduced. Also if the constant term has a t-value less than arma limit in magnitude, it is removed from the set of regressors.

automdl.reducecv
numeric, ReduceCV. The percentage by which the outlier's critical value will be reduced when an identified model is found to have a Ljung-Box statistic with an unacceptable confidence coefficient. The parameter should be between 0 and 1, and will only be active when automatic outlier identification is enabled. The reduced critical value will be set to (1-ReduceCV)xCV, where CV is the original critical value.

automdl.ljungboxlimit
numeric, Ljung Box limit. Acceptance criterion for the confidence intervals of the Ljung-Box Q statistic. If the LjungBox Q statistics for the residuals of a final model is greater than Ljung Box limit, the model is rejected, the outlier critical value is reduced, and model and outlier identification (if specified) is redone with a reduced value.

automdl.ubfinal
numeric, final unit root limit. The threshold value for the final unit root test. If the magnitude of an AR root for the final model is less than the final unit root limit, a unit root is assumed, the order of the AR polynomial is reduced by one, and the appropriate order of the differencing (non-seasonal, seasonal) is increased. The parameter value should be greater than one.

Control variables for the non-automatic modelling of the ARIMA model (automdl.enabled is set to FALSE):

arima.mu
logicals. If TRUE, the mean is considered as part of the ARIMA model.

arima.p
numeric. The order of the non-seasonal autoregressive (AR) polynomial.

arima.d
numeric. Regular differencing order.

arima.q
numeric. The order of the non-seasonal moving average (MA) polynomial.

arima.bp
numeric. The order of the seasonal autoregressive (AR) polynomial.

arima.bd
numeric. Seasonal differencing order.

arima.bq
numeric. The order of the seasonal moving average (MA) polynomial.

Control variables for the user-defined ARMA coefficients. Coefficients can be defined for the regular and seasonal autoregressive (AR) polynomials and moving average (MA) polynomials. The model considers the coefficients only if the procedure for their estimation (arima.coefType) is provided, and the number of provided coefficients matches the sum of (regular and seasonal) AR and MA orders (p,q,bp,bq).

arima.coefEnabled
logicals. If TRUE the program uses the user-defined ARMA coefficients.

arima.coef
vector providing the coefficients for the regular and seasonal AR and MA polynominals. The length of the vector must equal the sum of the regular and seasonal AR and MA orders. The coefficients shall be provided in the order: regular AR (Phi - p elements), regular MA (Theta - q elements), seasonal AR (BPhi - bp elements) and seasonal MA (BTheta - bq elements). E.g.: arima.coef=c(0.6,0.7) with arima.p=1, arima.q=0,arima.bp=1 and arima.bq=0.

arima.coefType
vector defining ARMA coefficients estimation procedure. Possible procedures are: "Undefined" - no use of user-defined input (i.e. coefficients are estimated), "Fixed" - fixes the coefficients at the value provided by the user, "Initial" - the value defined by the user is used as initial condition. For orders for which the coefficients shall not be defined, the arima.coef can be set to NA or 0 or the arima.coefType can be set to "Undefined". E.g.: arima.coef = c(-0.8,-0.6,NA), arima.coefType = c("Fixed","Fixed","Undefined").

fcst.horizon
numeric, forecasts horizon. Length of the forecasts generated by the RegARIMA model in periods (positive values) or years (negative values). By default the program generates two years forecasts (fcst.horizon set to -2).

x11.mode
character, decomposition mode. Determines the mode of the seasonal adjustment decomposition to be performed: "Undefined" - no assumption concerning the relationship between the time series components is made; "Additive" - assumes an additive relationship; "Multiplicative" - assumes a multiplicative relationship; "LogAdditive" - performs an additive decomposition of the logarithms of the series being adjusted; "PseudoAdditive" - assumes an pseudo-additive relationship. Could be changed by the program, if needed.

x11.seasonalComp
logicals. If TRUE the program computes a seasonal component. Otherwise, the seasonal component is not estimated and its values are all set to 0 (additive decomposition) or 1 (multiplicative decomposition).

x11.lsigma
numeric, lower sigma boundary for the detection of extreme values.

x11.usigma
numeric, upper sigma boundary for the detection of extreme values.

x11.trendAuto
logicals, automatic Henderson filter. If TRUE an automatic selection of the Henderson filter's length for the trend estimation is enabled.

x11.trendma
numeric, length of the Henderson filter. The user-defined length of the Henderson filter. The option is available when the automatic Henderson filter selection is disabled (x11.trendAuto=FALSE). Should be an odd number in the range (1, 101].

x11.seasonalma
vector of character(s) specifying which seasonal moving average (i.e. seasonal filter) will be used to estimate the seasonal factors for the entire series. The vector can be of length: 1 - same seasonal filters for all periods (e.g.: seasonalma=c("Msr")); or period's number - a seasonal filer is defined for each period (e.g. for quarterly series: seasonalma=c("S3X3","Msr","S3X3","Msr")). Possible filters are: "Msr", "Stable", "X11Default", "S3X1", "S3X3", "S3X5", "S3X9", "S3X15". "Msr" - the program chooses the final seasonal filter automatically.

x11.fcasts
numeric, RegARIMA forecast. Length of the forecasts generated by the RegARIMA model in periods (positive values) or years (negative values).

x11.bcasts
numeric, backcast. Length of the backcasts used in X11. Negative figures are translated in years of backcasts.

x11.calendarSigma
character to specify if the standard errors used for extreme values detection and adjustment are computed from 5 year spans of irregulars ("None", the default); separately for each calendar month/quarter ("All"); separately for each period only if Cochran’s hypothesis test determines that the irregular component is heteroskedastic by calendar month/quarter ("Signif"); separately for two complementary sets of calendar months/quarters specified by the x11.sigmaVector parameter ("Select", see parameter x11.sigmaVector).

x11.sigmaVector
vector to specifies one of the two groups of periods for whose standard errors used for extreme values detection and adjustment will be computed. Only used if x11.calendarSigma = "Select". Possibe values are: "Group1" and "Group2".

x11.excludeFcasts
logicals, exclude forecats and backcasts. If TRUE forecasts and backcasts from the RegARIMA model are not used in the generation of extreme values in the seasonal adjustment routines.

Value

A two-elements list of class c("SA_spec", "X13"): (1) object of class c("regarima_spec", "X13") with the RegARIMA model specification, (2) object of class c("X11_spec", "data.frame") with the X11 algorithm specification. Each component refers to different part of the SA model specification, mirroring the arguments of the function (for details see arguments description). Each of the lowest-level component (except span, pre-specified outliers, user-defined variables and pre-specified ARMA coefficients) is structured within a data frame with columns denoting different variables of the model specification and rows referring to: first row - base specification, as provided within the argument spec; second row - user modifications as specified by the remaining arguments of the function (e.g.: arima.d); and third row - final model specification. The final specification (third row) shall include user modifications (row two) unless they were wrongly specified. The pre-specified outliers, user-defined variables and pre-specified ARMA coefficients consist of a list with the Predefined (base model specification) and Final values.

regarima

object of class c("regarima_spec", "x13"). See Value of the function regarima_spec_x13

x11

data.frame of class c("X11_spec", "data.frame"), containing the x11 variables in line with the names of the arguments variables. The final values can be also accessed with the function s_x11.

Details

The available predefined 'JDemetra+' model specifications are described in the table below.

Identifier |	Log/level detection |	Outliers detection |	Calendar effects |	ARIMA	RSA0 |	NA |

NA |	NA |	Airline(+mean)	RSA1 |	automatic |	AO/LS/TC |	NA |

Airline(+mean)	RSA2c |	automatic |	AO/LS/TC |	2 td vars + Easter |	Airline(+mean)	RSA3 |

automatic |	AO/LS/TC |	NA |	automatic	RSA4c |	automatic |	AO/LS/TC |

2 td vars + Easter |	automatic	RSA5c |	automatic |	AO/LS/TC |	7 td vars + Easter |	automatic

References

Info on 'JDemetra+', usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en BOX G.E.P. and JENKINS G.M. (1970), "Time Series Analysis: Forecasting and Control", Holden-Day, San Francisco.

BOX G.E.P., JENKINS G.M., REINSEL G.C. and LJUNG G.M. (2015), "Time Series Analysis: Forecasting and Control", John Wiley & Sons, Hoboken, N. J., 5th edition.

See also

Examples

# \donttest{
myseries <- ipi_c_eu[, "FR"]
myspec1 <- x13_spec(spec = "RSA5c")
myreg1 <- x13(myseries, spec = myspec1)

# Modify a pre-specified model specification
myspec2 <- x13_spec(spec = "RSA5c", tradingdays.option = "WorkingDays")
myreg2 <- x13(myseries, spec = myspec2)

# Modify the model specification from a "X13" object
 myspec3 <- x13_spec(myreg1, tradingdays.option = "WorkingDays")
 myreg3 <- x13(myseries, myspec3)

# Modify the model specification from a "X13_spec" object
 myspec4 <- x13_spec(myspec1, tradingdays.option = "WorkingDays")
 myreg4 <- x13(myseries, myspec4)

# Pre-specified outliers
 myspec1 <- x13_spec(spec = "RSA5c", usrdef.outliersEnabled = TRUE,
             usrdef.outliersType = c("LS", "AO"),
             usrdef.outliersDate = c("2008-10-01", "2002-01-01"),
             usrdef.outliersCoef = c(36, 14),
             transform.function = "None")

 myreg1 <- x13(myseries, myspec1)
 myreg1

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.1057      0.114
#> Phi(2)      0.2141      0.076
#> Theta(1)   -0.4709      0.111
#> BTheta(1)  -0.6642      0.043
#> 
#>               Estimate Std. Error
#> Monday         0.58039      0.234
#> Tuesday        0.82241      0.235
#> Wednesday      1.10804      0.234
#> Thursday       0.09871      0.232
#> Friday         0.89917      0.233
#> Saturday      -1.62006      0.231
#> Leap year      2.14549      0.731
#> Easter [1]    -2.34335      0.456
#> AO (9-2008)   32.07187      2.931
#> LS (9-2008)  -57.03600      2.631
#> AO (5-2011)   13.22101      1.841
#> TC (9-2008)   23.33640      3.967
#> TC (12-2001) -20.56517      2.911
#> AO (12-2001)  17.24940      2.981
#> TC (2-2002)   10.70687      1.939
#> 
#> Fixed outliers: 
#>              Coefficients
#> LS (10-2008)           36
#> AO (1-2002)            14
#> 
#> 
#> Residual standard error: 2.185 on 327 degrees of freedom
#> Log likelihood = -767.3, aic =  1575 aicc =  1577, bic(corrected for length) = 1.883
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.141
#> M(2)    0.079
#> M(3)    1.235
#> M(4)    0.421
#> M(5)    1.083
#> M(6)    0.025
#> M(7)    0.084
#> M(8)    0.236
#> M(9)    0.047
#> M(10)   0.255
#> M(11)   0.240
#> Q       0.356
#> Q-M2    0.391
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t           s            i
#> Jan 2019 103.9 105.1283 105.1200  -1.2282924  0.008279642
#> Feb 2019 101.9 106.0190 105.4062  -4.1190475  0.612828917
#> Mar 2019 111.0 104.5647 105.5846   6.4353294 -1.019956247
#> Apr 2019 107.4 106.1904 105.6134   1.2095984  0.576997697
#> May 2019 105.5 106.0389 105.5526  -0.5388967  0.486297209
#> Jun 2019 105.8 100.9035 105.3689   4.8965398 -4.465403016
#> Jul 2019 110.1 105.4033 105.0100   4.6966957  0.393271648
#> Aug 2019  78.7 103.3862 104.5357 -24.6862487 -1.149431050
#> Sep 2019 108.5 104.5228 103.9942   3.9771575  0.528676270
#> Oct 2019 116.8 105.0244 103.4956  11.7756247  1.528791350
#> Nov 2019 103.8 101.7248 103.1841   2.0751974 -1.459264138
#> Dec 2019  97.7 101.9645 103.0531  -4.2645362 -1.088545538
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f        i_f
#> Jan 2020 102.21622 103.4251 103.0441  -1.2088796  0.3809940
#> Feb 2020  99.95203 103.4974 103.0612  -3.5453586  0.4361614
#> Mar 2020 112.07522 103.4829 103.0151   8.5923235  0.4677573
#> Apr 2020 104.38199 103.4163 102.9213   0.9656826  0.4950351
#> May 2020  96.83109 102.1021 102.7618  -5.2709823 -0.6597729
#> Jun 2020 111.87859 102.0150 102.5999   9.8635972 -0.5849235
#> Jul 2020 106.90231 102.4777 102.5807   4.4246394 -0.1030684
#> Aug 2020  76.02869 103.0997 102.7007 -27.0709810  0.3989725
#> Sep 2020 110.58476 103.4537 102.8614   7.1310799  0.5922947
#> Oct 2020 111.74768 102.4560 102.9954   9.2916534 -0.5393773
#> Nov 2020 104.38299 102.8633 103.1082   1.5196712 -0.2449275
#> Dec 2020 101.41574 103.4436 103.1822  -2.0278301  0.2613430
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.783
#>  Seasonal     46.184
#>  Irregular     0.810
#>  TD & Hol.     2.086
#>  Others       50.912
#>  Total       101.775
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.022
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          0.833
#>  qs test on i                           0.622
#>  f-test on sa (seasonal dummies)        0.948
#>  f-test on i (seasonal dummies)         0.875
#>  Residual seasonality (entire series)   0.807
#>  Residual seasonality (last 3 years)    0.856
#>  f-test on sa (td)                      0.974
#>  f-test on i (td)                       0.981
#> 
#> 
#> Additional output variables

 s_preOut(myreg1)

#>   type       date coeff
#> 1   LS 2008-10-01    36
#> 2   AO 2002-01-01    14

# User-defined calendar regressors
 var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
 var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
 var <- ts.union(var1, var2)
 myspec1 <- x13_spec(spec = "RSA5c", tradingdays.option = "UserDefined",
                     usrdef.varEnabled = TRUE,
                     usrdef.var = var,
                     usrdef.varType = c("Calendar", "Calendar"))

#> Warning: With tradingdays.option = "UserDefined", the parameters tradingdays.autoadjust, tradingdays.leapyear and tradingdays.stocktd are ignored.

 myreg1 <- x13(myseries, myspec1)
 myreg1

#> RegARIMA
#> y = regression model + arima (3, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.33648      0.378
#> Phi(2)     0.06353      0.378
#> Phi(3)    -0.41840      0.271
#> Theta(1)  -0.67609      0.364
#> BTheta(1) -0.73379      0.040
#> 
#>              Estimate Std. Error
#> Easter [8]   -0.02456      0.006
#> LS (11-2008) -0.09205      0.017
#> AO (5-2011)   0.08996      0.021
#> LS (1-2009)  -0.07635      0.017
#> 
#> 
#> Residual standard error: 0.02698 on 337 degrees of freedom
#> Log likelihood = 755.8, aic =  1714 aicc =  1715, bic(corrected for length) = -7.074
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.298
#> M(2)    0.276
#> M(3)    3.000
#> M(4)    0.454
#> M(5)    2.096
#> M(6)    0.558
#> M(7)    0.113
#> M(8)    0.273
#> M(9)    0.089
#> M(10)   0.396
#> M(11)   0.379
#> Q       0.759
#> Q-M2    0.819
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  23-Henderson
#> 
#> 
#> Final
#> Last observed values
#>              y        sa        t         s         i
#> Jan 2019 103.9 105.95263 105.9423 0.9806269 1.0000972
#> Feb 2019 101.9 106.72619 106.1401 0.9547797 1.0055216
#> Mar 2019 111.0 100.93443 106.2326 1.0997238 0.9501269
#> Apr 2019 107.4 108.35947 106.2125 0.9911454 1.0202144
#> May 2019 105.5 108.32879 106.0760 0.9738870 1.0212371
#> Jun 2019 105.8  98.10432 105.8235 1.0784439 0.9270558
#> Jul 2019 110.1 107.47489 105.4650 1.0244254 1.0190574
#> Aug 2019  78.7 101.71055 105.0329 0.7737644 0.9683683
#> Sep 2019 108.5 104.12987 104.5740 1.0419680 0.9957531
#> Oct 2019 116.8 105.84574 104.1333 1.1034927 1.0164450
#> Nov 2019 103.8 100.84420 103.7430 1.0293105 0.9720574
#> Dec 2019  97.7 103.23642 103.4390 0.9463714 0.9980419
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 103.19553 104.8378 103.2285 0.9843354 1.0155894
#> Feb 2020  97.97326 102.7994 103.0967 0.9530532 0.9971160
#> Mar 2020 112.53987 102.3923 103.0368 1.0991044 0.9937453
#> Apr 2020 104.35876 105.1246 103.0187 0.9927153 1.0204412
#> May 2020  99.40669 102.1157 103.0034 0.9734708 0.9913823
#> Jun 2020 110.49273 102.4854 103.0048 1.0781315 0.9949574
#> Jul 2020 105.57679 102.7670 103.0011 1.0273415 0.9977271
#> Aug 2020  78.93925 102.1948 103.0550 0.7724389 0.9916528
#> Sep 2020 108.57592 104.3315 103.1229 1.0406824 1.0117201
#> Oct 2020 113.30516 102.4371 103.2548 1.1060945 0.9920810
#> Nov 2020 105.67402 102.8250 103.2991 1.0277075 0.9954104
#> Dec 2020  99.22963 105.1491 103.3814 0.9437036 1.0170996
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.881
#>  Seasonal     59.583
#>  Irregular     2.746
#>  TD & Hol.     0.106
#>  Others       36.375
#>  Total       100.691
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.553
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          0.000
#>  qs test on i                           0.001
#>  f-test on sa (seasonal dummies)        0.862
#>  f-test on i (seasonal dummies)         0.809
#>  Residual seasonality (entire series)   0.733
#>  Residual seasonality (last 3 years)    0.831
#>  f-test on sa (td)                      0.000
#>  f-test on i (td)                       0.000
#> 
#> 
#> Additional output variables

 myspec2 <- x13_spec(spec = "RSA5c", usrdef.varEnabled = TRUE,
             usrdef.var = var1, usrdef.varCoef = 2,
             transform.function = "None")
 myreg2 <- x13(myseries, myspec2)
 s_preVar(myreg2)

#> $series
#>               Jan          Feb          Mar          Apr          May
#> 1990  -7.37468113  -1.25015343   2.79824147  -4.47372026   0.47957680
#> 1991   3.26711666  -9.63575111   4.78657978   7.80564917  -1.85425499
#> 1992  -4.19990291  -5.14581743  -4.81189141  -0.22896117  -2.35817375
#> 1993  -1.83718896   5.64549112  -6.45663305 -19.84197517  -9.32810794
#> 1994  12.35461267 -10.96546214  -6.78783706  -5.59867238   2.15476356
#> 1995   0.35119899 -11.20976025  -3.31965872 -12.77255127   5.38651328
#> 1996   2.25022902  -6.88411870  -6.48787614   4.71731346   1.74606248
#> 1997  11.90267693   1.91977818 -18.27600531  -8.40076820  16.23393990
#> 1998  -7.80711972  -6.82293210 -22.43426371   1.80276393   7.81506635
#> 1999  10.15744026  -3.95593431  21.59511068  22.15581578  -2.55717069
#> 2000  -5.91062251   8.56944610   5.08893390   3.97103898   3.86757251
#> 2001  -5.03572195 -17.13177445   7.52451631  -8.83402109   1.36518779
#> 2002  -4.93380941 -12.28615839  12.84563150  12.11780172   0.17867075
#> 2003  17.91430812  -8.01034578  -5.57933180   7.56117489  -0.65654280
#> 2004   6.99530402  -4.78932287 -13.14202741  -9.03645819  -6.82677769
#> 2005   6.85714751 -15.74190181 -10.39202637  10.62892692 -12.79380891
#> 2006   6.55348285   2.08748369   4.79304525  11.19721889   3.58970059
#> 2007 -16.27036866  10.95973351 -11.30615770   7.49407597   5.07040903
#> 2008 -16.41368825   1.85576753   8.08139505   4.62741902  -1.40862948
#> 2009  21.10958544   1.06319181   2.88638282  -8.28129022   2.59246188
#> 2010   7.49492399  13.68161796   8.09041925  10.52043916  15.76710173
#> 2011  14.61477127  -3.93734253   5.88577012 -10.59461309   7.02083603
#> 2012  -4.70589637  26.18843475  17.81970431   3.11227898   6.04652922
#> 2013   6.97447152   8.50194263  -2.82469962  10.56820511  -1.64563751
#> 2014   6.45915651   2.86538618 -19.55302496  -0.70122520 -19.93476957
#> 2015  -4.71600217  -3.86758443  -0.32921492 -15.89925204 -16.31630380
#> 2016 -15.66243530  -8.27262343  10.43432242  -5.61818713   4.58990225
#> 2017 -13.44497531  -0.07853537  -7.09360740  14.19910316  21.00753869
#> 2018  16.38748991   7.14250476   0.21730272  -7.10864161   0.67650247
#> 2019  19.42670865   4.54756408  -4.35878269  16.42250741  -1.19391302
#>               Jun          Jul          Aug          Sep          Oct
#> 1990  10.92852535   5.08009676  -8.69910799  -7.02676476 -13.59407270
#> 1991  -3.71415762  -3.75906415 -10.49773092   1.35509887 -10.41934249
#> 1992  -0.42915889  -5.33472185   2.40923657   2.75787999   9.87337885
#> 1993  18.64559750   3.07668334 -18.73476807 -13.81586373  12.14281005
#> 1994 -20.55435201  -5.20408628   2.52263906   1.53156424   4.92124424
#> 1995  13.48841284   6.01111502  -0.12719586   7.16138996   4.05634420
#> 1996  -3.93274869   1.23240578 -20.06667650  12.09783540   1.40372442
#> 1997   1.60041454  16.00151269   1.14274708 -13.65424549   4.38538734
#> 1998  -9.57231341   0.91407055  -4.15741978  -6.61423552   6.83971635
#> 1999   3.02291700  -6.65805050  -6.96061207   6.83341934  -5.89274870
#> 2000   3.70136185  10.36723098  23.58765058  -9.13602075  12.75599423
#> 2001   4.26281885  -9.25139751  -9.69095736  -2.26582778  -9.78223964
#> 2002  17.00247776  13.44115353   2.47249503  -0.95850995  -1.05894771
#> 2003  14.03825101   8.06123109  -1.25334438  -5.64422544   9.40777992
#> 2004   3.74599658   8.09784422 -10.27311894   2.71204716   3.72163386
#> 2005 -17.65259960 -23.94684943   6.61851420   2.79202139 -19.75798133
#> 2006  11.98583489   5.63019465   3.60364311   8.25363990  -0.40392584
#> 2007   1.49227768   7.56914894  -4.27224530 -13.47831121   0.82595585
#> 2008 -18.30167300   4.04087877  -0.13750674  -1.02241238   2.02454445
#> 2009  -4.95230738  -2.89501667   6.40683953  -1.41462861  -8.79847386
#> 2010  11.52029911  -5.17742857  -6.35012824  -4.00123911   5.93241461
#> 2011  -2.54390092  -3.80451079   7.22051799  13.27842786  -1.25492499
#> 2012   8.02925361  -7.42182079 -10.86486271  15.85888347  -1.40704169
#> 2013  -5.05197318  -4.02419900  22.95251358  -0.58040167   1.22377660
#> 2014   1.21132684   4.35595856   1.38904324   9.82497501  -2.86391397
#> 2015  -6.19216932  -4.28563857  21.43623439  13.81055332  -9.81823126
#> 2016   4.61175487   5.88209678  13.65187644   2.37584966 -13.01932383
#> 2017   0.56429659   8.05545122  -1.92729984  -1.20578134  -2.41185306
#> 2018  18.67312976   7.39326267   9.65687441   4.69334931   7.32928056
#> 2019   0.03403357  -2.87662480  10.04914840 -15.19681487  10.80588866
#>               Nov          Dec
#> 1990   3.45646991   4.21805481
#> 1991  13.71543461   2.11723020
#> 1992  -6.07175018  -2.43155234
#> 1993   1.41481904   0.63103851
#> 1994   5.33822382  -2.07132362
#> 1995  -3.54852294  -8.31256082
#> 1996  -8.75827874   9.62364735
#> 1997  -9.61507272  12.16348887
#> 1998   8.92384341  -7.26907567
#> 1999  -0.74093549 -11.12271027
#> 2000 -14.39090092   0.38142207
#> 2001  -5.33354128   8.66392934
#> 2002 -12.61154714 -17.60210399
#> 2003  12.69808946  12.65038490
#> 2004 -14.73803514 -14.22799149
#> 2005   4.96506933  13.53031635
#> 2006  -6.05456736  -4.78648838
#> 2007  11.53728116  25.86841900
#> 2008   8.03320548  -5.87793245
#> 2009  -5.23639611  -6.13436596
#> 2010  -0.01094905 -18.90227763
#> 2011  11.39794704  -7.84406639
#> 2012  -9.75276634  -0.91846724
#> 2013  -2.00777681   4.60803569
#> 2014  16.94152064   5.21310799
#> 2015  13.12917276   3.72374042
#> 2016  17.37977175  14.64151538
#> 2017  -2.55729561  -2.08778684
#> 2018  13.33797734  13.30182641
#> 2019   4.35371426 -13.89147408
#> 
#> $description
#>              type coeff
#> userdef Undefined     2
#> 

# Pre-specified ARMA coefficients
 myspec1 <- x13_spec(spec = "RSA5c", automdl.enabled = FALSE,
             arima.p = 1, arima.q = 1, arima.bp = 0, arima.bq = 1,
             arima.coefEnabled = TRUE,
             arima.coef = c(-0.8, -0.6, 0),
             arima.coefType = c(rep("Fixed", 2), "Undefined"))

 s_arimaCoef(myspec1)

#>                Type Value
#> Phi(1)        Fixed  -0.8
#> Theta(1)      Fixed  -0.6
#> BTheta(1) Undefined   0.0

 myreg1 <- x13(myseries, myspec1)
 myreg1

#> RegARIMA
#> y = regression model + arima (1, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     -0.8000      0.000
#> Theta(1)   -0.6000      0.000
#> BTheta(1)  -0.6933      0.041
#> 
#>              Estimate Std. Error
#> Monday       0.005835      0.002
#> Tuesday      0.007964      0.002
#> Wednesday    0.010841      0.002
#> Thursday     0.001883      0.002
#> Friday       0.009833      0.002
#> Saturday    -0.018229      0.002
#> Easter [1]  -0.020470      0.004
#> AO (5-2011)  0.144191      0.016
#> 
#> 
#> Residual standard error: 0.02585 on 337 degrees of freedom
#> Log likelihood =   772, aic =  1681 aicc =  1682, bic(corrected for length) = -7.159
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.104
#> M(2)    0.050
#> M(3)    0.820
#> M(4)    0.810
#> M(5)    0.774
#> M(6)    0.144
#> M(7)    0.075
#> M(8)    0.213
#> M(9)    0.060
#> M(10)   0.281
#> M(11)   0.274
#> Q       0.317
#> Q-M2    0.350
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.4392 105.1420 0.9854022 1.0028269
#> Feb 2019 101.9 106.5484 105.5999 0.9563728 1.0089822
#> Mar 2019 111.0 104.9544 105.8893 1.0576025 0.9911709
#> Apr 2019 107.4 106.1896 105.9114 1.0113981 1.0026268
#> May 2019 105.5 106.3579 105.7434 0.9919334 1.0058112
#> Jun 2019 105.8 101.4914 105.4084 1.0424527 0.9628395
#> Jul 2019 110.1 105.3528 104.9476 1.0450604 1.0038607
#> Aug 2019  78.7 102.9607 104.4123 0.7643693 0.9860971
#> Sep 2019 108.5 104.2297 103.7658 1.0409699 1.0044713
#> Oct 2019 116.8 104.6445 103.0383 1.1161596 1.0155886
#> Nov 2019 103.8 101.4319 102.3438 1.0233467 0.9910901
#> Dec 2019  97.7 101.3468 101.6903 0.9640167 0.9966220
#> 
#> Forecasts:
#>                y_f      sa_f       t_f       s_f       i_f
#> Jan 2020  99.39834 100.81133 101.08186 0.9859837 0.9973237
#> Feb 2020  97.82021 100.50616 100.52964 0.9732758 0.9997664
#> Mar 2020 108.74299 100.55838 100.01880 1.0813916 1.0053948
#> Apr 2020 101.08116 100.17315  99.58960 1.0090645 1.0058596
#> May 2020  93.07749  98.52290  99.19467 0.9447295 0.9932278
#> Jun 2020 108.00847  98.50199  98.83464 1.0965105 0.9966343
#> Jul 2020 102.83686  98.50394  98.60286 1.0439872 0.9989968
#> Aug 2020  73.50899  98.60445  98.48637 0.7454936 1.0011989
#> Sep 2020 106.31917  99.06688  98.43519 1.0732059 1.0064173
#> Oct 2020 106.67440  98.00056  98.38190 1.0885080 0.9961239
#> Nov 2020  99.86793  97.97595  98.33449 1.0193106 0.9963538
#> Dec 2020  97.16631  98.52982  98.27442 0.9861614 1.0025989
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         5.744
#>  Seasonal     89.850
#>  Irregular     1.230
#>  TD & Hol.     3.698
#>  Others        0.489
#>  Total       101.012
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.272
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          0.992
#>  qs test on i                           0.858
#>  f-test on sa (seasonal dummies)        0.988
#>  f-test on i (seasonal dummies)         0.970
#>  Residual seasonality (entire series)   0.984
#>  Residual seasonality (last 3 years)    0.952
#>  f-test on sa (td)                      0.989
#>  f-test on i (td)                       0.946
#> 
#> 
#> Additional output variables

# Defined seasonal filters
 myspec1 <- x13_spec("RSA5c", x11.seasonalma = rep("S3X1", 12))
 mysa1 <- x13(myseries, myspec1)
# }