Compute a workspace multi-processing(s)

Function to compute all the multi-processings or only a given one from a workspace. By default, the workspace only contains definitions: computation is needed to recalculate and access the adjusted model (with get_model).

compute(workspace, i)

Arguments

workspace
the workspace to compute.

i
a character or numeric indicating the name or the index of the multiprocessing to compute. By default, all multi-processings are computed.

See also

Examples

# \donttest{
spec_x13 <- x13_spec(spec = "RSA5c", easter.enabled = FALSE)
sa_x13 <- x13(ipi_c_eu[, "FR"], spec = spec_x13)

wk <- new_workspace()
mp <- new_multiprocessing(wk, "sa1")
add_sa_item(wk, "sa1", sa_x13, "X13")
sa_item1 <- get_object(mp, 1)

get_model(sa_item1, wk) # Returns NULL
#> Warning: The result of the object is NULL: have you computed the workspace after importing it?
#> See ?compute for more information.
#> NULL

compute(wk)

get_model(sa_item1, wk) # Returns the SA model sa_x13
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.03368      0.108
#> Phi(2)     0.17542      0.075
#> Theta(1)  -0.53601      0.102
#> BTheta(1) -0.67991      0.042
#> 
#>                Estimate Std. Error
#> Monday         0.547498      0.233
#> Tuesday        0.915020      0.232
#> Wednesday      1.073764      0.234
#> Thursday       0.003317      0.233
#> Friday         0.866778      0.235
#> Saturday      -1.535790      0.232
#> Leap year      2.177766      0.715
#> Easter [8]    -2.276133      0.467
#> TC (4-2020)  -20.782385      2.187
#> TC (3-2020)  -20.996390      2.185
#> AO (5-2011)   13.229655      1.879
#> LS (11-2008) -12.597511      1.632
#> 
#> 
#> Residual standard error: 2.225 on 342 degrees of freedom
#> Log likelihood = -800.5, aic =  1635 aicc =  1637, bic(corrected for length) = 1.862
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.172
#> M(2)    0.098
#> M(3)    1.248
#> M(4)    0.367
#> M(5)    1.095
#> M(6)    0.074
#> M(7)    0.083
#> M(8)    0.242
#> M(9)    0.062
#> M(10)   0.259
#> M(11)   0.246
#> Q       0.366
#> Q-M2    0.400
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y        sa        t            s            i
#> Jan 2020 101.0 102.99160 102.9294  -1.99160498   0.06217355
#> Feb 2020 100.1 103.45426 102.9247  -3.35426344   0.52959521
#> Mar 2020  91.8  82.65612 103.1192   9.14388018 -20.46308279
#> Apr 2020  66.7  66.67651 103.5232   0.02349271 -36.84671169
#> May 2020  73.7  78.87377 103.9445  -5.17376662 -25.07074264
#> Jun 2020  98.2  87.34056 104.2875  10.85944310 -16.94698952
#> Jul 2020  97.4  92.53355 104.4507   4.86645017 -11.91712826
#> Aug 2020  71.7  97.50941 104.3140 -25.80940727  -6.80457315
#> Sep 2020 104.7  97.33065 103.8836   7.36935330  -6.55299506
#> Oct 2020 106.7  98.43014 103.2962   8.26986026  -4.86603247
#> Nov 2020 101.6 100.25464 102.7833   1.34535858  -2.52865083
#> Dec 2020  96.6  99.64539 102.5212  -3.04538644  -2.87581385
#> 
#> Forecasts:
#>                y_f     sa_f      t_f        s_f        i_f
#> Jan 2021  94.58042 101.1230 102.4688  -6.542572 -1.3458176
#> Feb 2021  97.88130 101.7320 102.4840  -3.850691 -0.7520506
#> Mar 2021 112.76674 102.1569 102.4403  10.609843 -0.2833876
#> Apr 2021 103.42922 102.1620 102.2951   1.267216 -0.1330803
#> May 2021  95.97032 101.4546 102.1330  -5.484286 -0.6783867
#> Jun 2021 112.78593 101.3141 102.0099  11.471876 -0.6958026
#> Jul 2021 104.00357 101.5966 101.9915   2.406945 -0.3948747
#> Aug 2021  79.06777 102.3509 102.1229 -23.283168  0.2280395
#> Sep 2021 109.03147 102.4766 102.3299   6.554824  0.1467741
#> Oct 2021 108.59415 102.2487 102.5371   6.345501 -0.2884769
#> Nov 2021 106.48447 102.5342 102.7285   3.950244 -0.1943016
#> Dec 2021  99.79090 103.1269 102.8914  -3.336042  0.2355813
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.174
#>  Seasonal     59.871
#>  Irregular     1.096
#>  TD & Hol.     2.568
#>  Others       33.844
#>  Total        99.553
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.044
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           0.994
#>  f-test on sa (seasonal dummies)        0.990
#>  f-test on i (seasonal dummies)         0.968
#>  Residual seasonality (entire series)   0.976
#>  Residual seasonality (last 3 years)    0.932
#>  f-test on sa (td)                      0.978
#>  f-test on i (td)                       0.997
#> 
#> 
#> Additional output variables
# }