Compute the multi-processing from a workspace

Function to compute all the multi-processings or a given one from a workspace. By default the workspace only contains definitions: computation is needed to get the seasonal adjustment model (with get_model).

compute(workspace, i)

Arguments

workspace
the workspace to compute.

i
a character or numeric indicating the name or the index of the multiprocessing to compute. By default all the multi-processings are compute.

See also

Examples

# \donttest{
spec_x13 <- x13_spec(spec = "RSA5c", easter.enabled = FALSE)
sa_x13 <- x13(ipi_c_eu[, "FR"], spec = spec_x13)

wk <- new_workspace()
mp <- new_multiprocessing(wk, "sa1")
add_sa_item(wk, "sa1", sa_x13, "X13")
sa_item1 <- get_object(mp, 1)

get_model(sa_item1, wk) # Returns NULL

#> Warning: The result of the object is NULL: have you compute the workspace importing?
#> See ?compute for more information

#> NULL

compute(wk)

get_model(sa_item1, wk) # Returns the SA model sa_x13

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>            Estimate Std. Error
#> Phi(1)     0.008774      0.119
#> Phi(2)     0.098387      0.081
#> Theta(1)  -0.562495      0.110
#> BTheta(1) -0.738153      0.039
#> 
#>                Estimate Std. Error
#> Monday        0.0058623      0.002
#> Tuesday       0.0091349      0.002
#> Wednesday     0.0109267      0.002
#> Thursday      0.0005156      0.002
#> Friday        0.0091500      0.002
#> Saturday     -0.0161713      0.002
#> AO (5-2011)   0.1278938      0.018
#> LS (11-2008) -0.1157755      0.016
#> 
#> 
#> Residual standard error: 0.02153 on 334 degrees of freedom
#> Log likelihood = 834.5, aic =  1562 aicc =  1564, bic(corrected for length) = -7.474
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.094
#> M(2)    0.049
#> M(3)    0.978
#> M(4)    0.421
#> M(5)    0.969
#> M(6)    0.180
#> M(7)    0.075
#> M(8)    0.211
#> M(9)    0.063
#> M(10)   0.250
#> M(11)   0.232
#> Q       0.323
#> Q-M2    0.357
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.3314 105.1139 0.9864104 1.0020689
#> Feb 2019 101.9 106.3742 105.3665 0.9579387 1.0095639
#> Mar 2019 111.0 104.5953 105.4822 1.0612327 0.9915923
#> Apr 2019 107.4 105.0934 105.4192 1.0219484 0.9969091
#> May 2019 105.5 106.8159 105.2936 0.9876808 1.0144578
#> Jun 2019 105.8 101.4580 105.1024 1.0427961 0.9653249
#> Jul 2019 110.1 105.3379 104.7923 1.0452083 1.0052062
#> Aug 2019  78.7 103.0406 104.3781 0.7637769 0.9871852
#> Sep 2019 108.5 104.4959 103.8759 1.0383182 1.0059689
#> Oct 2019 116.8 104.9663 103.3582 1.1127377 1.0155587
#> Nov 2019 103.8 101.6218 102.9772 1.0214349 0.9868375
#> Dec 2019  97.7 101.8080 102.7604 0.9596494 0.9907318
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.59660 103.0784 102.6789 0.9856249 1.0038905
#> Feb 2020 100.87436 103.2029 102.7042 0.9774377 1.0048553
#> Mar 2020 111.27482 102.6228 102.7383 1.0843091 0.9988753
#> Apr 2020 105.12395 103.3465 102.7663 1.0171986 1.0056464
#> May 2020  96.44225 102.2715 102.7176 0.9430021 0.9956572
#> Jun 2020 112.46775 102.4907 102.6115 1.0973459 0.9988229
#> Jul 2020 106.58633 102.4070 102.5858 1.0408106 0.9982572
#> Aug 2020  76.49372 102.5838 102.6507 0.7456707 0.9993479
#> Sep 2020 111.06093 103.4916 102.7527 1.0731400 1.0071907
#> Oct 2020 111.06175 102.3655 102.8697 1.0849527 0.9950991
#> Nov 2020 104.14664 102.6519 102.9955 1.0145617 0.9966639
#> Dec 2020 101.63644 103.5595 103.0930 0.9814307 1.0045245
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.426
#>  Seasonal     72.514
#>  Irregular     1.136
#>  TD & Hol.     2.703
#>  Others       21.274
#>  Total       100.053
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.498
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.964
#>  f-test on i (seasonal dummies)         0.936
#>  Residual seasonality (entire series)   0.970
#>  Residual seasonality (last 3 years)    0.994
#>  f-test on sa (td)                      0.996
#>  f-test on i (td)                       0.993
#> 
#> 
#> Additional output variables

# }