Compute the multi-processing from a workspace

Function to compute all the multi-processings or a given one from a workspace. By default the workspace only contains definitions: computation is needed to get the seasonal adjustment model (with get_model).

compute(workspace, i)

Arguments

workspace
the workspace to compute.

i
a character or numeric indicating the name or the index of the multiprocessing to compute. By default all the multi-processings are compute.

See also

Examples

spec_x13 <- x13_spec_def(spec = "RSA5c", easter.enabled = FALSE)
sa_x13 <- x13(ipi_c_eu[, "FR"], spec = spec_x13)

wk <- new_workspace()
mp <- new_multiprocessing(wk, "sa1")
add_sa_item(wk, "sa1", sa_x13, "X13")
sa_item1 <- get_object(mp, 1)

get_model(sa_item1, wk) # Returns NULL

#> Warning: The result of the object is NULL: have you compute the workspace importing?
#> See ?compute for more information

#> NULL

compute(wk)

get_model(sa_item1, wk) # Returns the SA model sa_x13

#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5393      0.048
#> BTheta(1)  -0.5076      0.050
#> 
#>              Estimate Std. Error
#> Monday       -0.07677      0.168
#> Tuesday       0.06278      0.167
#> Wednesday     0.20386      0.168
#> Thursday     -0.56900      0.168
#> Friday        0.48643      0.168
#> Saturday      0.16823      0.168
#> Leap year    -0.04202      0.516
#> LS (11-2008) -8.47862      1.328
#> LS (1-2009)  -7.24743      1.328
#> LS (5-2008)  -5.17890      1.271
#> 
#> 
#> Residual standard error: 1.692 on 323 degrees of freedom
#> Log likelihood = -630.1, aic =  1286 aicc =  1287, bic(corrected for length) = 1.266
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.070
#> M(2)    0.044
#> M(3)    0.908
#> M(4)    0.621
#> M(5)    0.702
#> M(6)    0.225
#> M(7)    0.075
#> M(8)    0.212
#> M(9)    0.057
#> M(10)   0.186
#> M(11)   0.160
#> Q       0.298
#> Q-M2    0.330
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t           s          i
#> Jan 2017  97.4 100.8343 100.7013  -3.4343306  0.1330766
#> Feb 2017  97.5 100.3650 101.0249  -2.8649615 -0.6599802
#> Mar 2017 112.0 102.2535 101.4647   9.7464542  0.7888578
#> Apr 2017 103.0 101.0035 101.9216   1.9964916 -0.9181391
#> May 2017 100.4 103.0243 102.3451  -2.6243031  0.6791988
#> Jun 2017 111.2 102.5088 102.7541   8.6911964 -0.2453078
#> Jul 2017 103.4 103.6337 103.1399  -0.2336808  0.4937705
#> Aug 2017  79.3 103.2251 103.5264 -23.9251325 -0.3013005
#> Sep 2017 109.7 103.2506 103.9215   6.4494326 -0.6709239
#> Oct 2017 114.0 107.0447 104.3063   6.9552773  2.7383859
#> Nov 2017 107.7 105.5642 104.6206   2.1358354  0.9435229
#> Dec 2017 101.4 104.5036 104.8670  -3.1036149 -0.3633398
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f           i_f
#> Jan 2018 102.12660 104.9825 105.0158  -2.8558846 -0.0333189370
#> Feb 2018 102.18467 105.1651 105.1322  -2.9804296  0.0328501611
#> Mar 2018 114.75482 104.9927 105.2650   9.7620667 -0.2722128214
#> Apr 2018 107.12347 105.4543 105.4551   1.6692056 -0.0008294846
#> May 2018 102.76596 105.8226 105.7251  -3.0566117  0.0974629726
#> Jun 2018 115.36179 106.0357 106.0400   9.3261226 -0.0043589056
#> Jul 2018 106.15595 106.5844 106.3555  -0.4284186  0.2288559841
#> Aug 2018  82.90307 106.3248 106.6641 -23.4217780 -0.3392719434
#> Sep 2018 112.28236 106.6926 106.9366   5.5897789 -0.2440563282
#> Oct 2018 115.41162 107.9129 107.1741   7.4986863  0.7388293101
#> Nov 2018 109.85751 107.2445 107.3673   2.6130640 -0.1228820216
#> Dec 2018 103.52884 107.3040 107.5048  -3.7751434 -0.2008278423
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.648
#>  Seasonal     49.383
#>  Irregular     0.419
#>  TD & Hol.     0.039
#>  Others       50.184
#>  Total       101.672
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.011
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.970
#>  f-test on i (seasonal dummies)         0.887
#>  Residual seasonality (entire series)   0.967
#>  Residual seasonality (last 3 years)    0.931
#>  f-test on sa (td)                      0.974
#>  f-test on i (td)                       0.999
#> 
#> 
#> Additional output variables