Get the seasonally adjusted model from a workspace

Generics functions to get seasonally adjusted model(s) from workspace, multiprocessing or sa_item object. get_model returns a "SA" objects while get_jmodel returns the Java objects of the models.

get_jmodel(x, workspace, userdefined = NULL, progress_bar = TRUE)

get_model(x, workspace, userdefined = NULL, progress_bar = TRUE)

Arguments

x
the object to get the seasonally adjusted model.

workspace
the workspace object where models are stored. If x is a workspace object this parameter is not used.

userdefined
vector with characters for additional output variables. (see x13 or tramoseats).

progress_bar
boolean: if TRUE a progress bar is printed.

Value

get_model() returns a seasonally adjust object (class c("SA", "X13") or c("SA", "TRAMO_SEATS") or list of seasonally adjust objects:

if x is a sa_item object, get_model(x) returns a "SA" object (or a jSA object with get_jmodel(x));

if x is a multiprocessing object, get_ts(x) returns list of length the number of sa_items, each element containing a "SA" object (or a jSA object with get_jmodel(x));

if x is a workspace object, get_ts(x) returns list of length the number of multiprocessing, each element containing a list of a "SA" object (or a jSA object with get_jmodel(x)).

See also

Examples

# \donttest{
spec_x13 <- x13_spec(spec = "RSA5c", easter.enabled = FALSE)
sa_x13 <- x13(ipi_c_eu[, "FR"], spec = spec_x13)
spec_ts <- tramoseats_spec(spec = "RSA5")
sa_ts <- tramoseats(ipi_c_eu[, "FR"], spec = spec_ts)

wk <- new_workspace()
mp <- new_multiprocessing(wk, "sa1")
add_sa_item(wk, "sa1", sa_x13, "X13")
add_sa_item(wk, "sa1", sa_ts, "TramoSeats")

compute(wk) # It's important to compute the workspace to get the SA model
sa_item1 <- get_object(mp, 1)

get_model(sa_item1, wk) # Extract the model of the sa_item1: its the object sa_x13

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.03368      0.108
#> Phi(2)     0.17542      0.075
#> Theta(1)  -0.53601      0.102
#> BTheta(1) -0.67991      0.042
#> 
#>                Estimate Std. Error
#> Monday         0.547498      0.233
#> Tuesday        0.915020      0.232
#> Wednesday      1.073764      0.234
#> Thursday       0.003317      0.233
#> Friday         0.866778      0.235
#> Saturday      -1.535790      0.232
#> Leap year      2.177766      0.715
#> Easter [8]    -2.276133      0.467
#> TC (4-2020)  -20.782385      2.187
#> TC (3-2020)  -20.996390      2.185
#> AO (5-2011)   13.229655      1.879
#> LS (11-2008) -12.597511      1.632
#> 
#> 
#> Residual standard error: 2.225 on 342 degrees of freedom
#> Log likelihood = -800.5, aic =  1635 aicc =  1637, bic(corrected for length) = 1.862
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.172
#> M(2)    0.098
#> M(3)    1.248
#> M(4)    0.367
#> M(5)    1.095
#> M(6)    0.074
#> M(7)    0.083
#> M(8)    0.242
#> M(9)    0.062
#> M(10)   0.259
#> M(11)   0.246
#> Q       0.366
#> Q-M2    0.400
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y        sa        t            s            i
#> Jan 2020 101.0 102.99160 102.9294  -1.99160498   0.06217355
#> Feb 2020 100.1 103.45426 102.9247  -3.35426344   0.52959521
#> Mar 2020  91.8  82.65612 103.1192   9.14388018 -20.46308279
#> Apr 2020  66.7  66.67651 103.5232   0.02349271 -36.84671169
#> May 2020  73.7  78.87377 103.9445  -5.17376662 -25.07074264
#> Jun 2020  98.2  87.34056 104.2875  10.85944310 -16.94698952
#> Jul 2020  97.4  92.53355 104.4507   4.86645017 -11.91712826
#> Aug 2020  71.7  97.50941 104.3140 -25.80940727  -6.80457315
#> Sep 2020 104.7  97.33065 103.8836   7.36935330  -6.55299506
#> Oct 2020 106.7  98.43014 103.2962   8.26986026  -4.86603247
#> Nov 2020 101.6 100.25464 102.7833   1.34535858  -2.52865083
#> Dec 2020  96.6  99.64539 102.5212  -3.04538644  -2.87581385
#> 
#> Forecasts:
#>                y_f     sa_f      t_f        s_f        i_f
#> Jan 2021  94.58042 101.1230 102.4688  -6.542572 -1.3458176
#> Feb 2021  97.88130 101.7320 102.4840  -3.850691 -0.7520506
#> Mar 2021 112.76674 102.1569 102.4403  10.609843 -0.2833876
#> Apr 2021 103.42922 102.1620 102.2951   1.267216 -0.1330803
#> May 2021  95.97032 101.4546 102.1330  -5.484286 -0.6783867
#> Jun 2021 112.78593 101.3141 102.0099  11.471876 -0.6958026
#> Jul 2021 104.00357 101.5966 101.9915   2.406945 -0.3948747
#> Aug 2021  79.06777 102.3509 102.1229 -23.283168  0.2280395
#> Sep 2021 109.03147 102.4766 102.3299   6.554824  0.1467741
#> Oct 2021 108.59415 102.2487 102.5371   6.345501 -0.2884769
#> Nov 2021 106.48447 102.5342 102.7285   3.950244 -0.1943016
#> Dec 2021  99.79090 103.1269 102.8914  -3.336042  0.2355813
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.174
#>  Seasonal     59.871
#>  Irregular     1.096
#>  TD & Hol.     2.568
#>  Others       33.844
#>  Total        99.553
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.044
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           0.994
#>  f-test on sa (seasonal dummies)        0.990
#>  f-test on i (seasonal dummies)         0.968
#>  Residual seasonality (entire series)   0.976
#>  Residual seasonality (last 3 years)    0.932
#>  f-test on sa (td)                      0.978
#>  f-test on i (td)                       0.997
#> 
#> 
#> Additional output variables

# To get all the models of the multiprocessing mp:
get_model(mp, wk)

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> $X13
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.03368      0.108
#> Phi(2)     0.17542      0.075
#> Theta(1)  -0.53601      0.102
#> BTheta(1) -0.67991      0.042
#> 
#>                Estimate Std. Error
#> Monday         0.547498      0.233
#> Tuesday        0.915020      0.232
#> Wednesday      1.073764      0.234
#> Thursday       0.003317      0.233
#> Friday         0.866778      0.235
#> Saturday      -1.535790      0.232
#> Leap year      2.177766      0.715
#> Easter [8]    -2.276133      0.467
#> TC (4-2020)  -20.782385      2.187
#> TC (3-2020)  -20.996390      2.185
#> AO (5-2011)   13.229655      1.879
#> LS (11-2008) -12.597511      1.632
#> 
#> 
#> Residual standard error: 2.225 on 342 degrees of freedom
#> Log likelihood = -800.5, aic =  1635 aicc =  1637, bic(corrected for length) = 1.862
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.172
#> M(2)    0.098
#> M(3)    1.248
#> M(4)    0.367
#> M(5)    1.095
#> M(6)    0.074
#> M(7)    0.083
#> M(8)    0.242
#> M(9)    0.062
#> M(10)   0.259
#> M(11)   0.246
#> Q       0.366
#> Q-M2    0.400
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y        sa        t            s            i
#> Jan 2020 101.0 102.99160 102.9294  -1.99160498   0.06217355
#> Feb 2020 100.1 103.45426 102.9247  -3.35426344   0.52959521
#> Mar 2020  91.8  82.65612 103.1192   9.14388018 -20.46308279
#> Apr 2020  66.7  66.67651 103.5232   0.02349271 -36.84671169
#> May 2020  73.7  78.87377 103.9445  -5.17376662 -25.07074264
#> Jun 2020  98.2  87.34056 104.2875  10.85944310 -16.94698952
#> Jul 2020  97.4  92.53355 104.4507   4.86645017 -11.91712826
#> Aug 2020  71.7  97.50941 104.3140 -25.80940727  -6.80457315
#> Sep 2020 104.7  97.33065 103.8836   7.36935330  -6.55299506
#> Oct 2020 106.7  98.43014 103.2962   8.26986026  -4.86603247
#> Nov 2020 101.6 100.25464 102.7833   1.34535858  -2.52865083
#> Dec 2020  96.6  99.64539 102.5212  -3.04538644  -2.87581385
#> 
#> Forecasts:
#>                y_f     sa_f      t_f        s_f        i_f
#> Jan 2021  94.58042 101.1230 102.4688  -6.542572 -1.3458176
#> Feb 2021  97.88130 101.7320 102.4840  -3.850691 -0.7520506
#> Mar 2021 112.76674 102.1569 102.4403  10.609843 -0.2833876
#> Apr 2021 103.42922 102.1620 102.2951   1.267216 -0.1330803
#> May 2021  95.97032 101.4546 102.1330  -5.484286 -0.6783867
#> Jun 2021 112.78593 101.3141 102.0099  11.471876 -0.6958026
#> Jul 2021 104.00357 101.5966 101.9915   2.406945 -0.3948747
#> Aug 2021  79.06777 102.3509 102.1229 -23.283168  0.2280395
#> Sep 2021 109.03147 102.4766 102.3299   6.554824  0.1467741
#> Oct 2021 108.59415 102.2487 102.5371   6.345501 -0.2884769
#> Nov 2021 106.48447 102.5342 102.7285   3.950244 -0.1943016
#> Dec 2021  99.79090 103.1269 102.8914  -3.336042  0.2355813
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.174
#>  Seasonal     59.871
#>  Irregular     1.096
#>  TD & Hol.     2.568
#>  Others       33.844
#>  Total        99.553
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.044
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           0.994
#>  f-test on sa (seasonal dummies)        0.990
#>  f-test on i (seasonal dummies)         0.968
#>  Residual seasonality (entire series)   0.976
#>  Residual seasonality (last 3 years)    0.932
#>  f-test on sa (td)                      0.978
#>  f-test on i (td)                       0.997
#> 
#> 
#> Additional output variables
#> 
#> $TramoSeats
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.02572      0.108
#> Phi(2)     0.16988      0.075
#> Theta(1)  -0.53992      0.102
#> BTheta(1) -0.67459      0.042
#> 
#>               Estimate Std. Error
#> Monday         0.56114      0.231
#> Tuesday        0.89966      0.231
#> Wednesday      1.06703      0.232
#> Thursday       0.02403      0.232
#> Friday         0.86378      0.234
#> Saturday      -1.55222      0.231
#> Leap year      2.19751      0.712
#> Easter [6]    -2.31734      0.461
#> TC (4-2020)  -20.83317      2.183
#> TC (3-2020)  -20.98304      2.183
#> AO (5-2011)   13.28128      1.872
#> LS (11-2008) -12.57486      1.631
#> 
#> 
#> Residual standard error: 2.221 on 342 degrees of freedom
#> Log likelihood = -799.8, aic =  1634 aicc =  1635, bic(corrected for length) = 1.858
#> 
#> 
#> 
#> Decomposition
#> Model

#> Warning: one argument not used by format '%+f'

#> AR :  1 + 0.025720 B + 0.169879 B^2 

#> Warning: one argument not used by format '%+f'

#> D :  1 - B - B^12 + B^13 

#> Warning: one argument not used by format '%+f'

#> MA :  1 - 0.539919 B - 0.674594 B^12 + 0.364227 B^13 
#> 
#> 
#> SA

#> Warning: one argument not used by format '%+f'

#> D :  1 - 2.000000 B + B^2 

#> Warning: one argument not used by format '%+f'

#> MA :  1 - 1.531766 B + 0.545834 B^2 
#> Innovation variance:  0.5502022 
#> 
#> Trend

#> Warning: one argument not used by format '%+f'

#> D :  1 - 2.000000 B + B^2 

#> Warning: one argument not used by format '%+f'

#> MA :  1 + 0.032214 B - 0.967786 B^2 
#> Innovation variance:  0.02623268 
#> 
#> Seasonal

#> Warning: one argument not used by format '%+f'

#> AR :  1 + 0.025720 B + 0.169879 B^2 

#> Warning: one argument not used by format '%+f'

#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 

#> Warning: one argument not used by format '%+f'

#> MA :  1 + 1.245052 B + 0.658623 B^2 + 0.593270 B^3 + 0.481574 B^4 + 0.366624 B^5 + 0.304938 B^6 + 0.184553 B^7 + 0.190108 B^8 + 0.095090 B^9 + 0.221559 B^10 + 0.186281 B^11 - 0.207005 B^12 - 0.348481 B^13 
#> Innovation variance:  0.1464012 
#> 
#> Irregular
#> Innovation variance:  0.3257068 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y        sa        t           s           i
#> Jan 2020 101.0 102.72726 103.1221  -1.7272634  -0.3948220
#> Feb 2020 100.1 102.95124 103.1797  -2.8512352  -0.2284534
#> Mar 2020  91.8  82.31999 103.3030   9.4800115 -20.9830366
#> Apr 2020  66.7  66.85029 103.5340  -0.1502858 -36.6836970
#> May 2020  73.7  79.46771 103.9124  -5.7677083 -24.4446419
#> Jun 2020  98.2  87.82311 104.2371  10.3768872 -16.4139594
#> Jul 2020  97.4  92.74114 104.3260   4.6588650 -11.5848433
#> Aug 2020  71.7  97.04194 104.1300 -25.3419376  -7.0880934
#> Sep 2020 104.7  97.32774 103.6918   7.3722629  -6.3640867
#> Oct 2020 106.7  98.66536 103.2317   8.0346351  -4.5663227
#> Nov 2020 101.6 100.05441 102.8519   1.5455901  -2.7974884
#> Dec 2020  96.6  99.35191 102.5735  -2.7519144  -3.2215659
#> 
#> Forecasts:
#>                y_f     sa_f      t_f        s_f         i_f
#> Jan 2021  94.55690 101.0346 102.4680  -6.477689 -1.43341231
#> Feb 2021  97.86506 101.4506 102.4540  -3.585540 -1.00338862
#> Mar 2021 113.01303 101.7376 102.4400  11.275402 -0.70237203
#> Apr 2021 103.19842 101.9344 102.4260   1.264068 -0.49166042
#> May 2021  96.00309 102.0679 102.4120  -6.064764 -0.34416230
#> Jun 2021 112.77599 102.1571 102.3980  10.618882 -0.24091361
#> Jul 2021 104.03135 102.2154 102.3840   1.815959 -0.16863953
#> Aug 2021  79.09787 102.2520 102.3700 -23.154116 -0.11804767
#> Sep 2021 109.04546 102.2734 102.3560   6.772048 -0.08263337
#> Oct 2021 108.59337 102.2842 102.3421   6.309157 -0.05784336
#> Nov 2021 106.47776 102.2876 102.3281   4.190193 -0.04049035
#> Dec 2021  99.77303 102.2857 102.3141  -2.512692 -0.02834325
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.887
#>  Seasonal     58.928
#>  Irregular     0.509
#>  TD & Hol.     2.598
#>  Others       33.908
#>  Total        97.830
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.064
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.968
#>  f-test on sa (td)                      0.984
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables
#> 

# To get all the models of the workspace wk:
get_model(wk)

#> Multiprocessing 1 on 1:
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> $sa1
#> $sa1$X13
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.03368      0.108
#> Phi(2)     0.17542      0.075
#> Theta(1)  -0.53601      0.102
#> BTheta(1) -0.67991      0.042
#> 
#>                Estimate Std. Error
#> Monday         0.547498      0.233
#> Tuesday        0.915020      0.232
#> Wednesday      1.073764      0.234
#> Thursday       0.003317      0.233
#> Friday         0.866778      0.235
#> Saturday      -1.535790      0.232
#> Leap year      2.177766      0.715
#> Easter [8]    -2.276133      0.467
#> TC (4-2020)  -20.782385      2.187
#> TC (3-2020)  -20.996390      2.185
#> AO (5-2011)   13.229655      1.879
#> LS (11-2008) -12.597511      1.632
#> 
#> 
#> Residual standard error: 2.225 on 342 degrees of freedom
#> Log likelihood = -800.5, aic =  1635 aicc =  1637, bic(corrected for length) = 1.862
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.172
#> M(2)    0.098
#> M(3)    1.248
#> M(4)    0.367
#> M(5)    1.095
#> M(6)    0.074
#> M(7)    0.083
#> M(8)    0.242
#> M(9)    0.062
#> M(10)   0.259
#> M(11)   0.246
#> Q       0.366
#> Q-M2    0.400
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y        sa        t            s            i
#> Jan 2020 101.0 102.99160 102.9294  -1.99160498   0.06217355
#> Feb 2020 100.1 103.45426 102.9247  -3.35426344   0.52959521
#> Mar 2020  91.8  82.65612 103.1192   9.14388018 -20.46308279
#> Apr 2020  66.7  66.67651 103.5232   0.02349271 -36.84671169
#> May 2020  73.7  78.87377 103.9445  -5.17376662 -25.07074264
#> Jun 2020  98.2  87.34056 104.2875  10.85944310 -16.94698952
#> Jul 2020  97.4  92.53355 104.4507   4.86645017 -11.91712826
#> Aug 2020  71.7  97.50941 104.3140 -25.80940727  -6.80457315
#> Sep 2020 104.7  97.33065 103.8836   7.36935330  -6.55299506
#> Oct 2020 106.7  98.43014 103.2962   8.26986026  -4.86603247
#> Nov 2020 101.6 100.25464 102.7833   1.34535858  -2.52865083
#> Dec 2020  96.6  99.64539 102.5212  -3.04538644  -2.87581385
#> 
#> Forecasts:
#>                y_f     sa_f      t_f        s_f        i_f
#> Jan 2021  94.58042 101.1230 102.4688  -6.542572 -1.3458176
#> Feb 2021  97.88130 101.7320 102.4840  -3.850691 -0.7520506
#> Mar 2021 112.76674 102.1569 102.4403  10.609843 -0.2833876
#> Apr 2021 103.42922 102.1620 102.2951   1.267216 -0.1330803
#> May 2021  95.97032 101.4546 102.1330  -5.484286 -0.6783867
#> Jun 2021 112.78593 101.3141 102.0099  11.471876 -0.6958026
#> Jul 2021 104.00357 101.5966 101.9915   2.406945 -0.3948747
#> Aug 2021  79.06777 102.3509 102.1229 -23.283168  0.2280395
#> Sep 2021 109.03147 102.4766 102.3299   6.554824  0.1467741
#> Oct 2021 108.59415 102.2487 102.5371   6.345501 -0.2884769
#> Nov 2021 106.48447 102.5342 102.7285   3.950244 -0.1943016
#> Dec 2021  99.79090 103.1269 102.8914  -3.336042  0.2355813
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.174
#>  Seasonal     59.871
#>  Irregular     1.096
#>  TD & Hol.     2.568
#>  Others       33.844
#>  Total        99.553
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.044
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           0.994
#>  f-test on sa (seasonal dummies)        0.990
#>  f-test on i (seasonal dummies)         0.968
#>  Residual seasonality (entire series)   0.976
#>  Residual seasonality (last 3 years)    0.932
#>  f-test on sa (td)                      0.978
#>  f-test on i (td)                       0.997
#> 
#> 
#> Additional output variables
#> 
#> $sa1$TramoSeats
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     0.02572      0.108
#> Phi(2)     0.16988      0.075
#> Theta(1)  -0.53992      0.102
#> BTheta(1) -0.67459      0.042
#> 
#>               Estimate Std. Error
#> Monday         0.56114      0.231
#> Tuesday        0.89966      0.231
#> Wednesday      1.06703      0.232
#> Thursday       0.02403      0.232
#> Friday         0.86378      0.234
#> Saturday      -1.55222      0.231
#> Leap year      2.19751      0.712
#> Easter [6]    -2.31734      0.461
#> TC (4-2020)  -20.83317      2.183
#> TC (3-2020)  -20.98304      2.183
#> AO (5-2011)   13.28128      1.872
#> LS (11-2008) -12.57486      1.631
#> 
#> 
#> Residual standard error: 2.221 on 342 degrees of freedom
#> Log likelihood = -799.8, aic =  1634 aicc =  1635, bic(corrected for length) = 1.858
#> 
#> 
#> 
#> Decomposition
#> Model

#> Warning: one argument not used by format '%+f'

#> AR :  1 + 0.025720 B + 0.169879 B^2 

#> Warning: one argument not used by format '%+f'

#> D :  1 - B - B^12 + B^13 

#> Warning: one argument not used by format '%+f'

#> MA :  1 - 0.539919 B - 0.674594 B^12 + 0.364227 B^13 
#> 
#> 
#> SA

#> Warning: one argument not used by format '%+f'

#> D :  1 - 2.000000 B + B^2 

#> Warning: one argument not used by format '%+f'

#> MA :  1 - 1.531766 B + 0.545834 B^2 
#> Innovation variance:  0.5502022 
#> 
#> Trend

#> Warning: one argument not used by format '%+f'

#> D :  1 - 2.000000 B + B^2 

#> Warning: one argument not used by format '%+f'

#> MA :  1 + 0.032214 B - 0.967786 B^2 
#> Innovation variance:  0.02623268 
#> 
#> Seasonal

#> Warning: one argument not used by format '%+f'

#> AR :  1 + 0.025720 B + 0.169879 B^2 

#> Warning: one argument not used by format '%+f'

#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 

#> Warning: one argument not used by format '%+f'

#> MA :  1 + 1.245052 B + 0.658623 B^2 + 0.593270 B^3 + 0.481574 B^4 + 0.366624 B^5 + 0.304938 B^6 + 0.184553 B^7 + 0.190108 B^8 + 0.095090 B^9 + 0.221559 B^10 + 0.186281 B^11 - 0.207005 B^12 - 0.348481 B^13 
#> Innovation variance:  0.1464012 
#> 
#> Irregular
#> Innovation variance:  0.3257068 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y        sa        t           s           i
#> Jan 2020 101.0 102.72726 103.1221  -1.7272634  -0.3948220
#> Feb 2020 100.1 102.95124 103.1797  -2.8512352  -0.2284534
#> Mar 2020  91.8  82.31999 103.3030   9.4800115 -20.9830366
#> Apr 2020  66.7  66.85029 103.5340  -0.1502858 -36.6836970
#> May 2020  73.7  79.46771 103.9124  -5.7677083 -24.4446419
#> Jun 2020  98.2  87.82311 104.2371  10.3768872 -16.4139594
#> Jul 2020  97.4  92.74114 104.3260   4.6588650 -11.5848433
#> Aug 2020  71.7  97.04194 104.1300 -25.3419376  -7.0880934
#> Sep 2020 104.7  97.32774 103.6918   7.3722629  -6.3640867
#> Oct 2020 106.7  98.66536 103.2317   8.0346351  -4.5663227
#> Nov 2020 101.6 100.05441 102.8519   1.5455901  -2.7974884
#> Dec 2020  96.6  99.35191 102.5735  -2.7519144  -3.2215659
#> 
#> Forecasts:
#>                y_f     sa_f      t_f        s_f         i_f
#> Jan 2021  94.55690 101.0346 102.4680  -6.477689 -1.43341231
#> Feb 2021  97.86506 101.4506 102.4540  -3.585540 -1.00338862
#> Mar 2021 113.01303 101.7376 102.4400  11.275402 -0.70237203
#> Apr 2021 103.19842 101.9344 102.4260   1.264068 -0.49166042
#> May 2021  96.00309 102.0679 102.4120  -6.064764 -0.34416230
#> Jun 2021 112.77599 102.1571 102.3980  10.618882 -0.24091361
#> Jul 2021 104.03135 102.2154 102.3840   1.815959 -0.16863953
#> Aug 2021  79.09787 102.2520 102.3700 -23.154116 -0.11804767
#> Sep 2021 109.04546 102.2734 102.3560   6.772048 -0.08263337
#> Oct 2021 108.59337 102.2842 102.3421   6.309157 -0.05784336
#> Nov 2021 106.47776 102.2876 102.3281   4.190193 -0.04049035
#> Dec 2021  99.77303 102.2857 102.3141  -2.512692 -0.02834325
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.887
#>  Seasonal     58.928
#>  Irregular     0.509
#>  TD & Hol.     2.598
#>  Others       33.908
#>  Total        97.830
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.064
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.968
#>  f-test on sa (td)                      0.984
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables
#> 
#> 

# }