Get the seasonally adjusted model from a workspace

Generics functions to get seasonally adjusted model(s) from workspace, multiprocessing or sa_item object. get_model returns a "SA" objects while get_jmodel returns the Java objects of the models.

get_jmodel(x, workspace, userdefined = NULL, progress_bar = TRUE)

get_model(x, workspace, userdefined = NULL, progress_bar = TRUE)

Arguments

x
the object to get the seasonally adjusted model.

workspace
the workspace object where models are stored. If x is a workspace object this parameter is not used.

userdefined
vector with characters for additional output variables. (see x13 or tramoseats).

progress_bar
boolean: if TRUE a progress bar is printed.

Value

get_model() returns a seasonally adjust object (class c("SA", "X13") or c("SA", "TRAMO_SEATS") or list of seasonally adjust objects:

if x is a sa_item object, get_model(x) returns a "SA" object (or a jSA object with get_jmodel(x));

if x is a multiprocessing object, get_ts(x) returns list of length the number of sa_items, each element containing a "SA" object (or a jSA object with get_jmodel(x));

if x is a workspace object, get_ts(x) returns list of length the number of multiprocessing, each element containing a list of a "SA" object (or a jSA object with get_jmodel(x)).

See also

Examples

# \donttest{
spec_x13 <- x13_spec(spec = "RSA5c", easter.enabled = FALSE)
sa_x13 <- x13(ipi_c_eu[, "FR"], spec = spec_x13)
spec_ts <- tramoseats_spec(spec = "RSA5")
sa_ts <- tramoseats(ipi_c_eu[, "FR"], spec = spec_ts)

wk <- new_workspace()
mp <- new_multiprocessing(wk, "sa1")
add_sa_item(wk, "sa1", sa_x13, "X13")
add_sa_item(wk, "sa1", sa_ts, "TramoSeats")

compute(wk) # It's important to compute the workspace to get the SA model
sa_item1 <- get_object(mp, 1)

get_model(sa_item1, wk) # Extract the model of the sa_item1: its the object sa_x13

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>            Estimate Std. Error
#> Phi(1)     0.008774      0.119
#> Phi(2)     0.098387      0.081
#> Theta(1)  -0.562495      0.110
#> BTheta(1) -0.738153      0.039
#> 
#>                Estimate Std. Error
#> Monday        0.0058623      0.002
#> Tuesday       0.0091349      0.002
#> Wednesday     0.0109267      0.002
#> Thursday      0.0005156      0.002
#> Friday        0.0091500      0.002
#> Saturday     -0.0161713      0.002
#> AO (5-2011)   0.1278938      0.018
#> LS (11-2008) -0.1157755      0.016
#> 
#> 
#> Residual standard error: 0.02153 on 334 degrees of freedom
#> Log likelihood = 834.5, aic =  1562 aicc =  1564, bic(corrected for length) = -7.474
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.094
#> M(2)    0.049
#> M(3)    0.978
#> M(4)    0.421
#> M(5)    0.969
#> M(6)    0.180
#> M(7)    0.075
#> M(8)    0.211
#> M(9)    0.063
#> M(10)   0.250
#> M(11)   0.232
#> Q       0.323
#> Q-M2    0.357
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.3314 105.1139 0.9864104 1.0020689
#> Feb 2019 101.9 106.3742 105.3665 0.9579387 1.0095639
#> Mar 2019 111.0 104.5953 105.4822 1.0612327 0.9915923
#> Apr 2019 107.4 105.0934 105.4192 1.0219484 0.9969091
#> May 2019 105.5 106.8159 105.2936 0.9876808 1.0144578
#> Jun 2019 105.8 101.4580 105.1024 1.0427961 0.9653249
#> Jul 2019 110.1 105.3379 104.7923 1.0452083 1.0052062
#> Aug 2019  78.7 103.0406 104.3781 0.7637769 0.9871852
#> Sep 2019 108.5 104.4959 103.8759 1.0383182 1.0059689
#> Oct 2019 116.8 104.9663 103.3582 1.1127377 1.0155587
#> Nov 2019 103.8 101.6218 102.9772 1.0214349 0.9868375
#> Dec 2019  97.7 101.8080 102.7604 0.9596494 0.9907318
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.59660 103.0784 102.6789 0.9856249 1.0038905
#> Feb 2020 100.87436 103.2029 102.7042 0.9774377 1.0048553
#> Mar 2020 111.27482 102.6228 102.7383 1.0843091 0.9988753
#> Apr 2020 105.12395 103.3465 102.7663 1.0171986 1.0056464
#> May 2020  96.44225 102.2715 102.7176 0.9430021 0.9956572
#> Jun 2020 112.46775 102.4907 102.6115 1.0973459 0.9988229
#> Jul 2020 106.58633 102.4070 102.5858 1.0408106 0.9982572
#> Aug 2020  76.49372 102.5838 102.6507 0.7456707 0.9993479
#> Sep 2020 111.06093 103.4916 102.7527 1.0731400 1.0071907
#> Oct 2020 111.06175 102.3655 102.8697 1.0849527 0.9950991
#> Nov 2020 104.14664 102.6519 102.9955 1.0145617 0.9966639
#> Dec 2020 101.63644 103.5595 103.0930 0.9814307 1.0045245
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.426
#>  Seasonal     72.514
#>  Irregular     1.136
#>  TD & Hol.     2.703
#>  Others       21.274
#>  Total       100.053
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.498
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.964
#>  f-test on i (seasonal dummies)         0.936
#>  Residual seasonality (entire series)   0.970
#>  Residual seasonality (last 3 years)    0.994
#>  f-test on sa (td)                      0.996
#>  f-test on i (td)                       0.993
#> 
#> 
#> Additional output variables

# To get all the models of the multiprocessing mp:
get_model(mp, wk)

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> $X13
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>            Estimate Std. Error
#> Phi(1)     0.008774      0.119
#> Phi(2)     0.098387      0.081
#> Theta(1)  -0.562495      0.110
#> BTheta(1) -0.738153      0.039
#> 
#>                Estimate Std. Error
#> Monday        0.0058623      0.002
#> Tuesday       0.0091349      0.002
#> Wednesday     0.0109267      0.002
#> Thursday      0.0005156      0.002
#> Friday        0.0091500      0.002
#> Saturday     -0.0161713      0.002
#> AO (5-2011)   0.1278938      0.018
#> LS (11-2008) -0.1157755      0.016
#> 
#> 
#> Residual standard error: 0.02153 on 334 degrees of freedom
#> Log likelihood = 834.5, aic =  1562 aicc =  1564, bic(corrected for length) = -7.474
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.094
#> M(2)    0.049
#> M(3)    0.978
#> M(4)    0.421
#> M(5)    0.969
#> M(6)    0.180
#> M(7)    0.075
#> M(8)    0.211
#> M(9)    0.063
#> M(10)   0.250
#> M(11)   0.232
#> Q       0.323
#> Q-M2    0.357
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.3314 105.1139 0.9864104 1.0020689
#> Feb 2019 101.9 106.3742 105.3665 0.9579387 1.0095639
#> Mar 2019 111.0 104.5953 105.4822 1.0612327 0.9915923
#> Apr 2019 107.4 105.0934 105.4192 1.0219484 0.9969091
#> May 2019 105.5 106.8159 105.2936 0.9876808 1.0144578
#> Jun 2019 105.8 101.4580 105.1024 1.0427961 0.9653249
#> Jul 2019 110.1 105.3379 104.7923 1.0452083 1.0052062
#> Aug 2019  78.7 103.0406 104.3781 0.7637769 0.9871852
#> Sep 2019 108.5 104.4959 103.8759 1.0383182 1.0059689
#> Oct 2019 116.8 104.9663 103.3582 1.1127377 1.0155587
#> Nov 2019 103.8 101.6218 102.9772 1.0214349 0.9868375
#> Dec 2019  97.7 101.8080 102.7604 0.9596494 0.9907318
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.59660 103.0784 102.6789 0.9856249 1.0038905
#> Feb 2020 100.87436 103.2029 102.7042 0.9774377 1.0048553
#> Mar 2020 111.27482 102.6228 102.7383 1.0843091 0.9988753
#> Apr 2020 105.12395 103.3465 102.7663 1.0171986 1.0056464
#> May 2020  96.44225 102.2715 102.7176 0.9430021 0.9956572
#> Jun 2020 112.46775 102.4907 102.6115 1.0973459 0.9988229
#> Jul 2020 106.58633 102.4070 102.5858 1.0408106 0.9982572
#> Aug 2020  76.49372 102.5838 102.6507 0.7456707 0.9993479
#> Sep 2020 111.06093 103.4916 102.7527 1.0731400 1.0071907
#> Oct 2020 111.06175 102.3655 102.8697 1.0849527 0.9950991
#> Nov 2020 104.14664 102.6519 102.9955 1.0145617 0.9966639
#> Dec 2020 101.63644 103.5595 103.0930 0.9814307 1.0045245
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.426
#>  Seasonal     72.514
#>  Irregular     1.136
#>  TD & Hol.     2.703
#>  Others       21.274
#>  Total       100.053
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.498
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.964
#>  f-test on i (seasonal dummies)         0.936
#>  Residual seasonality (entire series)   0.970
#>  Residual seasonality (last 3 years)    0.994
#>  f-test on sa (td)                      0.996
#>  f-test on i (td)                       0.993
#> 
#> 
#> Additional output variables
#> 
#> $TramoSeats
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.0241      0.123
#> Phi(2)      0.1482      0.080
#> Theta(1)   -0.5102      0.118
#> BTheta(1)  -0.7193      0.040
#> 
#>               Estimate Std. Error
#> Monday        0.004951      0.002
#> Tuesday       0.008599      0.002
#> Wednesday     0.011454      0.002
#> Thursday      0.001241      0.002
#> Friday        0.007779      0.002
#> Saturday     -0.015954      0.002
#> Leap year     0.022096      0.007
#> Easter [6]   -0.019882      0.004
#> AO (5-2011)   0.128286      0.018
#> LS (11-2008) -0.114105      0.016
#> 
#> 
#> Residual standard error: 0.02084 on 332 degrees of freedom
#> Log likelihood = 846.3, aic =  1543 aicc =  1544, bic(corrected for length) = -7.506
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 + 0.024102 B + 0.148206 B^2 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.510230 B - 0.719325 B^12 + 0.367021 B^13 
#> 
#> 
#> SA
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.505621 B + 0.518275 B^2 
#> Innovation variance:  0.5963745 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.027051 B - 0.972949 B^2 
#> Innovation variance:  0.03262266 
#> 
#> Seasonal
#> AR :  1 + 0.024102 B + 0.148206 B^2 
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 1.307845 B + 0.758576 B^2 + 0.703126 B^3 + 0.603379 B^4 + 0.494040 B^5 + 0.433104 B^6 + 0.314376 B^7 + 0.311439 B^8 + 0.215331 B^9 + 0.323556 B^10 + 0.288284 B^11 - 0.162781 B^12 - 0.347675 B^13 
#> Innovation variance:  0.131756 
#> 
#> Irregular
#> Innovation variance:  0.3408261 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.1230 104.9050 0.9883662 1.0020778
#> Feb 2019 101.9 105.5664 105.0034 0.9652688 1.0053621
#> Mar 2019 111.0 104.5184 105.0206 1.0620135 0.9952184
#> Apr 2019 107.4 105.2891 105.0542 1.0200482 1.0022362
#> May 2019 105.5 106.9239 104.9010 0.9866832 1.0192835
#> Jun 2019 105.8 102.2691 104.5675 1.0345256 0.9780200
#> Jul 2019 110.1 105.8170 104.3421 1.0404759 1.0141350
#> Aug 2019  78.7 103.4109 104.1245 0.7610418 0.9931464
#> Sep 2019 108.5 104.0283 103.8893 1.0429858 1.0013378
#> Oct 2019 116.8 104.6908 103.5909 1.1156663 1.0106173
#> Nov 2019 103.8 102.2897 103.1567 1.0147647 0.9915958
#> Dec 2019  97.7 102.0473 102.8588 0.9573990 0.9921103
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.59890 102.7892 102.7929 0.9884197 0.9999634
#> Feb 2020  99.81313 102.7988 102.8025 0.9709572 0.9999634
#> Mar 2020 112.05983 102.8084 102.8121 1.0899876 0.9999634
#> Apr 2020 104.38189 102.8180 102.8217 1.0152106 0.9999634
#> May 2020  96.45762 102.8276 102.8313 0.9380522 0.9999634
#> Jun 2020 112.13530 102.8372 102.8409 1.0904162 0.9999634
#> Jul 2020 106.56822 102.8468 102.8505 1.0361846 0.9999634
#> Aug 2020  76.56202 102.8564 102.8601 0.7443586 0.9999634
#> Sep 2020 110.92049 102.8660 102.8697 1.0783011 0.9999634
#> Oct 2020 111.00686 102.8756 102.8793 1.0790401 0.9999634
#> Nov 2020 104.15494 102.8852 102.8890 1.0123416 0.9999634
#> Dec 2020 101.56243 102.8948 102.8986 0.9870513 0.9999634
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.126
#>  Seasonal     71.716
#>  Irregular     0.501
#>  TD & Hol.     2.739
#>  Others       20.785
#>  Total        97.867
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.381
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          0.995
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.964
#>  f-test on sa (td)                      0.984
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables
#> 

# To get all the models of the workspace wk:
get_model(wk)

#> Multiprocessing 1 on 1:
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

#> $sa1
#> $sa1$X13
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>            Estimate Std. Error
#> Phi(1)     0.008774      0.119
#> Phi(2)     0.098387      0.081
#> Theta(1)  -0.562495      0.110
#> BTheta(1) -0.738153      0.039
#> 
#>                Estimate Std. Error
#> Monday        0.0058623      0.002
#> Tuesday       0.0091349      0.002
#> Wednesday     0.0109267      0.002
#> Thursday      0.0005156      0.002
#> Friday        0.0091500      0.002
#> Saturday     -0.0161713      0.002
#> AO (5-2011)   0.1278938      0.018
#> LS (11-2008) -0.1157755      0.016
#> 
#> 
#> Residual standard error: 0.02153 on 334 degrees of freedom
#> Log likelihood = 834.5, aic =  1562 aicc =  1564, bic(corrected for length) = -7.474
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.094
#> M(2)    0.049
#> M(3)    0.978
#> M(4)    0.421
#> M(5)    0.969
#> M(6)    0.180
#> M(7)    0.075
#> M(8)    0.211
#> M(9)    0.063
#> M(10)   0.250
#> M(11)   0.232
#> Q       0.323
#> Q-M2    0.357
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.3314 105.1139 0.9864104 1.0020689
#> Feb 2019 101.9 106.3742 105.3665 0.9579387 1.0095639
#> Mar 2019 111.0 104.5953 105.4822 1.0612327 0.9915923
#> Apr 2019 107.4 105.0934 105.4192 1.0219484 0.9969091
#> May 2019 105.5 106.8159 105.2936 0.9876808 1.0144578
#> Jun 2019 105.8 101.4580 105.1024 1.0427961 0.9653249
#> Jul 2019 110.1 105.3379 104.7923 1.0452083 1.0052062
#> Aug 2019  78.7 103.0406 104.3781 0.7637769 0.9871852
#> Sep 2019 108.5 104.4959 103.8759 1.0383182 1.0059689
#> Oct 2019 116.8 104.9663 103.3582 1.1127377 1.0155587
#> Nov 2019 103.8 101.6218 102.9772 1.0214349 0.9868375
#> Dec 2019  97.7 101.8080 102.7604 0.9596494 0.9907318
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.59660 103.0784 102.6789 0.9856249 1.0038905
#> Feb 2020 100.87436 103.2029 102.7042 0.9774377 1.0048553
#> Mar 2020 111.27482 102.6228 102.7383 1.0843091 0.9988753
#> Apr 2020 105.12395 103.3465 102.7663 1.0171986 1.0056464
#> May 2020  96.44225 102.2715 102.7176 0.9430021 0.9956572
#> Jun 2020 112.46775 102.4907 102.6115 1.0973459 0.9988229
#> Jul 2020 106.58633 102.4070 102.5858 1.0408106 0.9982572
#> Aug 2020  76.49372 102.5838 102.6507 0.7456707 0.9993479
#> Sep 2020 111.06093 103.4916 102.7527 1.0731400 1.0071907
#> Oct 2020 111.06175 102.3655 102.8697 1.0849527 0.9950991
#> Nov 2020 104.14664 102.6519 102.9955 1.0145617 0.9966639
#> Dec 2020 101.63644 103.5595 103.0930 0.9814307 1.0045245
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.426
#>  Seasonal     72.514
#>  Irregular     1.136
#>  TD & Hol.     2.703
#>  Others       21.274
#>  Total       100.053
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.498
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.964
#>  f-test on i (seasonal dummies)         0.936
#>  Residual seasonality (entire series)   0.970
#>  Residual seasonality (last 3 years)    0.994
#>  f-test on sa (td)                      0.996
#>  f-test on i (td)                       0.993
#> 
#> 
#> Additional output variables
#> 
#> $sa1$TramoSeats
#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.0241      0.123
#> Phi(2)      0.1482      0.080
#> Theta(1)   -0.5102      0.118
#> BTheta(1)  -0.7193      0.040
#> 
#>               Estimate Std. Error
#> Monday        0.004951      0.002
#> Tuesday       0.008599      0.002
#> Wednesday     0.011454      0.002
#> Thursday      0.001241      0.002
#> Friday        0.007779      0.002
#> Saturday     -0.015954      0.002
#> Leap year     0.022096      0.007
#> Easter [6]   -0.019882      0.004
#> AO (5-2011)   0.128286      0.018
#> LS (11-2008) -0.114105      0.016
#> 
#> 
#> Residual standard error: 0.02084 on 332 degrees of freedom
#> Log likelihood = 846.3, aic =  1543 aicc =  1544, bic(corrected for length) = -7.506
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 + 0.024102 B + 0.148206 B^2 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.510230 B - 0.719325 B^12 + 0.367021 B^13 
#> 
#> 
#> SA
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.505621 B + 0.518275 B^2 
#> Innovation variance:  0.5963745 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.027051 B - 0.972949 B^2 
#> Innovation variance:  0.03262266 
#> 
#> Seasonal
#> AR :  1 + 0.024102 B + 0.148206 B^2 
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 1.307845 B + 0.758576 B^2 + 0.703126 B^3 + 0.603379 B^4 + 0.494040 B^5 + 0.433104 B^6 + 0.314376 B^7 + 0.311439 B^8 + 0.215331 B^9 + 0.323556 B^10 + 0.288284 B^11 - 0.162781 B^12 - 0.347675 B^13 
#> Innovation variance:  0.131756 
#> 
#> Irregular
#> Innovation variance:  0.3408261 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2019 103.9 105.1230 104.9050 0.9883662 1.0020778
#> Feb 2019 101.9 105.5664 105.0034 0.9652688 1.0053621
#> Mar 2019 111.0 104.5184 105.0206 1.0620135 0.9952184
#> Apr 2019 107.4 105.2891 105.0542 1.0200482 1.0022362
#> May 2019 105.5 106.9239 104.9010 0.9866832 1.0192835
#> Jun 2019 105.8 102.2691 104.5675 1.0345256 0.9780200
#> Jul 2019 110.1 105.8170 104.3421 1.0404759 1.0141350
#> Aug 2019  78.7 103.4109 104.1245 0.7610418 0.9931464
#> Sep 2019 108.5 104.0283 103.8893 1.0429858 1.0013378
#> Oct 2019 116.8 104.6908 103.5909 1.1156663 1.0106173
#> Nov 2019 103.8 102.2897 103.1567 1.0147647 0.9915958
#> Dec 2019  97.7 102.0473 102.8588 0.9573990 0.9921103
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2020 101.59890 102.7892 102.7929 0.9884197 0.9999634
#> Feb 2020  99.81313 102.7988 102.8025 0.9709572 0.9999634
#> Mar 2020 112.05983 102.8084 102.8121 1.0899876 0.9999634
#> Apr 2020 104.38189 102.8180 102.8217 1.0152106 0.9999634
#> May 2020  96.45762 102.8276 102.8313 0.9380522 0.9999634
#> Jun 2020 112.13530 102.8372 102.8409 1.0904162 0.9999634
#> Jul 2020 106.56822 102.8468 102.8505 1.0361846 0.9999634
#> Aug 2020  76.56202 102.8564 102.8601 0.7443586 0.9999634
#> Sep 2020 110.92049 102.8660 102.8697 1.0783011 0.9999634
#> Oct 2020 111.00686 102.8756 102.8793 1.0790401 0.9999634
#> Nov 2020 104.15494 102.8852 102.8890 1.0123416 0.9999634
#> Dec 2020 101.56243 102.8948 102.8986 0.9870513 0.9999634
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.126
#>  Seasonal     71.716
#>  Irregular     0.501
#>  TD & Hol.     2.739
#>  Others       20.785
#>  Total        97.867
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.381
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          0.995
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.964
#>  f-test on sa (td)                      0.984
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables
#> 
#> 

# }