Industrial Production Indices in manufacturing in the European Union

A dataset containing on monthly industrial production indices in manufacturing in the European Union (from sts_inpr_m dataset of Eurostat). Data are based 100 in 2015 and are unadjusted, i.e. neither seasonally adjusted nor calendar adjusted.

ipi_c_eu

Format

A monthly ts object from january 1990 to december 2017 with 37 variables.

Source

http://ec.europa.eu/eurostat/wdds/rest/data/v2.1/json/en/sts_inpr_m?nace_r2=C&precision=1&sinceTimePeriod=1980M01&unit=I15&s_adj=NSA

Details

The dataset contains 37 time series corresponding to the following geographical area

EU28

European Union (current composition)

EU27_2019	European Union (without United Kingdom)

EA19

Euro area (19 countries)

BE	Belgium

BG

Bulgaria

CZ	Czechia

DK

Denmark

DE	Germany (until 1990 former territory of the FRG)

EE

Estonia

IE	Ireland

EL

Greece

ES	Spain

FR

France

HR	Croatia

IT

Italy

CY	Cyprus

LV

Latvia

LT	Lithuania

LU

Luxembourg

HU	Hungary

MT

Malta

NL	Netherlands

AT

Austria

PL	Poland

PT

Portugal

RO	Romania

SI

Slovenia

SK	Slovakia

FI

Finland

SE	Sweden

UK

United Kingdom

NO	Norway

ME

Montenegro

MK	Former Yugoslav Republic of Macedonia, the

RS

Serbia

TR	Turkey

BA

Bosnia and Herzegovina