Functions around 'jSA' objects

get_dictionary returns the indicators that can be extracted from "jSA" objects, get_indicators extract a list of indicators and .

get_dictionary(x)

get_indicators(x, ...)

jSA2R(x, userdefined = NULL)

Arguments

x
a "jSA" object.

...
characters containing the names of the indicators to extract.

userdefined
userdefined vector with characters for additional output variables (see user_defined_variables). Only used for "SA" objects.

Value

get_dictionary a vector of characters, get_indicators returns a list containing the indicators that are extracted and jSA2R returns a "SA" or a "regarima" object.

Details

A "jSA" object is a list with three elements:

"result": the Java object with the results of a seasonal adjustment or a pre-adjustment method.

"spec": the Java object with the specification of a seasonal adjustment or a pre-adjustment method.

"dictionary": the Java object with dictionnary of a seasonal adjustment or a pre-adjustment method. In particular, it contains all the user-defined regressors.

get_dictionary returns the list of indicators that can be extracted from a jSA object by the function get_indicators.

jSA2R returns the corresponding formatted seasonal adjustment ("SA" object) or RegARIMA ("regarima" object) model.

Examples

myseries <- ipi_c_eu[, "FR"]
mysa <- jx13(myseries, spec = "RSA5c")
get_dictionary(mysa)

#>   [1] "y"                                              
#>   [2] "y_f"                                            
#>   [3] "t"                                              
#>   [4] "t_f"                                            
#>   [5] "sa"                                             
#>   [6] "sa_f"                                           
#>   [7] "s"                                              
#>   [8] "s_f"                                            
#>   [9] "i"                                              
#>  [10] "i_f"                                            
#>  [11] "mode"                                           
#>  [12] "preprocessing.model.span.start"                 
#>  [13] "preprocessing.model.span.end"                   
#>  [14] "preprocessing.model.span.n"                     
#>  [15] "preprocessing.model.espan.start"                
#>  [16] "preprocessing.model.espan.end"                  
#>  [17] "preprocessing.model.espan.n"                    
#>  [18] "preprocessing.model.log"                        
#>  [19] "preprocessing.model.adjust"                     
#>  [20] "preprocessing.model.y"                          
#>  [21] "preprocessing.model.y_f"                        
#>  [22] "preprocessing.model.y_ef"                       
#>  [23] "preprocessing.model.yc"                         
#>  [24] "preprocessing.model.yc_f"                       
#>  [25] "preprocessing.model.yc_ef"                      
#>  [26] "preprocessing.model.l"                          
#>  [27] "preprocessing.model.y_lin"                      
#>  [28] "preprocessing.model.y_lin_f"                    
#>  [29] "preprocessing.model.ycal"                       
#>  [30] "preprocessing.model.ycal_f"                     
#>  [31] "preprocessing.model.det"                        
#>  [32] "preprocessing.model.det_f"                      
#>  [33] "preprocessing.model.l_f"                        
#>  [34] "preprocessing.model.l_b"                        
#>  [35] "preprocessing.model.cal"                        
#>  [36] "preprocessing.model.cal_f"                      
#>  [37] "preprocessing.model.tde"                        
#>  [38] "preprocessing.model.tde_f"                      
#>  [39] "preprocessing.model.mhe"                        
#>  [40] "preprocessing.model.mhe_f"                      
#>  [41] "preprocessing.model.ee"                         
#>  [42] "preprocessing.model.ee_f"                       
#>  [43] "preprocessing.model.omhe"                       
#>  [44] "preprocessing.model.omhe_f"                     
#>  [45] "preprocessing.model.out(*)"                     
#>  [46] "preprocessing.model.out_f"                      
#>  [47] "preprocessing.model.out_i"                      
#>  [48] "preprocessing.model.out_i_f"                    
#>  [49] "preprocessing.model.out_t"                      
#>  [50] "preprocessing.model.out_t_f"                    
#>  [51] "preprocessing.model.out_s"                      
#>  [52] "preprocessing.model.out_s_f"                    
#>  [53] "preprocessing.model.reg"                        
#>  [54] "preprocessing.model.reg_f"                      
#>  [55] "preprocessing.model.reg_t"                      
#>  [56] "preprocessing.model.reg_t_f"                    
#>  [57] "preprocessing.model.reg_s"                      
#>  [58] "preprocessing.model.reg_s_f"                    
#>  [59] "preprocessing.model.reg_i"                      
#>  [60] "preprocessing.model.reg_i_f"                    
#>  [61] "preprocessing.model.reg_sa"                     
#>  [62] "preprocessing.model.reg_sa_f"                   
#>  [63] "preprocessing.model.reg_y"                      
#>  [64] "preprocessing.model.reg_y_f"                    
#>  [65] "preprocessing.model.reg_u"                      
#>  [66] "preprocessing.model.reg_u_f"                    
#>  [67] "preprocessing.model.fullresiduals"              
#>  [68] "preprocessing.model.lp"                         
#>  [69] "preprocessing.model.ntd"                        
#>  [70] "preprocessing.model.nmh"                        
#>  [71] "preprocessing.model.td(*)"                      
#>  [72] "preprocessing.model.easter"                     
#>  [73] "preprocessing.model.nout"                       
#>  [74] "preprocessing.model.noutao"                     
#>  [75] "preprocessing.model.noutls"                     
#>  [76] "preprocessing.model.nouttc"                     
#>  [77] "preprocessing.model.noutso"                     
#>  [78] "preprocessing.model.coefficients"               
#>  [79] "preprocessing.model.description"                
#>  [80] "preprocessing.model.covar"                      
#>  [81] "preprocessing.model.pcovar"                     
#>  [82] "preprocessing.model.fcasts(?)"                  
#>  [83] "preprocessing.model.bcasts(?)"                  
#>  [84] "preprocessing.model.lin_fcasts(?)"              
#>  [85] "preprocessing.model.lin_bcasts(?)"              
#>  [86] "preprocessing.model.efcasts(?)"                 
#>  [87] "preprocessing.arima.parameters"                 
#>  [88] "preprocessing.arima.p"                          
#>  [89] "preprocessing.arima.d"                          
#>  [90] "preprocessing.arima.q"                          
#>  [91] "preprocessing.arima.bp"                         
#>  [92] "preprocessing.arima.bd"                         
#>  [93] "preprocessing.arima.bq"                         
#>  [94] "preprocessing.likelihood.neffectiveobs"         
#>  [95] "preprocessing.likelihood.np"                    
#>  [96] "preprocessing.likelihood.logvalue"              
#>  [97] "preprocessing.likelihood.adjustedlogvalue"      
#>  [98] "preprocessing.likelihood.ssqerr"                
#>  [99] "preprocessing.likelihood.aic"                   
#> [100] "preprocessing.likelihood.aicc"                  
#> [101] "preprocessing.likelihood.bic"                   
#> [102] "preprocessing.likelihood.bicc"                  
#> [103] "preprocessing.likelihood.ser"                   
#> [104] "preprocessing.likelihood.ser-ml"                
#> [105] "preprocessing.residuals.res"                    
#> [106] "preprocessing.residuals.mean"                   
#> [107] "preprocessing.residuals.skewness"               
#> [108] "preprocessing.residuals.kurtosis"               
#> [109] "preprocessing.residuals.dh"                     
#> [110] "preprocessing.residuals.lb"                     
#> [111] "preprocessing.residuals.lb2"                    
#> [112] "preprocessing.residuals.seaslb"                 
#> [113] "preprocessing.residuals.bp"                     
#> [114] "preprocessing.residuals.bp2"                    
#> [115] "preprocessing.residuals.seasbp"                 
#> [116] "preprocessing.residuals.nruns"                  
#> [117] "preprocessing.residuals.lruns"                  
#> [118] "mstats.M(*)"                                    
#> [119] "mstats.Q"                                       
#> [120] "mstats.Q-M2"                                    
#> [121] "decomposition.a1"                               
#> [122] "decomposition.a1a"                              
#> [123] "decomposition.a1b"                              
#> [124] "decomposition.a6"                               
#> [125] "decomposition.a7"                               
#> [126] "decomposition.a8"                               
#> [127] "decomposition.a8t"                              
#> [128] "decomposition.a8s"                              
#> [129] "decomposition.a8i"                              
#> [130] "decomposition.a9"                               
#> [131] "decomposition.a9sa"                             
#> [132] "decomposition.a9u"                              
#> [133] "decomposition.a9ser"                            
#> [134] "decomposition.b1"                               
#> [135] "decomposition.b2"                               
#> [136] "decomposition.b3"                               
#> [137] "decomposition.b4"                               
#> [138] "decomposition.b5"                               
#> [139] "decomposition.b6"                               
#> [140] "decomposition.b7"                               
#> [141] "decomposition.b8"                               
#> [142] "decomposition.b9"                               
#> [143] "decomposition.b10"                              
#> [144] "decomposition.b11"                              
#> [145] "decomposition.b12"                              
#> [146] "decomposition.b13"                              
#> [147] "decomposition.b14"                              
#> [148] "decomposition.b15"                              
#> [149] "decomposition.b16"                              
#> [150] "decomposition.b17"                              
#> [151] "decomposition.b18"                              
#> [152] "decomposition.b19"                              
#> [153] "decomposition.b20"                              
#> [154] "decomposition.c1"                               
#> [155] "decomposition.c2"                               
#> [156] "decomposition.c3"                               
#> [157] "decomposition.c4"                               
#> [158] "decomposition.c5"                               
#> [159] "decomposition.c6"                               
#> [160] "decomposition.c7"                               
#> [161] "decomposition.c8"                               
#> [162] "decomposition.c9"                               
#> [163] "decomposition.c10"                              
#> [164] "decomposition.c11"                              
#> [165] "decomposition.c12"                              
#> [166] "decomposition.c13"                              
#> [167] "decomposition.c14"                              
#> [168] "decomposition.c15"                              
#> [169] "decomposition.c16"                              
#> [170] "decomposition.c17"                              
#> [171] "decomposition.c18"                              
#> [172] "decomposition.c19"                              
#> [173] "decomposition.c20"                              
#> [174] "decomposition.d1"                               
#> [175] "decomposition.d2"                               
#> [176] "decomposition.d3"                               
#> [177] "decomposition.d4"                               
#> [178] "decomposition.d5"                               
#> [179] "decomposition.d6"                               
#> [180] "decomposition.d7"                               
#> [181] "decomposition.d8"                               
#> [182] "decomposition.d9"                               
#> [183] "decomposition.d10"                              
#> [184] "decomposition.d10a"                             
#> [185] "decomposition.d10b"                             
#> [186] "decomposition.d11"                              
#> [187] "decomposition.d11a"                             
#> [188] "decomposition.d12"                              
#> [189] "decomposition.d12a"                             
#> [190] "decomposition.d13"                              
#> [191] "decomposition.d14"                              
#> [192] "decomposition.d15"                              
#> [193] "decomposition.d16"                              
#> [194] "decomposition.d16a"                             
#> [195] "decomposition.d16b"                             
#> [196] "decomposition.d18"                              
#> [197] "decomposition.d19"                              
#> [198] "decomposition.d20"                              
#> [199] "decomposition.e1"                               
#> [200] "decomposition.e2"                               
#> [201] "decomposition.e3"                               
#> [202] "decomposition.e11"                              
#> [203] "decomposition.d9filter"                         
#> [204] "decomposition.slen"                             
#> [205] "decomposition.d12filter"                        
#> [206] "decomposition.tlen"                             
#> [207] "diagnostics.qs"                                 
#> [208] "diagnostics.ftest"                              
#> [209] "diagnostics.qs.on.i"                            
#> [210] "diagnostics.ftest.on.i"                         
#> [211] "diagnostics.combined.all.kruskalwallis"         
#> [212] "diagnostics.combined.all.stable"                
#> [213] "diagnostics.combined.all.evolutive"             
#> [214] "diagnostics.combined.all.summary"               
#> [215] "diagnostics.combined.all.stable.ssm"            
#> [216] "diagnostics.combined.all.stable.ssr"            
#> [217] "diagnostics.combined.all.stable.ssq"            
#> [218] "diagnostics.combined.all.evolutive.ssm"         
#> [219] "diagnostics.combined.all.evolutive.ssr"         
#> [220] "diagnostics.combined.all.evolutive.ssq"         
#> [221] "diagnostics.combined.end.kruskalwallis"         
#> [222] "diagnostics.combined.end.stable"                
#> [223] "diagnostics.combined.end.evolutive"             
#> [224] "diagnostics.combined.end.summary"               
#> [225] "diagnostics.combined.end.stable.ssm"            
#> [226] "diagnostics.combined.end.stable.ssr"            
#> [227] "diagnostics.combined.end.stable.ssq"            
#> [228] "diagnostics.combined.end.evolutive.ssm"         
#> [229] "diagnostics.combined.end.evolutive.ssr"         
#> [230] "diagnostics.combined.end.evolutive.ssq"         
#> [231] "diagnostics.combined.residual.all.kruskalwallis"
#> [232] "diagnostics.combined.residual.all.stable"       
#> [233] "diagnostics.combined.residual.all.evolutive"    
#> [234] "diagnostics.combined.residual.all.summary"      
#> [235] "diagnostics.combined.residual.all.stable.ssm"   
#> [236] "diagnostics.combined.residual.all.stable.ssr"   
#> [237] "diagnostics.combined.residual.all.stable.ssq"   
#> [238] "diagnostics.combined.residual.all.evolutive.ssm"
#> [239] "diagnostics.combined.residual.all.evolutive.ssr"
#> [240] "diagnostics.combined.residual.all.evolutive.ssq"
#> [241] "diagnostics.combined.residual.end.kruskalwallis"
#> [242] "diagnostics.combined.residual.end.stable"       
#> [243] "diagnostics.combined.residual.end.evolutive"    
#> [244] "diagnostics.combined.residual.end.summary"      
#> [245] "diagnostics.combined.residual.end.stable.ssm"   
#> [246] "diagnostics.combined.residual.end.stable.ssr"   
#> [247] "diagnostics.combined.residual.end.stable.ssq"   
#> [248] "diagnostics.combined.residual.end.evolutive.ssm"
#> [249] "diagnostics.combined.residual.end.evolutive.ssr"
#> [250] "diagnostics.combined.residual.end.evolutive.ssq"
#> [251] "diagnostics.residual.all"                       
#> [252] "diagnostics.residual.end"                       
#> [253] "diagnostics.residualtd"                         
#> [254] "diagnostics.residualtd.on.i"                    
#> [255] "diagnostics.variancedecomposition"              
#> [256] "diagnostics.logstat"                            
#> [257] "diagnostics.levelstat"                          
#> [258] "diagnostics.fcast-insample-mean"                
#> [259] "diagnostics.fcast-outsample-mean"               
#> [260] "diagnostics.fcast-outsample-variance"           
#> [261] "diagnostics.seas-lin-f"                         
#> [262] "diagnostics.seas-lin-qs"                        
#> [263] "diagnostics.seas-lin-kw"                        
#> [264] "diagnostics.seas-lin-friedman"                  
#> [265] "diagnostics.seas-lin-periodogram"               
#> [266] "diagnostics.seas-lin-spectralpeaks"             
#> [267] "diagnostics.seas-si-combined"                   
#> [268] "diagnostics.seas-si-evolutive"                  
#> [269] "diagnostics.seas-si-stable"                     
#> [270] "diagnostics.seas-res-f"                         
#> [271] "diagnostics.seas-res-qs"                        
#> [272] "diagnostics.seas-res-kw"                        
#> [273] "diagnostics.seas-res-friedman"                  
#> [274] "diagnostics.seas-res-periodogram"               
#> [275] "diagnostics.seas-res-spectralpeaks"             
#> [276] "diagnostics.seas-res-combined"                  
#> [277] "diagnostics.seas-res-combined3"                 
#> [278] "diagnostics.seas-res-evolutive"                 
#> [279] "diagnostics.seas-res-stable"                    
#> [280] "diagnostics.seas-i-f"                           
#> [281] "diagnostics.seas-i-qs"                          
#> [282] "diagnostics.seas-i-kw"                          
#> [283] "diagnostics.seas-i-periodogram"                 
#> [284] "diagnostics.seas-i-spectralpeaks"               
#> [285] "diagnostics.seas-i-combined"                    
#> [286] "diagnostics.seas-i-combined3"                   
#> [287] "diagnostics.seas-i-evolutive"                   
#> [288] "diagnostics.seas-i-stable"                      
#> [289] "diagnostics.seas-sa-f"                          
#> [290] "diagnostics.seas-sa-qs"                         
#> [291] "diagnostics.seas-sa-kw"                         
#> [292] "diagnostics.seas-sa-friedman"                   
#> [293] "diagnostics.seas-sa-periodogram"                
#> [294] "diagnostics.seas-sa-spectralpeaks"              
#> [295] "diagnostics.seas-sa-combined"                   
#> [296] "diagnostics.seas-sa-combined3"                  
#> [297] "diagnostics.seas-sa-evolutive"                  
#> [298] "diagnostics.seas-sa-stable"                     
#> [299] "diagnostics.seas-sa-ac1"                        
#> [300] "diagnostics.td-sa-all"                          
#> [301] "diagnostics.td-sa-last"                         
#> [302] "diagnostics.td-i-all"                           
#> [303] "diagnostics.td-i-last"                          
#> [304] "diagnostics.td-res-all"                         
#> [305] "diagnostics.td-res-last"                        
#> [306] "diagnostics.ic-ratio-henderson"                 
#> [307] "diagnostics.ic-ratio"                           
#> [308] "diagnostics.msr-global"                         
#> [309] "diagnostics.msr(*)"                             

get_indicators(mysa, "decomposition.b2", "decomposition.d10")

#> $decomposition.b2
#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1990                                                              72.81606
#> 1991  71.79743  71.68024  71.64589  71.55773  71.52495  71.56958  71.49960
#> 1992  71.27281  71.22519  71.03194  70.76615  70.54668  70.32064  70.02536
#> 1993  67.38024  67.09041  66.83558  66.60880  66.29927  66.01210  65.90924
#> 1994  67.05880  67.40160  67.75609  68.18424  68.76302  69.43857  70.01124
#> 1995  72.30924  72.51606  72.66661  72.79874  72.77725  72.77547  72.81410
#> 1996  72.75332  72.73698  72.85085  72.92897  73.00902  73.06096  73.06031
#> 1997  74.92023  75.42547  75.89510  76.49960  77.13115  77.70993  78.40107
#> 1998  81.44960  81.73007  81.89219  82.07607  82.27505  82.43690  82.47964
#> 1999  83.10107  83.24874  83.50225  83.89630  84.31177  84.69126  85.07174
#> 2000  88.09186  88.54981  88.91735  89.22763  89.55159  89.92536  90.30749
#> 2001  91.55291  91.66626  91.83948  91.83076  91.66442  91.30507  90.99106
#> 2002  90.14326  89.99274  89.90005  89.80356  89.75411  89.67374  89.49543
#> 2003  88.47023  88.28380  88.06177  88.03710  88.00198  87.98076  88.04690
#> 2004  89.05782  89.19430  89.38909  89.64202  89.77313  89.84781  90.01269
#> 2005  89.97964  89.94326  90.04359  89.96757  89.95885  90.13857  90.13208
#> 2006  91.05924  91.29523  91.42078  91.57791  91.68975  91.71297  91.85160
#> 2007  92.83624  93.07043  93.16865  93.25160  93.40609  93.45507  93.54523
#> 2008  94.13732  94.27705  94.36816  94.27784  94.15919  94.18357  94.21296
#> 2009  93.57673  93.80434  94.15943  94.44461  94.80983  95.11117  95.29450
#> 2010  97.65343  97.93443  98.14211  98.40283  98.65579  98.97279  99.46577
#> 2011 102.08200 102.32896 102.49595 102.62410 102.84038 102.99510 102.91777
#> 2012 101.36332 101.32517 101.28227 101.00568 100.54285 100.17339  99.92300
#> 2013  99.15396  98.93260  98.73158  98.71426  98.81064  98.83860  98.83529
#> 2014  98.73092  98.72783  98.74329  98.72279  98.59184  98.56229  98.62843
#> 2015  99.12696  99.21160  99.39455  99.56176  99.82741 100.01426 100.16950
#> 2016 100.24438 100.23485 100.22618 100.14839 100.17180 100.22684 100.18832
#> 2017 101.19493 101.46842 101.67731 102.11360 102.60044 102.93200 103.27099
#> 2018 105.06921 105.32232 105.59445 105.77455 105.89967 106.09710          
#>            Aug       Sep       Oct       Nov       Dec
#> 1990  72.72429  72.54606  72.36152  72.05376  71.88948
#> 1991  71.38952  71.32788  71.33517  71.39619  71.35230
#> 1992  69.65136  69.18076  68.67692  68.24257  67.77135
#> 1993  65.90367  65.94523  66.06661  66.27791  66.65225
#> 1994  70.51771  71.03293  71.48531  71.79743  72.03109
#> 1995  72.74643  72.70036  72.67730  72.71447  72.80074
#> 1996  73.15969  73.45899  73.85027  74.15107  74.46255
#> 1997  79.07560  79.56982  79.96083  80.43380  80.94510
#> 1998  82.50165  82.56826  82.69776  82.79257  82.93385
#> 1999  85.53902  86.09781  86.63408  87.14499  87.64049
#> 2000  90.64105  91.02190  91.30422  91.41606  91.51245
#> 2001  90.82429  90.66689  90.59069  90.47876  90.29781
#> 2002  89.37314  89.24282  89.12577  88.93857  88.64172
#> 2003  88.13000  88.17832  88.21966  88.40853  88.74027
#> 2004  90.13724  90.09082  90.04970  90.08208  90.09808
#> 2005  90.04688  90.13318  90.26473  90.44326  90.76859
#> 2006  92.08071  92.31774  92.48339  92.65773  92.73328
#> 2007  93.71652  93.81245  93.87336  93.93365  93.99503
#> 2008  94.21345  94.05591  93.87489  93.75417  93.64177
#> 2009  95.49151  95.88312  96.41678  96.94077  97.31955
#> 2010 100.12270 100.71009 101.07314 101.45943 101.83794
#> 2011 102.62442 102.34351 102.23238 101.96467 101.57805
#> 2012  99.71433  99.51925  99.34305  99.27279  99.24601
#> 2013  98.90050  98.97851  98.99896  98.86593  98.72741
#> 2014  98.61963  98.65943  98.75461  98.89033  99.08496
#> 2015 100.34900 100.39775 100.36530 100.39249 100.35395
#> 2016 100.14840 100.25220 100.43375 100.61646 100.87928
#> 2017 103.63979 103.96288 104.26253 104.55039 104.80925
#> 2018                                                  
#> 
#> $decomposition.d10
#>               Jan          Feb          Mar          Apr          May
#> 1990  -3.43358181  -2.79298747   7.08534396   1.15547739  -3.15912292
#> 1991  -3.38973288  -2.86104178   6.94156257   1.15454975  -3.10124288
#> 1992  -3.28924983  -3.02463814   6.74631613   1.14998270  -3.12045113
#> 1993  -3.20004794  -3.25185274   6.54683478   1.13249128  -3.07979167
#> 1994  -3.22613188  -3.48248715   6.45424261   1.13158590  -3.14237932
#> 1995  -3.34699418  -3.66714193   6.47060341   1.07583133  -3.12235496
#> 1996  -3.49228704  -3.85977963   6.68000234   0.98777740  -3.22776804
#> 1997  -3.60522411  -4.07767133   7.01262728   0.83957060  -3.19720540
#> 1998  -3.75318599  -4.22302426   7.56107921   0.69707690  -3.22935820
#> 1999  -3.82120522  -4.22766264   8.21618404   0.57735011  -3.23654940
#> 2000  -3.84140732  -4.11717526   9.01090077   0.65738366  -3.49096106
#> 2001  -3.84191645  -3.89071887   9.64355061   0.88112428  -3.80274730
#> 2002  -3.86778789  -3.65405953  10.05930721   1.21886237  -4.10935198
#> 2003  -3.88858662  -3.53206090  10.07825045   1.45242285  -4.13464469
#> 2004  -3.89788703  -3.57201707   9.86750554   1.53069825  -3.86110323
#> 2005  -3.88893870  -3.65843012   9.48950368   1.35924211  -3.32582926
#> 2006  -3.74565774  -3.69203207   9.12288933   1.04668756  -2.76282900
#> 2007  -3.57832988  -3.71553140   8.86929130   0.80580254  -2.29601083
#> 2008  -3.33383519  -3.59070066   8.73239526   0.82722816  -2.09485285
#> 2009  -3.15670299  -3.42854170   8.66822753   1.08952294  -2.11969725
#> 2010  -3.01087959  -3.20402352   8.57843931   1.53013436  -2.25048607
#> 2011  -3.12280609  -3.04232609   8.55361380   2.01859496  -2.35521385
#> 2012  -3.30382307  -2.82394045   8.58316526   2.42080797  -2.52142275
#> 2013  -3.45660000  -2.65220181   8.63728171   2.63887860  -2.68066152
#> 2014  -3.46391762  -2.56932638   8.73680863   2.61181107  -2.79209786
#> 2015  -3.41326157  -2.59518040   8.87002893   2.44504925  -2.78918835
#> 2016  -3.28221865  -2.70285431   9.07330749   2.22439199  -2.76573298
#> 2017  -3.11552030  -2.82441560   9.21538185   2.03400186  -2.71091112
#>               Jun          Jul          Aug          Sep          Oct
#> 1990  10.49951731  -1.63339905 -26.84050439   5.77297230   9.27917342
#> 1991  10.50094865  -1.49653032 -26.81587901   5.73136912   9.22661744
#> 1992  10.51294649  -1.20786150 -26.76878406   5.67486280   9.09029874
#> 1993  10.48075785  -0.83251540 -26.75860820   5.60469322   8.92993944
#> 1994  10.49422821  -0.45118417 -26.87818678   5.63462666   8.90934652
#> 1995  10.57384811  -0.08855159 -27.26017730   5.68872234   8.98588392
#> 1996  10.78809784   0.18699742 -27.81160551   5.84101663   9.18957132
#> 1997  11.04750759   0.40950708 -28.41348859   5.91593721   9.37841396
#> 1998  11.34981381   0.49502718 -28.95631652   6.00966513   9.66306265
#> 1999  11.62613706   0.50417547 -29.42785255   6.15095519   9.78464384
#> 2000  11.73925810   0.45079186 -29.72454625   6.41297342   9.93271869
#> 2001  11.75261560   0.39812001 -29.89969837   6.81478679  10.00414769
#> 2002  11.74467585   0.26974944 -29.97939291   7.27733039  10.15166538
#> 2003  11.86552757   0.18222396 -30.06281652   7.77904040  10.13642328
#> 2004  11.97851731   0.21120961 -30.02564517   8.06062883  10.14322028
#> 2005  12.15869941   0.38845330 -29.87613816   8.15702299   9.96237025
#> 2006  12.21686454   0.62862980 -29.48131670   8.05429784   9.67204131
#> 2007  12.06604912   0.92080807 -28.87821963   7.91988889   9.11229452
#> 2008  11.52046007   1.29734548 -28.10164930   7.60833142   8.63499734
#> 2009  10.78815013   1.72886315 -27.30101643   7.39530178   8.05111660
#> 2010  10.05650230   1.98531661 -26.52189418   7.14150110   7.62276838
#> 2011   9.47080223   2.11822913 -25.88967683   6.97824123   7.18499463
#> 2012   9.16013327   1.96654892 -25.32602810   6.71025439   7.03235621
#> 2013   9.02452500   1.66049064 -24.85691676   6.57446426   6.86066742
#> 2014   9.04306286   1.15969912 -24.44743516   6.37359795   6.89526551
#> 2015   8.99577499   0.71506334 -24.15851931   6.22463886   6.97110028
#> 2016   8.95565407   0.30973812 -23.93881234   6.00031998   7.19182379
#> 2017   8.83465222   0.08980872 -23.80461653   5.89143898   7.32997666
#>               Nov          Dec
#> 1990   2.64044096   1.46417473
#> 1991   2.64217676   1.56389728
#> 1992   2.61562779   1.76896549
#> 1993   2.56632661   2.07398830
#> 1994   2.40812856   2.34240502
#> 1995   2.33271806   2.51796293
#> 1996   2.29831587   2.46958234
#> 1997   2.37867358   2.21884553
#> 1998   2.33795467   1.67993279
#> 1999   2.36130737   0.91174553
#> 2000   2.26210749   0.05515349
#> 2001   2.12493560  -0.76138566
#> 2002   1.91440985  -1.37661506
#> 2003   1.77144148  -1.80077824
#> 2004   1.59028084  -2.07321344
#> 2005   1.41494197  -2.31457950
#> 2006   1.34424536  -2.58715497
#> 2007   1.36745029  -2.95402241
#> 2008   1.45503749  -3.30266398
#> 2009   1.44373201  -3.50721786
#> 2010   1.48426543  -3.55091968
#> 2011   1.40200885  -3.40929912
#> 2012   1.33393308  -3.24003971
#> 2013   1.33348139  -3.06132289
#> 2014   1.60011132  -3.05589201
#> 2015   1.94332424  -3.14822333
#> 2016   2.28758920  -3.31650128
#> 2017   2.52680901  -3.43568125
#> 

# To convert to the R object
jSA2R(mysa)

#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5270      0.048
#> BTheta(1)  -0.4865      0.051
#> 
#>               Estimate Std. Error
#> Monday       -0.133839      0.164
#> Tuesday      -0.002384      0.163
#> Wednesday     0.241712      0.163
#> Thursday     -0.531275      0.163
#> Friday        0.432474      0.164
#> Saturday      0.152956      0.163
#> Leap year    -0.045977      0.501
#> Easter [1]   -1.094082      0.335
#> LS (11-2008) -8.441602      1.307
#> LS (1-2009)  -7.274012      1.306
#> LS (5-2008)  -5.020079      1.257
#> 
#> 
#> Residual standard error: 1.665 on 323 degrees of freedom
#> Log likelihood = -624.7, aic =  1277 aicc =  1279, bic(corrected for length) = 1.252
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.061
#> M(2)    0.034
#> M(3)    0.840
#> M(4)    0.420
#> M(5)    0.697
#> M(6)    0.236
#> M(7)    0.075
#> M(8)    0.206
#> M(9)    0.055
#> M(10)   0.166
#> M(11)   0.137
#> Q       0.272
#> Q-M2    0.301
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t            s          i
#> Jan 2017  97.4 100.8114 100.6915  -3.41138724  0.1198412
#> Feb 2017  97.5 100.3129 101.0507  -2.81292129 -0.7377402
#> Mar 2017 112.0 102.3507 101.5023   9.64931855  0.8483317
#> Apr 2017 103.0 101.2637 101.9608   1.73628825 -0.6971084
#> May 2017 100.4 103.0054 102.3694  -2.60542262  0.6359951
#> Jun 2017 111.2 102.4641 102.7599   8.73585142 -0.2957736
#> Jul 2017 103.4 103.4507 103.1325  -0.05071812  0.3182150
#> Aug 2017  79.3 103.3966 103.5091 -24.09656399 -0.1125714
#> Sep 2017 109.7 103.2231 103.8871   6.47686943 -0.6640036
#> Oct 2017 114.0 106.9659 104.2566   7.03410972  2.7093305
#> Nov 2017 107.7 105.4628 104.5657   2.23724550  0.8970898
#> Dec 2017 101.4 104.4099 104.8237  -3.00989466 -0.4138486
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f         i_f
#> Jan 2018 102.00877 104.9065 105.0038  -2.8977705 -0.09720660
#> Feb 2018 102.14378 105.0529 105.1506  -2.9091625 -0.09761841
#> Mar 2018 113.92764 105.4072 105.3021   8.5204140  0.10515354
#> Apr 2018 107.88844 105.4548 105.4869   2.4336221 -0.03210021
#> May 2018 102.83006 105.7518 105.7343  -2.9217368  0.01749135
#> Jun 2018 115.26921 105.9309 106.0232   9.3383460 -0.09234653
#> Jul 2018 106.09888 106.4237 106.3248  -0.3248322  0.09889228
#> Aug 2018  82.95525 106.5440 106.6390 -23.5887165 -0.09503822
#> Sep 2018 112.41847 106.6179 106.9320   5.8005883 -0.31411494
#> Oct 2018 115.41327 107.8929 107.1812   7.5203393  0.71172901
#> Nov 2018 109.88167 107.2964 107.3766   2.5852407 -0.08013772
#> Dec 2018 103.58107 107.2605 107.5126  -3.6794200 -0.25212937
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.646
#>  Seasonal     49.667
#>  Irregular     0.411
#>  TD & Hol.     0.057
#>  Others       49.814
#>  Total       101.596
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.013
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.980
#>  f-test on i (seasonal dummies)         0.922
#>  Residual seasonality (entire series)   0.969
#>  Residual seasonality (last 3 years)    0.897
#>  f-test on sa (td)                      0.980
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables