Saving and loading a model specification, SA and pre-adjustment in X13 and TRAMO-SEATS

save_spec saves a SA or RegARIMA model specification. load_spec loads the previously saved model specification.

save_spec(object, file = file.path(tempdir(), "spec.RData"))

load_spec(file = "spec.RData")

Arguments

object
object of one of the classes: c("SA_spec","X13"), c("SA_spec","TRAMO_SEATS"), c("SA","X13"), c("SA","TRAMO_SEATS"), c("regarima_spec","X13"), c("regarima_spec","TRAMO_SEATS"), c("regarima","X13"), c("regarima","TRAMO_SEATS").

file
(path and) name of the file where the model specification will be saved (have been saved).

Value

load_spec returns an object of class "SA_spec" or "regarima_spec".

Details

save_spec saves the final model specification of a "SA_spec", "SA", "regarima_spec" or "regarima" class object. load_spec loads the previously saved model specification. It creates a c("SA_spec","X13"), c("sA_spec","TRAMO_SEATS"), c("regarima_spec","X13") or c("regarima_spec","TRAMO_SEATS") class object, in line with the class of the previously saved model specification.

References

Info on JDemtra+, usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en

Examples

# \donttest{
myseries <- ipi_c_eu[, "FR"]
myreg1 <- regarima_x13(myseries, spec = "RG5c")
myspec2 <- regarima_spec_x13(myreg1, estimate.from = "2005-10-01", outlier.from = "2010-03-01")
myreg2 <- regarima(myseries, myspec2)

myreg3 <- regarima_tramoseats(myseries, spec = "TRfull")
myspec4 <-regarima_spec_tramoseats(myreg3, tradingdays.mauto = "Unused",
                                  tradingdays.option ="WorkingDays",
                                  easter.type = "Standard",
                                  automdl.enabled = FALSE, arima.mu = TRUE)
myreg4 <-regarima(myseries, myspec4)

myspec6 <- x13_spec("RSA5c")
mysa6 <- x13(myseries, myspec6)

myspec7 <- tramoseats_spec("RSAfull")
mysa7 <- tramoseats(myseries, myspec7)

dir <- tempdir()

 # Save the model specification from a c("regarima_spec","X13") class object
save_spec(myspec2, file.path(dir, "specx13.RData"))
 # Save the model specification from a c("regarima","X13") class object
save_spec(myreg2, file.path(dir,"regx13.RData"))
 # Save the model specification from a c("regarima_spec","TRAMO_SEATS") class object
save_spec(myspec4, file.path(dir,"specTS.RData"))
 # Save the model specification from a c("regarima","TRAMO_SEATS") class object
save_spec(myreg4, file.path(dir,"regTS.RData"))
 # Save model from a c("SA_spec","X13") class object
save_spec(myspec6, file.path(dir,"specFullx13.RData"))
 # Save model from a c("SA","X13") class object
save_spec(mysa6, file.path(dir,"sax13.RData"))
 # Save model from a c("SA_spec","TRAMO_SEATS") class object
save_spec(myspec7, file.path(dir,"specFullTS.RData"))
 # Save model from a c("SA","TRAMO_SEATS") class object
save_spec(mysa7, file.path(dir,"saTS.RData"))

 # Load the model specification
myspec2a <- load_spec(file.path(dir,"specx13.RData"))
myspec2b <- load_spec(file.path(dir,"regx13.RData"))
myspec4a <- load_spec(file.path(dir,"specTS.RData"))
myspec4b <- load_spec(file.path(dir,"regTS.RData"))
myspec6a <- load_spec(file.path(dir,"specFullx13.RData"))
myspec6b <- load_spec(file.path(dir,"sax13.RData"))
myspec7a <- load_spec(file.path(dir,"specFullTS.RData"))
myspec7b <- load_spec(file.path(dir,"saTS.RData"))

regarima(myseries, myspec2a)

#> y = regression model + arima (1, 1, 0, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.2650      0.084
#> BTheta(1)  -0.6913      0.067
#> 
#>            Estimate Std. Error
#> Easter [1] -0.01809      0.006
#> 
#> 
#> Residual standard error: 0.01978 on 130 degrees of freedom
#> Log likelihood = 331.6, aic = 583.3 aicc = 583.6, bic(corrected for length) = -7.737
#> 

x13(myseries, myspec6a)

#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5270      0.048
#> BTheta(1)  -0.4865      0.051
#> 
#>               Estimate Std. Error
#> Monday       -0.133839      0.164
#> Tuesday      -0.002384      0.163
#> Wednesday     0.241712      0.163
#> Thursday     -0.531275      0.163
#> Friday        0.432474      0.164
#> Saturday      0.152956      0.163
#> Leap year    -0.045977      0.501
#> Easter [1]   -1.094082      0.335
#> LS (11-2008) -8.441602      1.307
#> LS (1-2009)  -7.274012      1.306
#> LS (5-2008)  -5.020079      1.257
#> 
#> 
#> Residual standard error: 1.665 on 309 degrees of freedom
#> Log likelihood = -624.7, aic =  1277 aicc =  1279, bic(corrected for length) = 1.252
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.061
#> M(2)    0.034
#> M(3)    0.840
#> M(4)    0.420
#> M(5)    0.697
#> M(6)    0.236
#> M(7)    0.075
#> M(8)    0.206
#> M(9)    0.055
#> M(10)   0.166
#> M(11)   0.137
#> Q       0.272
#> Q-M2    0.301
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t            s          i
#> Jan 2017  97.4 100.8114 100.6915  -3.41138724  0.1198412
#> Feb 2017  97.5 100.3129 101.0507  -2.81292129 -0.7377402
#> Mar 2017 112.0 102.3507 101.5023   9.64931855  0.8483317
#> Apr 2017 103.0 101.2637 101.9608   1.73628825 -0.6971084
#> May 2017 100.4 103.0054 102.3694  -2.60542262  0.6359951
#> Jun 2017 111.2 102.4641 102.7599   8.73585142 -0.2957736
#> Jul 2017 103.4 103.4507 103.1325  -0.05071812  0.3182150
#> Aug 2017  79.3 103.3966 103.5091 -24.09656399 -0.1125714
#> Sep 2017 109.7 103.2231 103.8871   6.47686943 -0.6640036
#> Oct 2017 114.0 106.9659 104.2566   7.03410972  2.7093305
#> Nov 2017 107.7 105.4628 104.5657   2.23724550  0.8970898
#> Dec 2017 101.4 104.4099 104.8237  -3.00989466 -0.4138486
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f         i_f
#> Jan 2018 102.00877 104.9065 105.0038  -2.8977705 -0.09720660
#> Feb 2018 102.14378 105.0529 105.1506  -2.9091625 -0.09761841
#> Mar 2018 113.92764 105.4072 105.3021   8.5204140  0.10515354
#> Apr 2018 107.88844 105.4548 105.4869   2.4336221 -0.03210021
#> May 2018 102.83006 105.7518 105.7343  -2.9217368  0.01749135
#> Jun 2018 115.26921 105.9309 106.0232   9.3383460 -0.09234653
#> Jul 2018 106.09888 106.4237 106.3248  -0.3248322  0.09889228
#> Aug 2018  82.95525 106.5440 106.6390 -23.5887165 -0.09503822
#> Sep 2018 112.41847 106.6179 106.9320   5.8005883 -0.31411494
#> Oct 2018 115.41327 107.8929 107.1812   7.5203393  0.71172901
#> Nov 2018 109.88167 107.2964 107.3766   2.5852407 -0.08013772
#> Dec 2018 103.58107 107.2605 107.5126  -3.6794200 -0.25212937
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.646
#>  Seasonal     49.667
#>  Irregular     0.411
#>  TD & Hol.     0.057
#>  Others       49.814
#>  Total       101.596
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.013
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.980
#>  f-test on i (seasonal dummies)         0.922
#>  Residual seasonality (entire series)   0.969
#>  Residual seasonality (last 3 years)    0.897
#>  f-test on sa (td)                      0.980
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables

tramoseats(myseries, myspec7a)

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.3525      0.256
#> Phi(2)      0.1336      0.130
#> Theta(1)   -0.1868      0.256
#> BTheta(1)  -0.6109      0.047
#> 
#>                Estimate Std. Error
#> Monday       -0.0022941      0.002
#> Tuesday       0.0005841      0.002
#> Wednesday     0.0032202      0.002
#> Thursday     -0.0054984      0.002
#> Friday        0.0033421      0.002
#> Saturday      0.0018654      0.002
#> Easter [6]   -0.0110283      0.003
#> LS (11-2008) -0.0794304      0.013
#> LS (1-2009)  -0.0711852      0.013
#> AO (8-2001)   0.0447084      0.013
#> 
#> 
#> Residual standard error: 0.01593 on 308 degrees of freedom
#> Log likelihood = 875.8, aic =  1261 aicc =  1263, bic(corrected for length) = -8.028
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 + 0.352498 B + 0.133616 B^2 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.186819 B - 0.610856 B^12 + 0.114119 B^13 
#> 
#> 
#> SA
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.314459 B + 0.340427 B^2 
#> Innovation variance:  0.4669153 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.040206 B - 0.959794 B^2 
#> Innovation variance:  0.04869563 
#> 
#> Seasonal
#> AR :  1 + 0.352498 B + 0.133616 B^2 
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 0.717848 B + 0.460721 B^2 + 0.310085 B^3 + 0.132447 B^4 - 0.049053 B^5 - 0.216655 B^6 - 0.354556 B^7 - 0.445030 B^8 - 0.469587 B^9 - 0.376625 B^10 - 0.166397 B^11 - 0.410618 B^12 - 0.132580 B^13 
#> Innovation variance:  0.1601924 
#> 
#> Irregular
#> Innovation variance:  0.2056884 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2017  97.4 100.9393 100.9657 0.9649365 0.9997384
#> Feb 2017  97.5 100.7654 101.2610 0.9675939 0.9951058
#> Mar 2017 112.0 102.0879 101.7064 1.0970937 1.0037513
#> Apr 2017 103.0 101.8416 102.1713 1.0113745 0.9967732
#> May 2017 100.4 103.0277 102.6087 0.9744949 1.0040835
#> Jun 2017 111.2 102.7958 102.9753 1.0817564 0.9982567
#> Jul 2017 103.4 103.3099 103.3367 1.0008726 0.9997398
#> Aug 2017  79.3 103.7805 103.7458 0.7641131 1.0003339
#> Sep 2017 109.7 103.7154 104.2747 1.0577017 0.9946372
#> Oct 2017 114.0 105.6674 104.8749 1.0788572 1.0075559
#> Nov 2017 107.7 105.4730 105.1795 1.0211146 1.0027905
#> Dec 2017 101.4 104.9614 105.2317 0.9660690 0.9974316
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2018 102.19076 105.3388 105.3397 0.9701155 0.9999914
#> Feb 2018 102.32758 105.5082 105.5091 0.9698541 0.9999914
#> Mar 2018 114.29051 105.6780 105.6789 1.0814979 0.9999914
#> Apr 2018 107.91102 105.8480 105.8489 1.0194904 0.9999914
#> May 2018 102.68809 106.0183 106.0192 0.9685884 0.9999914
#> Jun 2018 115.89624 106.1889 106.1898 1.0914161 0.9999914
#> Jul 2018 106.23593 106.3597 106.3606 0.9988362 0.9999914
#> Aug 2018  81.47084 106.5308 106.5317 0.7647630 0.9999914
#> Sep 2018 112.88837 106.7022 106.7031 1.0579757 0.9999914
#> Oct 2018 115.34277 106.8739 106.8748 1.0792417 0.9999914
#> Nov 2018 109.29540 107.0458 107.0468 1.0210148 0.9999914
#> Dec 2018 103.27711 107.2181 107.2190 0.9632435 0.9999914
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.181
#>  Seasonal     66.768
#>  Irregular     0.106
#>  TD & Hol.     0.069
#>  Others       30.507
#>  Total        99.630
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.185
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.977
#>  f-test on sa (td)                      0.927
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables

regarima(myseries, myspec4a)

#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.4987      0.049
#> BTheta(1)  -0.6179      0.046
#> 
#>                Estimate Std. Error
#> Mean          9.636e-05      0.000
#> Monday       -2.104e-03      0.002
#> Tuesday       1.781e-04      0.002
#> Wednesday     3.680e-03      0.002
#> Thursday     -5.706e-03      0.002
#> Friday        4.013e-03      0.002
#> Saturday      1.271e-03      0.002
#> Easter [6]   -1.063e-02      0.003
#> LS (11-2008) -8.174e-02      0.013
#> LS (1-2009)  -7.134e-02      0.013
#> 
#> 
#> Residual standard error: 0.01633 on 310 degrees of freedom
#> Log likelihood = 867.8, aic =  1273 aicc =  1274, bic(corrected for length) = -8.015
#> 

x13(myseries, myspec6b)

#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5270      0.048
#> BTheta(1)  -0.4865      0.051
#> 
#>               Estimate Std. Error
#> Monday       -0.133839      0.164
#> Tuesday      -0.002384      0.163
#> Wednesday     0.241712      0.163
#> Thursday     -0.531275      0.163
#> Friday        0.432474      0.164
#> Saturday      0.152956      0.163
#> Leap year    -0.045977      0.501
#> Easter [1]   -1.094082      0.335
#> LS (11-2008) -8.441602      1.307
#> LS (1-2009)  -7.274012      1.306
#> LS (5-2008)  -5.020079      1.257
#> 
#> 
#> Residual standard error: 1.665 on 309 degrees of freedom
#> Log likelihood = -624.7, aic =  1277 aicc =  1279, bic(corrected for length) = 1.252
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.061
#> M(2)    0.034
#> M(3)    0.840
#> M(4)    0.420
#> M(5)    0.697
#> M(6)    0.236
#> M(7)    0.075
#> M(8)    0.206
#> M(9)    0.055
#> M(10)   0.166
#> M(11)   0.137
#> Q       0.272
#> Q-M2    0.301
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t            s          i
#> Jan 2017  97.4 100.8114 100.6915  -3.41138724  0.1198412
#> Feb 2017  97.5 100.3129 101.0507  -2.81292129 -0.7377402
#> Mar 2017 112.0 102.3507 101.5023   9.64931855  0.8483317
#> Apr 2017 103.0 101.2637 101.9608   1.73628825 -0.6971084
#> May 2017 100.4 103.0054 102.3694  -2.60542262  0.6359951
#> Jun 2017 111.2 102.4641 102.7599   8.73585142 -0.2957736
#> Jul 2017 103.4 103.4507 103.1325  -0.05071812  0.3182150
#> Aug 2017  79.3 103.3966 103.5091 -24.09656399 -0.1125714
#> Sep 2017 109.7 103.2231 103.8871   6.47686943 -0.6640036
#> Oct 2017 114.0 106.9659 104.2566   7.03410972  2.7093305
#> Nov 2017 107.7 105.4628 104.5657   2.23724550  0.8970898
#> Dec 2017 101.4 104.4099 104.8237  -3.00989466 -0.4138486
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f         i_f
#> Jan 2018 102.00877 104.9065 105.0038  -2.8977705 -0.09720660
#> Feb 2018 102.14378 105.0529 105.1506  -2.9091625 -0.09761841
#> Mar 2018 113.92764 105.4072 105.3021   8.5204140  0.10515354
#> Apr 2018 107.88844 105.4548 105.4869   2.4336221 -0.03210021
#> May 2018 102.83006 105.7518 105.7343  -2.9217368  0.01749135
#> Jun 2018 115.26921 105.9309 106.0232   9.3383460 -0.09234653
#> Jul 2018 106.09888 106.4237 106.3248  -0.3248322  0.09889228
#> Aug 2018  82.95525 106.5440 106.6390 -23.5887165 -0.09503822
#> Sep 2018 112.41847 106.6179 106.9320   5.8005883 -0.31411494
#> Oct 2018 115.41327 107.8929 107.1812   7.5203393  0.71172901
#> Nov 2018 109.88167 107.2964 107.3766   2.5852407 -0.08013772
#> Dec 2018 103.58107 107.2605 107.5126  -3.6794200 -0.25212937
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.646
#>  Seasonal     49.667
#>  Irregular     0.411
#>  TD & Hol.     0.057
#>  Others       49.814
#>  Total       101.596
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.013
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.980
#>  f-test on i (seasonal dummies)         0.922
#>  Residual seasonality (entire series)   0.969
#>  Residual seasonality (last 3 years)    0.897
#>  f-test on sa (td)                      0.980
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables

tramoseats(myseries, myspec7b)

#> RegARIMA
#> y = regression model + arima (2, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      0.3525      0.256
#> Phi(2)      0.1336      0.130
#> Theta(1)   -0.1868      0.256
#> BTheta(1)  -0.6109      0.047
#> 
#>                Estimate Std. Error
#> Monday       -0.0022941      0.002
#> Tuesday       0.0005841      0.002
#> Wednesday     0.0032202      0.002
#> Thursday     -0.0054984      0.002
#> Friday        0.0033421      0.002
#> Saturday      0.0018654      0.002
#> Easter [6]   -0.0110283      0.003
#> LS (11-2008) -0.0794304      0.013
#> LS (1-2009)  -0.0711852      0.013
#> AO (8-2001)   0.0447084      0.013
#> 
#> 
#> Residual standard error: 0.01593 on 308 degrees of freedom
#> Log likelihood = 875.8, aic =  1261 aicc =  1263, bic(corrected for length) = -8.028
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 + 0.352498 B + 0.133616 B^2 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.186819 B - 0.610856 B^12 + 0.114119 B^13 
#> 
#> 
#> SA
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.314459 B + 0.340427 B^2 
#> Innovation variance:  0.4669153 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.040206 B - 0.959794 B^2 
#> Innovation variance:  0.04869563 
#> 
#> Seasonal
#> AR :  1 + 0.352498 B + 0.133616 B^2 
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 0.717848 B + 0.460721 B^2 + 0.310085 B^3 + 0.132447 B^4 - 0.049053 B^5 - 0.216655 B^6 - 0.354556 B^7 - 0.445030 B^8 - 0.469587 B^9 - 0.376625 B^10 - 0.166397 B^11 - 0.410618 B^12 - 0.132580 B^13 
#> Innovation variance:  0.1601924 
#> 
#> Irregular
#> Innovation variance:  0.2056884 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2017  97.4 100.9393 100.9657 0.9649365 0.9997384
#> Feb 2017  97.5 100.7654 101.2610 0.9675939 0.9951058
#> Mar 2017 112.0 102.0879 101.7064 1.0970937 1.0037513
#> Apr 2017 103.0 101.8416 102.1713 1.0113745 0.9967732
#> May 2017 100.4 103.0277 102.6087 0.9744949 1.0040835
#> Jun 2017 111.2 102.7958 102.9753 1.0817564 0.9982567
#> Jul 2017 103.4 103.3099 103.3367 1.0008726 0.9997398
#> Aug 2017  79.3 103.7805 103.7458 0.7641131 1.0003339
#> Sep 2017 109.7 103.7154 104.2747 1.0577017 0.9946372
#> Oct 2017 114.0 105.6674 104.8749 1.0788572 1.0075559
#> Nov 2017 107.7 105.4730 105.1795 1.0211146 1.0027905
#> Dec 2017 101.4 104.9614 105.2317 0.9660690 0.9974316
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2018 102.19076 105.3388 105.3397 0.9701155 0.9999914
#> Feb 2018 102.32758 105.5082 105.5091 0.9698541 0.9999914
#> Mar 2018 114.29051 105.6780 105.6789 1.0814979 0.9999914
#> Apr 2018 107.91102 105.8480 105.8489 1.0194904 0.9999914
#> May 2018 102.68809 106.0183 106.0192 0.9685884 0.9999914
#> Jun 2018 115.89624 106.1889 106.1898 1.0914161 0.9999914
#> Jul 2018 106.23593 106.3597 106.3606 0.9988362 0.9999914
#> Aug 2018  81.47084 106.5308 106.5317 0.7647630 0.9999914
#> Sep 2018 112.88837 106.7022 106.7031 1.0579757 0.9999914
#> Oct 2018 115.34277 106.8739 106.8748 1.0792417 0.9999914
#> Nov 2018 109.29540 107.0458 107.0468 1.0210148 0.9999914
#> Dec 2018 103.27711 107.2181 107.2190 0.9632435 0.9999914
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.181
#>  Seasonal     66.768
#>  Irregular     0.106
#>  TD & Hol.     0.069
#>  Others       30.507
#>  Total        99.630
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.185
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.977
#>  f-test on sa (td)                      0.927
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables

# }