Access model specification, SA and pre-adjustment in X13 and TRAMO-SEATS

Below functions access different parts of the final model specification as included in the "SA", "regarima", "SA_spec" and "regarima_spec" S3 class objects.

s_estimate(object = NA)

s_transform(object = NA)

s_usrdef(object = NA)

s_preOut(object = NA)

s_preVar(object = NA)

s_td(object = NA)

s_easter(object = NA)

s_out(object = NA)

s_arima(object = NA)

s_arimaCoef(object = NA)

s_fcst(object = NA)

s_span(object = NA)

s_x11(object = NA)

s_seats(object = NA)

Arguments

object
object of one of the classes: c("SA","X13"), c("SA","TRAMO_SEATS"), c("SA_spec","X13"), c("SA_spec","TRAMO_SEATS"), c("regarima","X13"), c("regarima","TRAMO_SEATS"), c("regarima_spec","X13"), c("regarima_spec","TRAMO_SEATS").

Value

s_estimate returns a data.frame with the estimate variables

s_transform returns a data.frame with the transform varaibles

s_usrdef returns a data.frame with the user-defined regressors (outliers and variables) model specification, indicating if those variables are included in the model and if coefficients are pre-specified

s_preOut returns a data.frame with the pre-specified outliers

s_preVar returns a list with the information on the user-defined variables, including: series - the time series and description - data.frame with the variable type and coefficients

s_td returns a data.frame with the trading.days variables

s_easter returns a data.frame with the easter variables

s_out returns a data.frame with the outliers detection variables

s_arima returns a data.frame with the arima variables

s_arimaCoef returns a data.frame with the user-specified ARMA coefficients

s_fcst returns a data.frame with the forecast horizon

s_span returns a data.frame with the span variables

s_x11 returns a data.frame with the x11 variables

s_seats returns a data.frame with the seats variables

References

Info on JDemetra+, usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en

Examples

myseries <- ipi_c_eu[, "FR"]
myreg1 <- regarima_def_x13(myseries, spec = c("RG5c"))
myspec1 <- regarima_spec_x13(myreg1,
             estimate.from = "2005-10-01",
             outlier.from = "2010-03-01")

s_estimate(myreg1)

#>  span tolerance
#>   All     1e-07

s_estimate(myspec1)

#>             span tolerance
#>  From 2005-10-01     1e-07

s_transform(myreg1)

#>  tfunction adjust aicdiff
#>       Auto   None      -2

s_transform(myspec1)

#>  tfunction adjust aicdiff
#>       Auto   None      -2

s_usrdef(myreg1)

#>   outlier outlier.coef variables variables.coef
#> 1   FALSE        FALSE     FALSE          FALSE

s_usrdef(myspec1)

#>  outlier outlier.coef variables variables.coef
#>    FALSE        FALSE     FALSE          FALSE
# NOT RUN {
myspec2 <- regarima_spec_x13(myreg1, usrdef.outliersEnabled = TRUE,
             usrdef.outliersType = c("LS", "AO"),
             usrdef.outliersDate = c("2009-10-01", "2005-02-01"))
myreg2 <- regarima(myseries, myspec2)

s_preOut(myreg2)
s_preOut(myspec2)
# }# NOT RUN {
var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
var3 <- ts.union(var1, var2)
myspec3 <- regarima_spec_def_x13(spec = "RG5c",
                                 usrdef.varEnabled = TRUE,
                                 usrdef.var = var3)
myreg3 <- regarima(myseries, myspec3)

s_preVar(myspec3)
s_preVar(myreg3)
# }

s_td(myreg1)

#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove

s_td(myspec1)

#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove

s_easter(myreg1)

#>  enabled julian duration test
#>     TRUE  FALSE        8  Add

s_easter(myspec1)

#>  enabled julian duration test
#>     TRUE  FALSE        8  Add

s_out(myreg1)

#>  enabled span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE  All TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7

s_out(myspec1)

#>  enabled            span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE From 2010-03-01 TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7
# NOT RUN {
s_arima(myreg1)
s_arima(myspec1)
# }# NOT RUN {
myspec4 <- regarima_spec_x13(myreg1, automdl.enabled = FALSE,
             arima.coefEnabled = TRUE,
             arima.p = 1,arima.q = 1, arima.bp = 1, arima.bq = 1,
             arima.coef = rep(0.2, 4),
             arima.coefType = rep("Initial", 4))
myreg4 <- regarima(myseries, myspec4)

s_arimaCoef(myreg4)
s_arimaCoef(myspec4)
# }

s_fcst(myreg1)

#>  horizon
#>       -2

s_fcst(myspec1)

#>  horizon
#>       -2
# NOT RUN {
s_span(myreg1)
s_span(myspec1)
# }# NOT RUN {
myspec5 <- x13_spec_def(spec = c("RSA5c"), x11.seasonalComp = FALSE)
mysa5 <- x13(myseries, myspec5)

s_x11(mysa5)
s_x11(myspec5)
# }# NOT RUN {
myspec6 <- tramoseats_spec_def(spec = c("RSAfull"), seats.approx = "Noisy")
mysa6 <- tramoseats(myseries, myspec6)

s_seats(mysa6)
s_seats(mysa6)
# }