Access model specification, SA and pre-adjustment in X13 and TRAMO-SEATS

Below functions access different parts of the final model specification as included in the "SA", "regarima", "SA_spec" and "regarima_spec" S3 class objects.

s_estimate(object = NA)

s_transform(object = NA)

s_usrdef(object = NA)

s_preOut(object = NA)

s_preVar(object = NA)

s_td(object = NA)

s_easter(object = NA)

s_out(object = NA)

s_arima(object = NA)

s_arimaCoef(object = NA)

s_fcst(object = NA)

s_span(object = NA)

s_x11(object = NA)

s_seats(object = NA)

Arguments

object
object of one of the classes: c("SA","X13"), c("SA","TRAMO_SEATS"), c("SA_spec","X13"), c("SA_spec","TRAMO_SEATS"), c("regarima","X13"), c("regarima","TRAMO_SEATS"), c("regarima_spec","X13"), c("regarima_spec","TRAMO_SEATS").

Value

s_estimate returns a data.frame with the estimate variables

s_transform returns a data.frame with the transform varaibles

s_usrdef returns a data.frame with the user-defined regressors (outliers and variables) model specification, indicating if those variables are included in the model and if coefficients are pre-specified

s_preOut returns a data.frame with the pre-specified outliers

s_preVar returns a list with the information on the user-defined variables, including: series - the time series and description - data.frame with the variable type and coefficients

s_td returns a data.frame with the trading.days variables

s_easter returns a data.frame with the easter variables

s_out returns a data.frame with the outliers detection variables

s_arima returns a data.frame with the arima variables

s_arimaCoef returns a data.frame with the user-specified ARMA coefficients

s_fcst returns a data.frame with the forecast horizon

s_span returns a data.frame with the span variables

s_x11 returns a data.frame with the x11 variables

s_seats returns a data.frame with the seats variables

References

Info on 'JDemetra+', usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en

Examples

myseries <- ipi_c_eu[, "FR"]
myreg1 <- regarima_x13(myseries, spec = "RG5c")
myspec1 <- regarima_spec_x13(myreg1,
             estimate.from = "2005-10-01",
             outlier.from = "2010-03-01")

s_estimate(myreg1)

#>  preliminary.check span tolerance
#>               TRUE  All     1e-07

s_estimate(myspec1)

#>  preliminary.check            span tolerance
#>               TRUE From 2005-10-01     1e-07

s_transform(myreg1)

#>  tfunction adjust aicdiff
#>       Auto   None      -2

s_transform(myspec1)

#>  tfunction adjust aicdiff
#>       Auto   None      -2

s_usrdef(myreg1)

#>   outlier outlier.coef variables variables.coef
#> 1   FALSE        FALSE     FALSE          FALSE

s_usrdef(myspec1)

#>  outlier outlier.coef variables variables.coef
#>    FALSE        FALSE     FALSE          FALSE

myspec2 <- regarima_spec_x13(myreg1, usrdef.outliersEnabled = TRUE,
             usrdef.outliersType = c("LS", "AO"),
             usrdef.outliersDate = c("2009-10-01", "2005-02-01"))
myreg2 <- regarima(myseries, myspec2)

s_preOut(myreg2)

#>   type       date coeff
#> 1   LS 2009-10-01    NA
#> 2   AO 2005-02-01    NA

s_preOut(myspec2)

#>   type       date coeff
#> 1   LS 2009-10-01    NA
#> 2   AO 2005-02-01    NA

var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
var3 <- ts.union(var1, var2)
myspec3 <- regarima_spec_x13(spec = "RG5c",
                             usrdef.varEnabled = TRUE,
                             usrdef.var = var3)
myreg3 <- regarima(myseries, myspec3)

s_preVar(myspec3)

#> $series
#>                  var1         var2
#> Jan 1990  -0.18087120 -135.3779986
#> Feb 1990  -1.69198739   38.7122228
#> Mar 1990  -1.64743004  -96.7667515
#> Apr 1990   7.13622416  -35.8785321
#> May 1990   2.37209036  -58.6397434
#> Jun 1990 -23.22686419  161.2913478
#> Jul 1990   0.79044275   50.5671343
#> Aug 1990  -4.28806351    2.4106090
#> Sep 1990  -3.72714386   29.5359647
#> Oct 1990  -0.44918615  127.5350360
#> Nov 1990  14.01343027   20.4477180
#> Dec 1990  -3.00872112   18.6747913
#> Jan 1991  -9.65754406   70.9749842
#> Feb 1991  16.83218482   37.5846729
#> Mar 1991  -7.76550575   40.5413237
#> Apr 1991   2.47301367   33.0985753
#> May 1991  -9.67146956  159.3760511
#> Jun 1991 -10.37854416   55.5075757
#> Jul 1991 -10.53599982 -120.9478318
#> Aug 1991  -2.05507436  -22.5557951
#> Sep 1991  13.08266555  -34.2381835
#> Oct 1991   5.13721896 -173.6559496
#> Nov 1991 -13.85355230   16.3560903
#> Dec 1991 -10.01238810   29.8293513
#> Jan 1992   3.01325680   17.5870029
#> Feb 1992   4.37682054 -140.5601011
#> Mar 1992  -6.50275146  -21.5255891
#> Apr 1992 -24.19678943  -50.2680266
#> May 1992   2.75546070  168.3881702
#> Jun 1992  -2.17346388 -172.5036625
#> Jul 1992   4.20810277  -25.2554758
#> Aug 1992   5.14492552 -160.4868007
#> Sep 1992  -9.76120254  115.3376302
#> Oct 1992   8.21322767 -126.6363058
#> Nov 1992  14.87457423  -25.0127880
#> Dec 1992   5.29521792  -53.8988584
#> Jan 1993  18.85181767  -61.1159385
#> Feb 1993   7.23419932   80.7837347
#> Mar 1993  -1.15305742  135.7967557
#> Apr 1993   7.81349665  100.0928446
#> May 1993  18.02937166 -125.8705305
#> Jun 1993  -3.86689222 -139.5429138
#> Jul 1993  27.15540230   32.6775093
#> Aug 1993  12.49770616  -56.6269784
#> Sep 1993 -19.56745628   -4.0093207
#> Oct 1993   6.66756093  100.7425936
#> Nov 1993  -5.36932307 -137.3520862
#> Dec 1993  11.21786862   33.0858231
#> Jan 1994  13.94158721  114.3623543
#> Feb 1994   1.12209490 -138.1713624
#> Mar 1994   2.63123346  -84.4069552
#> Apr 1994  10.06751657 -183.6583479
#> May 1994  -0.46200909  104.4170831
#> Jun 1994  -6.52330648   21.5210107
#> Jul 1994  -9.77249153   80.0349614
#> Aug 1994  -2.59937146   71.4933031
#> Sep 1994 -15.53235983  -31.7171209
#> Oct 1994  -0.45252811 -225.2045124
#> Nov 1994   6.77355665  150.5248197
#> Dec 1994   5.36125012  -41.9147872
#> Jan 1995   6.16873488  -20.8179987
#> Feb 1995  -6.61436541  102.5008622
#> Mar 1995  13.38586709 -210.9352248
#> Apr 1995 -11.17817883   58.1155983
#> May 1995   6.84733535   42.7476044
#> Jun 1995  -6.10556343   30.9712183
#> Jul 1995  -3.49306827  -34.8680801
#> Aug 1995  11.33969304   47.2684011
#> Sep 1995  15.33756578   70.2742649
#> Oct 1995  11.06669286 -154.3797498
#> Nov 1995   3.44061750 -113.9419424
#> Dec 1995 -20.38579747   37.0802634
#> Jan 1996  -4.93457634   24.9291812
#> Feb 1996  -0.60792957  124.0415284
#> Mar 1996   4.38221794  -37.1594045
#> Apr 1996 -12.14891231   90.8843016
#> May 1996   4.98646877   77.2073385
#> Jun 1996 -10.82869957   36.6647135
#> Jul 1996   4.22166451   23.4063054
#> Aug 1996  -2.22888718  -82.4637494
#> Sep 1996  -5.00324694   79.3613760
#> Oct 1996 -11.09585791   81.5375576
#> Nov 1996   9.09209005  -64.0412559
#> Dec 1996  -2.33965281   15.4668564
#> Jan 1997   8.42443753 -131.9062104
#> Feb 1997  12.64094183  -55.9444239
#> Mar 1997  -7.10347269 -127.8683713
#> Apr 1997  -3.53829620  -16.1854635
#> May 1997 -10.83490986 -121.2698055
#> Jun 1997   9.66857761  137.1345622
#> Jul 1997  -4.80530784    8.2627704
#> Aug 1997 -10.41036179 -126.5853773
#> Sep 1997   1.99629954  117.4275795
#> Oct 1997  17.12906789   69.5991675
#> Nov 1997   5.52736362   80.0070376
#> Dec 1997  -8.68735805 -100.2196891
#> Jan 1998   0.98637158  -67.7115667
#> Feb 1998 -18.34874715   72.0963599
#> Mar 1998  14.11194917   96.0168144
#> Apr 1998   3.79831016  -14.5665667
#> May 1998  -0.15671859  150.5255540
#> Jun 1998 -15.74581458   67.8180406
#> Jul 1998  -3.52542777  -26.5552196
#> Aug 1998   6.21810434  -45.8132537
#> Sep 1998   4.13433266  173.8431533
#> Oct 1998  -0.99195265   -1.3105949
#> Nov 1998  -1.96023354  -79.4353354
#> Dec 1998   6.21162356   25.5333393
#> Jan 1999  -0.47497088  -87.8777648
#> Feb 1999 -10.95678714  -50.2384014
#> Mar 1999   5.97835715   91.8235108
#> Apr 1999  -8.10128008  -41.5556110
#> May 1999   8.19117555  -49.7454127
#> Jun 1999  11.04700628   12.2983974
#> Jul 1999  14.19988400  -99.4093841
#> Aug 1999   5.01813908   48.5937859
#> Sep 1999 -12.92365689  -55.1041758
#> Oct 1999   8.17587784   16.6600822
#> Nov 1999  -7.00872260    9.5564919
#> Dec 1999  -3.60897569   93.7361737
#> Jan 2000  -6.82321746  -77.8592244
#> Feb 2000 -11.36635355  -34.6014963
#> Mar 2000  -5.64223368   14.5927947
#> Apr 2000 -13.49055622  -67.9378277
#> May 2000  -6.50880106  -61.3698253
#> Jun 2000 -25.69596903  220.2935147
#> Jul 2000   3.31580338  -32.8772531
#> Aug 2000   5.08728668  175.1776850
#> Sep 2000  -4.97105594  -97.1651737
#> Oct 2000  -9.44389799 -168.9546894
#> Nov 2000 -11.94459690 -148.1386178
#> Dec 2000  13.51632313  242.6894783
#> Jan 2001  -6.13883787  -22.8460020
#> Feb 2001  -6.66701676  111.3283521
#> Mar 2001 -12.92547529  -43.7063852
#> Apr 2001   8.22722389  126.4884610
#> May 2001  30.08092737  143.6597731
#> Jun 2001  11.19241756  -91.8664322
#> Jul 2001 -28.21799077 -191.4602512
#> Aug 2001  16.17287527   11.4969531
#> Sep 2001   5.13820257  151.0348121
#> Oct 2001   5.02789311   -0.9052613
#> Nov 2001  -7.35291577  -13.5856744
#> Dec 2001 -11.08518180  -33.6250668
#> Jan 2002   7.17057189   35.7796527
#> Feb 2002   4.05856898   34.0498580
#> Mar 2002 -14.99181516  309.7230704
#> Apr 2002 -12.23578511   67.7352741
#> May 2002  14.25347975   38.5994964
#> Jun 2002   5.35483990  -83.8842632
#> Jul 2002  -4.79124096  -53.0027590
#> Aug 2002   7.08755575 -229.3182084
#> Sep 2002  -0.10940710  185.5263642
#> Oct 2002 -13.92609039  -62.4714964
#> Nov 2002   6.24947544   17.5872582
#> Dec 2002   6.80719557 -136.5788853
#> Jan 2003  22.51140001 -134.5518474
#> Feb 2003 -18.42120983   94.7265271
#> Mar 2003  -0.88932029  -11.5739671
#> Apr 2003  -6.54185827  -42.9404494
#> May 2003  -3.71089697    7.9631976
#> Jun 2003   7.32142682  100.6463994
#> Jul 2003   1.58182934  175.4518747
#> Aug 2003  10.05473158  -11.9361311
#> Sep 2003  -2.97361367  164.3998625
#> Oct 2003  -2.35435164 -131.0997765
#> Nov 2003  -2.63170847   36.2701424
#> Dec 2003  -3.20464012  136.9665491
#> Jan 2004  19.30659239  -34.8512400
#> Feb 2004 -21.87774508  -54.5470652
#> Mar 2004   7.00974005   -2.0519678
#> Apr 2004  -3.76301495  -33.8849889
#> May 2004   3.35181016   18.5592100
#> Jun 2004  -2.93525659   34.5862657
#> Jul 2004   0.11142179  -79.3332374
#> Aug 2004  11.98017338  -74.8631612
#> Sep 2004  -4.67733450 -184.6666644
#> Oct 2004   9.16382866 -155.6824415
#> Nov 2004  -7.06060331    0.2354941
#> Dec 2004   7.86935538  147.0206284
#> Jan 2005   3.99883945  -77.7484495
#> Feb 2005  -6.38705322   70.4613433
#> Mar 2005  -3.43121408   69.6606398
#> Apr 2005 -11.18278418   93.1972929
#> May 2005   5.78376820  261.6178688
#> Jun 2005   1.13186948 -147.7492275
#> Jul 2005   4.40206334    7.4677229
#> Aug 2005 -20.02320938  -64.0770175
#> Sep 2005  10.41633836   11.4925465
#> Oct 2005 -10.13748921   74.7021300
#> Nov 2005  -5.04802858 -111.8479256
#> Dec 2005   0.89953043   34.1004819
#> Jan 2006 -12.92000906   12.5031186
#> Feb 2006 -17.30530080   31.1369721
#> Mar 2006  11.32424775   42.6443804
#> Apr 2006   9.55398518 -126.9579318
#> May 2006  -7.14008278   86.9811052
#> Jun 2006  -1.30393939   -2.3574630
#> Jul 2006   9.50804885   64.6068260
#> Aug 2006   4.70794223  123.2672999
#> Sep 2006   3.34838125 -128.0410543
#> Oct 2006 -15.49423258  -17.6086170
#> Nov 2006  -6.21150422  114.8912321
#> Dec 2006 -16.15618228   10.6409425
#> Jan 2007   8.53465374  105.2599358
#> Feb 2007  18.88819161    6.3508668
#> Mar 2007  -8.67441159   76.0288668
#> Apr 2007   7.14420496  -58.3391234
#> May 2007   6.14427022  -83.3176971
#> Jun 2007   3.05098743  -95.5279258
#> Jul 2007  -9.25999799   46.8197456
#> Aug 2007  13.59534396  148.9361598
#> Sep 2007   2.03698336   84.5637803
#> Oct 2007  12.88543508   -9.0330359
#> Nov 2007  -2.12971834  -76.0974092
#> Dec 2007   4.87549814  -42.2703485
#> Jan 2008 -10.67709165 -111.6990551
#> Feb 2008 -11.80488666 -137.3205437
#> Mar 2008  21.59175760  -46.4627202
#> Apr 2008  -4.17812648  -75.1982369
#> May 2008   0.87654958  -40.7974543
#> Jun 2008  -5.43709068   17.7182569
#> Jul 2008  -1.19867022   99.9326344
#> Aug 2008  18.00624858   11.6269256
#> Sep 2008  -7.94247350 -183.5065315
#> Oct 2008  -3.60360559  -29.8695682
#> Nov 2008   6.31460646   31.4652720
#> Dec 2008 -16.63279352 -108.2466160
#> Jan 2009  -4.88005245   22.1793014
#> Feb 2009  -0.20752219   -4.9851303
#> Mar 2009   9.33871038  110.7534869
#> Apr 2009  -2.32670878  -99.3360619
#> May 2009 -10.19630064  192.4843474
#> Jun 2009 -16.80328321  -55.6762404
#> Jul 2009  22.81394917   32.9474662
#> Aug 2009 -10.42285303  175.8286351
#> Sep 2009   7.12307192  127.3909790
#> Oct 2009  12.20747981  -58.3885195
#> Nov 2009   3.36018683  -32.7692378
#> Dec 2009  -2.79147937   14.5241262
#> Jan 2010   0.76422131   21.4101138
#> Feb 2010  -3.54509030    1.1554669
#> Mar 2010  14.32658547  121.9699084
#> Apr 2010   0.02671474 -103.0643641
#> May 2010   3.44012511   76.4734975
#> Jun 2010  -8.30323380  -51.4982604
#> Jul 2010   4.82115591  174.0986972
#> Aug 2010  -6.88660369  100.5845241
#> Sep 2010 -10.93359720 -110.7352945
#> Oct 2010   8.41060978  -62.7882790
#> Nov 2010  14.54136238 -147.9501948
#> Dec 2010   1.37240272   15.4343173
#> Jan 2011   3.25544925 -229.9484409
#> Feb 2011  12.23728019  -43.8727684
#> Mar 2011 -10.53314136  -64.9680607
#> Apr 2011  18.40332640   42.4040947
#> May 2011 -22.31484275    4.7049728
#> Jun 2011  -5.95844719 -128.5044066
#> Jul 2011  -2.30472754    4.5532358
#> Aug 2011  17.56315261 -144.9429700
#> Sep 2011   4.75705566 -163.3704403
#> Oct 2011   7.82644366  -51.7284962
#> Nov 2011 -11.46245649  -99.1151430
#> Dec 2011   9.50180682  -72.8485045
#> Jan 2012  -3.45227990   45.1951634
#> Feb 2012 -24.73551488   88.7181335
#> Mar 2012  13.76264164  126.4573401
#> Apr 2012  24.77145067 -129.1419391
#> May 2012 -13.37234198   18.2510306
#> Jun 2012  20.24155203   -8.9809675
#> Jul 2012   8.59593181  -11.2629276
#> Aug 2012   7.28471344   16.2801338
#> Sep 2012  17.02971138   55.5861006
#> Oct 2012 -13.49671966   85.3394680
#> Nov 2012   0.85960972  -19.4899361
#> Dec 2012   2.04207089   -4.3411909
#> Jan 2013  -9.79050445  -52.9438316
#> Feb 2013   5.94650797  -59.8510599
#> Mar 2013  10.39317096   14.9297682
#> Apr 2013 -12.78834279  125.5820916
#> May 2013  -3.19978820  -23.4369528
#> Jun 2013 -18.72828606  142.2340276
#> Jul 2013   0.91322657   29.2127542
#> Aug 2013   7.12363770  -73.6886751
#> Sep 2013  -1.10119494  -12.9996552
#> Oct 2013 -10.51575816  220.0885652
#> Nov 2013  -8.91401342  161.3081286
#> Dec 2013  -6.19138043  -26.7499583
#> Jan 2014   2.55199046   12.7603168
#> Feb 2014  -4.39315917  -37.0892736
#> Mar 2014  14.68149212  -45.2402281
#> Apr 2014   8.32506126  119.7478248
#> May 2014  -9.13787407   96.1339696
#> Jun 2014  -2.72595426  109.9244552
#> Jul 2014   7.51086555    5.3534542
#> Aug 2014  -0.76856255   -7.1082557
#> Sep 2014   1.47101528  -26.5212622
#> Oct 2014  -3.21238201 -108.7472787
#> Nov 2014 -17.03140073   11.2995341
#> Dec 2014  -7.61038444  -36.8761608
#> Jan 2015  -7.59025393 -181.3818422
#> Feb 2015   2.92608648  -72.7994368
#> Mar 2015   3.32236386  -32.3343456
#> Apr 2015  -2.31934768   11.8178061
#> May 2015  -8.15807168   52.2721179
#> Jun 2015   2.16074797  -39.4712044
#> Jul 2015 -11.38865785  -73.8262361
#> Aug 2015   1.40105007  131.9700103
#> Sep 2015  -6.62133906  -54.7717472
#> Oct 2015  -5.34391766  211.6530743
#> Nov 2015 -14.12830881   23.8889683
#> Dec 2015  12.17203500  -51.5969374
#> Jan 2016   4.25531888  115.8752180
#> Feb 2016  -1.20928613  -63.4357812
#> Mar 2016   6.51430723   86.2710827
#> Apr 2016   4.90205565 -168.1573026
#> May 2016   7.05192456 -163.6792119
#> Jun 2016   8.05650671  143.0290161
#> Jul 2016  -1.79990583  153.9376495
#> Aug 2016  10.72861877   -7.7210862
#> Sep 2016  -0.82523465   72.2232423
#> Oct 2016  10.72598758   76.2012389
#> Nov 2016  -0.02079741 -110.5216744
#> Dec 2016  -8.59905321 -146.1168585
#> Jan 2017   7.48074130   29.5561895
#> Feb 2017  13.56515448   65.8526506
#> Mar 2017   4.00001599   64.5142725
#> Apr 2017   6.36256611   62.3653458
#> May 2017 -18.66035907  231.5849301
#> Jun 2017  -5.80927235  -38.1598877
#> Jul 2017  10.93896134    3.1688164
#> Aug 2017  -5.32691097   77.6625400
#> Sep 2017   6.83984092 -198.3047831
#> Oct 2017  -6.84263667   -0.5446710
#> Nov 2017  16.25943271   70.2411713
#> Dec 2017  -6.31734454  122.3164016
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 

s_preVar(myreg3)

#> $series
#>                  var1         var2
#> Jan 1990  -0.18087120 -135.3779986
#> Feb 1990  -1.69198739   38.7122228
#> Mar 1990  -1.64743004  -96.7667515
#> Apr 1990   7.13622416  -35.8785321
#> May 1990   2.37209036  -58.6397434
#> Jun 1990 -23.22686419  161.2913478
#> Jul 1990   0.79044275   50.5671343
#> Aug 1990  -4.28806351    2.4106090
#> Sep 1990  -3.72714386   29.5359647
#> Oct 1990  -0.44918615  127.5350360
#> Nov 1990  14.01343027   20.4477180
#> Dec 1990  -3.00872112   18.6747913
#> Jan 1991  -9.65754406   70.9749842
#> Feb 1991  16.83218482   37.5846729
#> Mar 1991  -7.76550575   40.5413237
#> Apr 1991   2.47301367   33.0985753
#> May 1991  -9.67146956  159.3760511
#> Jun 1991 -10.37854416   55.5075757
#> Jul 1991 -10.53599982 -120.9478318
#> Aug 1991  -2.05507436  -22.5557951
#> Sep 1991  13.08266555  -34.2381835
#> Oct 1991   5.13721896 -173.6559496
#> Nov 1991 -13.85355230   16.3560903
#> Dec 1991 -10.01238810   29.8293513
#> Jan 1992   3.01325680   17.5870029
#> Feb 1992   4.37682054 -140.5601011
#> Mar 1992  -6.50275146  -21.5255891
#> Apr 1992 -24.19678943  -50.2680266
#> May 1992   2.75546070  168.3881702
#> Jun 1992  -2.17346388 -172.5036625
#> Jul 1992   4.20810277  -25.2554758
#> Aug 1992   5.14492552 -160.4868007
#> Sep 1992  -9.76120254  115.3376302
#> Oct 1992   8.21322767 -126.6363058
#> Nov 1992  14.87457423  -25.0127880
#> Dec 1992   5.29521792  -53.8988584
#> Jan 1993  18.85181767  -61.1159385
#> Feb 1993   7.23419932   80.7837347
#> Mar 1993  -1.15305742  135.7967557
#> Apr 1993   7.81349665  100.0928446
#> May 1993  18.02937166 -125.8705305
#> Jun 1993  -3.86689222 -139.5429138
#> Jul 1993  27.15540230   32.6775093
#> Aug 1993  12.49770616  -56.6269784
#> Sep 1993 -19.56745628   -4.0093207
#> Oct 1993   6.66756093  100.7425936
#> Nov 1993  -5.36932307 -137.3520862
#> Dec 1993  11.21786862   33.0858231
#> Jan 1994  13.94158721  114.3623543
#> Feb 1994   1.12209490 -138.1713624
#> Mar 1994   2.63123346  -84.4069552
#> Apr 1994  10.06751657 -183.6583479
#> May 1994  -0.46200909  104.4170831
#> Jun 1994  -6.52330648   21.5210107
#> Jul 1994  -9.77249153   80.0349614
#> Aug 1994  -2.59937146   71.4933031
#> Sep 1994 -15.53235983  -31.7171209
#> Oct 1994  -0.45252811 -225.2045124
#> Nov 1994   6.77355665  150.5248197
#> Dec 1994   5.36125012  -41.9147872
#> Jan 1995   6.16873488  -20.8179987
#> Feb 1995  -6.61436541  102.5008622
#> Mar 1995  13.38586709 -210.9352248
#> Apr 1995 -11.17817883   58.1155983
#> May 1995   6.84733535   42.7476044
#> Jun 1995  -6.10556343   30.9712183
#> Jul 1995  -3.49306827  -34.8680801
#> Aug 1995  11.33969304   47.2684011
#> Sep 1995  15.33756578   70.2742649
#> Oct 1995  11.06669286 -154.3797498
#> Nov 1995   3.44061750 -113.9419424
#> Dec 1995 -20.38579747   37.0802634
#> Jan 1996  -4.93457634   24.9291812
#> Feb 1996  -0.60792957  124.0415284
#> Mar 1996   4.38221794  -37.1594045
#> Apr 1996 -12.14891231   90.8843016
#> May 1996   4.98646877   77.2073385
#> Jun 1996 -10.82869957   36.6647135
#> Jul 1996   4.22166451   23.4063054
#> Aug 1996  -2.22888718  -82.4637494
#> Sep 1996  -5.00324694   79.3613760
#> Oct 1996 -11.09585791   81.5375576
#> Nov 1996   9.09209005  -64.0412559
#> Dec 1996  -2.33965281   15.4668564
#> Jan 1997   8.42443753 -131.9062104
#> Feb 1997  12.64094183  -55.9444239
#> Mar 1997  -7.10347269 -127.8683713
#> Apr 1997  -3.53829620  -16.1854635
#> May 1997 -10.83490986 -121.2698055
#> Jun 1997   9.66857761  137.1345622
#> Jul 1997  -4.80530784    8.2627704
#> Aug 1997 -10.41036179 -126.5853773
#> Sep 1997   1.99629954  117.4275795
#> Oct 1997  17.12906789   69.5991675
#> Nov 1997   5.52736362   80.0070376
#> Dec 1997  -8.68735805 -100.2196891
#> Jan 1998   0.98637158  -67.7115667
#> Feb 1998 -18.34874715   72.0963599
#> Mar 1998  14.11194917   96.0168144
#> Apr 1998   3.79831016  -14.5665667
#> May 1998  -0.15671859  150.5255540
#> Jun 1998 -15.74581458   67.8180406
#> Jul 1998  -3.52542777  -26.5552196
#> Aug 1998   6.21810434  -45.8132537
#> Sep 1998   4.13433266  173.8431533
#> Oct 1998  -0.99195265   -1.3105949
#> Nov 1998  -1.96023354  -79.4353354
#> Dec 1998   6.21162356   25.5333393
#> Jan 1999  -0.47497088  -87.8777648
#> Feb 1999 -10.95678714  -50.2384014
#> Mar 1999   5.97835715   91.8235108
#> Apr 1999  -8.10128008  -41.5556110
#> May 1999   8.19117555  -49.7454127
#> Jun 1999  11.04700628   12.2983974
#> Jul 1999  14.19988400  -99.4093841
#> Aug 1999   5.01813908   48.5937859
#> Sep 1999 -12.92365689  -55.1041758
#> Oct 1999   8.17587784   16.6600822
#> Nov 1999  -7.00872260    9.5564919
#> Dec 1999  -3.60897569   93.7361737
#> Jan 2000  -6.82321746  -77.8592244
#> Feb 2000 -11.36635355  -34.6014963
#> Mar 2000  -5.64223368   14.5927947
#> Apr 2000 -13.49055622  -67.9378277
#> May 2000  -6.50880106  -61.3698253
#> Jun 2000 -25.69596903  220.2935147
#> Jul 2000   3.31580338  -32.8772531
#> Aug 2000   5.08728668  175.1776850
#> Sep 2000  -4.97105594  -97.1651737
#> Oct 2000  -9.44389799 -168.9546894
#> Nov 2000 -11.94459690 -148.1386178
#> Dec 2000  13.51632313  242.6894783
#> Jan 2001  -6.13883787  -22.8460020
#> Feb 2001  -6.66701676  111.3283521
#> Mar 2001 -12.92547529  -43.7063852
#> Apr 2001   8.22722389  126.4884610
#> May 2001  30.08092737  143.6597731
#> Jun 2001  11.19241756  -91.8664322
#> Jul 2001 -28.21799077 -191.4602512
#> Aug 2001  16.17287527   11.4969531
#> Sep 2001   5.13820257  151.0348121
#> Oct 2001   5.02789311   -0.9052613
#> Nov 2001  -7.35291577  -13.5856744
#> Dec 2001 -11.08518180  -33.6250668
#> Jan 2002   7.17057189   35.7796527
#> Feb 2002   4.05856898   34.0498580
#> Mar 2002 -14.99181516  309.7230704
#> Apr 2002 -12.23578511   67.7352741
#> May 2002  14.25347975   38.5994964
#> Jun 2002   5.35483990  -83.8842632
#> Jul 2002  -4.79124096  -53.0027590
#> Aug 2002   7.08755575 -229.3182084
#> Sep 2002  -0.10940710  185.5263642
#> Oct 2002 -13.92609039  -62.4714964
#> Nov 2002   6.24947544   17.5872582
#> Dec 2002   6.80719557 -136.5788853
#> Jan 2003  22.51140001 -134.5518474
#> Feb 2003 -18.42120983   94.7265271
#> Mar 2003  -0.88932029  -11.5739671
#> Apr 2003  -6.54185827  -42.9404494
#> May 2003  -3.71089697    7.9631976
#> Jun 2003   7.32142682  100.6463994
#> Jul 2003   1.58182934  175.4518747
#> Aug 2003  10.05473158  -11.9361311
#> Sep 2003  -2.97361367  164.3998625
#> Oct 2003  -2.35435164 -131.0997765
#> Nov 2003  -2.63170847   36.2701424
#> Dec 2003  -3.20464012  136.9665491
#> Jan 2004  19.30659239  -34.8512400
#> Feb 2004 -21.87774508  -54.5470652
#> Mar 2004   7.00974005   -2.0519678
#> Apr 2004  -3.76301495  -33.8849889
#> May 2004   3.35181016   18.5592100
#> Jun 2004  -2.93525659   34.5862657
#> Jul 2004   0.11142179  -79.3332374
#> Aug 2004  11.98017338  -74.8631612
#> Sep 2004  -4.67733450 -184.6666644
#> Oct 2004   9.16382866 -155.6824415
#> Nov 2004  -7.06060331    0.2354941
#> Dec 2004   7.86935538  147.0206284
#> Jan 2005   3.99883945  -77.7484495
#> Feb 2005  -6.38705322   70.4613433
#> Mar 2005  -3.43121408   69.6606398
#> Apr 2005 -11.18278418   93.1972929
#> May 2005   5.78376820  261.6178688
#> Jun 2005   1.13186948 -147.7492275
#> Jul 2005   4.40206334    7.4677229
#> Aug 2005 -20.02320938  -64.0770175
#> Sep 2005  10.41633836   11.4925465
#> Oct 2005 -10.13748921   74.7021300
#> Nov 2005  -5.04802858 -111.8479256
#> Dec 2005   0.89953043   34.1004819
#> Jan 2006 -12.92000906   12.5031186
#> Feb 2006 -17.30530080   31.1369721
#> Mar 2006  11.32424775   42.6443804
#> Apr 2006   9.55398518 -126.9579318
#> May 2006  -7.14008278   86.9811052
#> Jun 2006  -1.30393939   -2.3574630
#> Jul 2006   9.50804885   64.6068260
#> Aug 2006   4.70794223  123.2672999
#> Sep 2006   3.34838125 -128.0410543
#> Oct 2006 -15.49423258  -17.6086170
#> Nov 2006  -6.21150422  114.8912321
#> Dec 2006 -16.15618228   10.6409425
#> Jan 2007   8.53465374  105.2599358
#> Feb 2007  18.88819161    6.3508668
#> Mar 2007  -8.67441159   76.0288668
#> Apr 2007   7.14420496  -58.3391234
#> May 2007   6.14427022  -83.3176971
#> Jun 2007   3.05098743  -95.5279258
#> Jul 2007  -9.25999799   46.8197456
#> Aug 2007  13.59534396  148.9361598
#> Sep 2007   2.03698336   84.5637803
#> Oct 2007  12.88543508   -9.0330359
#> Nov 2007  -2.12971834  -76.0974092
#> Dec 2007   4.87549814  -42.2703485
#> Jan 2008 -10.67709165 -111.6990551
#> Feb 2008 -11.80488666 -137.3205437
#> Mar 2008  21.59175760  -46.4627202
#> Apr 2008  -4.17812648  -75.1982369
#> May 2008   0.87654958  -40.7974543
#> Jun 2008  -5.43709068   17.7182569
#> Jul 2008  -1.19867022   99.9326344
#> Aug 2008  18.00624858   11.6269256
#> Sep 2008  -7.94247350 -183.5065315
#> Oct 2008  -3.60360559  -29.8695682
#> Nov 2008   6.31460646   31.4652720
#> Dec 2008 -16.63279352 -108.2466160
#> Jan 2009  -4.88005245   22.1793014
#> Feb 2009  -0.20752219   -4.9851303
#> Mar 2009   9.33871038  110.7534869
#> Apr 2009  -2.32670878  -99.3360619
#> May 2009 -10.19630064  192.4843474
#> Jun 2009 -16.80328321  -55.6762404
#> Jul 2009  22.81394917   32.9474662
#> Aug 2009 -10.42285303  175.8286351
#> Sep 2009   7.12307192  127.3909790
#> Oct 2009  12.20747981  -58.3885195
#> Nov 2009   3.36018683  -32.7692378
#> Dec 2009  -2.79147937   14.5241262
#> Jan 2010   0.76422131   21.4101138
#> Feb 2010  -3.54509030    1.1554669
#> Mar 2010  14.32658547  121.9699084
#> Apr 2010   0.02671474 -103.0643641
#> May 2010   3.44012511   76.4734975
#> Jun 2010  -8.30323380  -51.4982604
#> Jul 2010   4.82115591  174.0986972
#> Aug 2010  -6.88660369  100.5845241
#> Sep 2010 -10.93359720 -110.7352945
#> Oct 2010   8.41060978  -62.7882790
#> Nov 2010  14.54136238 -147.9501948
#> Dec 2010   1.37240272   15.4343173
#> Jan 2011   3.25544925 -229.9484409
#> Feb 2011  12.23728019  -43.8727684
#> Mar 2011 -10.53314136  -64.9680607
#> Apr 2011  18.40332640   42.4040947
#> May 2011 -22.31484275    4.7049728
#> Jun 2011  -5.95844719 -128.5044066
#> Jul 2011  -2.30472754    4.5532358
#> Aug 2011  17.56315261 -144.9429700
#> Sep 2011   4.75705566 -163.3704403
#> Oct 2011   7.82644366  -51.7284962
#> Nov 2011 -11.46245649  -99.1151430
#> Dec 2011   9.50180682  -72.8485045
#> Jan 2012  -3.45227990   45.1951634
#> Feb 2012 -24.73551488   88.7181335
#> Mar 2012  13.76264164  126.4573401
#> Apr 2012  24.77145067 -129.1419391
#> May 2012 -13.37234198   18.2510306
#> Jun 2012  20.24155203   -8.9809675
#> Jul 2012   8.59593181  -11.2629276
#> Aug 2012   7.28471344   16.2801338
#> Sep 2012  17.02971138   55.5861006
#> Oct 2012 -13.49671966   85.3394680
#> Nov 2012   0.85960972  -19.4899361
#> Dec 2012   2.04207089   -4.3411909
#> Jan 2013  -9.79050445  -52.9438316
#> Feb 2013   5.94650797  -59.8510599
#> Mar 2013  10.39317096   14.9297682
#> Apr 2013 -12.78834279  125.5820916
#> May 2013  -3.19978820  -23.4369528
#> Jun 2013 -18.72828606  142.2340276
#> Jul 2013   0.91322657   29.2127542
#> Aug 2013   7.12363770  -73.6886751
#> Sep 2013  -1.10119494  -12.9996552
#> Oct 2013 -10.51575816  220.0885652
#> Nov 2013  -8.91401342  161.3081286
#> Dec 2013  -6.19138043  -26.7499583
#> Jan 2014   2.55199046   12.7603168
#> Feb 2014  -4.39315917  -37.0892736
#> Mar 2014  14.68149212  -45.2402281
#> Apr 2014   8.32506126  119.7478248
#> May 2014  -9.13787407   96.1339696
#> Jun 2014  -2.72595426  109.9244552
#> Jul 2014   7.51086555    5.3534542
#> Aug 2014  -0.76856255   -7.1082557
#> Sep 2014   1.47101528  -26.5212622
#> Oct 2014  -3.21238201 -108.7472787
#> Nov 2014 -17.03140073   11.2995341
#> Dec 2014  -7.61038444  -36.8761608
#> Jan 2015  -7.59025393 -181.3818422
#> Feb 2015   2.92608648  -72.7994368
#> Mar 2015   3.32236386  -32.3343456
#> Apr 2015  -2.31934768   11.8178061
#> May 2015  -8.15807168   52.2721179
#> Jun 2015   2.16074797  -39.4712044
#> Jul 2015 -11.38865785  -73.8262361
#> Aug 2015   1.40105007  131.9700103
#> Sep 2015  -6.62133906  -54.7717472
#> Oct 2015  -5.34391766  211.6530743
#> Nov 2015 -14.12830881   23.8889683
#> Dec 2015  12.17203500  -51.5969374
#> Jan 2016   4.25531888  115.8752180
#> Feb 2016  -1.20928613  -63.4357812
#> Mar 2016   6.51430723   86.2710827
#> Apr 2016   4.90205565 -168.1573026
#> May 2016   7.05192456 -163.6792119
#> Jun 2016   8.05650671  143.0290161
#> Jul 2016  -1.79990583  153.9376495
#> Aug 2016  10.72861877   -7.7210862
#> Sep 2016  -0.82523465   72.2232423
#> Oct 2016  10.72598758   76.2012389
#> Nov 2016  -0.02079741 -110.5216744
#> Dec 2016  -8.59905321 -146.1168585
#> Jan 2017   7.48074130   29.5561895
#> Feb 2017  13.56515448   65.8526506
#> Mar 2017   4.00001599   64.5142725
#> Apr 2017   6.36256611   62.3653458
#> May 2017 -18.66035907  231.5849301
#> Jun 2017  -5.80927235  -38.1598877
#> Jul 2017  10.93896134    3.1688164
#> Aug 2017  -5.32691097   77.6625400
#> Sep 2017   6.83984092 -198.3047831
#> Oct 2017  -6.84263667   -0.5446710
#> Nov 2017  16.25943271   70.2411713
#> Dec 2017  -6.31734454  122.3164016
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 

s_td(myreg1)

#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove

s_td(myspec1)

#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove

s_easter(myreg1)

#>  enabled julian duration test
#>     TRUE  FALSE        8  Add

s_easter(myspec1)

#>  enabled julian duration test
#>     TRUE  FALSE        8  Add

s_out(myreg1)

#>  enabled span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE  All TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7

s_out(myspec1)

#>  enabled            span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE From 2010-03-01 TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7

s_arima(myreg1)

#>  enabled automdl.acceptdefault automdl.cancel automdl.ub1 automdl.ub2
#>     TRUE                 FALSE            0.1    1.041667        0.88
#>  automdl.mixed automdl.balanced automdl.armalimit automdl.reducecv
#>           TRUE            FALSE                 1          0.14286
#>  automdl.ljungboxlimit automdl.ubfinal arima.mu arima.p arima.d arima.q
#>                   0.95            1.05    FALSE       0       1       1
#>  arima.bp arima.bd arima.bq arima.coef
#>         0        1        1      FALSE

s_arima(myspec1)

#>  enabled automdl.acceptdefault automdl.cancel automdl.ub1 automdl.ub2
#>     TRUE                 FALSE            0.1    1.041667        0.88
#>  automdl.mixed automdl.balanced automdl.armalimit automdl.reducecv
#>           TRUE            FALSE                 1          0.14286
#>  automdl.ljungboxlimit automdl.ubfinal arima.mu arima.p arima.d arima.q
#>                   0.95            1.05    FALSE       0       1       1
#>  arima.bp arima.bd arima.bq arima.coef
#>         0        1        1      FALSE

myspec4 <- regarima_spec_x13(myreg1, automdl.enabled = FALSE,
             arima.coefEnabled = TRUE,
             arima.p = 1,arima.q = 1, arima.bp = 1, arima.bq = 1,
             arima.coef = rep(0.2, 4),
             arima.coefType = rep("Initial", 4))
myreg4 <- regarima(myseries, myspec4)

s_arimaCoef(myreg4)

#>              Type Value
#> Phi(1)    Initial   0.2
#> Theta(1)  Initial   0.2
#> BPhi(1)   Initial   0.2
#> BTheta(1) Initial   0.2

s_arimaCoef(myspec4)

#>              Type Value
#> Phi(1)    Initial   0.2
#> Theta(1)  Initial   0.2
#> BPhi(1)   Initial   0.2
#> BTheta(1) Initial   0.2

s_fcst(myreg1)

#>  horizon
#>       -2

s_fcst(myspec1)

#>  horizon
#>       -2

s_span(myreg1)

#>          type         d0         d1 n0 n1
#> estimate  All 1900-01-01 2020-12-31  0  0
#> outlier   All 1900-01-01 2020-12-31  0  0

s_span(myspec1)

#>          type         d0         d1 n0 n1
#> estimate From 2005-10-01 2020-12-31  0  0
#> outlier  From 2010-03-01 2020-12-31  0  0

myspec5 <- x13_spec(spec = "RSA5c", x11.seasonalComp = FALSE)
mysa5 <- x13(myseries, myspec5)

s_x11(mysa5)

#>   x11.mode x11.seasonalComp x11.lsigma x11.usigma x11.trendAuto x11.trendma
#>  Undefined            FALSE        1.5        2.5          TRUE          13
#>  x11.seasonalma x11.fcasts x11.bcasts x11.excludeFcasts
#>             Msr         -1          0             FALSE

s_x11(myspec5)

#>   x11.mode x11.seasonalComp x11.lsigma x11.usigma x11.trendAuto x11.trendma
#>  Undefined            FALSE        1.5        2.5          TRUE          13
#>  x11.seasonalma x11.fcasts x11.bcasts x11.excludeFcasts
#>             Msr         -1          0             FALSE

myspec6 <- tramoseats_spec(spec = "RSAfull", seats.approx = "Noisy")
mysa6 <- tramoseats(myseries, myspec6)

s_seats(mysa6)

#>  seats.approx seats.maBoundary seats.trendBoundary seats.seasdBoundary
#>         Noisy             0.95                 0.5                 0.8
#>  seats.seasdBoundary1 seats.seasTol seats.method
#>                   0.8             2       Burman

s_seats(mysa6)

#>  seats.approx seats.maBoundary seats.trendBoundary seats.seasdBoundary
#>         Noisy             0.95                 0.5                 0.8
#>  seats.seasdBoundary1 seats.seasTol seats.method
#>                   0.8             2       Burman