Access model specification, SA and pre-adjustment in X13 and TRAMO-SEATS

Below functions access different parts of the final model specification as included in the "SA", "regarima", "SA_spec" and "regarima_spec" S3 class objects.

s_estimate(object = NA)

s_transform(object = NA)

s_usrdef(object = NA)

s_preOut(object = NA)

s_preVar(object = NA)

s_td(object = NA)

s_easter(object = NA)

s_out(object = NA)

s_arima(object = NA)

s_arimaCoef(object = NA)

s_fcst(object = NA)

s_span(object = NA)

s_x11(object = NA)

s_seats(object = NA)

Arguments

object
object of one of the classes: c("SA","X13"), c("SA","TRAMO_SEATS"), c("SA_spec","X13"), c("SA_spec","TRAMO_SEATS"), c("regarima","X13"), c("regarima","TRAMO_SEATS"), c("regarima_spec","X13"), c("regarima_spec","TRAMO_SEATS").

Value

s_estimate returns a data.frame with the estimate variables

s_transform returns a data.frame with the transform varaibles

s_usrdef returns a data.frame with the user-defined regressors (outliers and variables) model specification, indicating if those variables are included in the model and if coefficients are pre-specified

s_preOut returns a data.frame with the pre-specified outliers

s_preVar returns a list with the information on the user-defined variables, including: series - the time series and description - data.frame with the variable type and coefficients

s_td returns a data.frame with the trading.days variables

s_easter returns a data.frame with the easter variables

s_out returns a data.frame with the outliers detection variables

s_arima returns a data.frame with the arima variables

s_arimaCoef returns a data.frame with the user-specified ARMA coefficients

s_fcst returns a data.frame with the forecast horizon

s_span returns a data.frame with the span variables

s_x11 returns a data.frame with the x11 variables

s_seats returns a data.frame with the seats variables

References

Info on JDemetra+, usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en

Examples

myseries <- ipi_c_eu[, "FR"]
myreg1 <- regarima_def_x13(myseries, spec = c("RG5c"))
myspec1 <- regarima_spec_x13(myreg1,
             estimate.from = "2005-10-01",
             outlier.from = "2010-03-01")

s_estimate(myreg1)

#>  span tolerance
#>   All     1e-07

s_estimate(myspec1)

#>             span tolerance
#>  From 2005-10-01     1e-07

s_transform(myreg1)

#>  tfunction adjust aicdiff
#>       Auto   None      -2

s_transform(myspec1)

#>  tfunction adjust aicdiff
#>       Auto   None      -2

s_usrdef(myreg1)

#>   outlier outlier.coef variables variables.coef
#> 1   FALSE        FALSE     FALSE          FALSE

s_usrdef(myspec1)

#>  outlier outlier.coef variables variables.coef
#>    FALSE        FALSE     FALSE          FALSE

myspec2 <- regarima_spec_x13(myreg1, usrdef.outliersEnabled = TRUE,
             usrdef.outliersType = c("LS", "AO"),
             usrdef.outliersDate = c("2009-10-01", "2005-02-01"))
myreg2 <- regarima(myseries, myspec2)

s_preOut(myreg2)

#>   type       date coeff
#> 1   LS 2009-10-01    NA
#> 2   AO 2005-02-01    NA

s_preOut(myspec2)

#>   type       date coeff
#> 1   LS 2009-10-01    NA
#> 2   AO 2005-02-01    NA

var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
var3 <- ts.union(var1, var2)
myspec3 <- regarima_spec_def_x13(spec = "RG5c",
                                 usrdef.varEnabled = TRUE,
                                 usrdef.var = var3)
myreg3 <- regarima(myseries, myspec3)

s_preVar(myspec3)

#> $series
#>                  var1         var2
#> Jan 1990 -14.00043517 -113.5343257
#> Feb 1990   2.55317055 -113.1053798
#> Mar 1990 -24.37263611   10.9993384
#> Apr 1990  -0.05571287   85.2905410
#> May 1990   6.21552721  -23.4337862
#> Jun 1990  11.48411606  208.6688567
#> Jul 1990 -18.21817661  -11.0919371
#> Aug 1990  -2.47325302 -139.2847056
#> Sep 1990  -2.44199607 -114.2290768
#> Oct 1990  -2.82705449  170.4608737
#> Nov 1990  -5.53699384   -8.0073634
#> Dec 1990   6.28982042  -43.7281240
#> Jan 1991  20.65024895  -11.9215094
#> Feb 1991 -16.30989402   78.6462865
#> Mar 1991   5.12426950  -57.8945246
#> Apr 1991 -18.63011492  -14.5426885
#> May 1991  -5.22012515   52.6457991
#> Jun 1991  -0.52601910  173.3578110
#> Jul 1991   5.42996343  144.8657220
#> Aug 1991  -9.14074827  151.8193149
#> Sep 1991   4.68154420  -38.4007254
#> Oct 1991   3.62951256  182.7125177
#> Nov 1991 -13.04543545  -55.1491750
#> Dec 1991   7.37776321  -86.5753541
#> Jan 1992  18.88504929  -34.3831481
#> Feb 1992  -0.97445104  106.2876458
#> Mar 1992  -9.35847354   81.3058204
#> Apr 1992  -0.15950311  180.3483361
#> May 1992  -8.26788954  -10.5068694
#> Jun 1992 -15.12399651   98.2453362
#> Jul 1992   9.35363190 -171.3302622
#> Aug 1992   1.76488611  -83.2019527
#> Sep 1992   2.43685465  110.0491882
#> Oct 1992  16.23548883  -17.3820106
#> Nov 1992   1.12038083   17.8812018
#> Dec 1992  -1.33997013  -69.8429449
#> Jan 1993 -19.10087468  -96.0449159
#> Feb 1993  -2.79237242  -97.5423042
#> Mar 1993  -3.13445978  -33.8576503
#> Apr 1993  10.67307879  115.2347074
#> May 1993   0.70034850   40.5101183
#> Jun 1993  -6.39123324  -47.0922497
#> Jul 1993  -0.49964899  -13.3251019
#> Aug 1993  -2.51483443  122.6682356
#> Sep 1993   4.44797116   33.2943995
#> Oct 1993  27.55417575  -34.7088466
#> Nov 1993   0.46531380   -9.8550690
#> Dec 1993   5.77709069    3.4766060
#> Jan 1994   1.18194874   38.6127022
#> Feb 1994 -19.11720491    2.0831228
#> Mar 1994   8.62086482    0.7586777
#> Apr 1994  -2.43236740   93.0844030
#> May 1994  -2.06087195  -68.4749941
#> Jun 1994   0.19177592   33.7401513
#> Jul 1994   0.29560754  -41.2137704
#> Aug 1994   5.49827542   93.4261130
#> Sep 1994 -22.74114857  184.0316741
#> Oct 1994  26.82557184  -70.4819663
#> Nov 1994  -3.61221255    0.8510312
#> Dec 1994   2.13355750  203.4189886
#> Jan 1995  10.74345882 -134.1686068
#> Feb 1995  -6.65088249  115.8979182
#> Mar 1995  11.13952419  -20.3208958
#> Apr 1995  -2.45896412  -37.8028555
#> May 1995 -11.77563309  173.6111043
#> Jun 1995  -9.75850616  -84.5247816
#> Jul 1995  10.65057320  -96.1571493
#> Aug 1995   1.31670635  101.7491053
#> Sep 1995   4.88628809 -149.6053742
#> Oct 1995 -16.99450568 -118.4818730
#> Nov 1995 -14.70736306   63.0234373
#> Dec 1995   2.84150344  210.1252514
#> Jan 1996  13.37320413  -61.3736810
#> Feb 1996   2.36696283 -163.4638272
#> Mar 1996  13.18293384   -1.0441117
#> Apr 1996   5.23909788  -65.6506139
#> May 1996   6.06748047  -66.9533441
#> Jun 1996  -1.09935672  -47.8589028
#> Jul 1996   1.72181715  131.9456316
#> Aug 1996  -0.90327287   63.6562761
#> Sep 1996  19.24343341   51.4327782
#> Oct 1996  12.98392759 -175.1375113
#> Nov 1996   7.48791268   89.3597518
#> Dec 1996   5.56224329   22.3038372
#> Jan 1997  -5.48257264   58.0816593
#> Feb 1997  11.10534893  -17.7821421
#> Mar 1997 -26.12334333   74.0966708
#> Apr 1997  -1.55693776  -99.7443079
#> May 1997   4.33889790 -293.8977561
#> Jun 1997  -3.81951112   71.9015661
#> Jul 1997   4.24187575  -69.8005041
#> Aug 1997  10.63101996 -189.4125843
#> Sep 1997  10.48712620    7.6299249
#> Oct 1997  -0.38102895   87.5308501
#> Nov 1997   4.86148920   45.3827393
#> Dec 1997  16.72882611  -85.0716906
#> Jan 1998  -3.54361164   56.6201613
#> Feb 1998   9.46347886  115.2211954
#> Mar 1998  13.16826356  -75.6197377
#> Apr 1998  -2.96640025  -48.9258334
#> May 1998  -3.87213575 -116.6052337
#> Jun 1998  -7.85432656  -47.9668950
#> Jul 1998 -10.56736867   11.5348218
#> Aug 1998  -7.95541430 -176.8048407
#> Sep 1998 -17.56275428 -140.7638919
#> Oct 1998  -6.90537897   70.9178461
#> Nov 1998  -5.58541994 -124.0842940
#> Dec 1998  -5.36663326  -36.8327348
#> Jan 1999   2.27127133   46.2080093
#> Feb 1999   9.78454920  -32.2833101
#> Mar 1999  -2.08882651 -128.7214810
#> Apr 1999 -13.99410460 -103.0040247
#> May 1999   2.58537288  151.4089316
#> Jun 1999  -4.41799453   34.6903586
#> Jul 1999   5.68599861  177.9441542
#> Aug 1999  21.26850459   38.6630924
#> Sep 1999   4.24858441  -91.8695239
#> Oct 1999 -16.84281532 -158.4336488
#> Nov 1999   2.49401784   -8.4058892
#> Dec 1999  10.72838252 -208.5070889
#> Jan 2000  20.39369263    0.3567992
#> Feb 2000   4.49453778  -35.5770822
#> Mar 2000  13.91814046  114.6359751
#> Apr 2000   4.26566547  -22.1188446
#> May 2000   1.07583992  101.8179021
#> Jun 2000   0.22294733  -26.3719295
#> Jul 2000   6.03611011  165.8542305
#> Aug 2000  -2.62650573  -77.4086771
#> Sep 2000  -5.28264082  -92.3937880
#> Oct 2000   1.92149422  -27.5533378
#> Nov 2000 -11.46199669  -59.3399688
#> Dec 2000   8.46184665  -12.2285891
#> Jan 2001   0.81719629  117.9784246
#> Feb 2001 -13.05117010   64.1037374
#> Mar 2001  -9.44912060  -62.9588508
#> Apr 2001   4.54341594  -80.7734971
#> May 2001  -8.55202501  -86.0489929
#> Jun 2001  -2.86895219 -216.9238693
#> Jul 2001   8.94961626 -137.5836518
#> Aug 2001   0.67304440  -49.3132472
#> Sep 2001  -1.62676337  -58.1652027
#> Oct 2001  -8.27310169  -16.7229304
#> Nov 2001  18.76505621   48.5993129
#> Dec 2001   7.66440199 -133.3395796
#> Jan 2002   9.79956696  -26.1965625
#> Feb 2002  13.21780992   65.2386303
#> Mar 2002 -11.19710829   74.8854971
#> Apr 2002   5.14599819   89.6560285
#> May 2002 -15.09099836  148.9300424
#> Jun 2002  15.32741480  -65.9403481
#> Jul 2002   4.29147371   53.7283179
#> Aug 2002   1.22103414   74.6803067
#> Sep 2002 -11.38012401  189.6317084
#> Oct 2002  -5.58015129 -206.0070725
#> Nov 2002  10.52538537    6.4543870
#> Dec 2002   6.77683644  -26.5147403
#> Jan 2003   0.38499547  -44.7344531
#> Feb 2003  -3.56381187 -141.0700927
#> Mar 2003   7.82844102  -50.6418882
#> Apr 2003   8.04411616  -26.9761838
#> May 2003 -19.00060823 -108.5154918
#> Jun 2003   9.35784286   36.2159127
#> Jul 2003  -3.09051503  -33.5672143
#> Aug 2003   2.63066677  136.3804498
#> Sep 2003 -17.90591856  -71.1524136
#> Oct 2003  -7.88258845   66.2178797
#> Nov 2003 -11.33021669   29.1130223
#> Dec 2003   3.63652568   19.7958000
#> Jan 2004  -2.85887914 -120.3566106
#> Feb 2004   5.17669134   -3.9817044
#> Mar 2004  -1.02908670   68.6982465
#> Apr 2004  -9.74069593   70.5267007
#> May 2004  12.70672301   99.1441680
#> Jun 2004   9.60864787  114.4248971
#> Jul 2004   7.68721370 -123.8910243
#> Aug 2004  10.35930771  265.4898333
#> Sep 2004  -4.73887074  -15.6917189
#> Oct 2004 -12.75334875  -42.3490117
#> Nov 2004  -3.05620674  -19.8387058
#> Dec 2004  22.11769487  -89.4802407
#> Jan 2005 -10.41668381   90.4269119
#> Feb 2005 -11.46523850    7.9649210
#> Mar 2005 -16.75327303 -125.8827223
#> Apr 2005  15.25938655  102.5685106
#> May 2005   5.54185515  -73.0778603
#> Jun 2005  19.93110265  -19.0145507
#> Jul 2005  -1.54120740   52.8864693
#> Aug 2005  25.64408338   55.0210535
#> Sep 2005  10.61999145   54.9684337
#> Oct 2005  11.42694878  -65.9542372
#> Nov 2005  11.23838843    5.7421706
#> Dec 2005  -3.97001493 -280.8010508
#> Jan 2006  -8.23261151  -91.2259753
#> Feb 2006  -5.78884625  -78.2379163
#> Mar 2006  17.63789378  -66.4104924
#> Apr 2006   1.32992146   62.6309770
#> May 2006   3.76499328  -50.7248206
#> Jun 2006  11.38707653   27.0361335
#> Jul 2006  12.41263075   46.7476865
#> Aug 2006   6.12090945   72.3994958
#> Sep 2006  -4.29380087   61.3836939
#> Oct 2006  13.60461327  -61.7869202
#> Nov 2006  -0.70857431   22.0724902
#> Dec 2006  -2.72153684  112.7926598
#> Jan 2007 -24.46680029  181.3454336
#> Feb 2007   0.65486641   -8.3825685
#> Mar 2007 -10.98508902  136.7706666
#> Apr 2007  -6.33178176  -62.7434620
#> May 2007 -20.63654451  -21.6629150
#> Jun 2007  26.48932029  -68.3713824
#> Jul 2007 -11.53398386  -44.4702734
#> Aug 2007  -3.40637876   60.6489806
#> Sep 2007   7.86362576   62.4183075
#> Oct 2007 -12.70513110  -69.5431074
#> Nov 2007   5.42141549  -78.3639078
#> Dec 2007   0.75105900  -95.3123859
#> Jan 2008   5.58514422  179.2756071
#> Feb 2008   4.15406399   34.8976696
#> Mar 2008 -14.52299769   25.9103768
#> Apr 2008   9.41206122  -80.5951897
#> May 2008  -3.38935872   10.5664701
#> Jun 2008  -0.75574247  -33.3599682
#> Jul 2008   0.40204392  164.1847970
#> Aug 2008   1.24301066  -64.3905859
#> Sep 2008  -9.98432551   58.7020562
#> Oct 2008  12.33390065  -15.0403088
#> Nov 2008   3.40424488 -171.0821848
#> Dec 2008  -4.72702482  143.1032558
#> Jan 2009   7.08753061 -264.5212268
#> Feb 2009 -15.28958715 -103.2457405
#> Mar 2009   2.37425345  -70.7466431
#> Apr 2009 -13.12814246  -70.0560014
#> May 2009   7.47028587   53.7885439
#> Jun 2009 -15.62518435  -31.6332175
#> Jul 2009   0.71053360  -83.9622754
#> Aug 2009  -6.39534770 -135.4928062
#> Sep 2009  -8.45195739  -81.7568272
#> Oct 2009   6.75244698  -63.4400003
#> Nov 2009  11.53375794   81.5949433
#> Dec 2009 -16.86504742   30.2795706
#> Jan 2010  -9.02814949  180.7086625
#> Feb 2010  13.17633698  -89.4026756
#> Mar 2010  11.00189745   -4.6428211
#> Apr 2010  12.03767839  -47.1179138
#> May 2010 -14.31270777  -52.6692630
#> Jun 2010  13.82910861   -9.5134908
#> Jul 2010   0.03125940 -249.5364809
#> Aug 2010  -0.77886824   16.6889217
#> Sep 2010   4.41428226   35.0492384
#> Oct 2010   1.28922896  143.3701009
#> Nov 2010  -8.30214260   76.5906803
#> Dec 2010  -5.03592910  116.7520670
#> Jan 2011 -11.93641182  -13.6943429
#> Feb 2011  -7.51723323  -51.4902044
#> Mar 2011  14.55841403  151.9744468
#> Apr 2011  -8.28603533  -32.8491678
#> May 2011   2.89774460   -5.3671506
#> Jun 2011  -4.80053484  -56.3524635
#> Jul 2011  -6.04829354  -74.3908963
#> Aug 2011  14.60110180  -10.9041651
#> Sep 2011   1.49679354  -56.0829227
#> Oct 2011 -14.33321100   18.8001549
#> Nov 2011  -0.10303319   74.8850942
#> Dec 2011  -2.12236035 -191.6538316
#> Jan 2012  -9.06340179   23.6095847
#> Feb 2012 -21.02152479   62.8953415
#> Mar 2012  18.93360464   41.7925676
#> Apr 2012  -9.68125837  197.6758477
#> May 2012  -1.02603036  -50.6286298
#> Jun 2012   2.39959572 -110.9968853
#> Jul 2012   0.60898893  -94.8705723
#> Aug 2012 -21.77576028   47.6843757
#> Sep 2012  -1.17860143  -79.5201560
#> Oct 2012   1.12294787   23.4326923
#> Nov 2012   0.07886198 -122.2451097
#> Dec 2012  18.77743872 -245.3647354
#> Jan 2013  21.58756554 -148.9260814
#> Feb 2013   7.09714522  -43.2147734
#> Mar 2013   7.66983379  -94.2554006
#> Apr 2013  -3.08211421  -12.1450799
#> May 2013  10.12001849  133.6446798
#> Jun 2013  -9.19051597  -86.0356182
#> Jul 2013   5.63380077   66.6537820
#> Aug 2013   3.22482749 -142.1534746
#> Sep 2013   3.66674363  117.0056168
#> Oct 2013  11.29835153 -140.4714543
#> Nov 2013  -9.41498076  110.1708096
#> Dec 2013   2.17837643   69.7986263
#> Jan 2014  14.15412293  -86.4349803
#> Feb 2014  -3.83733048 -109.1470351
#> Mar 2014  -1.74086374   -3.7051465
#> Apr 2014  -2.21744517   81.0053792
#> May 2014 -10.09528722  -49.9355412
#> Jun 2014   4.80725266   94.8031588
#> Jul 2014  16.04407328  -17.4245957
#> Aug 2014 -15.15024529 -110.6235952
#> Sep 2014 -14.16023914  -94.5985005
#> Oct 2014   8.76777327   28.9089591
#> Nov 2014   6.24132413   87.6913145
#> Dec 2014  21.12277288 -114.8903940
#> Jan 2015  -3.56124416 -113.7612756
#> Feb 2015 -10.64464209 -143.7246735
#> Mar 2015  10.77116538  -49.4143476
#> Apr 2015  11.81575567   84.0801808
#> May 2015   1.98392095   79.1534124
#> Jun 2015  -4.00405249  -16.8848948
#> Jul 2015   6.16154281   61.2722104
#> Aug 2015  19.74156748  -77.1158924
#> Sep 2015  18.84662324   88.8628993
#> Oct 2015 -15.88620547    1.3214477
#> Nov 2015  -5.39923164   22.5339515
#> Dec 2015 -11.69461464  -72.9915210
#> Jan 2016   5.59105989 -122.2487070
#> Feb 2016 -18.19347247   40.6805171
#> Mar 2016   3.93343972  -75.1012223
#> Apr 2016   0.42134106  -16.2116540
#> May 2016  11.79664177   35.2010126
#> Jun 2016  -2.56921176  -28.9058300
#> Jul 2016 -10.56336098   10.4662227
#> Aug 2016   1.98777205   72.0186531
#> Sep 2016   6.50533552  -61.1046082
#> Oct 2016   3.43913337 -110.6914072
#> Nov 2016  14.77532312   53.4803326
#> Dec 2016   0.72025698   73.6067968
#> Jan 2017  21.26444534 -122.2501574
#> Feb 2017 -14.76196906  102.1415310
#> Mar 2017   4.07888500   46.5165158
#> Apr 2017  13.93977798   79.0472705
#> May 2017   3.60278296  -13.0264801
#> Jun 2017   6.54550251  -93.0285334
#> Jul 2017  10.52155422  -36.4851004
#> Aug 2017 -19.79555125   15.3872493
#> Sep 2017  12.08385605   41.3154818
#> Oct 2017  -1.69280084  248.0823360
#> Nov 2017   2.95029753 -217.9956742
#> Dec 2017  12.66340587   42.0874578
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 

s_preVar(myreg3)

#> $series
#>                  var1         var2
#> Jan 1990 -14.00043517 -113.5343257
#> Feb 1990   2.55317055 -113.1053798
#> Mar 1990 -24.37263611   10.9993384
#> Apr 1990  -0.05571287   85.2905410
#> May 1990   6.21552721  -23.4337862
#> Jun 1990  11.48411606  208.6688567
#> Jul 1990 -18.21817661  -11.0919371
#> Aug 1990  -2.47325302 -139.2847056
#> Sep 1990  -2.44199607 -114.2290768
#> Oct 1990  -2.82705449  170.4608737
#> Nov 1990  -5.53699384   -8.0073634
#> Dec 1990   6.28982042  -43.7281240
#> Jan 1991  20.65024895  -11.9215094
#> Feb 1991 -16.30989402   78.6462865
#> Mar 1991   5.12426950  -57.8945246
#> Apr 1991 -18.63011492  -14.5426885
#> May 1991  -5.22012515   52.6457991
#> Jun 1991  -0.52601910  173.3578110
#> Jul 1991   5.42996343  144.8657220
#> Aug 1991  -9.14074827  151.8193149
#> Sep 1991   4.68154420  -38.4007254
#> Oct 1991   3.62951256  182.7125177
#> Nov 1991 -13.04543545  -55.1491750
#> Dec 1991   7.37776321  -86.5753541
#> Jan 1992  18.88504929  -34.3831481
#> Feb 1992  -0.97445104  106.2876458
#> Mar 1992  -9.35847354   81.3058204
#> Apr 1992  -0.15950311  180.3483361
#> May 1992  -8.26788954  -10.5068694
#> Jun 1992 -15.12399651   98.2453362
#> Jul 1992   9.35363190 -171.3302622
#> Aug 1992   1.76488611  -83.2019527
#> Sep 1992   2.43685465  110.0491882
#> Oct 1992  16.23548883  -17.3820106
#> Nov 1992   1.12038083   17.8812018
#> Dec 1992  -1.33997013  -69.8429449
#> Jan 1993 -19.10087468  -96.0449159
#> Feb 1993  -2.79237242  -97.5423042
#> Mar 1993  -3.13445978  -33.8576503
#> Apr 1993  10.67307879  115.2347074
#> May 1993   0.70034850   40.5101183
#> Jun 1993  -6.39123324  -47.0922497
#> Jul 1993  -0.49964899  -13.3251019
#> Aug 1993  -2.51483443  122.6682356
#> Sep 1993   4.44797116   33.2943995
#> Oct 1993  27.55417575  -34.7088466
#> Nov 1993   0.46531380   -9.8550690
#> Dec 1993   5.77709069    3.4766060
#> Jan 1994   1.18194874   38.6127022
#> Feb 1994 -19.11720491    2.0831228
#> Mar 1994   8.62086482    0.7586777
#> Apr 1994  -2.43236740   93.0844030
#> May 1994  -2.06087195  -68.4749941
#> Jun 1994   0.19177592   33.7401513
#> Jul 1994   0.29560754  -41.2137704
#> Aug 1994   5.49827542   93.4261130
#> Sep 1994 -22.74114857  184.0316741
#> Oct 1994  26.82557184  -70.4819663
#> Nov 1994  -3.61221255    0.8510312
#> Dec 1994   2.13355750  203.4189886
#> Jan 1995  10.74345882 -134.1686068
#> Feb 1995  -6.65088249  115.8979182
#> Mar 1995  11.13952419  -20.3208958
#> Apr 1995  -2.45896412  -37.8028555
#> May 1995 -11.77563309  173.6111043
#> Jun 1995  -9.75850616  -84.5247816
#> Jul 1995  10.65057320  -96.1571493
#> Aug 1995   1.31670635  101.7491053
#> Sep 1995   4.88628809 -149.6053742
#> Oct 1995 -16.99450568 -118.4818730
#> Nov 1995 -14.70736306   63.0234373
#> Dec 1995   2.84150344  210.1252514
#> Jan 1996  13.37320413  -61.3736810
#> Feb 1996   2.36696283 -163.4638272
#> Mar 1996  13.18293384   -1.0441117
#> Apr 1996   5.23909788  -65.6506139
#> May 1996   6.06748047  -66.9533441
#> Jun 1996  -1.09935672  -47.8589028
#> Jul 1996   1.72181715  131.9456316
#> Aug 1996  -0.90327287   63.6562761
#> Sep 1996  19.24343341   51.4327782
#> Oct 1996  12.98392759 -175.1375113
#> Nov 1996   7.48791268   89.3597518
#> Dec 1996   5.56224329   22.3038372
#> Jan 1997  -5.48257264   58.0816593
#> Feb 1997  11.10534893  -17.7821421
#> Mar 1997 -26.12334333   74.0966708
#> Apr 1997  -1.55693776  -99.7443079
#> May 1997   4.33889790 -293.8977561
#> Jun 1997  -3.81951112   71.9015661
#> Jul 1997   4.24187575  -69.8005041
#> Aug 1997  10.63101996 -189.4125843
#> Sep 1997  10.48712620    7.6299249
#> Oct 1997  -0.38102895   87.5308501
#> Nov 1997   4.86148920   45.3827393
#> Dec 1997  16.72882611  -85.0716906
#> Jan 1998  -3.54361164   56.6201613
#> Feb 1998   9.46347886  115.2211954
#> Mar 1998  13.16826356  -75.6197377
#> Apr 1998  -2.96640025  -48.9258334
#> May 1998  -3.87213575 -116.6052337
#> Jun 1998  -7.85432656  -47.9668950
#> Jul 1998 -10.56736867   11.5348218
#> Aug 1998  -7.95541430 -176.8048407
#> Sep 1998 -17.56275428 -140.7638919
#> Oct 1998  -6.90537897   70.9178461
#> Nov 1998  -5.58541994 -124.0842940
#> Dec 1998  -5.36663326  -36.8327348
#> Jan 1999   2.27127133   46.2080093
#> Feb 1999   9.78454920  -32.2833101
#> Mar 1999  -2.08882651 -128.7214810
#> Apr 1999 -13.99410460 -103.0040247
#> May 1999   2.58537288  151.4089316
#> Jun 1999  -4.41799453   34.6903586
#> Jul 1999   5.68599861  177.9441542
#> Aug 1999  21.26850459   38.6630924
#> Sep 1999   4.24858441  -91.8695239
#> Oct 1999 -16.84281532 -158.4336488
#> Nov 1999   2.49401784   -8.4058892
#> Dec 1999  10.72838252 -208.5070889
#> Jan 2000  20.39369263    0.3567992
#> Feb 2000   4.49453778  -35.5770822
#> Mar 2000  13.91814046  114.6359751
#> Apr 2000   4.26566547  -22.1188446
#> May 2000   1.07583992  101.8179021
#> Jun 2000   0.22294733  -26.3719295
#> Jul 2000   6.03611011  165.8542305
#> Aug 2000  -2.62650573  -77.4086771
#> Sep 2000  -5.28264082  -92.3937880
#> Oct 2000   1.92149422  -27.5533378
#> Nov 2000 -11.46199669  -59.3399688
#> Dec 2000   8.46184665  -12.2285891
#> Jan 2001   0.81719629  117.9784246
#> Feb 2001 -13.05117010   64.1037374
#> Mar 2001  -9.44912060  -62.9588508
#> Apr 2001   4.54341594  -80.7734971
#> May 2001  -8.55202501  -86.0489929
#> Jun 2001  -2.86895219 -216.9238693
#> Jul 2001   8.94961626 -137.5836518
#> Aug 2001   0.67304440  -49.3132472
#> Sep 2001  -1.62676337  -58.1652027
#> Oct 2001  -8.27310169  -16.7229304
#> Nov 2001  18.76505621   48.5993129
#> Dec 2001   7.66440199 -133.3395796
#> Jan 2002   9.79956696  -26.1965625
#> Feb 2002  13.21780992   65.2386303
#> Mar 2002 -11.19710829   74.8854971
#> Apr 2002   5.14599819   89.6560285
#> May 2002 -15.09099836  148.9300424
#> Jun 2002  15.32741480  -65.9403481
#> Jul 2002   4.29147371   53.7283179
#> Aug 2002   1.22103414   74.6803067
#> Sep 2002 -11.38012401  189.6317084
#> Oct 2002  -5.58015129 -206.0070725
#> Nov 2002  10.52538537    6.4543870
#> Dec 2002   6.77683644  -26.5147403
#> Jan 2003   0.38499547  -44.7344531
#> Feb 2003  -3.56381187 -141.0700927
#> Mar 2003   7.82844102  -50.6418882
#> Apr 2003   8.04411616  -26.9761838
#> May 2003 -19.00060823 -108.5154918
#> Jun 2003   9.35784286   36.2159127
#> Jul 2003  -3.09051503  -33.5672143
#> Aug 2003   2.63066677  136.3804498
#> Sep 2003 -17.90591856  -71.1524136
#> Oct 2003  -7.88258845   66.2178797
#> Nov 2003 -11.33021669   29.1130223
#> Dec 2003   3.63652568   19.7958000
#> Jan 2004  -2.85887914 -120.3566106
#> Feb 2004   5.17669134   -3.9817044
#> Mar 2004  -1.02908670   68.6982465
#> Apr 2004  -9.74069593   70.5267007
#> May 2004  12.70672301   99.1441680
#> Jun 2004   9.60864787  114.4248971
#> Jul 2004   7.68721370 -123.8910243
#> Aug 2004  10.35930771  265.4898333
#> Sep 2004  -4.73887074  -15.6917189
#> Oct 2004 -12.75334875  -42.3490117
#> Nov 2004  -3.05620674  -19.8387058
#> Dec 2004  22.11769487  -89.4802407
#> Jan 2005 -10.41668381   90.4269119
#> Feb 2005 -11.46523850    7.9649210
#> Mar 2005 -16.75327303 -125.8827223
#> Apr 2005  15.25938655  102.5685106
#> May 2005   5.54185515  -73.0778603
#> Jun 2005  19.93110265  -19.0145507
#> Jul 2005  -1.54120740   52.8864693
#> Aug 2005  25.64408338   55.0210535
#> Sep 2005  10.61999145   54.9684337
#> Oct 2005  11.42694878  -65.9542372
#> Nov 2005  11.23838843    5.7421706
#> Dec 2005  -3.97001493 -280.8010508
#> Jan 2006  -8.23261151  -91.2259753
#> Feb 2006  -5.78884625  -78.2379163
#> Mar 2006  17.63789378  -66.4104924
#> Apr 2006   1.32992146   62.6309770
#> May 2006   3.76499328  -50.7248206
#> Jun 2006  11.38707653   27.0361335
#> Jul 2006  12.41263075   46.7476865
#> Aug 2006   6.12090945   72.3994958
#> Sep 2006  -4.29380087   61.3836939
#> Oct 2006  13.60461327  -61.7869202
#> Nov 2006  -0.70857431   22.0724902
#> Dec 2006  -2.72153684  112.7926598
#> Jan 2007 -24.46680029  181.3454336
#> Feb 2007   0.65486641   -8.3825685
#> Mar 2007 -10.98508902  136.7706666
#> Apr 2007  -6.33178176  -62.7434620
#> May 2007 -20.63654451  -21.6629150
#> Jun 2007  26.48932029  -68.3713824
#> Jul 2007 -11.53398386  -44.4702734
#> Aug 2007  -3.40637876   60.6489806
#> Sep 2007   7.86362576   62.4183075
#> Oct 2007 -12.70513110  -69.5431074
#> Nov 2007   5.42141549  -78.3639078
#> Dec 2007   0.75105900  -95.3123859
#> Jan 2008   5.58514422  179.2756071
#> Feb 2008   4.15406399   34.8976696
#> Mar 2008 -14.52299769   25.9103768
#> Apr 2008   9.41206122  -80.5951897
#> May 2008  -3.38935872   10.5664701
#> Jun 2008  -0.75574247  -33.3599682
#> Jul 2008   0.40204392  164.1847970
#> Aug 2008   1.24301066  -64.3905859
#> Sep 2008  -9.98432551   58.7020562
#> Oct 2008  12.33390065  -15.0403088
#> Nov 2008   3.40424488 -171.0821848
#> Dec 2008  -4.72702482  143.1032558
#> Jan 2009   7.08753061 -264.5212268
#> Feb 2009 -15.28958715 -103.2457405
#> Mar 2009   2.37425345  -70.7466431
#> Apr 2009 -13.12814246  -70.0560014
#> May 2009   7.47028587   53.7885439
#> Jun 2009 -15.62518435  -31.6332175
#> Jul 2009   0.71053360  -83.9622754
#> Aug 2009  -6.39534770 -135.4928062
#> Sep 2009  -8.45195739  -81.7568272
#> Oct 2009   6.75244698  -63.4400003
#> Nov 2009  11.53375794   81.5949433
#> Dec 2009 -16.86504742   30.2795706
#> Jan 2010  -9.02814949  180.7086625
#> Feb 2010  13.17633698  -89.4026756
#> Mar 2010  11.00189745   -4.6428211
#> Apr 2010  12.03767839  -47.1179138
#> May 2010 -14.31270777  -52.6692630
#> Jun 2010  13.82910861   -9.5134908
#> Jul 2010   0.03125940 -249.5364809
#> Aug 2010  -0.77886824   16.6889217
#> Sep 2010   4.41428226   35.0492384
#> Oct 2010   1.28922896  143.3701009
#> Nov 2010  -8.30214260   76.5906803
#> Dec 2010  -5.03592910  116.7520670
#> Jan 2011 -11.93641182  -13.6943429
#> Feb 2011  -7.51723323  -51.4902044
#> Mar 2011  14.55841403  151.9744468
#> Apr 2011  -8.28603533  -32.8491678
#> May 2011   2.89774460   -5.3671506
#> Jun 2011  -4.80053484  -56.3524635
#> Jul 2011  -6.04829354  -74.3908963
#> Aug 2011  14.60110180  -10.9041651
#> Sep 2011   1.49679354  -56.0829227
#> Oct 2011 -14.33321100   18.8001549
#> Nov 2011  -0.10303319   74.8850942
#> Dec 2011  -2.12236035 -191.6538316
#> Jan 2012  -9.06340179   23.6095847
#> Feb 2012 -21.02152479   62.8953415
#> Mar 2012  18.93360464   41.7925676
#> Apr 2012  -9.68125837  197.6758477
#> May 2012  -1.02603036  -50.6286298
#> Jun 2012   2.39959572 -110.9968853
#> Jul 2012   0.60898893  -94.8705723
#> Aug 2012 -21.77576028   47.6843757
#> Sep 2012  -1.17860143  -79.5201560
#> Oct 2012   1.12294787   23.4326923
#> Nov 2012   0.07886198 -122.2451097
#> Dec 2012  18.77743872 -245.3647354
#> Jan 2013  21.58756554 -148.9260814
#> Feb 2013   7.09714522  -43.2147734
#> Mar 2013   7.66983379  -94.2554006
#> Apr 2013  -3.08211421  -12.1450799
#> May 2013  10.12001849  133.6446798
#> Jun 2013  -9.19051597  -86.0356182
#> Jul 2013   5.63380077   66.6537820
#> Aug 2013   3.22482749 -142.1534746
#> Sep 2013   3.66674363  117.0056168
#> Oct 2013  11.29835153 -140.4714543
#> Nov 2013  -9.41498076  110.1708096
#> Dec 2013   2.17837643   69.7986263
#> Jan 2014  14.15412293  -86.4349803
#> Feb 2014  -3.83733048 -109.1470351
#> Mar 2014  -1.74086374   -3.7051465
#> Apr 2014  -2.21744517   81.0053792
#> May 2014 -10.09528722  -49.9355412
#> Jun 2014   4.80725266   94.8031588
#> Jul 2014  16.04407328  -17.4245957
#> Aug 2014 -15.15024529 -110.6235952
#> Sep 2014 -14.16023914  -94.5985005
#> Oct 2014   8.76777327   28.9089591
#> Nov 2014   6.24132413   87.6913145
#> Dec 2014  21.12277288 -114.8903940
#> Jan 2015  -3.56124416 -113.7612756
#> Feb 2015 -10.64464209 -143.7246735
#> Mar 2015  10.77116538  -49.4143476
#> Apr 2015  11.81575567   84.0801808
#> May 2015   1.98392095   79.1534124
#> Jun 2015  -4.00405249  -16.8848948
#> Jul 2015   6.16154281   61.2722104
#> Aug 2015  19.74156748  -77.1158924
#> Sep 2015  18.84662324   88.8628993
#> Oct 2015 -15.88620547    1.3214477
#> Nov 2015  -5.39923164   22.5339515
#> Dec 2015 -11.69461464  -72.9915210
#> Jan 2016   5.59105989 -122.2487070
#> Feb 2016 -18.19347247   40.6805171
#> Mar 2016   3.93343972  -75.1012223
#> Apr 2016   0.42134106  -16.2116540
#> May 2016  11.79664177   35.2010126
#> Jun 2016  -2.56921176  -28.9058300
#> Jul 2016 -10.56336098   10.4662227
#> Aug 2016   1.98777205   72.0186531
#> Sep 2016   6.50533552  -61.1046082
#> Oct 2016   3.43913337 -110.6914072
#> Nov 2016  14.77532312   53.4803326
#> Dec 2016   0.72025698   73.6067968
#> Jan 2017  21.26444534 -122.2501574
#> Feb 2017 -14.76196906  102.1415310
#> Mar 2017   4.07888500   46.5165158
#> Apr 2017  13.93977798   79.0472705
#> May 2017   3.60278296  -13.0264801
#> Jun 2017   6.54550251  -93.0285334
#> Jul 2017  10.52155422  -36.4851004
#> Aug 2017 -19.79555125   15.3872493
#> Sep 2017  12.08385605   41.3154818
#> Oct 2017  -1.69280084  248.0823360
#> Nov 2017   2.95029753 -217.9956742
#> Dec 2017  12.66340587   42.0874578
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 

s_td(myreg1)

#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove

s_td(myspec1)

#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove

s_easter(myreg1)

#>  enabled julian duration test
#>     TRUE  FALSE        8  Add

s_easter(myspec1)

#>  enabled julian duration test
#>     TRUE  FALSE        8  Add

s_out(myreg1)

#>  enabled span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE  All TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7

s_out(myspec1)

#>  enabled            span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE From 2010-03-01 TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7

s_arima(myreg1)

#>  enabled automdl.acceptdefault automdl.cancel automdl.ub1 automdl.ub2
#>     TRUE                 FALSE            0.1    1.041667        0.88
#>  automdl.mixed automdl.balanced automdl.armalimit automdl.reducecv
#>           TRUE            FALSE                 1          0.14286
#>  automdl.ljungboxlimit automdl.ubfinal arima.mu arima.p arima.d arima.q
#>                   0.95            1.05    FALSE       0       1       1
#>  arima.bp arima.bd arima.bq arima.coef
#>         0        1        1      FALSE

s_arima(myspec1)

#>  enabled automdl.acceptdefault automdl.cancel automdl.ub1 automdl.ub2
#>     TRUE                 FALSE            0.1    1.041667        0.88
#>  automdl.mixed automdl.balanced automdl.armalimit automdl.reducecv
#>           TRUE            FALSE                 1          0.14286
#>  automdl.ljungboxlimit automdl.ubfinal arima.mu arima.p arima.d arima.q
#>                   0.95            1.05    FALSE       0       1       1
#>  arima.bp arima.bd arima.bq arima.coef
#>         0        1        1      FALSE

myspec4 <- regarima_spec_x13(myreg1, automdl.enabled = FALSE,
             arima.coefEnabled = TRUE,
             arima.p = 1,arima.q = 1, arima.bp = 1, arima.bq = 1,
             arima.coef = rep(0.2, 4),
             arima.coefType = rep("Initial", 4))
myreg4 <- regarima(myseries, myspec4)

s_arimaCoef(myreg4)

#>              Type Value
#> Phi(1)    Initial   0.2
#> Theta(1)  Initial   0.2
#> BPhi(1)   Initial   0.2
#> BTheta(1) Initial   0.2

s_arimaCoef(myspec4)

#>              Type Value
#> Phi(1)    Initial   0.2
#> Theta(1)  Initial   0.2
#> BPhi(1)   Initial   0.2
#> BTheta(1) Initial   0.2

s_fcst(myreg1)

#>  horizon
#>       -2

s_fcst(myspec1)

#>  horizon
#>       -2

s_span(myreg1)

#>          type         d0         d1 n0 n1
#> estimate  All 1900-01-01 2020-12-31  0  0
#> outlier   All 1900-01-01 2020-12-31  0  0

s_span(myspec1)

#>          type         d0         d1 n0 n1
#> estimate From 2005-10-01 2020-12-31  0  0
#> outlier  From 2010-03-01 2020-12-31  0  0

myspec5 <- x13_spec_def(spec = c("RSA5c"), x11.seasonalComp = FALSE)
mysa5 <- x13(myseries, myspec5)

s_x11(mysa5)

#>   x11.mode x11.seasonalComp x11.lsigma x11.usigma x11.trendAuto x11.trendma
#>  Undefined            FALSE        1.5        2.5          TRUE          13
#>  x11.seasonalma x11.fcasts x11.bcasts x11.excludeFcasts
#>             Msr         -1          0             FALSE

s_x11(myspec5)

#>   x11.mode x11.seasonalComp x11.lsigma x11.usigma x11.trendAuto x11.trendma
#>  Undefined            FALSE        1.5        2.5          TRUE          13
#>  x11.seasonalma x11.fcasts x11.bcasts x11.excludeFcasts
#>             Msr         -1          0             FALSE

myspec6 <- tramoseats_spec_def(spec = c("RSAfull"), seats.approx = "Noisy")
mysa6 <- tramoseats(myseries, myspec6)

s_seats(mysa6)

#>  seats.approx seats.maBoundary seats.trendBoundary seats.seasdBoundary
#>         Noisy             0.95                 0.5                 0.8
#>  seats.seasdBoundary1 seats.seasTol seats.method
#>                   0.8             2       Burman

s_seats(mysa6)

#>  seats.approx seats.maBoundary seats.trendBoundary seats.seasdBoundary
#>         Noisy             0.95                 0.5                 0.8
#>  seats.seasdBoundary1 seats.seasTol seats.method
#>                   0.8             2       Burman