Access a model specification, a SA or a pre-adjustment model in X13 and TRAMO-SEATS

The following functions enable the access to different parts of the final model specification, as included in the "SA", "regarima", "SA_spec" and "regarima_spec" S3 class objects.

Usage

s_estimate(object = NA)

s_transform(object = NA)

s_usrdef(object = NA)

s_preOut(object = NA)

s_preVar(object = NA)

s_td(object = NA)

s_easter(object = NA)

s_out(object = NA)

s_arima(object = NA)

s_arimaCoef(object = NA)

s_fcst(object = NA)

s_span(object = NA)

s_x11(object = NA)

s_benchmarking(object = NA)

s_seats(object = NA)

Arguments

object

an object of one of the following classes: c("SA","X13"), c("SA","TRAMO_SEATS"), c("SA_spec","X13"), c("SA_spec","TRAMO_SEATS"), c("regarima","X13"), c("regarima","TRAMO_SEATS"), c("regarima_spec","X13"), c("regarima_spec","TRAMO_SEATS").

Value

s_estimate returns a data.frame with the estimate variables

s_transform returns a data.frame with the transform variables

s_usrdef returns a data.frame with the user-defined regressors (outliers and variables) model specification, indicating if those variables are included in the model and if coefficients are pre-specified

s_preOut returns a data.frame with the pre-specified outliers

s_preVar returns a list with information on the user-defined variables, including: series - the time series and description - data.frame with the variable type and coefficients

s_td returns a data.frame with the trading.days variables

s_easter returns a data.frame with the easter variable

s_out returns a data.frame with the outliers detection variables

s_arima returns a data.frame with the arima variables

s_arimaCoef returns a data.frame with the user-specified ARMA coefficients

s_fcst returns a data.frame with the forecast horizon

s_span returns a data.frame with the span variables

s_x11 returns a data.frame with the x11 variables

s_seats returns a data.frame with the seats variables

References

More information and examples related to 'JDemetra+' features in the online documentation: https://jdemetra-new-documentation.netlify.app/

Examples

# \donttest{
myseries <- ipi_c_eu[, "FR"]
myreg1 <- regarima_x13(myseries, spec = "RG5c")
myspec1 <- regarima_spec_x13(myreg1,
             estimate.from = "2005-10-01",
             outlier.from = "2010-03-01")

s_estimate(myreg1)
#>  preliminary.check span tolerance
#>               TRUE  All     1e-07
s_estimate(myspec1)
#>  preliminary.check            span tolerance
#>               TRUE From 2005-10-01     1e-07

s_transform(myreg1)
#>  tfunction adjust aicdiff
#>       Auto   None      -2
s_transform(myspec1)
#>  tfunction adjust aicdiff
#>       Auto   None      -2

s_usrdef(myreg1)
#>   outlier outlier.coef variables variables.coef
#> 1   FALSE        FALSE     FALSE          FALSE
s_usrdef(myspec1)
#>  outlier outlier.coef variables variables.coef
#>    FALSE        FALSE     FALSE          FALSE

myspec2 <- regarima_spec_x13(myreg1, usrdef.outliersEnabled = TRUE,
             usrdef.outliersType = c("LS", "AO"),
             usrdef.outliersDate = c("2009-10-01", "2005-02-01"))
myreg2 <- regarima(myseries, myspec2)

s_preOut(myreg2)
#>   type       date coeff
#> 1   LS 2009-10-01    NA
#> 2   AO 2005-02-01    NA
s_preOut(myspec2)
#>   type       date coeff
#> 1   LS 2009-10-01    NA
#> 2   AO 2005-02-01    NA

var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
var3 <- ts.union(var1, var2)
myspec3 <- regarima_spec_x13(spec = "RG5c",
                             usrdef.varEnabled = TRUE,
                             usrdef.var = var3)
myreg3 <- regarima(myseries, myspec3)

s_preVar(myspec3)
#> $series
#>                  var1         var2
#> Jan 1990 -10.67709165 -229.9484409
#> Feb 1990 -11.80488666  -43.8727684
#> Mar 1990  21.59175760  -64.9680607
#> Apr 1990  -4.17812648   42.4040947
#> May 1990   0.87654958    4.7049728
#> Jun 1990  -5.43709068 -128.5044066
#> Jul 1990  -1.19867022    4.5532358
#> Aug 1990  18.00624858 -144.9429700
#> Sep 1990  -7.94247350 -163.3704403
#> Oct 1990  -3.60360559  -51.7284962
#> Nov 1990   6.31460646  -99.1151430
#> Dec 1990 -16.63279352  -72.8485045
#> Jan 1991  -4.88005245   45.1951634
#> Feb 1991  -0.20752219   88.7181335
#> Mar 1991   9.33871038  126.4573401
#> Apr 1991  -2.32670878 -129.1419391
#> May 1991 -10.19630064   18.2510306
#> Jun 1991 -16.80328321   -8.9809675
#> Jul 1991  22.81394917  -11.2629276
#> Aug 1991 -10.42285303   16.2801338
#> Sep 1991   7.12307192   55.5861006
#> Oct 1991  12.20747981   85.3394680
#> Nov 1991   3.36018683  -19.4899361
#> Dec 1991  -2.79147937   -4.3411909
#> Jan 1992   0.76422131  -52.9438316
#> Feb 1992  -3.54509030  -59.8510599
#> Mar 1992  14.32658547   14.9297682
#> Apr 1992   0.02671474  125.5820916
#> May 1992   3.44012511  -23.4369528
#> Jun 1992  -8.30323380  142.2340276
#> Jul 1992   4.82115591   29.2127542
#> Aug 1992  -6.88660369  -73.6886751
#> Sep 1992 -10.93359720  -12.9996552
#> Oct 1992   8.41060978  220.0885652
#> Nov 1992  14.54136238  161.3081286
#> Dec 1992   1.37240272  -26.7499583
#> Jan 1993   3.25544925   12.7603168
#> Feb 1993  12.23728019  -37.0892736
#> Mar 1993 -10.53314136  -45.2402281
#> Apr 1993  18.40332640  119.7478248
#> May 1993 -22.31484275   96.1339696
#> Jun 1993  -5.95844719  109.9244552
#> Jul 1993  -2.30472754    5.3534542
#> Aug 1993  17.56315261   -7.1082557
#> Sep 1993   4.75705566  -26.5212622
#> Oct 1993   7.82644366 -108.7472787
#> Nov 1993 -11.46245649   11.2995341
#> Dec 1993   9.50180682  -36.8761608
#> Jan 1994  -3.45227990 -181.3818422
#> Feb 1994 -24.73551488  -72.7994368
#> Mar 1994  13.76264164  -32.3343456
#> Apr 1994  24.77145067   11.8178061
#> May 1994 -13.37234198   52.2721179
#> Jun 1994  20.24155203  -39.4712044
#> Jul 1994   8.59593181  -73.8262361
#> Aug 1994   7.28471344  131.9700103
#> Sep 1994  17.02971138  -54.7717472
#> Oct 1994 -13.49671966  211.6530743
#> Nov 1994   0.85960972   23.8889683
#> Dec 1994   2.04207089  -51.5969374
#> Jan 1995  -9.79050445  115.8752180
#> Feb 1995   5.94650797  -63.4357812
#> Mar 1995  10.39317096   86.2710827
#> Apr 1995 -12.78834279 -168.1573026
#> May 1995  -3.19978820 -163.6792119
#> Jun 1995 -18.72828606  143.0290161
#> Jul 1995   0.91322657  153.9376495
#> Aug 1995   7.12363770   -7.7210862
#> Sep 1995  -1.10119494   72.2232423
#> Oct 1995 -10.51575816   76.2012389
#> Nov 1995  -8.91401342 -110.5216744
#> Dec 1995  -6.19138043 -146.1168585
#> Jan 1996   2.55199046   29.5561895
#> Feb 1996  -4.39315917   65.8526506
#> Mar 1996  14.68149212   64.5142725
#> Apr 1996   8.32506126   62.3653458
#> May 1996  -9.13787407  231.5849301
#> Jun 1996  -2.72595426  -38.1598877
#> Jul 1996   7.51086555    3.1688164
#> Aug 1996  -0.76856255   77.6625400
#> Sep 1996   1.47101528 -198.3047831
#> Oct 1996  -3.21238201   -0.5446710
#> Nov 1996 -17.03140073   70.2411713
#> Dec 1996  -7.61038444  122.3164016
#> Jan 1997  -7.59025393  -14.1397453
#> Feb 1997   2.92608648  -49.5379010
#> Mar 1997   3.32236386   84.0522626
#> Apr 1997  -2.31934768   56.0050141
#> May 1997  -8.15807168  113.7045939
#> Jun 1997   2.16074797   47.3801310
#> Jul 1997 -11.38865785  -15.0440355
#> Aug 1997   1.40105007  213.8995235
#> Sep 1997  -6.62133906   21.7996549
#> Oct 1997  -5.34391766 -189.5969181
#> Nov 1997 -14.12830881  -14.2738393
#> Dec 1997  12.17203500  146.2425734
#> Jan 1998   4.25531888  107.4657122
#> Feb 1998  -1.20928613   65.5056672
#> Mar 1998   6.51430723  -12.3133658
#> Apr 1998   4.90205565  -92.9721805
#> May 1998   7.05192456  151.6556529
#> Jun 1998   8.05650671   -7.6792959
#> Jul 1998  -1.79990583  -36.9791935
#> Aug 1998  10.72861877   60.6232668
#> Sep 1998  -0.82523465  152.6737997
#> Oct 1998  10.72598758   -8.1510851
#> Nov 1998  -0.02079741  142.8358850
#> Dec 1998  -8.59905321  101.3571835
#> Jan 1999   7.48074130  -67.6577827
#> Feb 1999  13.56515448  -23.6565665
#> Mar 1999   4.00001599   50.5599115
#> Apr 1999   6.36256611 -252.7752928
#> May 1999 -18.66035907  -70.0668673
#> Jun 1999  -5.80927235    9.1607577
#> Jul 1999  10.93896134   85.2173585
#> Aug 1999  -5.32691097  112.3909322
#> Sep 1999   6.83984092   82.7869298
#> Oct 1999  -6.84263667 -136.8112810
#> Nov 1999  16.25943271   18.9759677
#> Dec 1999  -6.31734454  196.0755931
#> Jan 2000 -13.53779986  -80.6017715
#> Feb 2000   3.87122228   24.7350268
#> Mar 2000  -9.67667515   50.7828231
#> Apr 2000  -3.58785321  -49.5801751
#> May 2000  -5.86397434  -90.2102368
#> Jun 2000  16.12913478  101.2532538
#> Jul 2000   5.05671343  163.6540567
#> Aug 2000   0.24106090  -82.5520511
#> Sep 2000   2.95359647 -130.1284005
#> Oct 2000  12.75350360  -48.4042607
#> Nov 2000   2.04477180  -19.1561941
#> Dec 2000   1.86747913 -120.3991350
#> Jan 2001   7.09749842 -155.4565833
#> Feb 2001   3.75846729  -14.9289700
#> Mar 2001   4.05413237  -98.8248953
#> Apr 2001   3.30985753 -130.0530931
#> May 2001  15.93760511   76.4001921
#> Jun 2001   5.55075757  -76.4220023
#> Jul 2001 -12.09478318  105.9842219
#> Aug 2001  -2.25557951   57.4400444
#> Sep 2001  -3.42381835  145.4354816
#> Oct 2001 -17.36559496  192.6606505
#> Nov 2001   1.63560903   37.7693133
#> Dec 2001   2.98293513   45.7817317
#> Jan 2002   1.75870029  -31.4478166
#> Feb 2002 -14.05601011  -19.6337224
#> Mar 2002  -2.15255891   53.2207247
#> Apr 2002  -5.02680266 -129.6930815
#> May 2002  16.83881702  -11.6459881
#> Jun 2002 -17.25036625 -179.1435814
#> Jul 2002  -2.52554758  103.7638450
#> Aug 2002 -16.04868007  -75.1737315
#> Sep 2002  11.53376302  144.3320207
#> Oct 2002 -12.66363058   73.4189087
#> Nov 2002  -2.50127880 -105.4557986
#> Dec 2002  -5.38988584    3.0424727
#> Jan 2003  -6.11159385  -62.0591656
#> Feb 2003   8.07837347  -19.3502015
#> Mar 2003  13.57967557   39.3606874
#> Apr 2003  10.00928446   -3.3226158
#> May 2003 -12.58705305  -52.8722573
#> Jun 2003 -13.95429138  307.1007147
#> Jul 2003   3.26775093 -142.4723774
#> Aug 2003  -5.66269784   33.0701264
#> Sep 2003  -0.40093207 -175.9479023
#> Oct 2003  10.07425936 -101.6384201
#> Nov 2003 -13.73520862  103.6862040
#> Dec 2003   3.30858231    3.4712279
#> Jan 2004  11.43623543 -169.3691356
#> Feb 2004 -13.81713624  -37.4082127
#> Mar 2004  -8.44069552 -125.8362453
#> Apr 2004 -18.36583479  -60.8105630
#> May 2004  10.44170831 -136.8707861
#> Jun 2004   2.15210107  -94.2163446
#> Jul 2004   8.00349614   73.9857730
#> Aug 2004   7.14933031    2.2094341
#> Sep 2004  -3.17171209   54.8882732
#> Oct 2004 -22.52045124 -135.0698058
#> Nov 2004  15.05248197  -64.4467633
#> Dec 2004  -4.19147872  104.7844734
#> Jan 2005  -2.08179987  -32.1393364
#> Feb 2005  10.25008622    3.4325883
#> Mar 2005 -21.09352248  106.3948370
#> Apr 2005   5.81155983 -158.7608154
#> May 2005   4.27476044 -158.3462404
#> Jun 2005   3.09712183   11.0531327
#> Jul 2005  -3.48680801   24.6298837
#> Aug 2005   4.72684011    5.7149917
#> Sep 2005   7.02742649 -108.3303495
#> Oct 2005 -15.43797498   21.6301111
#> Nov 2005 -11.39419424  -82.3685525
#> Dec 2005   3.70802634 -150.0780397
#> Jan 2006   2.49291812  122.6567010
#> Feb 2006  12.40415284  -21.5867461
#> Mar 2006  -3.71594045  173.3784762
#> Apr 2006   9.08843016 -143.2535684
#> May 2006   7.72073385  -17.6478952
#> Jun 2006   3.66647135   75.6314868
#> Jul 2006   2.34063054  -98.0007712
#> Aug 2006  -8.24637494  -38.6424379
#> Sep 2006   7.93613760   38.1516254
#> Oct 2006   8.15375576  134.5768539
#> Nov 2006  -6.40412559   98.3748677
#> Dec 2006   1.54668564    8.6637298
#> Jan 2007 -13.19062104   46.9669925
#> Feb 2007  -5.59444239  -89.2900170
#> Mar 2007 -12.78683713    0.2229617
#> Apr 2007  -1.61854635   15.0747679
#> May 2007 -12.12698055  -57.4554552
#> Jun 2007  13.71345622   23.5160435
#> Jul 2007   0.82627704  126.5858408
#> Aug 2007 -12.65853773 -139.0298039
#> Sep 2007  11.74275795   11.5795724
#> Oct 2007   6.95991675   89.2577296
#> Nov 2007   8.00070376    3.8752879
#> Dec 2007 -10.02196891   -3.8372698
#> Jan 2008  -6.77115667   81.1453346
#> Feb 2008   7.20963599    3.3486905
#> Mar 2008   9.60168144   66.3318866
#> Apr 2008  -1.45665667  -22.0091795
#> May 2008  15.05255540   -1.5736545
#> Jun 2008   6.78180406  208.7713501
#> Jul 2008  -2.65552196  149.6586647
#> Aug 2008  -4.58132537  -30.9764025
#> Sep 2008  17.38431533  158.4290933
#> Oct 2008  -0.13105949 -114.7278380
#> Nov 2008  -7.94353354   93.7994138
#> Dec 2008   2.55333393 -117.6473378
#> Jan 2009  -8.78777648  140.2133012
#> Feb 2009  -5.02384014  -17.7175377
#> Mar 2009   9.18235108  -65.0604263
#> Apr 2009  -4.15556110 -122.9364408
#> May 2009  -4.97454127   25.1325895
#> Jun 2009   1.22983974 -156.3215351
#> Jul 2009  -9.94093841   81.3079991
#> Aug 2009   4.85937859  -11.0848678
#> Sep 2009  -5.51041758  -76.1172061
#> Oct 2009   1.66600822   82.9512582
#> Nov 2009   0.95564919 -105.5171964
#> Dec 2009   9.37361737   22.2967963
#> Jan 2010  -7.78592244  244.1195038
#> Feb 2010  -3.46014963  200.0641700
#> Mar 2010   1.45927947  -75.4805825
#> Apr 2010  -6.79378277 -130.3238961
#> May 2010  -6.13698253   24.6026158
#> Jun 2010  22.02935147   54.8893707
#> Jul 2010  -3.28772531   48.6414289
#> Aug 2010  17.51776850  -85.7244463
#> Sep 2010  -9.71651737  -20.8889627
#> Oct 2010 -16.89546894   -2.1828351
#> Nov 2010 -14.81386178   37.1728862
#> Dec 2010  24.26894783 -230.0614875
#> Jan 2011  -2.28460020  -68.6391490
#> Feb 2011  11.13283521   82.3954712
#> Mar 2011  -4.37063852   67.1560953
#> Apr 2011  12.64884610  -13.6811085
#> May 2011  14.36597731  -32.4855419
#> Jun 2011  -9.18664322  116.5364420
#> Jul 2011 -19.14602512  -90.3741440
#> Aug 2011   1.14969531   98.5612115
#> Sep 2011  15.10348121  -12.6323527
#> Oct 2011  -0.09052613 -144.3967173
#> Nov 2011  -1.35856744    3.2154077
#> Dec 2011  -3.36250668 -211.7716447
#> Jan 2012   3.57796527  166.5816928
#> Feb 2012   3.40498580   57.5671055
#> Mar 2012  30.97230704   -0.2235961
#> Apr 2012   6.77352741  -72.2014992
#> May 2012   3.85994964   44.8672361
#> Jun 2012  -8.38842632    5.1736240
#> Jul 2012  -5.30027590 -125.3649454
#> Aug 2012 -22.93182084   55.8374959
#> Sep 2012  18.55263642   29.8702402
#> Oct 2012  -6.24714964  -41.5528828
#> Nov 2012   1.75872582 -162.5374469
#> Dec 2012 -13.65788853   29.9647475
#> Jan 2013 -13.45518474  -14.6007948
#> Feb 2013   9.47265271   33.0470436
#> Mar 2013  -1.15739671  -34.4358479
#> Apr 2013  -4.29404494   89.1762344
#> May 2013   0.79631976  -13.6294335
#> Jun 2013  10.06463994   28.1191185
#> Jul 2013  17.54518747  -11.0812460
#> Aug 2013  -1.19361311   -3.6247432
#> Sep 2013  16.43998625    2.4977216
#> Oct 2013 -13.10997765   13.1053564
#> Nov 2013   3.62701424  259.5523922
#> Dec 2013  13.69665491   41.9283992
#> Jan 2014  -3.48512400  -28.1698421
#> Feb 2014  -5.45470652   54.4863114
#> Mar 2014  -0.20519678   86.8230689
#> Apr 2014  -3.38849889  -20.6675268
#> May 2014   1.85592100    7.4947320
#> Jun 2014   3.45862657 -225.0901330
#> Jul 2014  -7.93332374  150.0961160
#> Aug 2014  -7.48631612  -48.2768943
#> Sep 2014 -18.46666644   60.5210639
#> Oct 2014 -15.56824415   80.0685483
#> Nov 2014   0.02354941   31.8362191
#> Dec 2014  14.70206284   10.3472107
#> Jan 2015  -7.77484495  -65.7636600
#> Feb 2015   7.04613433    3.3874701
#> Mar 2015   6.96606398  -64.9756559
#> Apr 2015   9.31972929   91.1239650
#> May 2015  26.16178688   -4.7272921
#> Jun 2015 -14.77492275 -117.8515013
#> Jul 2015   0.74677229  224.2660967
#> Aug 2015  -6.40770175  146.7485607
#> Sep 2015   1.14925465   55.2061735
#> Oct 2015   7.47021300   35.9622515
#> Nov 2015 -11.18479256  183.7521669
#> Dec 2015   3.41004819   88.3273304
#> Jan 2016   1.25031186 -171.3201462
#> Feb 2016   3.11369721   29.3728600
#> Mar 2016   4.26443804  133.0807535
#> Apr 2016 -12.69579318    0.1025496
#> May 2016   8.69811052   94.4012124
#> Jun 2016  -0.23574630  -16.3110155
#> Jul 2016   6.46068260    2.0936334
#> Aug 2016  12.32672999   55.1059233
#> Sep 2016 -12.80410543   75.4814726
#> Oct 2016  -1.76086170 -101.7904646
#> Nov 2016  11.48912321  -32.0525550
#> Dec 2016   1.06409425  -45.7239570
#> Jan 2017  10.52599358 -216.8675688
#> Feb 2017   0.63508668    6.6409650
#> Mar 2017   7.60288668  151.3114967
#> Apr 2017  -5.83391234   24.2563643
#> May 2017  -8.33176971   19.6097401
#> Jun 2017  -9.55279258  -65.6023781
#> Jul 2017   4.68197456  -47.8355050
#> Aug 2017  14.89361598   56.8878744
#> Sep 2017   8.45637803 -210.2626810
#> Oct 2017  -0.90330359   -0.3301496
#> Nov 2017  -7.60974092   41.6918015
#> Dec 2017  -4.22703485   47.3867599
#> Jan 2018 -11.16990551   -6.2925119
#> Feb 2018 -13.73205437 -163.6219349
#> Mar 2018  -4.64627202  116.8229652
#> Apr 2018  -7.51982369  -40.8686195
#> May 2018  -4.07974543   -9.4117059
#> Jun 2018   1.77182569   46.7589309
#> Jul 2018   9.99326344  -82.8195377
#> Aug 2018   1.16269256   39.6813229
#> Sep 2018 -18.35065315   24.4475509
#> Oct 2018  -2.98695682   62.7437586
#> Nov 2018   3.14652720  -80.5206472
#> Dec 2018 -10.82466160 -193.2053070
#> Jan 2019   2.21793014   71.7082011
#> Feb 2019  -0.49851303  179.2486559
#> Mar 2019  11.07534869  112.3194865
#> Apr 2019  -9.93360619  160.9947684
#> May 2019  19.24843474  173.2799496
#> Jun 2019  -5.56762404  127.5703717
#> Jul 2019   3.29474662   42.4007988
#> Aug 2019  17.58286351  -13.7778131
#> Sep 2019  12.73909790  -59.7937324
#> Oct 2019  -5.83885195  -72.2073272
#> Nov 2019  -3.27692378    7.2762483
#> Dec 2019   1.45241262  147.3362511
#> Jan 2020   2.14101138  -31.6206077
#> Feb 2020   0.11554669 -141.4721408
#> Mar 2020  12.19699084 -198.6179040
#> Apr 2020 -10.30643641   70.5734028
#> May 2020   7.64734975  -79.0427631
#> Jun 2020  -5.14982604   75.8480158
#> Jul 2020  17.40986972  207.8375621
#> Aug 2020  10.05845241  191.9134900
#> Sep 2020 -11.07352945    7.5529822
#> Oct 2020  -6.27882790   26.7751527
#> Nov 2020 -14.79501948   43.6290736
#> Dec 2020   1.54343173  120.9748781
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 
s_preVar(myreg3)
#> $series
#>                  var1         var2
#> Jan 1990 -10.67709165 -229.9484409
#> Feb 1990 -11.80488666  -43.8727684
#> Mar 1990  21.59175760  -64.9680607
#> Apr 1990  -4.17812648   42.4040947
#> May 1990   0.87654958    4.7049728
#> Jun 1990  -5.43709068 -128.5044066
#> Jul 1990  -1.19867022    4.5532358
#> Aug 1990  18.00624858 -144.9429700
#> Sep 1990  -7.94247350 -163.3704403
#> Oct 1990  -3.60360559  -51.7284962
#> Nov 1990   6.31460646  -99.1151430
#> Dec 1990 -16.63279352  -72.8485045
#> Jan 1991  -4.88005245   45.1951634
#> Feb 1991  -0.20752219   88.7181335
#> Mar 1991   9.33871038  126.4573401
#> Apr 1991  -2.32670878 -129.1419391
#> May 1991 -10.19630064   18.2510306
#> Jun 1991 -16.80328321   -8.9809675
#> Jul 1991  22.81394917  -11.2629276
#> Aug 1991 -10.42285303   16.2801338
#> Sep 1991   7.12307192   55.5861006
#> Oct 1991  12.20747981   85.3394680
#> Nov 1991   3.36018683  -19.4899361
#> Dec 1991  -2.79147937   -4.3411909
#> Jan 1992   0.76422131  -52.9438316
#> Feb 1992  -3.54509030  -59.8510599
#> Mar 1992  14.32658547   14.9297682
#> Apr 1992   0.02671474  125.5820916
#> May 1992   3.44012511  -23.4369528
#> Jun 1992  -8.30323380  142.2340276
#> Jul 1992   4.82115591   29.2127542
#> Aug 1992  -6.88660369  -73.6886751
#> Sep 1992 -10.93359720  -12.9996552
#> Oct 1992   8.41060978  220.0885652
#> Nov 1992  14.54136238  161.3081286
#> Dec 1992   1.37240272  -26.7499583
#> Jan 1993   3.25544925   12.7603168
#> Feb 1993  12.23728019  -37.0892736
#> Mar 1993 -10.53314136  -45.2402281
#> Apr 1993  18.40332640  119.7478248
#> May 1993 -22.31484275   96.1339696
#> Jun 1993  -5.95844719  109.9244552
#> Jul 1993  -2.30472754    5.3534542
#> Aug 1993  17.56315261   -7.1082557
#> Sep 1993   4.75705566  -26.5212622
#> Oct 1993   7.82644366 -108.7472787
#> Nov 1993 -11.46245649   11.2995341
#> Dec 1993   9.50180682  -36.8761608
#> Jan 1994  -3.45227990 -181.3818422
#> Feb 1994 -24.73551488  -72.7994368
#> Mar 1994  13.76264164  -32.3343456
#> Apr 1994  24.77145067   11.8178061
#> May 1994 -13.37234198   52.2721179
#> Jun 1994  20.24155203  -39.4712044
#> Jul 1994   8.59593181  -73.8262361
#> Aug 1994   7.28471344  131.9700103
#> Sep 1994  17.02971138  -54.7717472
#> Oct 1994 -13.49671966  211.6530743
#> Nov 1994   0.85960972   23.8889683
#> Dec 1994   2.04207089  -51.5969374
#> Jan 1995  -9.79050445  115.8752180
#> Feb 1995   5.94650797  -63.4357812
#> Mar 1995  10.39317096   86.2710827
#> Apr 1995 -12.78834279 -168.1573026
#> May 1995  -3.19978820 -163.6792119
#> Jun 1995 -18.72828606  143.0290161
#> Jul 1995   0.91322657  153.9376495
#> Aug 1995   7.12363770   -7.7210862
#> Sep 1995  -1.10119494   72.2232423
#> Oct 1995 -10.51575816   76.2012389
#> Nov 1995  -8.91401342 -110.5216744
#> Dec 1995  -6.19138043 -146.1168585
#> Jan 1996   2.55199046   29.5561895
#> Feb 1996  -4.39315917   65.8526506
#> Mar 1996  14.68149212   64.5142725
#> Apr 1996   8.32506126   62.3653458
#> May 1996  -9.13787407  231.5849301
#> Jun 1996  -2.72595426  -38.1598877
#> Jul 1996   7.51086555    3.1688164
#> Aug 1996  -0.76856255   77.6625400
#> Sep 1996   1.47101528 -198.3047831
#> Oct 1996  -3.21238201   -0.5446710
#> Nov 1996 -17.03140073   70.2411713
#> Dec 1996  -7.61038444  122.3164016
#> Jan 1997  -7.59025393  -14.1397453
#> Feb 1997   2.92608648  -49.5379010
#> Mar 1997   3.32236386   84.0522626
#> Apr 1997  -2.31934768   56.0050141
#> May 1997  -8.15807168  113.7045939
#> Jun 1997   2.16074797   47.3801310
#> Jul 1997 -11.38865785  -15.0440355
#> Aug 1997   1.40105007  213.8995235
#> Sep 1997  -6.62133906   21.7996549
#> Oct 1997  -5.34391766 -189.5969181
#> Nov 1997 -14.12830881  -14.2738393
#> Dec 1997  12.17203500  146.2425734
#> Jan 1998   4.25531888  107.4657122
#> Feb 1998  -1.20928613   65.5056672
#> Mar 1998   6.51430723  -12.3133658
#> Apr 1998   4.90205565  -92.9721805
#> May 1998   7.05192456  151.6556529
#> Jun 1998   8.05650671   -7.6792959
#> Jul 1998  -1.79990583  -36.9791935
#> Aug 1998  10.72861877   60.6232668
#> Sep 1998  -0.82523465  152.6737997
#> Oct 1998  10.72598758   -8.1510851
#> Nov 1998  -0.02079741  142.8358850
#> Dec 1998  -8.59905321  101.3571835
#> Jan 1999   7.48074130  -67.6577827
#> Feb 1999  13.56515448  -23.6565665
#> Mar 1999   4.00001599   50.5599115
#> Apr 1999   6.36256611 -252.7752928
#> May 1999 -18.66035907  -70.0668673
#> Jun 1999  -5.80927235    9.1607577
#> Jul 1999  10.93896134   85.2173585
#> Aug 1999  -5.32691097  112.3909322
#> Sep 1999   6.83984092   82.7869298
#> Oct 1999  -6.84263667 -136.8112810
#> Nov 1999  16.25943271   18.9759677
#> Dec 1999  -6.31734454  196.0755931
#> Jan 2000 -13.53779986  -80.6017715
#> Feb 2000   3.87122228   24.7350268
#> Mar 2000  -9.67667515   50.7828231
#> Apr 2000  -3.58785321  -49.5801751
#> May 2000  -5.86397434  -90.2102368
#> Jun 2000  16.12913478  101.2532538
#> Jul 2000   5.05671343  163.6540567
#> Aug 2000   0.24106090  -82.5520511
#> Sep 2000   2.95359647 -130.1284005
#> Oct 2000  12.75350360  -48.4042607
#> Nov 2000   2.04477180  -19.1561941
#> Dec 2000   1.86747913 -120.3991350
#> Jan 2001   7.09749842 -155.4565833
#> Feb 2001   3.75846729  -14.9289700
#> Mar 2001   4.05413237  -98.8248953
#> Apr 2001   3.30985753 -130.0530931
#> May 2001  15.93760511   76.4001921
#> Jun 2001   5.55075757  -76.4220023
#> Jul 2001 -12.09478318  105.9842219
#> Aug 2001  -2.25557951   57.4400444
#> Sep 2001  -3.42381835  145.4354816
#> Oct 2001 -17.36559496  192.6606505
#> Nov 2001   1.63560903   37.7693133
#> Dec 2001   2.98293513   45.7817317
#> Jan 2002   1.75870029  -31.4478166
#> Feb 2002 -14.05601011  -19.6337224
#> Mar 2002  -2.15255891   53.2207247
#> Apr 2002  -5.02680266 -129.6930815
#> May 2002  16.83881702  -11.6459881
#> Jun 2002 -17.25036625 -179.1435814
#> Jul 2002  -2.52554758  103.7638450
#> Aug 2002 -16.04868007  -75.1737315
#> Sep 2002  11.53376302  144.3320207
#> Oct 2002 -12.66363058   73.4189087
#> Nov 2002  -2.50127880 -105.4557986
#> Dec 2002  -5.38988584    3.0424727
#> Jan 2003  -6.11159385  -62.0591656
#> Feb 2003   8.07837347  -19.3502015
#> Mar 2003  13.57967557   39.3606874
#> Apr 2003  10.00928446   -3.3226158
#> May 2003 -12.58705305  -52.8722573
#> Jun 2003 -13.95429138  307.1007147
#> Jul 2003   3.26775093 -142.4723774
#> Aug 2003  -5.66269784   33.0701264
#> Sep 2003  -0.40093207 -175.9479023
#> Oct 2003  10.07425936 -101.6384201
#> Nov 2003 -13.73520862  103.6862040
#> Dec 2003   3.30858231    3.4712279
#> Jan 2004  11.43623543 -169.3691356
#> Feb 2004 -13.81713624  -37.4082127
#> Mar 2004  -8.44069552 -125.8362453
#> Apr 2004 -18.36583479  -60.8105630
#> May 2004  10.44170831 -136.8707861
#> Jun 2004   2.15210107  -94.2163446
#> Jul 2004   8.00349614   73.9857730
#> Aug 2004   7.14933031    2.2094341
#> Sep 2004  -3.17171209   54.8882732
#> Oct 2004 -22.52045124 -135.0698058
#> Nov 2004  15.05248197  -64.4467633
#> Dec 2004  -4.19147872  104.7844734
#> Jan 2005  -2.08179987  -32.1393364
#> Feb 2005  10.25008622    3.4325883
#> Mar 2005 -21.09352248  106.3948370
#> Apr 2005   5.81155983 -158.7608154
#> May 2005   4.27476044 -158.3462404
#> Jun 2005   3.09712183   11.0531327
#> Jul 2005  -3.48680801   24.6298837
#> Aug 2005   4.72684011    5.7149917
#> Sep 2005   7.02742649 -108.3303495
#> Oct 2005 -15.43797498   21.6301111
#> Nov 2005 -11.39419424  -82.3685525
#> Dec 2005   3.70802634 -150.0780397
#> Jan 2006   2.49291812  122.6567010
#> Feb 2006  12.40415284  -21.5867461
#> Mar 2006  -3.71594045  173.3784762
#> Apr 2006   9.08843016 -143.2535684
#> May 2006   7.72073385  -17.6478952
#> Jun 2006   3.66647135   75.6314868
#> Jul 2006   2.34063054  -98.0007712
#> Aug 2006  -8.24637494  -38.6424379
#> Sep 2006   7.93613760   38.1516254
#> Oct 2006   8.15375576  134.5768539
#> Nov 2006  -6.40412559   98.3748677
#> Dec 2006   1.54668564    8.6637298
#> Jan 2007 -13.19062104   46.9669925
#> Feb 2007  -5.59444239  -89.2900170
#> Mar 2007 -12.78683713    0.2229617
#> Apr 2007  -1.61854635   15.0747679
#> May 2007 -12.12698055  -57.4554552
#> Jun 2007  13.71345622   23.5160435
#> Jul 2007   0.82627704  126.5858408
#> Aug 2007 -12.65853773 -139.0298039
#> Sep 2007  11.74275795   11.5795724
#> Oct 2007   6.95991675   89.2577296
#> Nov 2007   8.00070376    3.8752879
#> Dec 2007 -10.02196891   -3.8372698
#> Jan 2008  -6.77115667   81.1453346
#> Feb 2008   7.20963599    3.3486905
#> Mar 2008   9.60168144   66.3318866
#> Apr 2008  -1.45665667  -22.0091795
#> May 2008  15.05255540   -1.5736545
#> Jun 2008   6.78180406  208.7713501
#> Jul 2008  -2.65552196  149.6586647
#> Aug 2008  -4.58132537  -30.9764025
#> Sep 2008  17.38431533  158.4290933
#> Oct 2008  -0.13105949 -114.7278380
#> Nov 2008  -7.94353354   93.7994138
#> Dec 2008   2.55333393 -117.6473378
#> Jan 2009  -8.78777648  140.2133012
#> Feb 2009  -5.02384014  -17.7175377
#> Mar 2009   9.18235108  -65.0604263
#> Apr 2009  -4.15556110 -122.9364408
#> May 2009  -4.97454127   25.1325895
#> Jun 2009   1.22983974 -156.3215351
#> Jul 2009  -9.94093841   81.3079991
#> Aug 2009   4.85937859  -11.0848678
#> Sep 2009  -5.51041758  -76.1172061
#> Oct 2009   1.66600822   82.9512582
#> Nov 2009   0.95564919 -105.5171964
#> Dec 2009   9.37361737   22.2967963
#> Jan 2010  -7.78592244  244.1195038
#> Feb 2010  -3.46014963  200.0641700
#> Mar 2010   1.45927947  -75.4805825
#> Apr 2010  -6.79378277 -130.3238961
#> May 2010  -6.13698253   24.6026158
#> Jun 2010  22.02935147   54.8893707
#> Jul 2010  -3.28772531   48.6414289
#> Aug 2010  17.51776850  -85.7244463
#> Sep 2010  -9.71651737  -20.8889627
#> Oct 2010 -16.89546894   -2.1828351
#> Nov 2010 -14.81386178   37.1728862
#> Dec 2010  24.26894783 -230.0614875
#> Jan 2011  -2.28460020  -68.6391490
#> Feb 2011  11.13283521   82.3954712
#> Mar 2011  -4.37063852   67.1560953
#> Apr 2011  12.64884610  -13.6811085
#> May 2011  14.36597731  -32.4855419
#> Jun 2011  -9.18664322  116.5364420
#> Jul 2011 -19.14602512  -90.3741440
#> Aug 2011   1.14969531   98.5612115
#> Sep 2011  15.10348121  -12.6323527
#> Oct 2011  -0.09052613 -144.3967173
#> Nov 2011  -1.35856744    3.2154077
#> Dec 2011  -3.36250668 -211.7716447
#> Jan 2012   3.57796527  166.5816928
#> Feb 2012   3.40498580   57.5671055
#> Mar 2012  30.97230704   -0.2235961
#> Apr 2012   6.77352741  -72.2014992
#> May 2012   3.85994964   44.8672361
#> Jun 2012  -8.38842632    5.1736240
#> Jul 2012  -5.30027590 -125.3649454
#> Aug 2012 -22.93182084   55.8374959
#> Sep 2012  18.55263642   29.8702402
#> Oct 2012  -6.24714964  -41.5528828
#> Nov 2012   1.75872582 -162.5374469
#> Dec 2012 -13.65788853   29.9647475
#> Jan 2013 -13.45518474  -14.6007948
#> Feb 2013   9.47265271   33.0470436
#> Mar 2013  -1.15739671  -34.4358479
#> Apr 2013  -4.29404494   89.1762344
#> May 2013   0.79631976  -13.6294335
#> Jun 2013  10.06463994   28.1191185
#> Jul 2013  17.54518747  -11.0812460
#> Aug 2013  -1.19361311   -3.6247432
#> Sep 2013  16.43998625    2.4977216
#> Oct 2013 -13.10997765   13.1053564
#> Nov 2013   3.62701424  259.5523922
#> Dec 2013  13.69665491   41.9283992
#> Jan 2014  -3.48512400  -28.1698421
#> Feb 2014  -5.45470652   54.4863114
#> Mar 2014  -0.20519678   86.8230689
#> Apr 2014  -3.38849889  -20.6675268
#> May 2014   1.85592100    7.4947320
#> Jun 2014   3.45862657 -225.0901330
#> Jul 2014  -7.93332374  150.0961160
#> Aug 2014  -7.48631612  -48.2768943
#> Sep 2014 -18.46666644   60.5210639
#> Oct 2014 -15.56824415   80.0685483
#> Nov 2014   0.02354941   31.8362191
#> Dec 2014  14.70206284   10.3472107
#> Jan 2015  -7.77484495  -65.7636600
#> Feb 2015   7.04613433    3.3874701
#> Mar 2015   6.96606398  -64.9756559
#> Apr 2015   9.31972929   91.1239650
#> May 2015  26.16178688   -4.7272921
#> Jun 2015 -14.77492275 -117.8515013
#> Jul 2015   0.74677229  224.2660967
#> Aug 2015  -6.40770175  146.7485607
#> Sep 2015   1.14925465   55.2061735
#> Oct 2015   7.47021300   35.9622515
#> Nov 2015 -11.18479256  183.7521669
#> Dec 2015   3.41004819   88.3273304
#> Jan 2016   1.25031186 -171.3201462
#> Feb 2016   3.11369721   29.3728600
#> Mar 2016   4.26443804  133.0807535
#> Apr 2016 -12.69579318    0.1025496
#> May 2016   8.69811052   94.4012124
#> Jun 2016  -0.23574630  -16.3110155
#> Jul 2016   6.46068260    2.0936334
#> Aug 2016  12.32672999   55.1059233
#> Sep 2016 -12.80410543   75.4814726
#> Oct 2016  -1.76086170 -101.7904646
#> Nov 2016  11.48912321  -32.0525550
#> Dec 2016   1.06409425  -45.7239570
#> Jan 2017  10.52599358 -216.8675688
#> Feb 2017   0.63508668    6.6409650
#> Mar 2017   7.60288668  151.3114967
#> Apr 2017  -5.83391234   24.2563643
#> May 2017  -8.33176971   19.6097401
#> Jun 2017  -9.55279258  -65.6023781
#> Jul 2017   4.68197456  -47.8355050
#> Aug 2017  14.89361598   56.8878744
#> Sep 2017   8.45637803 -210.2626810
#> Oct 2017  -0.90330359   -0.3301496
#> Nov 2017  -7.60974092   41.6918015
#> Dec 2017  -4.22703485   47.3867599
#> Jan 2018 -11.16990551   -6.2925119
#> Feb 2018 -13.73205437 -163.6219349
#> Mar 2018  -4.64627202  116.8229652
#> Apr 2018  -7.51982369  -40.8686195
#> May 2018  -4.07974543   -9.4117059
#> Jun 2018   1.77182569   46.7589309
#> Jul 2018   9.99326344  -82.8195377
#> Aug 2018   1.16269256   39.6813229
#> Sep 2018 -18.35065315   24.4475509
#> Oct 2018  -2.98695682   62.7437586
#> Nov 2018   3.14652720  -80.5206472
#> Dec 2018 -10.82466160 -193.2053070
#> Jan 2019   2.21793014   71.7082011
#> Feb 2019  -0.49851303  179.2486559
#> Mar 2019  11.07534869  112.3194865
#> Apr 2019  -9.93360619  160.9947684
#> May 2019  19.24843474  173.2799496
#> Jun 2019  -5.56762404  127.5703717
#> Jul 2019   3.29474662   42.4007988
#> Aug 2019  17.58286351  -13.7778131
#> Sep 2019  12.73909790  -59.7937324
#> Oct 2019  -5.83885195  -72.2073272
#> Nov 2019  -3.27692378    7.2762483
#> Dec 2019   1.45241262  147.3362511
#> Jan 2020   2.14101138  -31.6206077
#> Feb 2020   0.11554669 -141.4721408
#> Mar 2020  12.19699084 -198.6179040
#> Apr 2020 -10.30643641   70.5734028
#> May 2020   7.64734975  -79.0427631
#> Jun 2020  -5.14982604   75.8480158
#> Jul 2020  17.40986972  207.8375621
#> Aug 2020  10.05845241  191.9134900
#> Sep 2020 -11.07352945    7.5529822
#> Oct 2020  -6.27882790   26.7751527
#> Nov 2020 -14.79501948   43.6290736
#> Dec 2020   1.54343173  120.9748781
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 

s_td(myreg1)
#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove
s_td(myspec1)
#>       option autoadjust leapyear stocktd   test
#>  TradingDays       TRUE LeapYear       0 Remove

s_easter(myreg1)
#>  enabled julian duration test
#>     TRUE  FALSE        8  Add
s_easter(myspec1)
#>  enabled julian duration test
#>     TRUE  FALSE        8  Add

s_out(myreg1)
#>  enabled span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE  All TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7
s_out(myspec1)
#>  enabled            span   ao   tc   ls    so usedefcv cv method tcrate
#>     TRUE From 2010-03-01 TRUE TRUE TRUE FALSE     TRUE  4 AddOne    0.7

s_arima(myreg1)
#>  enabled automdl.acceptdefault automdl.cancel automdl.ub1 automdl.ub2
#>     TRUE                 FALSE            0.1    1.041667        0.88
#>  automdl.mixed automdl.balanced automdl.armalimit automdl.reducecv
#>           TRUE            FALSE                 1          0.14286
#>  automdl.ljungboxlimit automdl.ubfinal arima.mu arima.p arima.d arima.q
#>                   0.95            1.05    FALSE       0       1       1
#>  arima.bp arima.bd arima.bq arima.coef
#>         0        1        1      FALSE
s_arima(myspec1)
#>  enabled automdl.acceptdefault automdl.cancel automdl.ub1 automdl.ub2
#>     TRUE                 FALSE            0.1    1.041667        0.88
#>  automdl.mixed automdl.balanced automdl.armalimit automdl.reducecv
#>           TRUE            FALSE                 1          0.14286
#>  automdl.ljungboxlimit automdl.ubfinal arima.mu arima.p arima.d arima.q
#>                   0.95            1.05    FALSE       0       1       1
#>  arima.bp arima.bd arima.bq arima.coef
#>         0        1        1      FALSE

myspec4 <- regarima_spec_x13(myreg1, automdl.enabled = FALSE,
             arima.coefEnabled = TRUE,
             arima.p = 1,arima.q = 1, arima.bp = 1, arima.bq = 1,
             arima.coef = rep(0.2, 4),
             arima.coefType = rep("Initial", 4))
myreg4 <- regarima(myseries, myspec4)

s_arimaCoef(myreg4)
#>              Type Value
#> Phi(1)    Initial   0.2
#> Theta(1)  Initial   0.2
#> BPhi(1)   Initial   0.2
#> BTheta(1) Initial   0.2
s_arimaCoef(myspec4)
#>              Type Value
#> Phi(1)    Initial   0.2
#> Theta(1)  Initial   0.2
#> BPhi(1)   Initial   0.2
#> BTheta(1) Initial   0.2

s_fcst(myreg1)
#>  horizon
#>       -2
s_fcst(myspec1)
#>  horizon
#>       -2

s_span(myreg1)
#>          type d0 d1 n0 n1
#> estimate  All NA NA  0  0
#> outlier   All NA NA  0  0
s_span(myspec1)
#>          type         d0 d1 n0 n1
#> estimate From 2005-10-01 NA  0  0
#> outlier  From 2010-03-01 NA  0  0

myspec5 <- x13_spec(spec = "RSA5c", x11.seasonalComp = FALSE)
mysa5 <- x13(myseries, myspec5)

s_x11(mysa5)
#>   x11.mode x11.seasonalComp x11.lsigma x11.usigma x11.trendAuto x11.trendma
#>  Undefined            FALSE        1.5        2.5          TRUE          13
#>  x11.seasonalma x11.fcasts x11.bcasts x11.calendarSigma x11.sigmaVector
#>             Msr         -1          0              None              NA
#>  x11.excludeFcasts
#>              FALSE
s_x11(myspec5)
#>   x11.mode x11.seasonalComp x11.lsigma x11.usigma x11.trendAuto x11.trendma
#>  Undefined            FALSE        1.5        2.5          TRUE          13
#>  x11.seasonalma x11.fcasts x11.bcasts x11.calendarSigma x11.sigmaVector
#>             Msr         -1          0              None              NA
#>  x11.excludeFcasts
#>              FALSE

myspec6 <- tramoseats_spec(spec = "RSAfull", seats.approx = "Noisy")
mysa6 <- tramoseats(myseries, myspec6)

s_seats(mysa6)
#>  seats.predictionLength seats.approx seats.maBoundary seats.trendBoundary
#>                      -1        Noisy             0.95                 0.5
#>  seats.seasdBoundary seats.seasdBoundary1 seats.seasTol seats.method
#>                  0.8                  0.8             2       Burman
s_seats(mysa6)
#>  seats.predictionLength seats.approx seats.maBoundary seats.trendBoundary
#>                      -1        Noisy             0.95                 0.5
#>  seats.seasdBoundary seats.seasdBoundary1 seats.seasTol seats.method
#>                  0.8                  0.8             2       Burman
# }