TRAMO-SEATS model specification, SA/TRAMO-SEATS

tramoseats_spec_def creates (and modifies), from a predefined 'JDemetra+' model specification, a c("SA_spec", "TRAMO_SEATS") class object with the SA model specification for the TRAMO-SEATS method.

tramoseats_spec creates (and/or modifies) a c("SA_spec", "TRAMO_SEATS") class object with the SA model specification for the TRAMO-SEATS method. The object is created from a c("SA", "TRAMO_SEATS") or c("SA_spec", "TRAMO_SEATS") class object.

tramoseats_spec_def(spec = c("RSAfull", "RSA0", "RSA1", "RSA2", "RSA3",
  "RSA4", "RSA5"), estimate.from = NA_character_,
  estimate.to = NA_character_, estimate.first = NA_integer_,
  estimate.last = NA_integer_, estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_, estimate.tol = NA_integer_,
  estimate.eml = NA, estimate.urfinal = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.fct = NA_integer_, usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA, usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA, usrdef.varEnabled = NA, usrdef.var = NA,
  usrdef.varType = NA, usrdef.varCoef = NA, tradingdays.mauto = c(NA,
  "Unused", "FTest", "WaldTest"), tradingdays.pftd = NA_integer_,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined",
  "None"), tradingdays.leapyear = NA,
  tradingdays.stocktd = NA_integer_, tradingdays.test = c(NA,
  "Separate_T", "Joint_F", "None"), easter.type = c(NA, "Unused",
  "Standard", "IncludeEaster", "IncludeEasterMonday"),
  easter.julian = NA, easter.duration = NA_integer_,
  easter.test = NA, outlier.enabled = NA,
  outlier.from = NA_character_, outlier.to = NA_character_,
  outlier.first = NA_integer_, outlier.last = NA_integer_,
  outlier.exclFirst = NA_integer_, outlier.exclLast = NA_integer_,
  outlier.ao = NA, outlier.tc = NA, outlier.ls = NA,
  outlier.so = NA, outlier.usedefcv = NA, outlier.cv = NA_integer_,
  outlier.eml = NA, outlier.tcrate = NA_integer_,
  automdl.enabled = NA, automdl.acceptdefault = NA,
  automdl.cancel = NA_integer_, automdl.ub1 = NA_integer_,
  automdl.ub2 = NA_integer_, automdl.armalimit = NA_integer_,
  automdl.reducecv = NA_integer_, automdl.ljungboxlimit = NA_integer_,
  automdl.compare = NA, arima.mu = NA, arima.p = NA_integer_,
  arima.d = NA_integer_, arima.q = NA_integer_,
  arima.bp = NA_integer_, arima.bd = NA_integer_,
  arima.bq = NA_integer_, arima.coefEnabled = NA, arima.coef = NA,
  arima.coefType = NA, fcst.horizon = NA_integer_,
  seats.approx = c(NA, "None", "Legacy", "Noisy"),
  seats.trendBoundary = NA_integer_, seats.seasdBoundary = NA_integer_,
  seats.seasdBoundary1 = NA_integer_, seats.seasTol = NA_integer_,
  seats.maBoundary = NA_integer_, seats.method = c(NA, "Burman",
  "KalmanSmoother", "McElroyMatrix"))

tramoseats_spec(object, estimate.from = NA_character_,
  estimate.to = NA_character_, estimate.first = NA_integer_,
  estimate.last = NA_integer_, estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_, estimate.tol = NA_integer_,
  estimate.eml = NA, estimate.urfinal = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.fct = NA_integer_, usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA, usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA, usrdef.varEnabled = NA, usrdef.var = NA,
  usrdef.varType = NA, usrdef.varCoef = NA, tradingdays.mauto = c(NA,
  "Unused", "FTest", "WaldTest"), tradingdays.pftd = NA_integer_,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined",
  "None"), tradingdays.leapyear = NA,
  tradingdays.stocktd = NA_integer_, tradingdays.test = c(NA,
  "Separate_T", "Joint_F", "None"), easter.type = c(NA, "Unused",
  "Standard", "IncludeEaster", "IncludeEasterMonday"),
  easter.julian = NA, easter.duration = NA_integer_,
  easter.test = NA, outlier.enabled = NA,
  outlier.from = NA_character_, outlier.to = NA_character_,
  outlier.first = NA_integer_, outlier.last = NA_integer_,
  outlier.exclFirst = NA_integer_, outlier.exclLast = NA_integer_,
  outlier.ao = NA, outlier.tc = NA, outlier.ls = NA,
  outlier.so = NA, outlier.usedefcv = NA, outlier.cv = NA_integer_,
  outlier.eml = NA, outlier.tcrate = NA_integer_,
  automdl.enabled = NA, automdl.acceptdefault = NA,
  automdl.cancel = NA_integer_, automdl.ub1 = NA_integer_,
  automdl.ub2 = NA_integer_, automdl.armalimit = NA_integer_,
  automdl.reducecv = NA_integer_, automdl.ljungboxlimit = NA_integer_,
  automdl.compare = NA, arima.mu = NA, arima.p = NA_integer_,
  arima.d = NA_integer_, arima.q = NA_integer_,
  arima.bp = NA_integer_, arima.bd = NA_integer_,
  arima.bq = NA_integer_, arima.coefEnabled = NA, arima.coef = NA,
  arima.coefType = NA, fcst.horizon = NA_integer_,
  seats.approx = c(NA, "None", "Legacy", "Noisy"),
  seats.trendBoundary = NA_integer_, seats.seasdBoundary = NA_integer_,
  seats.seasdBoundary1 = NA_integer_, seats.seasTol = NA_integer_,
  seats.maBoundary = NA_integer_, seats.method = c(NA, "Burman",
  "KalmanSmoother", "McElroyMatrix"))

Arguments

spec
predefined 'JDemetra+' model specification (see Details). The default is "RSAfull".

The time span of the series to be used for the estimation of the RegArima model coefficients (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: estimate.from, estimate.to, estimate.first, estimate.last, estimate.exclFirst and estimate.exclLast; where estimate.from and estimate.to have priority over remaining span control variables, estimate.last and estimate.first have priority over estimate.exclFirst and estimate.exclLast, and estimate.last has priority over estimate.first.

estimate.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with estimate.to.

estimate.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with estimate.from.

estimate.first
numeric specifying the number of periods considered at the beginning of the series.

estimate.last
numeric specifying the number of periods considered at the end of the series.

estimate.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with estimate.exclLast.

estimate.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with estimate.exclFirst.

estimate.tol
numeric, convergence tolerance. The absolute changes in the log-likelihood function are compared to this value to check for the convergence of the estimation iterations.

estimate.eml
logicals, exact maximum likelihood estimation. If TRUE the program performs an exact maximum likelihood estimation. If FASLE the Unconditional Least Squares method is used.

estimate.urfinal
numeric, final unit root limit. The threshold value for the final unit root test for identification of differencing orders. If the magnitude of an AR root for the final model is less than this number, a unit root is assumed, the order of the AR polynomial is reduced by one, and the appropriate order of the differencing (non-seasonal, seasonal) is increased.

transform.function
transformation of the input series: "None" - no transformation of the series; "Log" - takes the log of the series; "Auto" - the program tests for the log-level specification.

transform.fct
numeric controlling the bias in the log/level pre-test: transform.fct > 1 favours levels, transform.fct< 1 favours logs. Considered only when transform.function is set to "Auto".

Control variables for the pre-specified outliers. The pre-specified outliers are used in the model only if they are enabled (usrdef.outliersEnabled=TRUE) and the outliers' type (usrdef.outliersType) and date (usrdef.outliersDate) are provided.

usrdef.outliersEnabled
logicals. If TRUE the program uses the pre-specified outliers.

usrdef.outliersType
vector defining the outliers' type. Possible types are: ("AO") - additive, ("LS") - level shift, ("TC") - transitory change, ("SO") - seasonal outlier. E.g.: usrdef.outliersType =c("AO","AO","LS").

usrdef.outliersDate
vector defining the outliers' date. The dates should be characters in format "YYYY-MM-DD". E.g.: usrdef.outliersDate=c("2009-10-01","2005-02-01","2003-04-01").

usrdef.outliersCoef
vector providing fixed coefficients for the outliers. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined. E.g.: usrdef.outliersCoef=c(200,170,20).

Control variables for the user-defined variables:

usrdef.varEnabled
logicals. If TRUE the program uses the user-defined variables.

usrdef.var
time series (ts) or matrix of time series (mts) with the user-defined variables.

usrdef.varType
vector of character(s) defining the user-defined variables component type. Possible types are: "Undefined", "Series", "Trend", "Seasonal", "SeasonallyAdjusted", "Irregular". If not specified, the program will assign the "Undefined" type.

usrdef.varCoef
vector providing fixed coefficients for the user-defined variables. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined.

tradingdays.mauto
defines whether the calendar effects should be added to the model manually ("Unused") or automatically. In the automatic selection, the choice of the number of calendar variables can be based on the F-Test ("FTest") or the Wald Test ("WaldTest"); the model with higher F value is chosen, provided that it is higher than tradingdays.pftd).

tradingdays.pftd
numeric. P-value applied in the test specified by the automatic parameter (tradingdays.mauto) to assess the significance of the pre-tested calendar effects variables and whether they should be included in the RegArima model.

Control variables for the manual selection of calendar effects variables (tradingdays.mauto is set to "Unused"):

tradingdays.option
defines the type of the trading days regression variables: "TradingDays" - six day-of-the-week regression variables; "WorkingDays" - one working/non-working day contrast variable; "None" - no correction for trading days and working days effects; "UserDefined" - user-defined trading days regressors (currently not working). "None" has also to be chosen for the "day-of-week effects" correction (tradingdays.stocktd to be modified accordingly).

tradingdays.leapyear
logicals. Specifies if the leap-year correction should be included. If TRUE the model includes the leap-year effect.

tradingdays.stocktd
numeric indicating the day of the month when inventories and other stock are reported (to denote the last day of the month set the variable to 31). Modifications of this variable are taken into account only when tradingdays.option is set to "None".

tradingdays.test
defines the pre-tests of the trading day effects: "None" - calendar variables are used in the model without pre-testing; "Separate_T" - a t-test is applied to each trading day variable separately and the trading day variables are included in the RegArima model if at least one t-statistic is greater than 2.6 or if two t-statistics are greater than 2.0 (in absolute terms); "Joint_F" - a joint F-test of significance of all the trading day variables. The trading day effect is significant if the F statistic is greater than 0.95.

easter.type
specifies the presence and the length of the Easter effect: "Unused" - Easter effect is not considered; "Standard" - influences the period of n days strictly before Easter Sunday; "IncludeEaster" - influences the entire period (n) up to and including Easter Sunday; "IncludeEasterMonday" - influences the entire period (n) up to and including Easter Monday.

easter.julian
logicals. If TRUE the program uses the Julian Easter (expressed in Gregorian calendar).

easter.duration
numeric indicating the duration of the Easter effect (length in days, between 1 and 15).

easter.test
logicals. If TRUE the program performs a t-test for the significance of the Easter effect. The Easter effect is considered as significant if the modulus of t-statistic is greater than 1.96.

outlier.enabled
logicals. If TRUE the automatic detection of outliers is enabled in the defined time span.

The time span of the series to be searched for outliers (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: outlier.from, outlier.to, outlier.first, outlier.last, outlier.exclFirst and outlier.exclLast; where outlier.from and outlier.to have priority over remaining span control variables, outlier.last and outlier.first have priority over outlier.exclFirst and outlier.exclLast, and outlier.last has priority over outlier.first.

outlier.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with outlier.to.

outlier.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with outlier.from.

outlier.first
numeric specifying the number of periods considered at the beginning of the series.

outlier.last
numeric specifying the number of periods considered at the end of the series.

outlier.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with outlier.exclLast.

outlier.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with outlier.exclFirst.

outlier.ao
logicals. If TRUE the automatic detection of additive outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.tc
logicals. If TRUE the automatic detection of transitory changes is enabled (outlier.enabled must be also set to TRUE).

outlier.ls
logicals. If TRUE the automatic detection of level shifts is enabled (outlier.enabled must be also set to TRUE).

outlier.so
logicals. If TRUE the automatic detection of seasonal outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.usedefcv
logicals. If TRUE the critical value for the outliers' detection procedure is automatically determined by the number of observations in the outlier detection time span. If FALSE the procedure uses the inputted critical value (outlier.cv).

outlier.cv
numeric. Inputted critical value for the outliers' detection procedure. The modification of this variable is taken in to account only when outlier.usedefcv is set to FALSE.

outlier.eml
logicals, exact likelihood estimation method.Controls the method applied for a parameter estimation in the intermediate steps of the automatic detection and correction of outliers. If TRUE an exact likelihood estimation method is used, when FALSE the fast Hannan-Rissanen method is used.

outlier.tcrate
numeric. The rate of decay for the transitory change outlier.

automdl.enabled
logicals. If TRUE the automatic modelling of the ARIMA model is enabled. If FALSE the parameters of the ARIMA model can be specified.

Control variables for the automatic modelling of the ARIMA model (automdl.enabled is set to TRUE):

automdl.acceptdefault
logicals. If TRUE the default model (ARIMA(0,1,1)(0,1,1)) may be chosen in the first step of the automatic model identification. If the Ljung-Box Q statistics for the residuals is acceptable, the default model is accepted and no further attempt will be made to identify any other.

automdl.cancel
numeric, cancelation limit. If the difference in moduli of an AR and an MA roots (when estimating ARIMA(1,0,1)(1,0,1) models in the second step of the automatic identification of the differencing orders) is smaller than cancelation limit, the two roots are assumed equal and cancel out.

automdl.ub1
numeric, first unit root limit. It is the threshold value for the initial unit root test in the automatic differencing procedure. When one of the roots in the estimation of the ARIMA(2,0,0)(1,0,0) plus mean model, performed in the first step of the automatic model identification procedure, is larger than first unit root limit in modulus, it is set equal to unity.

automdl.ub2
numeric, second unit root limit. When one of the roots in the estimation of the ARIMA(1,0,1)(1,0,1) plus mean model, which is performed in the second step of the automatic model identification procedure, is larger than second unit root limit in modulus, it is checked if there is a common factor in the corresponding AR and MA polynomials of the ARMA model that can be cancelled (see automdl.cancel). If there is no cancellation, the AR root is set equal to unity (i.e. the differencing order changes).

automdl.armalimit
numeric, arma limit. It is the threshold value for t-statistics of ARMA coefficients and constant term used for the final test of model parsimony. If the highest order ARMA coefficient has a t-value less than this value in magnitude, the order of the model is reduced. Also if the constant term has a t-value less than arma limit in magnitude, it is removed from the set of regressors.

automdl.reducecv
numeric, ReduceCV. The percentage by which the outlier's critical value will be reduced when an identified model is found to have a Ljung-Box statistic with an unacceptable confidence coefficient. The parameter should be between 0 and 1, and will only be active when automatic outlier identification is enabled. The reduced critical value will be set to (1-ReduceCV)xCV, where CV is the original critical value.

automdl.ljungboxlimit
numeric, Ljung Box limit. Acceptance criterion for the confidence intervals of the Ljung-Box Q statistic. If the LjungBox Q statistics for the residuals of a final model is greater than Ljung Box limit, the model is rejected, the outlier critical value is reduced, and model and outlier identification (if specified) is redone with a reduced value.

automdl.compare
logicals. If TRUE the program compares the model identified by the automatic procedure to the default model (ARIMA(0,1,1)(0,1,1)) and the model with the best fit is selected. Criteria considered are residual diagnostics, the model structure and the number of outliers.

Control variables for the non-automatic modelling of the ARIMA model (automdl.enabled is set to FALSE):

arima.mu
logicals. If TRUE, the mean is considered as part of the ARIMA model.

arima.p
numeric. The order of the non-seasonal autoregressive (AR) polynomial.

arima.d
numeric. Regular differencing order.

arima.q
numeric. The order of the non-seasonal moving average (MA) polynomial.

arima.bp
numeric. The order of the seasonal autoregressive (AR) polynomial.

arima.bd
numeric. Seasonal differencing order.

arima.bq
numeric. The order of the seasonal moving average (MA) polynomial.

Control variables for the user-defined ARMA coefficients. Coefficients can be defined for the regular and seasonal autoregressive (AR) polynomials and moving average (MA) polynomials. The model considers the coefficients only if the procedure for their estimation (arima.coefType) is provided, and the number of provided coefficients matches the sum of (regular and seasonal) AR and MA orders (p,q,bp,bq).

arima.coefEnabled
logicals. If TRUE the program uses the user-defined ARMA coefficients.

arima.coef
vector providing the coefficients for the regular and seasonal AR and MA polynominals. The length of the vector must equal the sum of the regular and seasonal AR and MA orders. The coefficients shall be provided in the order: regular AR (Phi - p elements), regular MA (Theta - q elements), seasonal AR (BPhi - bp elements) and seasonal MA (BTheta - bq elements). E.g.: arima.coef=c(0.6,0.7) with arima.p=1, arima.q=0,arima.bp=1 and arima.bq=0.

arima.coefType
vector defining ARMA coefficients estimation procedure. Possible procedures are: "Undefined" - no use of user-defined input (i.e. coefficients are estimated), "Fixed" - fixes the coefficients at the value provided by the user, "Initial" - the value defined by the user is used as initial condition. For orders for which the coefficients shall not be defined, the arima.coef can be set to NA or 0 or the arima.coefType can be set to "Undefined". E.g.: arima.coef = c(-0.8,-0.6,NA), arima.coefType = c("Fixed","Fixed","Undefined").

fcst.horizon
numeric, forecasts horizon. Length of the forecasts generated by the RegARIMA model in periods (positive values) or years (negative values). By default the program generates two years forecasts (fcst.horizon set to -2).

seats.approx
character, approximation mode. When the ARIMA model estimated by TRAMO does not accept an admissible decomposition, SEATS: "None" - performs an approximation; "Legacy" - replaces the model with a decomposable one; "Noisy" - estimates a new model by adding a white noise to the non-admissible model estimated by TRAMO.

seats.trendBoundary
numeric, trend boundary. The boundary from which an AR root is integrated in the trend component. If the modulus of the inverse real root is greater than Trend boundary, the AR root is integrated in the trend component. Below this value the root is integrated in the transitory component.

seats.seasdBoundary
numeric, seasonal boundary. Boundary from which a negative AR root is integrated in the seasonal component.

seats.seasdBoundary1
numeric, seasonal boundary (unique). Boundary from which a negative AR root is integrated in the seasonal component when the root is the unique seasonal root.

seats.seasTol
numeric, seasonal tolerance. The tolerance (measured in degrees) to allocate the AR non-real roots to the seasonal component (if the modulus of the inverse complex AR root is greater than Trend boundary and the frequency of this root differs from one of the seasonal frequencies by less than Seasonal tolerance) or the transitory component (otherwise).

seats.maBoundary
numeric, MA unit root boundary. When the modulus of an estimated MA root falls in the range (xl, 1), it is set to xl.

seats.method
character, estimation method of the unobserved components. The choice can be made from: "Burman" (default, may result in a significant underestimation of the standard deviations of the components as it may become numerically unstable when some roots of the MA polynomial are near 1); "KalmanSmoother" (it is not disturbed by the (quasi-) unit roots in MA); "McElroyMatrix" (has the same stability issues as the Burman's algorithm).

object
model specification, object of class c("SA_spec","TRAMO_SEATS") or c("SA","TRAMO_SEATS").

Value

A two-elements list of class c("SA_spec", "TRAMO_SEATS"): (1) object of class c("regarima_spec", "TRAMO_SEATS") with the RegARIMA model specification, (2) object of class c("seats_spec", "data.frame") with the SEATS algorithm specification. Each component refers to different part of the SA model specification, mirroring the arguments of the function (for details see arguments description). Each of the lowest-level component (except span, pre-specified outliers, user-defined variables and pre-specified ARMA coefficients) is structured within a data frame with columns denoting different variables of the model specification and rows referring to: first row - base specification, as provided within the argument spec or object; second row - user modifications as specified by the remaining arguments of the function (e.g.: arima.d); and third row - final model specification. The final specification (third row) shall include user modifications (row two) unless they were wrongly specified. The pre-specified outliers, user-defined variables and pre-specified ARMA coefficients consist of a list with the Predefined (base model specification) and Final values.

regarima

object of class c("regarima_spec", "TRAMO_SEATS"). See Value of the function regarima_spec_tramoseats

seats

data.frame of class c("seats_spec", "data.frame"), containing the seats variables in line with the names of the arguments variables. The final values can be also accessed with the function s_seats.

Details

The available predefined 'JDemetra+' model specifications (for the function tramoseats_spec_def) are described in the table below.

Identifier	|		Log/level detection	|		Outliers detection

|		Calender effects	|		ARIMA

RSA0 |		NA	|		NA	|

NA	|	Airline(+mean)		RSA1 |	automatic |	AO/LS/TC |

NA	|	Airline(+mean)		RSA2 |	automatic |	AO/LS/TC |	2 td vars + Easter |

Airline(+mean)		RSA3 |	automatic |	AO/LS/TC |		NA	|

automatic		RSA4 |	automatic |	AO/LS/TC |	2 td vars + Easter |	automatic

RSA5 |	automatic |	AO/LS/TC |	7 td vars + Easter |	automatic		RSAfull |	automatic |

AO/LS/TC |	automatic |	automatic		Identifier	|		Log/level detection

References

Info on 'JDemetra+', usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en BOX G.E.P. and JENKINS G.M. (1970), "Time Series Analysis: Forecasting and Control", Holden-Day, San Francisco.

BOX G.E.P., JENKINS G.M., REINSEL G.C. and LJUNG G.M. (2015), "Time Series Analysis: Forecasting and Control", John Wiley & Sons, Hoboken, N. J., 5th edition.

See also

Examples

myseries <- ipi_c_eu[, "FR"]
myspec1 <- tramoseats_spec_def(spec = c("RSAfull"))
mysa1 <- tramoseats(myseries, spec = myspec1)

# Modify a pre-specified model specification
myspec2 <- tramoseats_spec_def(spec = "RSAfull", tradingdays.mauto = "Unused",
                               tradingdays.option = "WorkingDays",
                               easter.type = "Standard",
                               automdl.enabled = FALSE, arima.mu = TRUE)
mysa2 <- tramoseats(myseries, spec = myspec2)

# Modify the model specification from a "SA" object
myspec3 <- tramoseats_spec(mysa1, tradingdays.mauto = "Unused",
                           tradingdays.option = "WorkingDays",
                           easter.type = "Standard", automdl.enabled = FALSE, arima.mu = TRUE)
mysa3 <- tramoseats(myseries, myspec3)

# Modify the model specification from a "SA_spec" object
myspec4 <- tramoseats_spec(myspec1, tradingdays.mauto = "Unused",
                           tradingdays.option = "WorkingDays",
                           easter.type = "Standard", automdl.enabled = FALSE, arima.mu = TRUE)
mysa4 <- tramoseats(myseries, myspec4)

# Pre-specified outliers
myspec5 <- tramoseats_spec_def(spec = "RSAfull",
                               usrdef.outliersEnabled = TRUE,
                               usrdef.outliersType = c("LS", "LS"),
                               usrdef.outliersDate = c("2008-10-01", "2003-01-01"),
                               usrdef.outliersCoef = c(10,-8), transform.function = "None")
s_preOut(myspec5)

#>   type       date coeff
#> 1   LS 2008-10-01    10
#> 2   LS 2003-01-01    -8

mysa5 <- tramoseats(myseries, myspec5)
mysa5

#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5290      0.048
#> BTheta(1)  -0.5767      0.048
#> 
#>               Estimate Std. Error
#> Monday        -0.07228      0.192
#> Tuesday        0.11018      0.190
#> Wednesday      0.16251      0.191
#> Thursday      -0.55185      0.190
#> Friday         0.41498      0.192
#> Saturday       0.13543      0.191
#> Easter [6]    -1.12520      0.394
#> LS (11-2008) -12.25330      1.561
#> AO (12-2002)  -6.45816      1.493
#> LS (1-2009)   -7.07048      1.498
#> LS (9-2008)   -6.20208      1.500
#> 
#> Fixed outliers: 
#>              Coefficients
#> LS (10-2008)           10
#> LS (1-2003)            -8
#> 
#> 
#> Residual standard error: 1.858 on 323 degrees of freedom
#> Log likelihood =  -661, aic =  1350 aicc =  1351, bic(corrected for length) = 1.471
#> 
#> 
#> 
#> Decomposition
#> Model
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.529038 B - 0.576652 B^12 + 0.305071 B^13 
#> 
#> 
#> SA
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.493010 B + 0.513732 B^2 
#> Innovation variance:  0.6429039 
#> 
#> Trend
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 + 0.044711 B - 0.955289 B^2 
#> Innovation variance:  0.03452403 
#> 
#> Seasonal
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 1.222150 B + 1.160157 B^2 + 1.000327 B^3 + 0.763420 B^4 + 0.512045 B^5 + 0.268153 B^6 + 0.048532 B^7 - 0.122082 B^8 - 0.270259 B^9 - 0.360991 B^10 - 0.581084 B^11 
#> Innovation variance:  0.04338372 
#> 
#> Irregular
#> Innovation variance:  0.3632607 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t           s           i
#> Jan 2017  97.4 100.7979 100.9584  -3.3979068 -0.16047894
#> Feb 2017  97.5 100.4892 101.2959  -2.9892168 -0.80664833
#> Mar 2017 112.0 102.5769 101.7208   9.4231475  0.85605254
#> Apr 2017 103.0 101.4948 102.1242   1.5051912 -0.62944043
#> May 2017 100.4 103.1439 102.5098  -2.7438926  0.63412880
#> Jun 2017 111.2 102.5950 102.8643   8.6049570 -0.26928561
#> Jul 2017 103.4 103.2819 103.2055   0.1181372  0.07637373
#> Aug 2017  79.3 103.3246 103.5825 -24.0245752 -0.25791384
#> Sep 2017 109.7 103.1983 104.0809   6.5016545 -0.88258093
#> Oct 2017 114.0 106.2017 104.6202   7.7982962  1.58152641
#> Nov 2017 107.7 105.4460 104.9030   2.2539632  0.54299390
#> Dec 2017 101.4 104.4912 104.9886  -3.0911545 -0.49745990
#> 
#> Forecasts:
#>                y_f     sa_f      t_f          s_f          i_f
#> Jan 2018 102.06576 105.1151 105.1151  -3.04936263 1.220712e-11
#> Feb 2018 102.22450 105.2868 105.2868  -3.06228987 1.148237e-11
#> Mar 2018 113.95694 105.4585 105.4585   8.49847169 1.313083e-11
#> Apr 2018 107.79414 105.6301 105.6301   2.16400028 1.155342e-11
#> May 2018 102.59676 105.8018 105.8018  -3.20504699 1.095657e-11
#> Jun 2018 115.21892 105.9735 105.9735   9.24544124 1.355716e-11
#> Jul 2018 106.21789 106.1452 106.1452   0.07273947 1.278977e-11
#> Aug 2018  82.55735 106.3168 106.3168 -23.75946902 1.050182e-11
#> Sep 2018 112.30108 106.4885 106.4885   5.81258745 1.290346e-11
#> Oct 2018 114.93423 106.6602 106.6602   8.27405932 1.314504e-11
#> Nov 2018 109.38136 106.8318 106.8318   2.54952421 1.145395e-11
#> Dec 2018 103.39058 107.0035 107.0035  -3.61293045 1.223555e-11
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.499
#>  Seasonal     45.062
#>  Irregular     0.357
#>  TD & Hol.     0.056
#>  Others       52.353
#>  Total        99.327
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.054
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.974
#>  f-test on sa (td)                      0.998
#>  f-test on i (td)                       1.000
#> 
#> 
#> Additional output variables

s_preOut(mysa5)

#>   type       date coeff
#> 1   LS 2008-10-01    10
#> 2   LS 2003-01-01    -8

# User-defined variables
var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
var<- ts.union(var1, var2)

myspec6 <- tramoseats_spec_def(spec = "RSAfull",
                              usrdef.varEnabled = TRUE, usrdef.var = var)
s_preVar(myspec6)

#> $series
#>                  var1          var2
#> Jan 1990   3.50212641   26.53547903
#> Feb 1990   7.58213674   61.90669397
#> Mar 1990   2.90928722   47.10808367
#> Apr 1990   0.22859592   46.31442134
#> May 1990  -2.08143110  -16.01613628
#> Jun 1990  -2.63611625   -3.56854719
#> Jul 1990 -19.84135796  -57.46848158
#> Aug 1990  -3.57620220   84.53500168
#> Sep 1990  -4.54297352  -58.17661225
#> Oct 1990   2.86157701  -50.23571057
#> Nov 1990  -2.89179390  108.80762646
#> Dec 1990   5.11904574  -39.29530194
#> Jan 1991 -11.04334548  -66.04611191
#> Feb 1991  -2.22474789  -14.90251268
#> Mar 1991 -14.52151487  -87.54852976
#> Apr 1991   3.99298770  169.90015707
#> May 1991  14.28571193    6.81902616
#> Jun 1991  -4.82140994   53.31875802
#> Jul 1991   9.92388183   -8.09489931
#> Aug 1991 -11.63700651  244.67814279
#> Sep 1991  16.20789124   38.38990336
#> Oct 1991  -1.07448152 -105.21715779
#> Nov 1991  -7.73206773 -172.22338269
#> Dec 1991  -7.36278805  -63.08192495
#> Jan 1992   8.15706680  -95.35959793
#> Feb 1992  22.03036444   -7.99824652
#> Mar 1992  -6.71046628  106.26959023
#> Apr 1992  -4.21589604  135.10568359
#> May 1992  -7.77749125   23.32902205
#> Jun 1992   5.61542375   33.38987914
#> Jul 1992 -22.22507056  -76.47702581
#> Aug 1992  -6.95413671   76.93844601
#> Sep 1992  -3.83549957  -21.50129615
#> Oct 1992  20.58566983   38.33805853
#> Nov 1992 -11.07988264  -54.24953612
#> Dec 1992  -9.55938774 -209.45782198
#> Jan 1993  -6.35768417   37.80500587
#> Feb 1993   7.66757452  -70.59342281
#> Mar 1993   6.86958102  -58.71045298
#> Apr 1993  17.21426693   10.00863593
#> May 1993   8.51309252  120.20663588
#> Jun 1993   4.29908280   65.98439960
#> Jul 1993 -10.55277738  -62.67453556
#> Aug 1993  -0.45880971  -52.97761060
#> Sep 1993  -1.31881384   87.65031871
#> Oct 1993  -3.19463558   -5.01382118
#> Nov 1993   4.73606163  228.03903528
#> Dec 1993   8.12516099  137.98105639
#> Jan 1994   2.56612328  -57.36360343
#> Feb 1994 -21.48765835 -141.68374281
#> Mar 1994   6.90108551 -124.49074589
#> Apr 1994 -17.72443831   90.00824800
#> May 1994   5.46294039  198.94302565
#> Jun 1994   0.18945820 -278.69752864
#> Jul 1994  -7.48702215   -2.31353042
#> Aug 1994  -7.94321035   20.41608395
#> Sep 1994   3.00784394    4.42285247
#> Oct 1994   0.54665853   62.70974878
#> Nov 1994   3.64229108  -94.72495541
#> Dec 1994 -11.65186698 -112.28269438
#> Jan 1995   2.60555826   -2.03236495
#> Feb 1995 -24.17116277  122.50255435
#> Mar 1995  11.49085017    8.30589847
#> Apr 1995  12.06579247 -188.59151704
#> May 1995  -4.31982134  105.69271276
#> Jun 1995   1.06856739   61.29121886
#> Jul 1995   4.86195719   58.96500909
#> Aug 1995 -15.53458324   -6.64513411
#> Sep 1995 -10.83142387  -60.81078807
#> Oct 1995  -5.16012028   94.73045743
#> Nov 1995  -2.07770980   39.31699575
#> Dec 1995   9.07281411  -38.52059312
#> Jan 1996   8.11425127  -60.92631219
#> Feb 1996  -9.91193254  -38.43851460
#> Mar 1996  -4.54306958  -90.14236692
#> Apr 1996  -3.15728538   87.29458934
#> May 1996 -10.06333083   56.18212586
#> Jun 1996 -10.16594660 -175.72785220
#> Jul 1996  20.29964782  -81.21160950
#> Aug 1996  -6.87638434   23.86905873
#> Sep 1996 -15.87906854  250.60017195
#> Oct 1996  10.16938102  -98.05871650
#> Nov 1996  -7.23102186  156.61250555
#> Dec 1996 -14.53485040  -75.88674355
#> Jan 1997 -21.40815482   49.38451044
#> Feb 1997   1.34854427  -32.71504519
#> Mar 1997   6.61997776  -43.48790398
#> Apr 1997   4.94613170  151.97231060
#> May 1997 -15.26643655    0.36778033
#> Jun 1997  -1.10216005  -40.23399502
#> Jul 1997  24.68741282 -149.54664332
#> Aug 1997  -5.63543434  -94.59855132
#> Sep 1997 -14.45853422   90.64295082
#> Oct 1997  -4.76555776 -181.38676228
#> Nov 1997  10.14600468  -95.17754317
#> Dec 1997  -1.18020034  112.68520993
#> Jan 1998  -3.54321844   13.98890649
#> Feb 1998   2.87255843   10.03399767
#> Mar 1998   0.18012616 -120.45286294
#> Apr 1998  -8.98423976   -8.34582964
#> May 1998  14.44736492  -79.70822469
#> Jun 1998  -5.37799146  110.15245786
#> Jul 1998  -4.82359998   10.84104063
#> Aug 1998  -7.35619613 -119.05067990
#> Sep 1998  -0.15235256   27.99212761
#> Oct 1998  -2.85060838   77.74077731
#> Nov 1998  -7.70895994  -22.10897240
#> Dec 1998 -19.11522944  -13.76768354
#> Jan 1999   1.77899655   11.52300433
#> Feb 1999   1.96320685  153.32695729
#> Mar 1999   0.50044155  -65.57744602
#> Apr 1999 -10.45274438   85.83787007
#> May 1999  -1.36682386 -118.76705517
#> Jun 1999  12.27500609   53.95403804
#> Jul 1999  -1.93865193  -26.63621627
#> Aug 1999 -19.93071105   75.81722346
#> Sep 1999  -4.29905670 -103.14020292
#> Oct 1999  -6.51916325  102.49315104
#> Nov 1999 -13.22755316  -14.52615921
#> Dec 1999  -6.11002356   10.92238302
#> Jan 2000  -4.33484772 -147.85987947
#> Feb 2000   2.21131620 -227.91912223
#> Mar 2000   4.14635297   60.00768904
#> Apr 2000  10.03519632  -34.65781995
#> May 2000 -26.26042237  -28.51536719
#> Jun 2000  -8.73534926  -64.76521580
#> Jul 2000  -7.73710833 -127.22147004
#> Aug 2000   5.93058550  -17.51731809
#> Sep 2000   4.53563341 -220.13195775
#> Oct 2000  -4.89626711   57.98759794
#> Nov 2000  -6.66085502    6.27964711
#> Dec 2000 -10.89654228  177.51245181
#> Jan 2001 -10.40906952  138.36946999
#> Feb 2001   5.01606680  163.09602864
#> Mar 2001   0.41353765  -80.83912981
#> Apr 2001  -0.53081681    9.55654515
#> May 2001 -15.27399081   91.68081903
#> Jun 2001  -1.63568520   47.38435430
#> Jul 2001 -21.35094918  -39.52457160
#> Aug 2001  -1.36833591   42.93985438
#> Sep 2001 -29.04220640   85.09307909
#> Oct 2001   3.37012645  -17.03946179
#> Nov 2001   7.64510656   -0.07914038
#> Dec 2001  -6.69378190   86.06414955
#> Jan 2002  -1.60968872   64.03423802
#> Feb 2002  15.07261029   10.68762896
#> Mar 2002  -1.21693844 -146.04777644
#> Apr 2002  -2.43414776   45.51713753
#> May 2002   7.08157194  127.20846841
#> Jun 2002  -4.80039667 -191.51875806
#> Jul 2002  -3.16551268   23.55890946
#> Aug 2002   5.10471244  105.59143097
#> Sep 2002  -2.66264354   43.16772438
#> Oct 2002   6.36432097 -174.59432119
#> Nov 2002 -10.13955594  263.45754763
#> Dec 2002   9.32230747   -3.08035714
#> Jan 2003  -6.50788332 -225.61887639
#> Feb 2003   4.36178180  183.68225390
#> Mar 2003  -1.64089402   49.85355661
#> Apr 2003  12.68656386   15.62468467
#> May 2003   4.15432504   55.65016905
#> Jun 2003  -0.48104283   68.71069775
#> Jul 2003   1.98292951 -330.41822297
#> Aug 2003  -3.88687305   87.83299631
#> Sep 2003  -4.52232286  146.11756864
#> Oct 2003  -1.39421427  -10.94066808
#> Nov 2003 -13.25197658  132.56063791
#> Dec 2003   7.54069518  114.06805737
#> Jan 2004  13.57852830  -93.03642146
#> Feb 2004 -12.20784861  -19.46233382
#> Mar 2004  -5.67918662   92.40072693
#> Apr 2004 -11.88849569  -28.37713588
#> May 2004 -10.12275885   64.38188432
#> Jun 2004  -4.32531486  120.49299753
#> Jul 2004   3.45706392  -98.64334772
#> Aug 2004   1.90927596  136.29994009
#> Sep 2004  12.87488578  -63.71366390
#> Oct 2004  -2.13504255 -125.27051445
#> Nov 2004   5.93561305  -57.51377304
#> Dec 2004 -18.15799874 -167.63008174
#> Jan 2005  -4.60524573   26.28120182
#> Feb 2005   7.46337982 -142.08832929
#> Mar 2005   8.98995877   77.38467161
#> Apr 2005 -11.55462161  -63.01913980
#> May 2005  -2.82423821 -252.32229681
#> Jun 2005  -0.03256871  104.56761237
#> Jul 2005  15.15028276   75.13402330
#> Aug 2005   3.09380270   79.81608534
#> Sep 2005  -4.60958821 -156.76418337
#> Oct 2005   2.93447573 -114.74768636
#> Nov 2005   0.70972006   43.55867220
#> Dec 2005 -23.13759365  105.32004191
#> Jan 2006   1.92706032   96.35066048
#> Feb 2006  -3.52441451   69.39309379
#> Mar 2006 -15.33292787  -23.00709516
#> Apr 2006 -17.34069209  -93.55586095
#> May 2006   1.41892735  -90.26203778
#> Jun 2006  -2.19599601   -3.08600940
#> Jul 2006   5.91564494  -86.86782246
#> Aug 2006   2.40247909  218.77875041
#> Sep 2006  -3.72564160  153.85301545
#> Oct 2006   2.64545865 -188.20726709
#> Nov 2006   6.51544488   91.58750045
#> Dec 2006  -4.18319774  -40.17400140
#> Jan 2007 -20.78760627 -125.68740002
#> Feb 2007   5.90955041  118.43968692
#> Mar 2007  -7.05640744  100.85724636
#> Apr 2007   9.86688400 -106.61521156
#> May 2007   0.32359045 -108.02139977
#> Jun 2007  -4.96043573 -152.65903705
#> Jul 2007  16.73249059 -137.04004261
#> Aug 2007 -10.51266614 -129.35694331
#> Sep 2007   0.28015821 -296.10297132
#> Oct 2007   5.26890091  112.55956830
#> Nov 2007  12.70786920    0.40275376
#> Dec 2007  -5.87592446   16.26711541
#> Jan 2008  -4.30471444 -148.39223699
#> Feb 2008  -0.22966263  116.48603184
#> Mar 2008   0.74873923 -127.24193004
#> Apr 2008 -20.89995616 -111.37786657
#> May 2008  10.23897173 -130.69370448
#> Jun 2008   6.23240132  -17.42082303
#> Jul 2008 -13.67330008  -68.32515233
#> Aug 2008   0.44589523  -62.15682860
#> Sep 2008   3.66273799  -15.51632320
#> Oct 2008  -1.78469697 -108.64935828
#> Nov 2008  21.01677593   55.02220067
#> Dec 2008  -2.42409658  139.47806779
#> Jan 2009   0.11717215 -158.83192579
#> Feb 2009 -27.78433689  206.75835289
#> Mar 2009   2.74491395 -169.71130469
#> Apr 2009  -6.22731928   86.05076312
#> May 2009 -10.27831717  -37.96434887
#> Jun 2009   7.28990109  184.76301068
#> Jul 2009  -3.92461437  103.76561791
#> Aug 2009   3.57845575 -146.95747682
#> Sep 2009  -7.55189860 -277.69482100
#> Oct 2009   0.66689713   -7.69797878
#> Nov 2009   2.57355650  -95.44920721
#> Dec 2009  -0.90963056   70.86794196
#> Jan 2010  -7.30508027  177.70868082
#> Feb 2010  -3.95841246   86.63143079
#> Mar 2010 -14.15808515   15.30058172
#> Apr 2010  -1.08938918   -4.63890794
#> May 2010   0.74299910 -167.28474148
#> Jun 2010   4.50190145 -108.98906119
#> Jul 2010   3.87345818   -7.61757047
#> Aug 2010  -2.58295014 -130.58317158
#> Sep 2010  15.23490603   -6.23845628
#> Oct 2010   5.51343438   34.23715685
#> Nov 2010   3.29901773  -54.89306150
#> Dec 2010   1.72922695   91.50613267
#> Jan 2011   4.26641536   -8.76039035
#> Feb 2011  -8.13699585   -2.55794707
#> Mar 2011   6.51355094 -125.43574063
#> Apr 2011   9.19981796  188.01248228
#> May 2011   3.90101788  -14.09493686
#> Jun 2011   9.56724098  -10.66911374
#> Jul 2011  -1.69973563   72.45550904
#> Aug 2011  -3.79977916   42.88180835
#> Sep 2011   1.44207957 -158.02157511
#> Oct 2011  -7.18284089   77.47558144
#> Nov 2011   2.06108648 -116.42921245
#> Dec 2011  -3.67334361    0.88543033
#> Jan 2012   7.88442877  -45.50780017
#> Feb 2012  -3.10952795 -100.26900477
#> Mar 2012  -8.59140099   42.80695718
#> Apr 2012  10.85279663   20.25623016
#> May 2012  -7.84779324   18.15832754
#> Jun 2012 -12.20960234   56.56938676
#> Jul 2012 -10.56691432  -27.76569799
#> Aug 2012  -9.96890554 -136.42484813
#> Sep 2012  -6.72605085  129.92072495
#> Oct 2012   6.65268367 -199.28465574
#> Nov 2012  22.36077683  -10.37160872
#> Dec 2012  -6.01798169   29.97662573
#> Jan 2013  -2.91617742  -40.30004095
#> Feb 2013  -8.40974862   33.84909572
#> Mar 2013  20.31699515 -147.84270394
#> Apr 2013   4.29372125   -4.65149259
#> May 2013   8.11017521    3.16881860
#> Jun 2013  -1.74299949   28.15326072
#> Jul 2013  -3.16371483   80.31952537
#> Aug 2013  -4.93001895  -20.88583488
#> Sep 2013   1.90778474  -33.67281762
#> Oct 2013 -11.03744984  -45.63097517
#> Nov 2013  -1.06413314  102.42603097
#> Dec 2013  14.58992713  145.73951507
#> Jan 2014  -3.07474298  -22.49305874
#> Feb 2014   7.57181411   29.00685098
#> Mar 2014 -17.10825390   25.92765437
#> Apr 2014   5.95309608    0.98922998
#> May 2014 -20.67966650  -18.46739254
#> Jun 2014  -0.85739756  165.42465847
#> Jul 2014   0.32286550   43.68543558
#> Aug 2014   2.89686397   84.73161118
#> Sep 2014   8.08938069   99.62823002
#> Oct 2014   3.26262471  106.74524423
#> Nov 2014 -17.03857479   -5.33425785
#> Dec 2014   5.25504589   99.29262251
#> Jan 2015  -8.03033850  -49.93526848
#> Feb 2015  -2.37655721 -111.63933292
#> Mar 2015   8.55349059  -97.45167645
#> Apr 2015  -4.69487045  278.74035400
#> May 2015   1.18012784  -77.33134688
#> Jun 2015  12.47441555  348.05397484
#> Jul 2015  -6.88397299  -61.74219736
#> Aug 2015 -13.83569575   52.07824425
#> Sep 2015  -0.42932415   59.25402530
#> Oct 2015  -2.09234230  -90.03468488
#> Nov 2015 -14.66993454  147.42813210
#> Dec 2015  22.35778380  108.51326027
#> Jan 2016  -6.25144963 -170.66874385
#> Feb 2016 -11.67913785 -116.86962998
#> Mar 2016   1.49528831 -141.94533339
#> Apr 2016   8.61319900   64.59709494
#> May 2016 -10.46203407  -42.71327890
#> Jun 2016   1.72260011  118.98281468
#> Jul 2016  -3.07020751  -13.98591424
#> Aug 2016   9.03345610   18.93959765
#> Sep 2016   9.85248694    5.75844052
#> Oct 2016  16.43828484  -58.72461267
#> Nov 2016   8.14559724  -24.14145797
#> Dec 2016   8.25113161   76.38003907
#> Jan 2017  -3.45773922  -67.80967332
#> Feb 2017   4.11147335   83.69941394
#> Mar 2017  -6.43214383    4.19394664
#> Apr 2017   9.42059477   97.93095986
#> May 2017   4.69591784  149.60956543
#> Jun 2017  -6.62112238  138.11671565
#> Jul 2017  -3.09471102   17.71868508
#> Aug 2017   7.94574264  -60.45588857
#> Sep 2017  -9.14850956   -0.64125536
#> Oct 2017  -4.65520309 -103.09176045
#> Nov 2017   8.19669363  -59.66193013
#> Dec 2017  15.21035738  123.00053344
#> 
#> $description
#>           type coeff
#> var1 Undefined    NA
#> var2 Undefined    NA
#> 

mysa6 <- tramoseats(myseries, myspec6)

myspec7 <- tramoseats_spec_def(spec = "RSAfull", usrdef.varEnabled = TRUE,
                               usrdef.var = var, usrdef.varCoef = c(17,-1),
                               transform.function = "None")
mysa7 <- tramoseats(myseries, myspec7)

# Pre-specified ARMA coefficients
myspec8 <- tramoseats_spec_def(spec = "RSAfull",
                               arima.coefEnabled = TRUE, automdl.enabled = FALSE,
                               arima.p = 2, arima.q = 0,
                               arima.bp = 1, arima.bq = 1,
                               arima.coef = c(-0.12, -0.12, -0.3, -0.99),
                               arima.coefType = rep("Fixed", 4))
mysa8 <- tramoseats(myseries, myspec8)
mysa8

#> RegARIMA
#> y = regression model + arima (2, 1, 0, 1, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)       -0.12          0
#> Phi(2)       -0.12          0
#> BPhi(1)      -0.30          0
#> BTheta(1)    -0.99          0
#> 
#>                Estimate Std. Error
#> Mean         -3.039e-05      0.000
#> Monday       -2.430e-03      0.002
#> Tuesday       3.562e-04      0.002
#> Wednesday     4.238e-03      0.002
#> Thursday     -6.307e-03      0.002
#> Friday        4.267e-03      0.002
#> Saturday      5.834e-04      0.002
#> Easter [6]   -1.144e-02      0.003
#> LS (11-2008) -5.733e-02      0.020
#> LS (1-2009)  -4.622e-02      0.020
#> 
#> 
#> Residual standard error: 0.02054 on 323 degrees of freedom
#> Log likelihood = 781.7, aic =  1441 aicc =  1442, bic(corrected for length) = -7.591
#> 
#> 
#> 
#> Decomposition
#> Model
#> AR :  1 - 0.120000 B - 0.120000 B^2 - 0.300000 B^12 + 0.036000 B^13 + 0.036000 B^14 
#> D :  1 - B - B^12 + B^13 
#> MA :  1 - 0.950000 B^12 
#> 
#> 
#> SA
#> AR :  1 - 1.024538 B - 0.011455 B^2 + 0.108545 B^3 
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 1.941460 B + 0.995458 B^2 + 0.036596 B^3 - 0.098708 B^4 + 0.008921 B^5 
#> Innovation variance:  0.4957156 
#> 
#> Trend
#> AR :  1 - 0.904538 B 
#> D :  1 - 2.000000 B + B^2 
#> MA :  1 - 0.722193 B - 0.998833 B^2 + 0.723360 B^3 
#> Innovation variance:  0.1026625 
#> 
#> Seasonal
#> AR :  1 + 0.904538 B + 0.818189 B^2 + 0.740083 B^3 + 0.669433 B^4 + 0.605527 B^5 + 0.547723 B^6 + 0.495436 B^7 + 0.448140 B^8 + 0.405360 B^9 + 0.366664 B^10 + 0.331661 B^11 
#> D :  1 + B + B^2 + B^3 + B^4 + B^5 + B^6 + B^7 + B^8 + B^9 + B^10 + B^11 
#> MA :  1 + 2.915179 B + 5.617938 B^2 + 8.595374 B^3 + 11.505684 B^4 + 14.135214 B^5 + 16.348944 B^6 + 18.036964 B^7 + 19.197197 B^8 + 19.965174 B^9 + 20.301073 B^10 + 20.149548 B^11 + 19.110302 B^12 + 17.206568 B^13 + 14.558550 B^14 + 11.650923 B^15 + 8.806927 B^16 + 6.228869 B^17 + 4.044510 B^18 + 2.357557 B^19 + 1.170033 B^20 + 0.354583 B^21 - 0.052292 B^22 
#> Innovation variance:  0.2584115 
#> 
#> Transitory
#> AR :  1 - 0.120000 B - 0.120000 B^2 
#> MA :  1 + 1.287572 B + 0.287572 B^2 
#> Innovation variance:  9.66767e-06 
#> 
#> Irregular
#> Innovation variance:  0.1228651 
#> 
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2017  97.4 101.2668 101.0578 0.9618154 1.0020686
#> Feb 2017  97.5 101.2721 101.5876 0.9627533 0.9968935
#> Mar 2017 112.0 102.5609 102.1042 1.0920339 1.0044733
#> Apr 2017 103.0 102.0990 102.4310 1.0088251 0.9967584
#> May 2017 100.4 103.0163 102.5974 0.9746033 1.0040828
#> Jun 2017 111.2 102.2454 102.7571 1.0875793 0.9950202
#> Jul 2017 103.4 103.2943 103.3689 1.0010236 0.9992781
#> Aug 2017  79.3 104.7000 103.7753 0.7574017 1.0089114
#> Sep 2017 109.7 102.5846 103.6790 1.0693609 0.9894443
#> Oct 2017 114.0 104.5874 104.0358 1.0899973 1.0053026
#> Nov 2017 107.7 104.4425 104.1189 1.0311891 1.0031083
#> Dec 2017 101.4 103.2201 103.5819 0.9823665 0.9965070
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2018  99.15257 103.2978 103.2989 0.9635235 0.9999901
#> Feb 2018  98.49070 103.2623 103.2633 0.9585165 0.9999903
#> Mar 2018 110.09644 103.2323 103.2333 1.0713667 0.9999904
#> Apr 2018 103.83531 103.2074 103.2084 1.0108616 0.9999904
#> May 2018  99.26109 103.1870 103.1880 0.9666254 0.9999905
#> Jun 2018 112.97989 103.1708 103.1717 1.0975670 0.9999905
#> Jul 2018 102.73624 103.1582 103.1592 0.9999921 0.9999905
#> Aug 2018  76.60539 103.1489 103.1499 0.7477158 0.9999905
#> Sep 2018 109.12847 103.1426 103.1436 1.0619685 0.9999905
#> Oct 2018 112.33867 103.1390 103.1400 1.0926514 0.9999905
#> Nov 2018 105.35175 103.1377 103.1387 1.0250662 0.9999905
#> Dec 2018 101.11651 103.1387 103.1396 0.9811503 0.9999905
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.773
#>  Seasonal     77.478
#>  Irregular     0.138
#>  TD & Hol.     0.092
#>  Others       16.847
#>  Total        97.328
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.254
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        1.000
#>  f-test on i (seasonal dummies)         1.000
#>  Residual seasonality (entire series)   1.000
#>  Residual seasonality (last 3 years)    0.134
#>  f-test on sa (td)                      0.954
#>  f-test on i (td)                       0.987
#> 
#> 
#> Additional output variables

s_arimaCoef(myspec8)

#>            Type Value
#> Phi(1)    Fixed -0.12
#> Phi(2)    Fixed -0.12
#> BPhi(1)   Fixed -0.30
#> BTheta(1) Fixed -0.99

s_arimaCoef(mysa8)

#>            Type Value
#> Phi(1)    Fixed -0.12
#> Phi(2)    Fixed -0.12
#> BPhi(1)   Fixed -0.30
#> BTheta(1) Fixed -0.99