Display a list of all the available output objects (series, parameters, diagnostics)

Function generating a comprehensive list of available output variables (series, parameters, diagnostics) from the estimation process with x13 and tramoseats. Some items are available in the default estimation output but the remainder can be added using the userdefined parameter.

Usage

user_defined_variables(sa_object = c("X13-ARIMA", "TRAMO-SEATS"))

Arguments

sa_object

a character: "X13-ARIMA" to retrieve the additional output variables available for the X13-ARIMA method and "TRAMO-SEATS" for the TRAMO-SEATS method.

Value

a vector containing the names of all the available output objects (series, diagnostics, parameters)

References

More information and examples related to 'JDemetra+' features in the online documentation: https://jdemetra-new-documentation.netlify.app/

Examples

y<- ipi_c_eu[, "FR"]
user_defined_variables("X13-ARIMA")
#>   [1] "y"                                              
#>   [2] "y_f"                                            
#>   [3] "t"                                              
#>   [4] "t_f"                                            
#>   [5] "sa"                                             
#>   [6] "sa_f"                                           
#>   [7] "s"                                              
#>   [8] "s_f"                                            
#>   [9] "i"                                              
#>  [10] "i_f"                                            
#>  [11] "mode"                                           
#>  [12] "preprocessing.model.span.start"                 
#>  [13] "preprocessing.model.span.end"                   
#>  [14] "preprocessing.model.span.n"                     
#>  [15] "preprocessing.model.espan.start"                
#>  [16] "preprocessing.model.espan.end"                  
#>  [17] "preprocessing.model.espan.n"                    
#>  [18] "preprocessing.model.log"                        
#>  [19] "preprocessing.model.adjust"                     
#>  [20] "preprocessing.model.y"                          
#>  [21] "preprocessing.model.y_f"                        
#>  [22] "preprocessing.model.y_ef"                       
#>  [23] "preprocessing.model.yc"                         
#>  [24] "preprocessing.model.yc_f"                       
#>  [25] "preprocessing.model.yc_ef"                      
#>  [26] "preprocessing.model.l"                          
#>  [27] "preprocessing.model.y_lin"                      
#>  [28] "preprocessing.model.y_lin_f"                    
#>  [29] "preprocessing.model.ycal"                       
#>  [30] "preprocessing.model.ycal_f"                     
#>  [31] "preprocessing.model.det"                        
#>  [32] "preprocessing.model.det_f"                      
#>  [33] "preprocessing.model.l_f"                        
#>  [34] "preprocessing.model.l_b"                        
#>  [35] "preprocessing.model.cal"                        
#>  [36] "preprocessing.model.cal_f"                      
#>  [37] "preprocessing.model.tde"                        
#>  [38] "preprocessing.model.tde_f"                      
#>  [39] "preprocessing.model.mhe"                        
#>  [40] "preprocessing.model.mhe_f"                      
#>  [41] "preprocessing.model.ee"                         
#>  [42] "preprocessing.model.ee_f"                       
#>  [43] "preprocessing.model.omhe"                       
#>  [44] "preprocessing.model.omhe_f"                     
#>  [45] "preprocessing.model.out(*)"                     
#>  [46] "preprocessing.model.out_f"                      
#>  [47] "preprocessing.model.out_i"                      
#>  [48] "preprocessing.model.out_i_f"                    
#>  [49] "preprocessing.model.out_t"                      
#>  [50] "preprocessing.model.out_t_f"                    
#>  [51] "preprocessing.model.out_s"                      
#>  [52] "preprocessing.model.out_s_f"                    
#>  [53] "preprocessing.model.reg"                        
#>  [54] "preprocessing.model.reg_f"                      
#>  [55] "preprocessing.model.reg_t"                      
#>  [56] "preprocessing.model.reg_t_f"                    
#>  [57] "preprocessing.model.reg_s"                      
#>  [58] "preprocessing.model.reg_s_f"                    
#>  [59] "preprocessing.model.reg_i"                      
#>  [60] "preprocessing.model.reg_i_f"                    
#>  [61] "preprocessing.model.reg_sa"                     
#>  [62] "preprocessing.model.reg_sa_f"                   
#>  [63] "preprocessing.model.reg_y"                      
#>  [64] "preprocessing.model.reg_y_f"                    
#>  [65] "preprocessing.model.reg_u"                      
#>  [66] "preprocessing.model.reg_u_f"                    
#>  [67] "preprocessing.model.fullresiduals"              
#>  [68] "preprocessing.model.lp"                         
#>  [69] "preprocessing.model.ntd"                        
#>  [70] "preprocessing.model.nmh"                        
#>  [71] "preprocessing.model.td(*)"                      
#>  [72] "preprocessing.model.easter"                     
#>  [73] "preprocessing.model.nout"                       
#>  [74] "preprocessing.model.noutao"                     
#>  [75] "preprocessing.model.noutls"                     
#>  [76] "preprocessing.model.nouttc"                     
#>  [77] "preprocessing.model.noutso"                     
#>  [78] "preprocessing.model.coefficients"               
#>  [79] "preprocessing.model.description"                
#>  [80] "preprocessing.model.covar"                      
#>  [81] "preprocessing.model.pcovar"                     
#>  [82] "preprocessing.model.fcasts(?)"                  
#>  [83] "preprocessing.model.bcasts(?)"                  
#>  [84] "preprocessing.model.lin_fcasts(?)"              
#>  [85] "preprocessing.model.lin_bcasts(?)"              
#>  [86] "preprocessing.model.efcasts(?)"                 
#>  [87] "preprocessing.arima.parameters"                 
#>  [88] "preprocessing.arima.p"                          
#>  [89] "preprocessing.arima.d"                          
#>  [90] "preprocessing.arima.q"                          
#>  [91] "preprocessing.arima.bp"                         
#>  [92] "preprocessing.arima.bd"                         
#>  [93] "preprocessing.arima.bq"                         
#>  [94] "preprocessing.likelihood.neffectiveobs"         
#>  [95] "preprocessing.likelihood.np"                    
#>  [96] "preprocessing.likelihood.logvalue"              
#>  [97] "preprocessing.likelihood.adjustedlogvalue"      
#>  [98] "preprocessing.likelihood.ssqerr"                
#>  [99] "preprocessing.likelihood.aic"                   
#> [100] "preprocessing.likelihood.aicc"                  
#> [101] "preprocessing.likelihood.bic"                   
#> [102] "preprocessing.likelihood.bicc"                  
#> [103] "preprocessing.likelihood.ser"                   
#> [104] "preprocessing.likelihood.ser-ml"                
#> [105] "preprocessing.residuals.res"                    
#> [106] "preprocessing.residuals.mean"                   
#> [107] "preprocessing.residuals.skewness"               
#> [108] "preprocessing.residuals.kurtosis"               
#> [109] "preprocessing.residuals.dh"                     
#> [110] "preprocessing.residuals.lb"                     
#> [111] "preprocessing.residuals.lb2"                    
#> [112] "preprocessing.residuals.seaslb"                 
#> [113] "preprocessing.residuals.bp"                     
#> [114] "preprocessing.residuals.bp2"                    
#> [115] "preprocessing.residuals.seasbp"                 
#> [116] "preprocessing.residuals.nruns"                  
#> [117] "preprocessing.residuals.lruns"                  
#> [118] "mstats.M(*)"                                    
#> [119] "mstats.Q"                                       
#> [120] "mstats.Q-M2"                                    
#> [121] "decomposition.a1"                               
#> [122] "decomposition.a1a"                              
#> [123] "decomposition.a1b"                              
#> [124] "decomposition.a6"                               
#> [125] "decomposition.a7"                               
#> [126] "decomposition.a8"                               
#> [127] "decomposition.a8t"                              
#> [128] "decomposition.a8s"                              
#> [129] "decomposition.a8i"                              
#> [130] "decomposition.a9"                               
#> [131] "decomposition.a9sa"                             
#> [132] "decomposition.a9u"                              
#> [133] "decomposition.a9ser"                            
#> [134] "decomposition.b1"                               
#> [135] "decomposition.b2"                               
#> [136] "decomposition.b3"                               
#> [137] "decomposition.b4"                               
#> [138] "decomposition.b5"                               
#> [139] "decomposition.b6"                               
#> [140] "decomposition.b7"                               
#> [141] "decomposition.b8"                               
#> [142] "decomposition.b9"                               
#> [143] "decomposition.b10"                              
#> [144] "decomposition.b11"                              
#> [145] "decomposition.b12"                              
#> [146] "decomposition.b13"                              
#> [147] "decomposition.b14"                              
#> [148] "decomposition.b15"                              
#> [149] "decomposition.b16"                              
#> [150] "decomposition.b17"                              
#> [151] "decomposition.b18"                              
#> [152] "decomposition.b19"                              
#> [153] "decomposition.b20"                              
#> [154] "decomposition.c1"                               
#> [155] "decomposition.c2"                               
#> [156] "decomposition.c3"                               
#> [157] "decomposition.c4"                               
#> [158] "decomposition.c5"                               
#> [159] "decomposition.c6"                               
#> [160] "decomposition.c7"                               
#> [161] "decomposition.c8"                               
#> [162] "decomposition.c9"                               
#> [163] "decomposition.c10"                              
#> [164] "decomposition.c11"                              
#> [165] "decomposition.c12"                              
#> [166] "decomposition.c13"                              
#> [167] "decomposition.c14"                              
#> [168] "decomposition.c15"                              
#> [169] "decomposition.c16"                              
#> [170] "decomposition.c17"                              
#> [171] "decomposition.c18"                              
#> [172] "decomposition.c19"                              
#> [173] "decomposition.c20"                              
#> [174] "decomposition.d1"                               
#> [175] "decomposition.d2"                               
#> [176] "decomposition.d3"                               
#> [177] "decomposition.d4"                               
#> [178] "decomposition.d5"                               
#> [179] "decomposition.d6"                               
#> [180] "decomposition.d7"                               
#> [181] "decomposition.d8"                               
#> [182] "decomposition.d9"                               
#> [183] "decomposition.d10"                              
#> [184] "decomposition.d10a"                             
#> [185] "decomposition.d10b"                             
#> [186] "decomposition.d11"                              
#> [187] "decomposition.d11a"                             
#> [188] "decomposition.d12"                              
#> [189] "decomposition.d12a"                             
#> [190] "decomposition.d13"                              
#> [191] "decomposition.d14"                              
#> [192] "decomposition.d15"                              
#> [193] "decomposition.d16"                              
#> [194] "decomposition.d16a"                             
#> [195] "decomposition.d16b"                             
#> [196] "decomposition.d18"                              
#> [197] "decomposition.d19"                              
#> [198] "decomposition.d20"                              
#> [199] "decomposition.e1"                               
#> [200] "decomposition.e2"                               
#> [201] "decomposition.e3"                               
#> [202] "decomposition.e11"                              
#> [203] "decomposition.y_cmp"                            
#> [204] "decomposition.t_cmp"                            
#> [205] "decomposition.i_cmp"                            
#> [206] "decomposition.s_cmp"                            
#> [207] "decomposition.sa_cmp"                           
#> [208] "decomposition.y_cmp_f"                          
#> [209] "decomposition.t_cmp_f"                          
#> [210] "decomposition.i_cmp_f"                          
#> [211] "decomposition.s_cmp_f"                          
#> [212] "decomposition.sa_cmp_f"                         
#> [213] "decomposition.d9filter"                         
#> [214] "decomposition.slen"                             
#> [215] "decomposition.d12filter"                        
#> [216] "decomposition.tlen"                             
#> [217] "diagnostics.qs"                                 
#> [218] "diagnostics.ftest"                              
#> [219] "diagnostics.qs.on.i"                            
#> [220] "diagnostics.ftest.on.i"                         
#> [221] "diagnostics.combined.all.kruskalwallis"         
#> [222] "diagnostics.combined.all.stable"                
#> [223] "diagnostics.combined.all.evolutive"             
#> [224] "diagnostics.combined.all.summary"               
#> [225] "diagnostics.combined.all.stable.ssm"            
#> [226] "diagnostics.combined.all.stable.ssr"            
#> [227] "diagnostics.combined.all.stable.ssq"            
#> [228] "diagnostics.combined.all.evolutive.ssm"         
#> [229] "diagnostics.combined.all.evolutive.ssr"         
#> [230] "diagnostics.combined.all.evolutive.ssq"         
#> [231] "diagnostics.combined.end.kruskalwallis"         
#> [232] "diagnostics.combined.end.stable"                
#> [233] "diagnostics.combined.end.evolutive"             
#> [234] "diagnostics.combined.end.summary"               
#> [235] "diagnostics.combined.end.stable.ssm"            
#> [236] "diagnostics.combined.end.stable.ssr"            
#> [237] "diagnostics.combined.end.stable.ssq"            
#> [238] "diagnostics.combined.end.evolutive.ssm"         
#> [239] "diagnostics.combined.end.evolutive.ssr"         
#> [240] "diagnostics.combined.end.evolutive.ssq"         
#> [241] "diagnostics.combined.residual.all.kruskalwallis"
#> [242] "diagnostics.combined.residual.all.stable"       
#> [243] "diagnostics.combined.residual.all.evolutive"    
#> [244] "diagnostics.combined.residual.all.summary"      
#> [245] "diagnostics.combined.residual.all.stable.ssm"   
#> [246] "diagnostics.combined.residual.all.stable.ssr"   
#> [247] "diagnostics.combined.residual.all.stable.ssq"   
#> [248] "diagnostics.combined.residual.all.evolutive.ssm"
#> [249] "diagnostics.combined.residual.all.evolutive.ssr"
#> [250] "diagnostics.combined.residual.all.evolutive.ssq"
#> [251] "diagnostics.combined.residual.end.kruskalwallis"
#> [252] "diagnostics.combined.residual.end.stable"       
#> [253] "diagnostics.combined.residual.end.evolutive"    
#> [254] "diagnostics.combined.residual.end.summary"      
#> [255] "diagnostics.combined.residual.end.stable.ssm"   
#> [256] "diagnostics.combined.residual.end.stable.ssr"   
#> [257] "diagnostics.combined.residual.end.stable.ssq"   
#> [258] "diagnostics.combined.residual.end.evolutive.ssm"
#> [259] "diagnostics.combined.residual.end.evolutive.ssr"
#> [260] "diagnostics.combined.residual.end.evolutive.ssq"
#> [261] "diagnostics.residual.all"                       
#> [262] "diagnostics.residual.end"                       
#> [263] "diagnostics.residualtd"                         
#> [264] "diagnostics.residualtd.on.i"                    
#> [265] "diagnostics.variancedecomposition"              
#> [266] "diagnostics.logstat"                            
#> [267] "diagnostics.levelstat"                          
#> [268] "diagnostics.fcast-insample-mean"                
#> [269] "diagnostics.fcast-outsample-mean"               
#> [270] "diagnostics.fcast-outsample-variance"           
#> [271] "diagnostics.seas-lin-f"                         
#> [272] "diagnostics.seas-lin-qs"                        
#> [273] "diagnostics.seas-lin-kw"                        
#> [274] "diagnostics.seas-lin-friedman"                  
#> [275] "diagnostics.seas-lin-periodogram"               
#> [276] "diagnostics.seas-lin-spectralpeaks"             
#> [277] "diagnostics.seas-si-combined"                   
#> [278] "diagnostics.seas-si-evolutive"                  
#> [279] "diagnostics.seas-si-stable"                     
#> [280] "diagnostics.seas-res-f"                         
#> [281] "diagnostics.seas-res-qs"                        
#> [282] "diagnostics.seas-res-kw"                        
#> [283] "diagnostics.seas-res-friedman"                  
#> [284] "diagnostics.seas-res-periodogram"               
#> [285] "diagnostics.seas-res-spectralpeaks"             
#> [286] "diagnostics.seas-res-combined"                  
#> [287] "diagnostics.seas-res-combined3"                 
#> [288] "diagnostics.seas-res-evolutive"                 
#> [289] "diagnostics.seas-res-stable"                    
#> [290] "diagnostics.seas-i-f"                           
#> [291] "diagnostics.seas-i-qs"                          
#> [292] "diagnostics.seas-i-kw"                          
#> [293] "diagnostics.seas-i-periodogram"                 
#> [294] "diagnostics.seas-i-spectralpeaks"               
#> [295] "diagnostics.seas-i-combined"                    
#> [296] "diagnostics.seas-i-combined3"                   
#> [297] "diagnostics.seas-i-evolutive"                   
#> [298] "diagnostics.seas-i-stable"                      
#> [299] "diagnostics.seas-sa-f"                          
#> [300] "diagnostics.seas-sa-qs"                         
#> [301] "diagnostics.seas-sa-kw"                         
#> [302] "diagnostics.seas-sa-friedman"                   
#> [303] "diagnostics.seas-sa-periodogram"                
#> [304] "diagnostics.seas-sa-spectralpeaks"              
#> [305] "diagnostics.seas-sa-combined"                   
#> [306] "diagnostics.seas-sa-combined3"                  
#> [307] "diagnostics.seas-sa-evolutive"                  
#> [308] "diagnostics.seas-sa-stable"                     
#> [309] "diagnostics.seas-sa-ac1"                        
#> [310] "diagnostics.td-sa-all"                          
#> [311] "diagnostics.td-sa-last"                         
#> [312] "diagnostics.td-i-all"                           
#> [313] "diagnostics.td-i-last"                          
#> [314] "diagnostics.td-res-all"                         
#> [315] "diagnostics.td-res-last"                        
#> [316] "diagnostics.ic-ratio-henderson"                 
#> [317] "diagnostics.ic-ratio"                           
#> [318] "diagnostics.msr-global"                         
#> [319] "diagnostics.msr(*)"                             
#> [320] "coherence.annualtotals.value"                   
#> [321] "coherence.annualtotals"                         
#> [322] "coherence.definition.value"                     
#> [323] "coherence.definition"                           
#> [324] "residuals.normality.value"                      
#> [325] "residuals.normality"                            
#> [326] "residuals.independence.value"                   
#> [327] "residuals.independence"                         
#> [328] "residuals.tdpeaks.value"                        
#> [329] "residuals.tdpeaks"                              
#> [330] "residuals.seaspeaks.value"                      
#> [331] "residuals.seaspeaks"                            
#> [332] "benchmarking.original"                          
#> [333] "benchmarking.target"                            
#> [334] "benchmarking.result"                            
m <- x13(y,"RSA5c", userdefined=c("b20","ycal","residuals.kurtosis" ))
m$user_defined$b20
#> NULL
m$user_defined$ycal
#> NULL
m$user_defined$residuals.kurtosis
#> NULL
user_defined_variables("TRAMO-SEATS")
#>   [1] "y"                                               
#>   [2] "y_f"                                             
#>   [3] "t"                                               
#>   [4] "t_f"                                             
#>   [5] "sa"                                              
#>   [6] "sa_f"                                            
#>   [7] "s"                                               
#>   [8] "s_f"                                             
#>   [9] "i"                                               
#>  [10] "i_f"                                             
#>  [11] "mode"                                            
#>  [12] "preprocessing.model.span.start"                  
#>  [13] "preprocessing.model.span.end"                    
#>  [14] "preprocessing.model.span.n"                      
#>  [15] "preprocessing.model.espan.start"                 
#>  [16] "preprocessing.model.espan.end"                   
#>  [17] "preprocessing.model.espan.n"                     
#>  [18] "preprocessing.model.log"                         
#>  [19] "preprocessing.model.adjust"                      
#>  [20] "preprocessing.model.y"                           
#>  [21] "preprocessing.model.y_f"                         
#>  [22] "preprocessing.model.y_ef"                        
#>  [23] "preprocessing.model.yc"                          
#>  [24] "preprocessing.model.yc_f"                        
#>  [25] "preprocessing.model.yc_ef"                       
#>  [26] "preprocessing.model.l"                           
#>  [27] "preprocessing.model.y_lin"                       
#>  [28] "preprocessing.model.y_lin_f"                     
#>  [29] "preprocessing.model.ycal"                        
#>  [30] "preprocessing.model.ycal_f"                      
#>  [31] "preprocessing.model.det"                         
#>  [32] "preprocessing.model.det_f"                       
#>  [33] "preprocessing.model.l_f"                         
#>  [34] "preprocessing.model.l_b"                         
#>  [35] "preprocessing.model.cal"                         
#>  [36] "preprocessing.model.cal_f"                       
#>  [37] "preprocessing.model.tde"                         
#>  [38] "preprocessing.model.tde_f"                       
#>  [39] "preprocessing.model.mhe"                         
#>  [40] "preprocessing.model.mhe_f"                       
#>  [41] "preprocessing.model.ee"                          
#>  [42] "preprocessing.model.ee_f"                        
#>  [43] "preprocessing.model.omhe"                        
#>  [44] "preprocessing.model.omhe_f"                      
#>  [45] "preprocessing.model.out(*)"                      
#>  [46] "preprocessing.model.out_f"                       
#>  [47] "preprocessing.model.out_i"                       
#>  [48] "preprocessing.model.out_i_f"                     
#>  [49] "preprocessing.model.out_t"                       
#>  [50] "preprocessing.model.out_t_f"                     
#>  [51] "preprocessing.model.out_s"                       
#>  [52] "preprocessing.model.out_s_f"                     
#>  [53] "preprocessing.model.reg"                         
#>  [54] "preprocessing.model.reg_f"                       
#>  [55] "preprocessing.model.reg_t"                       
#>  [56] "preprocessing.model.reg_t_f"                     
#>  [57] "preprocessing.model.reg_s"                       
#>  [58] "preprocessing.model.reg_s_f"                     
#>  [59] "preprocessing.model.reg_i"                       
#>  [60] "preprocessing.model.reg_i_f"                     
#>  [61] "preprocessing.model.reg_sa"                      
#>  [62] "preprocessing.model.reg_sa_f"                    
#>  [63] "preprocessing.model.reg_y"                       
#>  [64] "preprocessing.model.reg_y_f"                     
#>  [65] "preprocessing.model.reg_u"                       
#>  [66] "preprocessing.model.reg_u_f"                     
#>  [67] "preprocessing.model.fullresiduals"               
#>  [68] "preprocessing.model.lp"                          
#>  [69] "preprocessing.model.ntd"                         
#>  [70] "preprocessing.model.nmh"                         
#>  [71] "preprocessing.model.td(*)"                       
#>  [72] "preprocessing.model.easter"                      
#>  [73] "preprocessing.model.nout"                        
#>  [74] "preprocessing.model.noutao"                      
#>  [75] "preprocessing.model.noutls"                      
#>  [76] "preprocessing.model.nouttc"                      
#>  [77] "preprocessing.model.noutso"                      
#>  [78] "preprocessing.model.coefficients"                
#>  [79] "preprocessing.model.description"                 
#>  [80] "preprocessing.model.covar"                       
#>  [81] "preprocessing.model.pcovar"                      
#>  [82] "preprocessing.model.fcasts(?)"                   
#>  [83] "preprocessing.model.bcasts(?)"                   
#>  [84] "preprocessing.model.lin_fcasts(?)"               
#>  [85] "preprocessing.model.lin_bcasts(?)"               
#>  [86] "preprocessing.model.efcasts(?)"                  
#>  [87] "preprocessing.arima.parameters"                  
#>  [88] "preprocessing.arima.p"                           
#>  [89] "preprocessing.arima.d"                           
#>  [90] "preprocessing.arima.q"                           
#>  [91] "preprocessing.arima.bp"                          
#>  [92] "preprocessing.arima.bd"                          
#>  [93] "preprocessing.arima.bq"                          
#>  [94] "preprocessing.likelihood.neffectiveobs"          
#>  [95] "preprocessing.likelihood.np"                     
#>  [96] "preprocessing.likelihood.logvalue"               
#>  [97] "preprocessing.likelihood.adjustedlogvalue"       
#>  [98] "preprocessing.likelihood.ssqerr"                 
#>  [99] "preprocessing.likelihood.aic"                    
#> [100] "preprocessing.likelihood.aicc"                   
#> [101] "preprocessing.likelihood.bic"                    
#> [102] "preprocessing.likelihood.bicc"                   
#> [103] "preprocessing.likelihood.ser"                    
#> [104] "preprocessing.likelihood.ser-ml"                 
#> [105] "preprocessing.residuals.res"                     
#> [106] "preprocessing.residuals.mean"                    
#> [107] "preprocessing.residuals.skewness"                
#> [108] "preprocessing.residuals.kurtosis"                
#> [109] "preprocessing.residuals.dh"                      
#> [110] "preprocessing.residuals.lb"                      
#> [111] "preprocessing.residuals.lb2"                     
#> [112] "preprocessing.residuals.seaslb"                  
#> [113] "preprocessing.residuals.bp"                      
#> [114] "preprocessing.residuals.bp2"                     
#> [115] "preprocessing.residuals.seasbp"                  
#> [116] "preprocessing.residuals.nruns"                   
#> [117] "preprocessing.residuals.lruns"                   
#> [118] "decomposition.y_lin"                             
#> [119] "decomposition.y_lin_f"                           
#> [120] "decomposition.y_lin_ef"                          
#> [121] "decomposition.t_lin"                             
#> [122] "decomposition.t_lin_f"                           
#> [123] "decomposition.t_lin_e"                           
#> [124] "decomposition.t_lin_ef"                          
#> [125] "decomposition.sa_lin"                            
#> [126] "decomposition.sa_lin_f"                          
#> [127] "decomposition.sa_lin_e"                          
#> [128] "decomposition.sa_lin_ef"                         
#> [129] "decomposition.s_lin"                             
#> [130] "decomposition.s_lin_f"                           
#> [131] "decomposition.s_lin_e"                           
#> [132] "decomposition.s_lin_ef"                          
#> [133] "decomposition.i_lin"                             
#> [134] "decomposition.i_lin_f"                           
#> [135] "decomposition.i_lin_e"                           
#> [136] "decomposition.i_lin_ef"                          
#> [137] "decomposition.y_cmp"                             
#> [138] "decomposition.y_cmp_f"                           
#> [139] "decomposition.t_cmp"                             
#> [140] "decomposition.t_cmp_f"                           
#> [141] "decomposition.sa_cmp"                            
#> [142] "decomposition.sa_cmp_f"                          
#> [143] "decomposition.s_cmp"                             
#> [144] "decomposition.s_cmp_f"                           
#> [145] "decomposition.i_cmp"                             
#> [146] "decomposition.i_cmp_f"                           
#> [147] "decomposition.i_cmp_e"                           
#> [148] "decomposition.t_cmp_e"                           
#> [149] "decomposition.s_cmp_e"                           
#> [150] "decomposition.sa_cmp_e"                          
#> [151] "decomposition.i_cmp_ef"                          
#> [152] "decomposition.t_cmp_ef"                          
#> [153] "decomposition.s_cmp_ef"                          
#> [154] "decomposition.sa_cmp_ef"                         
#> [155] "decomposition.parameterscutoff"                  
#> [156] "decomposition.modelchanged"                      
#> [157] "decomposition.model.ar"                          
#> [158] "decomposition.model.diff"                        
#> [159] "decomposition.model.ma"                          
#> [160] "decomposition.model.fullar"                      
#> [161] "decomposition.model.innovationvariance"          
#> [162] "decomposition.tmodel.ar"                         
#> [163] "decomposition.tmodel.diff"                       
#> [164] "decomposition.tmodel.ma"                         
#> [165] "decomposition.tmodel.fullar"                     
#> [166] "decomposition.tmodel.innovationvariance"         
#> [167] "decomposition.smodel.ar"                         
#> [168] "decomposition.smodel.diff"                       
#> [169] "decomposition.smodel.ma"                         
#> [170] "decomposition.smodel.fullar"                     
#> [171] "decomposition.smodel.innovationvariance"         
#> [172] "decomposition.samodel.ar"                        
#> [173] "decomposition.samodel.diff"                      
#> [174] "decomposition.samodel.ma"                        
#> [175] "decomposition.samodel.fullar"                    
#> [176] "decomposition.samodel.innovationvariance"        
#> [177] "decomposition.transitorymodel.ar"                
#> [178] "decomposition.transitorymodel.diff"              
#> [179] "decomposition.transitorymodel.ma"                
#> [180] "decomposition.transitorymodel.fullar"            
#> [181] "decomposition.transitorymodel.innovationvariance"
#> [182] "decomposition.imodel.ar"                         
#> [183] "decomposition.imodel.diff"                       
#> [184] "decomposition.imodel.ma"                         
#> [185] "decomposition.imodel.fullar"                     
#> [186] "decomposition.imodel.innovationvariance"         
#> [187] "diagnostics.qs"                                  
#> [188] "diagnostics.ftest"                               
#> [189] "diagnostics.qs.on.i"                             
#> [190] "diagnostics.ftest.on.i"                          
#> [191] "diagnostics.combined.all.kruskalwallis"          
#> [192] "diagnostics.combined.all.stable"                 
#> [193] "diagnostics.combined.all.evolutive"              
#> [194] "diagnostics.combined.all.summary"                
#> [195] "diagnostics.combined.all.stable.ssm"             
#> [196] "diagnostics.combined.all.stable.ssr"             
#> [197] "diagnostics.combined.all.stable.ssq"             
#> [198] "diagnostics.combined.all.evolutive.ssm"          
#> [199] "diagnostics.combined.all.evolutive.ssr"          
#> [200] "diagnostics.combined.all.evolutive.ssq"          
#> [201] "diagnostics.combined.end.kruskalwallis"          
#> [202] "diagnostics.combined.end.stable"                 
#> [203] "diagnostics.combined.end.evolutive"              
#> [204] "diagnostics.combined.end.summary"                
#> [205] "diagnostics.combined.end.stable.ssm"             
#> [206] "diagnostics.combined.end.stable.ssr"             
#> [207] "diagnostics.combined.end.stable.ssq"             
#> [208] "diagnostics.combined.end.evolutive.ssm"          
#> [209] "diagnostics.combined.end.evolutive.ssr"          
#> [210] "diagnostics.combined.end.evolutive.ssq"          
#> [211] "diagnostics.combined.residual.all.kruskalwallis" 
#> [212] "diagnostics.combined.residual.all.stable"        
#> [213] "diagnostics.combined.residual.all.evolutive"     
#> [214] "diagnostics.combined.residual.all.summary"       
#> [215] "diagnostics.combined.residual.all.stable.ssm"    
#> [216] "diagnostics.combined.residual.all.stable.ssr"    
#> [217] "diagnostics.combined.residual.all.stable.ssq"    
#> [218] "diagnostics.combined.residual.all.evolutive.ssm" 
#> [219] "diagnostics.combined.residual.all.evolutive.ssr" 
#> [220] "diagnostics.combined.residual.all.evolutive.ssq" 
#> [221] "diagnostics.combined.residual.end.kruskalwallis" 
#> [222] "diagnostics.combined.residual.end.stable"        
#> [223] "diagnostics.combined.residual.end.evolutive"     
#> [224] "diagnostics.combined.residual.end.summary"       
#> [225] "diagnostics.combined.residual.end.stable.ssm"    
#> [226] "diagnostics.combined.residual.end.stable.ssr"    
#> [227] "diagnostics.combined.residual.end.stable.ssq"    
#> [228] "diagnostics.combined.residual.end.evolutive.ssm" 
#> [229] "diagnostics.combined.residual.end.evolutive.ssr" 
#> [230] "diagnostics.combined.residual.end.evolutive.ssq" 
#> [231] "diagnostics.residual.all"                        
#> [232] "diagnostics.residual.end"                        
#> [233] "diagnostics.residualtd"                          
#> [234] "diagnostics.residualtd.on.i"                     
#> [235] "diagnostics.variancedecomposition"               
#> [236] "diagnostics.logstat"                             
#> [237] "diagnostics.levelstat"                           
#> [238] "diagnostics.fcast-insample-mean"                 
#> [239] "diagnostics.fcast-outsample-mean"                
#> [240] "diagnostics.fcast-outsample-variance"            
#> [241] "diagnostics.seas-lin-f"                          
#> [242] "diagnostics.seas-lin-qs"                         
#> [243] "diagnostics.seas-lin-kw"                         
#> [244] "diagnostics.seas-lin-friedman"                   
#> [245] "diagnostics.seas-lin-periodogram"                
#> [246] "diagnostics.seas-lin-spectralpeaks"              
#> [247] "diagnostics.seas-si-combined"                    
#> [248] "diagnostics.seas-si-evolutive"                   
#> [249] "diagnostics.seas-si-stable"                      
#> [250] "diagnostics.seas-res-f"                          
#> [251] "diagnostics.seas-res-qs"                         
#> [252] "diagnostics.seas-res-kw"                         
#> [253] "diagnostics.seas-res-friedman"                   
#> [254] "diagnostics.seas-res-periodogram"                
#> [255] "diagnostics.seas-res-spectralpeaks"              
#> [256] "diagnostics.seas-res-combined"                   
#> [257] "diagnostics.seas-res-combined3"                  
#> [258] "diagnostics.seas-res-evolutive"                  
#> [259] "diagnostics.seas-res-stable"                     
#> [260] "diagnostics.seas-i-f"                            
#> [261] "diagnostics.seas-i-qs"                           
#> [262] "diagnostics.seas-i-kw"                           
#> [263] "diagnostics.seas-i-periodogram"                  
#> [264] "diagnostics.seas-i-spectralpeaks"                
#> [265] "diagnostics.seas-i-combined"                     
#> [266] "diagnostics.seas-i-combined3"                    
#> [267] "diagnostics.seas-i-evolutive"                    
#> [268] "diagnostics.seas-i-stable"                       
#> [269] "diagnostics.seas-sa-f"                           
#> [270] "diagnostics.seas-sa-qs"                          
#> [271] "diagnostics.seas-sa-kw"                          
#> [272] "diagnostics.seas-sa-friedman"                    
#> [273] "diagnostics.seas-sa-periodogram"                 
#> [274] "diagnostics.seas-sa-spectralpeaks"               
#> [275] "diagnostics.seas-sa-combined"                    
#> [276] "diagnostics.seas-sa-combined3"                   
#> [277] "diagnostics.seas-sa-evolutive"                   
#> [278] "diagnostics.seas-sa-stable"                      
#> [279] "diagnostics.seas-sa-ac1"                         
#> [280] "diagnostics.td-sa-all"                           
#> [281] "diagnostics.td-sa-last"                          
#> [282] "diagnostics.td-i-all"                            
#> [283] "diagnostics.td-i-last"                           
#> [284] "diagnostics.td-res-all"                          
#> [285] "diagnostics.td-res-last"                         
#> [286] "diagnostics.ic-ratio-henderson"                  
#> [287] "diagnostics.ic-ratio"                            
#> [288] "diagnostics.msr-global"                          
#> [289] "diagnostics.msr(*)"                              
#> [290] "coherence.annualtotals.value"                    
#> [291] "coherence.annualtotals"                          
#> [292] "coherence.definition.value"                      
#> [293] "coherence.definition"                            
#> [294] "residuals.normality.value"                       
#> [295] "residuals.normality"                             
#> [296] "residuals.independence.value"                    
#> [297] "residuals.independence"                          
#> [298] "residuals.tdpeaks.value"                         
#> [299] "residuals.tdpeaks"                               
#> [300] "residuals.seaspeaks.value"                       
#> [301] "residuals.seaspeaks"                             
#> [302] "benchmarking.original"                           
#> [303] "benchmarking.target"                             
#> [304] "benchmarking.result"                             
m <- tramoseats(y,"RSAfull", userdefined=c("ycal","variancedecomposition.seasonality"))
m$user_defined$ycal
#> NULL
m$user_defined$variancedecomposition.seasonality
#> NULL