X-13ARIMA-SEATS model specification, SA/X13

Function to create (and/or modify) a c("SA_spec", "X13") class object with the SA model specification for the X13 method. It can be done from a pre-defined 'JDemetra+' model specification (a character), a previous specification (c("SA_spec", "X13") object) or a seasonal adjustment model (c("SA", "X13") object).

x13_spec(spec = c("RSA5c", "RSA0", "RSA1", "RSA2c", "RSA3", "RSA4c",
  "X11"), preliminary.check = NA, estimate.from = NA_character_,
  estimate.to = NA_character_, estimate.first = NA_integer_,
  estimate.last = NA_integer_, estimate.exclFirst = NA_integer_,
  estimate.exclLast = NA_integer_, estimate.tol = NA_integer_,
  transform.function = c(NA, "Auto", "None", "Log"),
  transform.adjust = c(NA, "None", "LeapYear", "LengthOfPeriod"),
  transform.aicdiff = NA_integer_, usrdef.outliersEnabled = NA,
  usrdef.outliersType = NA, usrdef.outliersDate = NA,
  usrdef.outliersCoef = NA, usrdef.varEnabled = NA, usrdef.var = NA,
  usrdef.varType = NA, usrdef.varCoef = NA,
  tradingdays.option = c(NA, "TradingDays", "WorkingDays", "UserDefined",
  "None"), tradingdays.autoadjust = NA, tradingdays.leapyear = c(NA,
  "LeapYear", "LengthOfPeriod", "None"),
  tradingdays.stocktd = NA_integer_, tradingdays.test = c(NA, "Remove",
  "Add", "None"), easter.enabled = NA, easter.julian = NA,
  easter.duration = NA_integer_, easter.test = c(NA, "Add", "Remove",
  "None"), outlier.enabled = NA, outlier.from = NA_character_,
  outlier.to = NA_character_, outlier.first = NA_integer_,
  outlier.last = NA_integer_, outlier.exclFirst = NA_integer_,
  outlier.exclLast = NA_integer_, outlier.ao = NA, outlier.tc = NA,
  outlier.ls = NA, outlier.so = NA, outlier.usedefcv = NA,
  outlier.cv = NA_integer_, outlier.method = c(NA, "AddOne", "AddAll"),
  outlier.tcrate = NA_integer_, automdl.enabled = NA,
  automdl.acceptdefault = NA, automdl.cancel = NA_integer_,
  automdl.ub1 = NA_integer_, automdl.ub2 = NA_integer_,
  automdl.mixed = NA, automdl.balanced = NA,
  automdl.armalimit = NA_integer_, automdl.reducecv = NA_integer_,
  automdl.ljungboxlimit = NA_integer_, automdl.ubfinal = NA_integer_,
  arima.mu = NA, arima.p = NA_integer_, arima.d = NA_integer_,
  arima.q = NA_integer_, arima.bp = NA_integer_,
  arima.bd = NA_integer_, arima.bq = NA_integer_,
  arima.coefEnabled = NA, arima.coef = NA, arima.coefType = NA,
  fcst.horizon = NA_integer_, x11.mode = c(NA, "Undefined", "Additive",
  "Multiplicative", "LogAdditive", "PseudoAdditive"),
  x11.seasonalComp = NA, x11.lsigma = NA_integer_,
  x11.usigma = NA_integer_, x11.trendAuto = NA,
  x11.trendma = NA_integer_, x11.seasonalma = NA_character_,
  x11.fcasts = NA_integer_, x11.bcasts = NA_integer_,
  x11.excludeFcasts = NA, object)

Arguments

spec
model specification X13. It can be a character of predefined X13 'JDemetra+' model specification (see Details), an object of class c("SA_spec","X13") or an object of class c("SA", "X13"). The default is "RSA5c".

The time span of the series to be used for the estimation of the RegArima model coefficients (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: estimate.from, estimate.to, estimate.first, estimate.last, estimate.exclFirst and estimate.exclLast; where estimate.from and estimate.to have priority over remaining span control variables, estimate.last and estimate.first have priority over estimate.exclFirst and estimate.exclLast, and estimate.last has priority over estimate.first.

preliminary.check
boolean to check the quality of the input series and exclude highly problematic ones: e.g. these with a number of identical observations and/or missing values above pre-specified threshold values.

estimate.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with estimate.to.

estimate.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with estimate.from.

estimate.first
numeric specifying the number of periods considered at the beginning of the series.

estimate.last
numeric specifying the number of periods considered at the end of the series.

estimate.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with estimate.exclLast.

estimate.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with estimate.exclFirst.

estimate.tol
numeric, convergence tolerance. The absolute changes in the log-likelihood function are compared to this value to check for the convergence of the estimation iterations.

transform.function
transformation of the input series: "None" - no transformation of the series; "Log" - takes the log of the series; "Auto" - the program tests for the log-level specification.

transform.adjust
pre-adjustment of the input series for length of period or leap year effects: "None" - no adjustment; "LeapYear" - leap year effect; "LengthOfPeriod" - length of period. Modifications of this variable are taken into account only when transform.function is set to "Log".

transform.aicdiff
numeric defining the difference in AICC needed to accept no transformation when the automatic transformation selection is chosen (considered only when transform.function is set to "Auto").

Control variables for the pre-specified outliers. The pre-specified outliers are used in the model only if they are enabled (usrdef.outliersEnabled=TRUE) and the outliers' type (usrdef.outliersType) and date (usrdef.outliersDate) are provided.

usrdef.outliersEnabled
logicals. If TRUE the program uses the pre-specified outliers.

usrdef.outliersType
vector defining the outliers' type. Possible types are: ("AO") - additive, ("LS") - level shift, ("TC") - transitory change, ("SO") - seasonal outlier. E.g.: usrdef.outliersType =c("AO","AO","LS").

usrdef.outliersDate
vector defining the outliers' date. The dates should be characters in format "YYYY-MM-DD". E.g.: usrdef.outliersDate=c("2009-10-01","2005-02-01","2003-04-01").

usrdef.outliersCoef
vector providing fixed coefficients for the outliers. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined. E.g.: usrdef.outliersCoef=c(200,170,20).

Control variables for the user-defined variables:

usrdef.varEnabled
logicals. If TRUE the program uses the user-defined variables.

usrdef.var
time series (ts) or matrix of time series (mts) with the user-defined variables.

usrdef.varType
vector of character(s) defining the user-defined variables component type. Possible types are: "Undefined", "Series", "Trend", "Seasonal", "SeasonallyAdjusted", "Irregular", "Calendar". The type "Calendar" has to be used with tradingdays.option = "UserDefined" to use user-defined calendar regressors. If not specified, the program will assign the "Undefined" type.

usrdef.varCoef
vector providing fixed coefficients for the user-defined variables. The coefficients can't be fixed if transform.function is set to "Auto" - the series transformation need to be pre-defined.

tradingdays.option
defines the type of the trading days regression variables: "TradingDays" - six day-of-the-week regression variables; "WorkingDays" - one working/non-working day contrast variable; "None" - no correction for trading days and working days effects; "UserDefined" - user-defined trading days regressors (regressors have to be defined by the usrdef.var argument with usrdef.varType set to "Calendar" and usrdef.varEnabled = TRUE). "None" has also to be chosen for the "day-of-week effects" correction (tradingdays.stocktd to be modified accordingly).

tradingdays.autoadjust
logicals. If TRUE the program corrects automatically for the leap year effect. Modifications of this variable are taken into account only when transform.function is set to "Auto".

tradingdays.leapyear
option for including the leap-year effect in the model: "LeapYear" - leap year effect; "LengthOfPeriod" - length of period, "None" - no effect included. The leap-year effect can be pre-specified in the model only if the input series was not pre-adjusted (transform.adjust set to "None") and the automatic correction for the leap-year effect was not selected (tradingdays.autoadjust set to FALSE).

tradingdays.stocktd
numeric indicating the day of the month when inventories and other stock are reported (to denote the last day of the month set the variable to 31). Modifications of this variable are taken into account only when tradingdays.option is set to "None".

tradingdays.test
defines the pre-tests for the significance of the trading day regression variables based on the AICC statistics: "Add" - the trading day variables are not included in the initial regression model but can be added to the RegARIMA model after the test; "Remove" - the trading day variables belong to the initial regression model but can be removed from the RegARIMA model after the test; "None" - the trading day variables are not pre-tested and are included in the model.

easter.enabled
logicals. If TRUE the program considers the Easter effect in the model.

easter.julian
logicals. If TRUE the program uses the Julian Easter (expressed in Gregorian calendar).

easter.duration
numeric indicating the duration of the Easter effect (length in days, between 1 and 20).

easter.test
defines the pre-tests for the significance of the Easter effect based on the t-statistic (Easter effect is considered as significant if the t-statistic is greater than 1.96): "Add" - the Easter effect variable is not included in the initial regression model but can be added to the RegARIMA model after the test; "Remove" - the Easter effect variable belong to the initial regression model but can be removed from the RegARIMA model after the test; "None" - the Easter effect variable is not pre-tested and is included in the model.

outlier.enabled
logicals. If TRUE the automatic detection of outliers is enabled in the defined time span.

The time span of the series to be searched for outliers (default from 1900-01-01 to 2020-12-31) is controlled by the following six variables: outlier.from, outlier.to, outlier.first, outlier.last, outlier.exclFirst and outlier.exclLast; where outlier.from and outlier.to have priority over remaining span control variables, outlier.last and outlier.first have priority over outlier.exclFirst and outlier.exclLast, and outlier.last has priority over outlier.first.

outlier.from
character in format "YYYY-MM-DD" indicating the start of the time span (e.g. "1900-01-01"). Can be combined with outlier.to.

outlier.to
character in format "YYYY-MM-DD" indicating the end of the time span (e.g. "2020-12-31"). Can be combined with outlier.from.

outlier.first
numeric specifying the number of periods considered at the beginning of the series.

outlier.last
numeric specifying the number of periods considered at the end of the series.

outlier.exclFirst
numeric specifying the number of periods excluded at the beginning of the series. Can be combined with outlier.exclLast.

outlier.exclLast
numeric specifying the number of periods excluded at the end of the series. Can be combined with outlier.exclFirst.

outlier.ao
logicals. If TRUE the automatic detection of additive outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.tc
logicals. If TRUE the automatic detection of transitory changes is enabled (outlier.enabled must be also set to TRUE).

outlier.ls
logicals. If TRUE the automatic detection of level shifts is enabled (outlier.enabled must be also set to TRUE).

outlier.so
logicals. If TRUE the automatic detection of seasonal outliers is enabled (outlier.enabled must be also set to TRUE).

outlier.usedefcv
logicals. If TRUE the critical value for the outliers' detection procedure is automatically determined by the number of observations in the outlier detection time span. If FALSE the procedure uses the inputted critical value (outlier.cv).

outlier.cv
numeric. Inputted critical value for the outliers' detection procedure. The modification of this variable is taken into account only when outlier.usedefcv is set to FALSE.

outlier.method
determines how the program successively adds detected outliers to the model. At present only the AddOne method is supported.

outlier.tcrate
numeric. The rate of decay for the transitory change outlier.

automdl.enabled
logicals. If TRUE the automatic modelling of the ARIMA model is enabled. If FALSE the parameters of the ARIMA model can be specified.

Control variables for the automatic modelling of the ARIMA model (automdl.enabled is set to TRUE):

automdl.acceptdefault
logicals. If TRUE the default model (ARIMA(0,1,1)(0,1,1)) may be chosen in the first step of the automatic model identification. If the Ljung-Box Q statistics for the residuals is acceptable, the default model is accepted and no further attempt will be made to identify any other.

automdl.cancel
numeric, cancelation limit. If the difference in moduli of an AR and an MA roots (when estimating ARIMA(1,0,1)(1,0,1) models in the second step of the automatic identification of the differencing orders) is smaller than cancelation limit, the two roots are assumed equal and cancel out.

automdl.ub1
numeric, first unit root limit. It is the threshold value for the initial unit root test in the automatic differencing procedure. When one of the roots in the estimation of the ARIMA(2,0,0)(1,0,0) plus mean model, performed in the first step of the automatic model identification procedure, is larger than first unit root limit in modulus, it is set equal to unity.

automdl.ub2
numeric, second unit root limit. When one of the roots in the estimation of the ARIMA(1,0,1)(1,0,1) plus mean model, which is performed in the second step of the automatic model identification procedure, is larger than second unit root limit in modulus, it is checked if there is a common factor in the corresponding AR and MA polynomials of the ARMA model that can be cancelled (see automdl.cancel). If there is no cancellation, the AR root is set equal to unity (i.e. the differencing order changes).

automdl.mixed
logicals. The variable controls whether ARIMA models with non-seasonal AR and MA terms or seasonal AR and MA terms will be considered in the automatic model identification procedure. If FALSE a model with AR and MA terms in both the seasonal and non-seasonal parts of the model can be acceptable, provided there are not AR and MA terms in either the seasonal or non-seasonal.

automdl.balanced
logicals. If TRUE, the automatic model identification procedure will have a preference for balanced models (i.e. models for which the order of the combined AR and differencing operator is equal to the order of the combined MA operator).

automdl.armalimit
numeric, arma limit. It is the threshold value for t-statistics of ARMA coefficients and constant term used for the final test of model parsimony. If the highest order ARMA coefficient has a t-value less than this value in magnitude, the order of the model is reduced. Also if the constant term has a t-value less than arma limit in magnitude, it is removed from the set of regressors.

automdl.reducecv
numeric, ReduceCV. The percentage by which the outlier's critical value will be reduced when an identified model is found to have a Ljung-Box statistic with an unacceptable confidence coefficient. The parameter should be between 0 and 1, and will only be active when automatic outlier identification is enabled. The reduced critical value will be set to (1-ReduceCV)xCV, where CV is the original critical value.

automdl.ljungboxlimit
numeric, Ljung Box limit. Acceptance criterion for the confidence intervals of the Ljung-Box Q statistic. If the LjungBox Q statistics for the residuals of a final model is greater than Ljung Box limit, the model is rejected, the outlier critical value is reduced, and model and outlier identification (if specified) is redone with a reduced value.

automdl.ubfinal
numeric, final unit root limit. The threshold value for the final unit root test. If the magnitude of an AR root for the final model is less than the final unit root limit, a unit root is assumed, the order of the AR polynomial is reduced by one, and the appropriate order of the differencing (non-seasonal, seasonal) is increased. The parameter value should be greater than one.

Control variables for the non-automatic modelling of the ARIMA model (automdl.enabled is set to FALSE):

arima.mu
logicals. If TRUE, the mean is considered as part of the ARIMA model.

arima.p
numeric. The order of the non-seasonal autoregressive (AR) polynomial.

arima.d
numeric. Regular differencing order.

arima.q
numeric. The order of the non-seasonal moving average (MA) polynomial.

arima.bp
numeric. The order of the seasonal autoregressive (AR) polynomial.

arima.bd
numeric. Seasonal differencing order.

arima.bq
numeric. The order of the seasonal moving average (MA) polynomial.

Control variables for the user-defined ARMA coefficients. Coefficients can be defined for the regular and seasonal autoregressive (AR) polynomials and moving average (MA) polynomials. The model considers the coefficients only if the procedure for their estimation (arima.coefType) is provided, and the number of provided coefficients matches the sum of (regular and seasonal) AR and MA orders (p,q,bp,bq).

arima.coefEnabled
logicals. If TRUE the program uses the user-defined ARMA coefficients.

arima.coef
vector providing the coefficients for the regular and seasonal AR and MA polynominals. The length of the vector must equal the sum of the regular and seasonal AR and MA orders. The coefficients shall be provided in the order: regular AR (Phi - p elements), regular MA (Theta - q elements), seasonal AR (BPhi - bp elements) and seasonal MA (BTheta - bq elements). E.g.: arima.coef=c(0.6,0.7) with arima.p=1, arima.q=0,arima.bp=1 and arima.bq=0.

arima.coefType
vector defining ARMA coefficients estimation procedure. Possible procedures are: "Undefined" - no use of user-defined input (i.e. coefficients are estimated), "Fixed" - fixes the coefficients at the value provided by the user, "Initial" - the value defined by the user is used as initial condition. For orders for which the coefficients shall not be defined, the arima.coef can be set to NA or 0 or the arima.coefType can be set to "Undefined". E.g.: arima.coef = c(-0.8,-0.6,NA), arima.coefType = c("Fixed","Fixed","Undefined").

fcst.horizon
numeric, forecasts horizon. Length of the forecasts generated by the RegARIMA model in periods (positive values) or years (negative values). By default the program generates two years forecasts (fcst.horizon set to -2).

x11.mode
character, decomposition mode. Determines the mode of the seasonal adjustment decomposition to be performed: "Undefined" - no assumption concerning the relationship between the time series components is made; "Additive" - assumes an additive relationship; "Multiplicative" - assumes a multiplicative relationship; "LogAdditive" - performs an additive decomposition of the logarithms of the series being adjusted; "PseudoAdditive" - assumes an pseudo-additive relationship. Could be changed by the program, if needed.

x11.seasonalComp
logicals. If TRUE the program computes a seasonal component. Otherwise, the seasonal component is not estimated and its values are all set to 0 (additive decomposition) or 1 (multiplicative decomposition).

x11.lsigma
numeric, lower sigma boundary for the detection of extreme values.

x11.usigma
numeric, upper sigma boundary for the detection of extreme values.

x11.trendAuto
logicals, automatic Henderson filter. If TRUE an automatic selection of the Henderson filter's length for the trend estimation is enabled.

x11.trendma
numeric, length of the Henderson filter. The user-defined length of the Henderson filter. The option is available when the automatic Henderson filter selection is disabled (x11.trendAuto=FALSE). Should be an odd number in the range (1, 101].

x11.seasonalma
vector of character(s) specifying which seasonal moving average (i.e. seasonal filter) will be used to estimate the seasonal factors for the entire series. The vector can be of length: 1 - same seasonal filters for all periods (e.g.: seasonalma=c("Msr")); or period's number - a seasonal filer is defined for each period (e.g. for quarterly series: seasonalma=c("S3X3","Msr","S3X3","Msr")). Possible filters are: "Msr", "Stable", "X11Default", "S3X1", "S3X3", "S3X5", "S3X9", "S3X15". "Msr" - the program chooses the final seasonal filter automatically.

x11.fcasts
numeric, RegARIMA forecast. Length of the forecasts generated by the RegARIMA model in periods (positive values) or years (negative values).

x11.bcasts
numeric, backcast. Length of the backcasts used in X11. Negative figures are translated in years of backcasts.

x11.excludeFcasts
logicals, exclude forecats and backcasts. If TRUE forecasts and backcasts from the RegARIMA model are not used in the generation of extreme values in the seasonal adjustment routines.

object
deprecated argument.

Value

A two-elements list of class c("SA_spec", "X13"): (1) object of class c("regarima_spec", "X13") with the RegARIMA model specification, (2) object of class c("X11_spec", "data.frame") with the X11 algorithm specification. Each component refers to different part of the SA model specification, mirroring the arguments of the function (for details see arguments description). Each of the lowest-level component (except span, pre-specified outliers, user-defined variables and pre-specified ARMA coefficients) is structured within a data frame with columns denoting different variables of the model specification and rows referring to: first row - base specification, as provided within the argument spec; second row - user modifications as specified by the remaining arguments of the function (e.g.: arima.d); and third row - final model specification. The final specification (third row) shall include user modifications (row two) unless they were wrongly specified. The pre-specified outliers, user-defined variables and pre-specified ARMA coefficients consist of a list with the Predefined (base model specification) and Final values.

regarima

object of class c("regarima_spec", "x13"). See Value of the function regarima_spec_x13

x11

data.frame of class c("X11_spec", "data.frame"), containing the x11 variables in line with the names of the arguments variables. The final values can be also accessed with the function s_x11.

Details

The available predefined 'JDemetra+' model specifications are described in the table below.

Identifier	|		Log/level detection	|

Outliers detection	|		Calendar effects	|		ARIMA

RSA0 |		NA	|

NA	|		NA	|	Airline(+mean)

RSA1 |	automatic |	AO/LS/TC |		NA	|	Airline(+mean)

RSA2c |	automatic |	AO/LS/TC |	2 td vars + Easter |	Airline(+mean)

RSA3 |	automatic |	AO/LS/TC |		NA	|	automatic

RSA4c |	automatic |	AO/LS/TC |	2 td vars + Easter |	automatic

RSA5c |	automatic |	AO/LS/TC |	7 td vars + Easter |	automatic		Identifier

References

Info on 'JDemetra+', usage and functions: https://ec.europa.eu/eurostat/cros/content/documentation_en BOX G.E.P. and JENKINS G.M. (1970), "Time Series Analysis: Forecasting and Control", Holden-Day, San Francisco.

BOX G.E.P., JENKINS G.M., REINSEL G.C. and LJUNG G.M. (2015), "Time Series Analysis: Forecasting and Control", John Wiley & Sons, Hoboken, N. J., 5th edition.

See also

Examples

myseries <- ipi_c_eu[, "FR"]
myspec1 <- x13_spec(spec = "RSA5c")
myreg1 <- x13(myseries, spec = myspec1)

# Modify a pre-specified model specification
myspec2 <- x13_spec(spec = "RSA5c", tradingdays.option = "WorkingDays")
myreg2 <- x13(myseries, spec = myspec2)

# Modify the model specification from a "X13" object
 myspec3 <- x13_spec(myreg1, tradingdays.option = "WorkingDays")
 myreg3 <- x13(myseries, myspec3)

# Modify the model specification from a "X13_spec" object
 myspec4 <- x13_spec(myspec1, tradingdays.option = "WorkingDays")
 myreg4 <- x13(myseries, myspec4)

# Pre-specified outliers
 myspec1 <- x13_spec(spec = "RSA5c", usrdef.outliersEnabled = TRUE,
             usrdef.outliersType = c("LS", "AO"),
             usrdef.outliersDate = c("2008-10-01", "2002-01-01"),
             usrdef.outliersCoef = c(36, 14),
             transform.function = "None")

 myreg1 <- x13(myseries, myspec1)
 myreg1

#> RegARIMA
#> y = regression model + arima (0, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Theta(1)   -0.5367      0.048
#> BTheta(1)  -0.5083      0.050
#> 
#>              Estimate Std. Error
#> Easter [1]     -1.169      0.338
#> AO (9-2008)    31.410      2.181
#> LS (9-2008)   -56.648      2.256
#> TC (9-2008)    24.181      3.256
#> LS (2-2002)    14.708      1.526
#> LS (12-2001)  -14.648      1.681
#> AO (12-2001)   13.030      1.731
#> LS (5-2008)    -5.485      1.296
#> 
#> Fixed outliers: 
#>              Coefficients
#> LS (10-2008)           36
#> AO (1-2002)            14
#> 
#> 
#> Residual standard error: 1.729 on 323 degrees of freedom
#> Log likelihood = -637.1, aic =  1296 aicc =  1297, bic(corrected for length) = 1.274
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.055
#> M(2)    0.041
#> M(3)    0.926
#> M(4)    0.621
#> M(5)    0.724
#> M(6)    0.215
#> M(7)    0.074
#> M(8)    0.208
#> M(9)    0.056
#> M(10)   0.158
#> M(11)   0.146
#> Q       0.297
#> Q-M2    0.329
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t           s             i
#> Jan 2017  97.4 100.6172 100.6174  -3.2172329 -0.0001992082
#> Feb 2017  97.5 100.3127 101.0283  -2.8126932 -0.7155966863
#> Mar 2017 112.0 102.5469 101.4894   9.4530696  1.0575376567
#> Apr 2017 103.0 101.0897 101.9282   1.9103111 -0.8385432983
#> May 2017 100.4 103.0319 102.3136  -2.6318733  0.7182480125
#> Jun 2017 111.2 102.4926 102.6921   8.7074293 -0.1994894034
#> Jul 2017 103.4 103.1596 103.0816   0.2404277  0.0779236963
#> Aug 2017  79.3 103.2483 103.5055 -23.9483256 -0.2572170473
#> Sep 2017 109.7 103.5536 103.9555   6.1464361 -0.4019376040
#> Oct 2017 114.0 106.6886 104.3955   7.3113786  2.2931579296
#> Nov 2017 107.7 105.4631 104.7505   2.2369236  0.7125546908
#> Dec 2017 101.4 104.7490 105.0214  -3.3490189 -0.2723590878
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f          i_f
#> Jan 2018 101.96630 105.0963 105.1795  -3.1299775 -0.083200162
#> Feb 2018 102.23632 105.1464 105.2838  -2.9100563 -0.137428535
#> Mar 2018 113.85794 105.5026 105.3966   8.3553336  0.105971540
#> Apr 2018 108.47477 105.4896 105.5573   2.9851827 -0.067754048
#> May 2018 103.22164 105.7963 105.7844  -2.5746309  0.011859024
#> Jun 2018 114.64042 106.0073 106.0629   8.6331483 -0.055612674
#> Jul 2018 106.53519 106.3942 106.3666   0.1410119  0.027594337
#> Aug 2018  82.77073 106.6890 106.6849 -23.9182264  0.004061745
#> Sep 2018 112.79551 106.7018 106.9859   6.0936895 -0.284129714
#> Oct 2018 115.13202 107.7516 107.2471   7.3803800  0.504589345
#> Nov 2018 109.87965 107.5136 107.4572   2.3660966  0.056314698
#> Dec 2018 103.97193 107.3744 107.6093  -3.4024325 -0.234923742
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         1.557
#>  Seasonal     39.219
#>  Irregular     0.362
#>  TD & Hol.     0.018
#>  Others       61.971
#>  Total       103.128
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.032
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.997
#>  f-test on i (seasonal dummies)         0.965
#>  Residual seasonality (entire series)   0.993
#>  Residual seasonality (last 3 years)    0.922
#>  f-test on sa (td)                      0.001
#>  f-test on i (td)                       0.006
#> 
#> 
#> Additional output variables

 s_preOut(myreg1)

#>   type       date coeff
#> 1   LS 2008-10-01    36
#> 2   AO 2002-01-01    14

# User-defined calendar regressors
 var1 <- ts(rnorm(length(myseries))*10, start = start(myseries), frequency = 12)
 var2 <- ts(rnorm(length(myseries))*100, start = start(myseries), frequency = 12)
 var <- ts.union(var1, var2)
 myspec1 <- x13_spec(spec = "RSA5c", tradingdays.option = "UserDefined",
                     usrdef.varEnabled = TRUE,
                     usrdef.var = var,
                     usrdef.varType = c("Calendar", "Calendar"))
 myreg1 <- x13(myseries, myspec1)
 myreg1

#> RegARIMA
#> y = regression model + arima (3, 1, 1, 0, 1, 1)
#> Log-transformation: yes
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)      1.4226      0.270
#> Phi(2)      0.7300      0.167
#> Phi(3)      0.2276      0.076
#> Theta(1)    0.8700      0.268
#> BTheta(1)  -0.6234      0.047
#> 
#>              Estimate Std. Error
#> Easter [1]   -0.01104      0.003
#> LS (11-2008) -0.07705      0.013
#> LS (1-2009)  -0.07318      0.013
#> 
#> 
#> Residual standard error: 0.0167 on 323 degrees of freedom
#> Log likelihood = 860.4, aic =  1280 aicc =  1280, bic(corrected for length) = -8.041
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.054
#> M(2)    0.029
#> M(3)    0.869
#> M(4)    0.319
#> M(5)    0.717
#> M(6)    0.135
#> M(7)    0.063
#> M(8)    0.170
#> M(9)    0.073
#> M(10)   0.154
#> M(11)   0.141
#> Q       0.254
#> Q-M2    0.281
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t         s         i
#> Jan 2017  97.4 100.6429 100.5846 0.9677779 1.0005803
#> Feb 2017  97.5 100.2558 100.9700 0.9725124 0.9929262
#> Mar 2017 112.0 102.3926 101.4214 1.0938286 1.0095759
#> Apr 2017 103.0 101.0271 101.8710 1.0195280 0.9917162
#> May 2017 100.4 103.1461 102.2956 0.9733763 1.0083149
#> Jun 2017 111.2 102.3476 102.7385 1.0864934 0.9961952
#> Jul 2017 103.4 103.1442 103.1947 1.0024800 0.9995110
#> Aug 2017  79.3 103.8828 103.6707 0.7633605 1.0020455
#> Sep 2017 109.7 103.4138 104.1600 1.0607871 0.9928357
#> Oct 2017 114.0 106.2481 104.6259 1.0729602 1.0155048
#> Nov 2017 107.7 105.5255 104.9979 1.0206060 1.0050256
#> Dec 2017 101.4 104.8633 105.2722 0.9669734 0.9961154
#> 
#> Forecasts:
#>                y_f     sa_f      t_f       s_f       i_f
#> Jan 2018 102.27419 105.5798 105.4209 0.9686908 1.0015077
#> Feb 2018 102.47193 105.4597 105.5188 0.9716695 0.9994397
#> Mar 2018 114.22955 105.5382 105.6198 1.0823526 0.9992279
#> Apr 2018 108.91456 105.7488 105.7651 1.0299363 0.9998460
#> May 2018 103.09937 105.8617 105.9783 0.9739058 0.9989007
#> Jun 2018 115.32196 106.2322 106.2191 1.0855653 1.0001232
#> Jul 2018 106.82216 106.6636 106.4682 1.0014865 1.0018348
#> Aug 2018  81.48952 106.6390 106.7136 0.7641622 0.9993008
#> Sep 2018 113.23087 106.8108 106.9289 1.0601067 0.9988960
#> Oct 2018 115.16593 107.2745 107.1163 1.0735629 1.0014773
#> Nov 2018 109.66981 107.3453 107.2748 1.0216546 1.0006567
#> Dec 2018 103.73900 107.3167 107.3854 0.9666618 0.9993609
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         2.191
#>  Seasonal     67.810
#>  Irregular     0.527
#>  TD & Hol.     0.024
#>  Others       30.505
#>  Total       101.056
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.615
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.960
#>  f-test on i (seasonal dummies)         0.929
#>  Residual seasonality (entire series)   0.996
#>  Residual seasonality (last 3 years)    0.858
#>  f-test on sa (td)                      0.001
#>  f-test on i (td)                       0.002
#> 
#> 
#> Additional output variables

 myspec2 <- x13_spec(spec = "RSA5c", usrdef.varEnabled = TRUE,
             usrdef.var = var1, usrdef.varCoef = 2,
             transform.function = "None")
 myreg2 <- x13(myseries, myspec2)
 s_preVar(myreg2)

#> $series
#>                Jan           Feb           Mar           Apr           May
#> 1990 -21.252639407   5.465412848  10.114553407  13.671746024   7.078379701
#> 1991  -6.423485204   6.623578138   2.458842772  -8.478936872  -4.347674107
#> 1992  11.425115700  -2.219401352  -2.083145925   7.499227118   1.176056177
#> 1993  10.454600553   7.204436071   5.029684073   9.558412135   4.921280559
#> 1994   6.047329872   7.375097630  -8.981034882   1.718213563   9.132347891
#> 1995  20.687405597 -13.150525575 -11.097503996   0.548967197  -8.625873906
#> 1996  -7.364807522  21.436128315  -3.381593680  -0.709703999  -0.882100204
#> 1997   7.793108467  -4.540641716 -20.286071844   1.374981227  -6.531652495
#> 1998  -1.968754162  -8.480613844 -18.774430764 -25.861413063   1.080672274
#> 1999  -4.413813498 -21.789991293  -6.781489679   7.288472658 -12.434327365
#> 2000   4.510162329  -3.525772249   3.664544857  -4.881248718   8.733900144
#> 2001  13.580460668   3.119454244   9.911976849  -3.818986462  -8.802615153
#> 2002  16.414782427 -25.369827268  15.149668186  13.444652874   3.422179953
#> 2003   8.261668965  11.562448319   4.109726943  -0.560486435   7.801771348
#> 2004 -20.168055585 -15.964884036   7.905133025  -0.559537712  27.846565380
#> 2005  -3.485907549  -2.132519004   7.486664662 -30.045498345  -2.057101950
#> 2006   1.753171579  13.969229710   8.254166479   5.547305107   5.453779827
#> 2007   2.631799404  -6.875861883   2.208611800  -5.541037774   5.518463880
#> 2008  -3.910052334  24.018101774   5.541959989   5.607759451 -14.307888081
#> 2009   0.923561855  -5.261090352  -9.412824980 -11.603851187  14.632167885
#> 2010  -7.196403767   3.172919743   1.667156540  25.026785326   4.004087093
#> 2011   2.275063071 -25.294187083  -8.185311433 -15.895299935 -10.824933233
#> 2012  16.485396371  -7.556829840  -2.881793527  -7.301505059   5.241485203
#> 2013 -18.604776905   2.061176045 -10.655260009   4.442791106   3.900637392
#> 2014  -7.374681135  -1.250153428   2.798241470  -4.473720260   0.479576801
#> 2015   3.267116657  -9.635751113   4.786579776   7.805649166  -1.854254995
#> 2016  -4.199902912  -5.145817426  -4.811891408  -0.228961166  -2.358173751
#> 2017  -1.837188962   5.645491125  -6.456633054 -19.841975169  -9.328107935
#>                Jun           Jul           Aug           Sep           Oct
#> 1990  -1.225158540  -7.994091347  -1.793600041   1.094528569   2.309265529
#> 1991   8.831759490   5.863186803  -2.566920407  12.201357048  23.258879673
#> 1992   0.990212037  13.616884687  -8.389882385 -10.649372190  21.131680157
#> 1993  -1.559494684  -4.330621776 -12.234203505  -2.903433729   3.402319889
#> 1994 -11.051534547   3.934501445 -13.698426711   7.394716153   6.383709370
#> 1995   9.307754912   5.711704007  12.348319265   4.620031264  -7.342081944
#> 1996   3.550907169 -12.048084555  14.560005038  -0.004656092  -1.176236571
#> 1997   7.200350979   8.084364277   9.624521304   1.031380723   6.863601515
#> 1998 -18.383419463  -1.230852580   1.474878788 -14.306464098  -1.059880850
#> 1999  -5.882719769  11.085391313  -5.023799876   3.472986863  17.481467948
#> 2000  17.987806220  10.507952142  21.752306395  -1.835628629  -6.867930022
#> 2001   0.230683213   7.083967956  14.710711674  14.231851069  -7.317572527
#> 2002  -3.246704812  -9.509212602   1.225963233  22.095975457  13.049280363
#> 2003 -19.231777839  -2.736494881   6.037177315 -13.173483134   9.555886301
#> 2004   3.954117043  26.979624187  10.566101329  17.818693119  13.186880388
#> 2005  -1.435167183  13.109571969   7.645732607   5.644685721  -1.753213851
#> 2006  -5.363845039 -18.392675089  -7.123421939  -0.945664557   2.430159689
#> 2007  -1.618329852  13.514035346  -0.053331422 -24.051334407   9.898093589
#> 2008   2.565309367   4.475129842   9.091545502  -6.576117921   2.547812926
#> 2009 -12.270357371  12.891223057   5.219834145   0.386469118   0.054809002
#> 2010  10.721119854   0.550009631  -0.613638601 -10.764435907   5.043042930
#> 2011   2.058043597 -10.811823198  -1.887925541   3.127564538   0.747169985
#> 2012  -5.917005292 -11.839119861  12.003531317  -9.080713269 -15.488176869
#> 2013  -1.194045102  14.749626484 -12.946582609 -14.676126825  -2.113007304
#> 2014  10.928525350   5.080096757  -8.699107994  -7.026764757 -13.594072699
#> 2015  -3.714157620  -3.759064151 -10.497730923   1.355098866 -10.419342492
#> 2016  -0.429158886  -5.334721852   2.409236569   2.757879994   9.873378846
#> 2017  18.645597497   3.076683344 -18.734768073 -13.815863726  12.142810047
#>                Nov           Dec
#> 1990 -12.696631620   6.205258113
#> 1991 -13.070423363   0.002714681
#> 1992   5.106671868  12.002630345
#> 1993   2.345520557   0.118239369
#> 1994 -22.803538615 -20.112246984
#> 1995  13.233606241   4.572947704
#> 1996   1.841482138  -8.612007223
#> 1997  10.951989336   1.818922153
#> 1998   5.031819093  25.126804489
#> 1999   4.628553413   7.678697074
#> 2000   0.252921842   7.770312224
#> 2001 -12.278623180  -8.465572264
#> 2002  -5.255140662  11.516359225
#> 2003   0.128535381   2.914360940
#> 2004  -7.240831537  -2.632911657
#> 2005  -0.025746856   5.887499380
#> 2006  10.204461058   9.232619914
#> 2007 -12.330316799   0.274105624
#> 2008 -13.629594240 -11.516580441
#> 2009  20.509627114   5.840034881
#> 2010   5.362245498 -13.549772917
#> 2011  -6.043448541   0.452466495
#> 2012   4.080449528   8.723428206
#> 2013   4.196187254  14.333403558
#> 2014   3.456469912   4.218054806
#> 2015  13.715434613   2.117230202
#> 2016  -6.071750184  -2.431552344
#> 2017   1.414819037   0.631038506
#> 
#> $description
#>              type coeff
#> userdef Undefined     2
#> 

# Pre-specified ARMA coefficients
 myspec1 <- x13_spec(spec = "RSA5c", automdl.enabled = FALSE,
             arima.p = 1, arima.q = 1, arima.bp = 0, arima.bq = 1,
             arima.coefEnabled = TRUE,
             arima.coef = c(-0.8, -0.6, 0),
             arima.coefType = c(rep("Fixed", 2), "Undefined"))

 s_arimaCoef(myspec1)

#>                Type Value
#> Phi(1)        Fixed  -0.8
#> Theta(1)      Fixed  -0.6
#> BTheta(1) Undefined   0.0

 myreg1 <- x13(myseries, myspec1)
 myreg1

#> RegARIMA
#> y = regression model + arima (1, 1, 1, 0, 1, 1)
#> Log-transformation: no
#> Coefficients:
#>           Estimate Std. Error
#> Phi(1)     -0.8000      0.000
#> Theta(1)   -0.6000      0.000
#> BTheta(1)  -0.4542      0.051
#> 
#>              Estimate Std. Error
#> Monday       -0.05826      0.139
#> Tuesday      -0.02508      0.139
#> Wednesday     0.30165      0.137
#> Thursday     -0.59487      0.138
#> Friday        0.50268      0.140
#> Saturday      0.03995      0.138
#> Leap year    -0.04011      0.451
#> Easter [1]   -1.20959      0.275
#> AO (10-2005) -4.95487      1.182
#> 
#> 
#> Residual standard error: 2.096 on 323 degrees of freedom
#> Log likelihood = -698.8, aic =  1420 aicc =  1420, bic(corrected for length) = 1.659
#> 
#> 
#> 
#> Decomposition
#>  Monitoring and Quality Assessment Statistics:  
#>       M stats
#> M(1)    0.062
#> M(2)    0.028
#> M(3)    0.543
#> M(4)    0.621
#> M(5)    0.536
#> M(6)    0.260
#> M(7)    0.078
#> M(8)    0.214
#> M(9)    0.055
#> M(10)   0.187
#> M(11)   0.171
#> Q       0.245
#> Q-M2    0.272
#> 
#> Final filters: 
#> Seasonal filter:  3x5
#> Trend filter:  13 terms Henderson moving average
#> 
#> 
#> Final
#> Last observed values
#>              y       sa        t            s           i
#> Jan 2017  97.4 100.8226 100.7235  -3.42255895  0.09907743
#> Feb 2017  97.5 100.4299 101.1028  -2.92989614 -0.67294637
#> Mar 2017 112.0 102.3393 101.5687   9.66069455  0.77061113
#> Apr 2017 103.0 101.4439 102.0275   1.55613149 -0.58358353
#> May 2017 100.4 102.9996 102.4280  -2.59955487  0.57159360
#> Jun 2017 111.2 102.5271 102.8002   8.67294536 -0.27318844
#> Jul 2017 103.4 103.4506 103.1564  -0.05064333  0.29424424
#> Aug 2017  79.3 103.3331 103.5205 -24.03313502 -0.18739845
#> Sep 2017 109.7 103.2505 103.8685   6.44950396 -0.61795652
#> Oct 2017 114.0 106.8134 104.1640   7.18655401  2.64946146
#> Nov 2017 107.7 105.3810 104.3398   2.31898284  1.04118208
#> Dec 2017 101.4 104.2789 104.4055  -2.87892623 -0.12656995
#> 
#> Forecasts:
#>                y_f     sa_f      t_f         s_f         i_f
#> Jan 2018 101.22600 104.0842 104.3608  -2.8581864 -0.27664562
#> Feb 2018 101.01524 104.0618 104.2820  -3.0465104 -0.22026718
#> Mar 2018 112.55419 104.3222 104.2357   8.2320286  0.08649224
#> Apr 2018 106.78270 104.2328 104.2647   2.5499347 -0.03188965
#> May 2018 101.40946 104.4771 104.3970  -3.0676654  0.08012361
#> Jun 2018 113.75858 104.5342 104.6051   9.2243530 -0.07091744
#> Jul 2018 104.66981 104.9101 104.8449  -0.2402895  0.06523367
#> Aug 2018  81.51533 104.9474 105.1025 -23.4320489 -0.15511871
#> Sep 2018 110.73413 105.0383 105.3493   5.6958578 -0.31102941
#> Oct 2018 114.06653 106.3250 105.5577   7.7415560  0.76728580
#> Nov 2018 108.38112 105.6755 105.7158   2.7056474 -0.04027530
#> Dec 2018 101.97830 105.4990 105.8238  -3.5207475 -0.32478623
#> 
#> 
#> Diagnostics
#>  Relative contribution of the components to the stationary
#>  portion of the variance in the original series,
#>  after the removal of the long term trend 
#>  Trend computed by Hodrick-Prescott filter (cycle length = 8.0 years)
#>            Component
#>  Cycle         7.755
#>  Seasonal     92.477
#>  Irregular     0.782
#>  TD & Hol.     0.118
#>  Others        0.073
#>  Total       101.204
#> 
#>  Combined test in the entire series 
#>  Non parametric tests for stable seasonality
#>                                                           P.value
#>    Kruskall-Wallis test                                      0.000
#>    Test for the presence of seasonality assuming stability   0.000
#>    Evolutive seasonality test                                0.004
#>  
#>  Identifiable seasonality present
#> 
#>  Residual seasonality tests 
#>                                       P.value
#>  qs test on sa                          1.000
#>  qs test on i                           1.000
#>  f-test on sa (seasonal dummies)        0.974
#>  f-test on i (seasonal dummies)         0.858
#>  Residual seasonality (entire series)   0.989
#>  Residual seasonality (last 3 years)    0.893
#>  f-test on sa (td)                      0.934
#>  f-test on i (td)                       0.983
#> 
#> 
#> Additional output variables

# Defined seasonal filters
 myspec1 <- x13_spec("RSA5c", x11.seasonalma = rep("S3X1", 12))
 mysa1 <- x13(myseries, myspec1)